% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{add to cast tab}
\alias{add to cast tab}
\alias{add_obs_to_cast_tab}
\title{Add Observations to a Cast Tab}
\usage{
add_obs_to_cast_tab(main = ".", cast_tab = NULL)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{cast_tab}{A \code{data.frame} of a cast's output. See \code{\link{read_cast_tab}}.}
}
\value{
\code{data.frame} of \code{cast_tab} with an additional column.
}
\description{
Appends a column of observations to a cast's cast tab.
}
\details{
If a model interpolated a data set, \code{add_obs_to_cast_tab} adds the true (non-interpolated) observations so that model predictions are all compared to the same data.
}
