% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{add_addl_future_moons}
\alias{add_addl_future_moons}
\title{Add more more moons to accomplish requested forecast length}
\usage{
add_addl_future_moons(future_moons, fdate)
}
\arguments{
\item{future_moons}{the table of moons forward from the moon table}

\item{fdate}{date from which future is defined, typically today's date.}
}
\value{
appended moon table
}
\description{
Add extra future moons to the moon table to achieve requested
  forecast length. Because the moon table might not have the most recent 
  moons, more rows than initially requested may need to be added to the 
  table.
}
