% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{add_future_moons}
\alias{add_future_moons}
\title{Add future moons to the moon table}
\usage{
add_future_moons(moons = prep_moons(), n_future_moons = 12,
  fdate = today())
}
\arguments{
\item{moons}{current newmoonnumber table}

\item{n_future_moons}{integer value for the number of future moons to add}

\item{fdate}{date from which future is defined, typically today's date.
Required if \code{n_future_moons > 0}}
}
\value{
appended moon table
}
\description{
Add future moons to the moon table, counting forward from
  \code{fdate}. Because the moon table might not have the most recent 
  moons, more rows than \code{n_future_moons} may be added to the table.
}
