% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{add_moons_from_date}
\alias{add_moons_from_date}
\title{Add a Newmoon Number to a Table that has a Date}
\usage{
add_moons_from_date(df, moons = NULL)
}
\arguments{
\item{df}{\code{data.frame} with column of \code{date}s.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}
}
\value{
\code{data.frame} \code{df} with column of \code{newmoonnumber}s added.
}
\description{
Add a \code{newmoonnumber} column to a table that has a \code{date} column.
}
