% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_newmoonnumbers_from_dates}
\alias{add_newmoonnumbers_from_dates}
\title{Add a Newmoon Number Column to a Table that has a Date Column}
\usage{
add_newmoonnumbers_from_dates(df, moons = NULL)
}
\arguments{
\item{df}{\code{data.frame} with column of \code{date}s.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prepare_moons}}.}
}
\value{
\code{data.frame} \code{df} with column of \code{newmoonnumber}s added.
}
\description{
Add a \code{newmoonnumber} column to a table that has a \code{date} column.
}
