% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{add_newmoons_from_date}
\alias{add_newmoons_from_date}
\title{Add a newmoon number to a table that has the date}
\usage{
add_newmoons_from_date(x, moons = NULL)
}
\arguments{
\item{x}{\code{data.frame} with column of newmoon \code{Date}s 
named \code{date}.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}
}
\value{
\code{data.frame} \code{x} with column of \code{newmoonnumber}s 
 added.
}
\description{
Add a newmoon number column to a table that has a 
 \code{date} (as a \code{Date}) column.
}
\examples{
 \donttest{
  create_dir()
  fill_dir()
  moons <- prep_moons()
  raw_path <- sub_paths(specific_subs = "raw")
  weather <- portalr::weather("daily", TRUE, raw_path)
  add_newmoons_from_date(weather, moons)
 }

}
