% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{add_obs_to_cast_tab}
\alias{add_obs_to_cast_tab}
\title{Add the associated observations to a cast tab}
\usage{
add_obs_to_cast_tab(main = ".", cast_tab = NULL, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{cast_tab}{A \code{data.frame} of a cast's output. See 
\code{\link{read_cast_tab}}.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
\code{data.frame} of \code{cast_tab} with an \code{obs} column.
}
\description{
Append a column of observations to a cast table. \cr
 If an interpolated data set is used to fit a model, the true data
 are appended here (so that model predictions are all compared to the
 same data).
}
\examples{
 \donttest{
  setup_dir()
  portalcast(models = c("AutoArima", "NaiveArima"), end_moons = 515:520)
  cast_tab <- read_cast_tab(cast_id = 1)
  add_obs_to_cast_tab(cast_tab = cast_tab)
}

}
