% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{add_to_cast_tab}
\alias{add_to_cast_tab}
\alias{add_lead_to_cast_tab}
\alias{add_err_to_cast_tab}
\alias{add_covered_to_cast_tab}
\alias{add_obs_to_cast_tab}
\title{Add the associated values to a cast tab}
\usage{
add_lead_to_cast_tab(main = ".", cast_tab = NUL)

add_err_to_cast_tab(main = ".", cast_tab = NULL)

add_covered_to_cast_tab(main = ".", cast_tab = NULL)

add_obs_to_cast_tab(main = ".", cast_tab = NULL)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{cast_tab}{A \code{data.frame} of a cast's output. See 
\code{\link{read_cast_tab}}.}
}
\value{
\code{data.frame} of \code{cast_tab} with an additional column or
 columns if needed.
}
\description{
Add values to a cast's cast tab. If necessary components are
 missing (such as no observations added yet and the user requests errors),
 the missing components are added. \cr \cr
 \code{add_lead_to_cast_tab} adds a column of lead times. \cr \cr
 \code{add_obs_to_cast_tab} appends a column of observations. \cr \cr
 \code{add_err_to_cast_tab} adds a column of raw error values. \cr \cr
 \code{add_covered_to_cast_tab} appends a \code{logical} column indicating
 if the observation was within the prediction interval.
}
\details{
If an interpolated data set is used to fit a model, 
 \code{add_obs_to_cast_tab} adds the true (non-interpolated) observations
 so that model predictions are all compared to the same data.
}
\examples{
 \donttest{
  setup_dir()
  portalcast(models = c("AutoArima", "NaiveArima"))
  cast_tab <- read_cast_tab(cast_id = 1)
  add_lead_to_cast_tab(cast_tab = cast_tab)
  add_obs_to_cast_tab(cast_tab = cast_tab)
  add_err_to_cast_tab(cast_tab = cast_tab)
  add_covered_to_cast_tab(cast_tab = cast_tab)
}

}
