% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{all_data_options}
\alias{all_data_options}
\title{Prepare all of the data options for}
\usage{
all_data_options(n_future_moons = 0, fdate = today(), m_save = TRUE,
  m_filename = "moons.csv", type = NULL, start = 217, drop_spp = "PI",
  min_plots = 24, level = "Site", treatment = NULL, length = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  historical = TRUE, forecasts = TRUE, yr = as.numeric(format(today(),
  "\%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL, nfcnm = 0,
  c_save = TRUE, c_filename = "covariates.csv")
}
\arguments{
\item{n_future_moons}{integer value for the number of future moons to add}

\item{fdate}{date from which future is defined, typically today's date.}

\item{m_save}{logical if the moons data should be saved out}

\item{m_filename}{the name of the file for saving moons data}

\item{type}{"all" or "controls"}

\item{start}{newmoon number of the first sample to be included. Default 
value is \code{217}, corresponding to 1995-01-01.}

\item{drop_spp}{species names to drop from the table}

\item{min_plots}{minimum number of plots surveyed for a survey to be used}

\item{level}{for get_rodent_data, automatically set by type}

\item{treatment}{for get_rodent_data, automatically set by type}

\item{length}{for get_rodent_data, automatically set by type}

\item{output}{for get_rodent_data, automatically set by type}

\item{r_save}{logical if the data should be saved out}

\item{r_filename}{the name of the file to save the rodent data in}

\item{historical}{logical indicator whether or historical covariates are
to be included}

\item{forecasts}{logical indicator whether or forecasted covariates are
to be included}

\item{yr}{year of today}

\item{lead_time}{number of moons into the future the rodents are forecast}

\item{min_lag}{the minimum lag time used in any model}

\item{fcast_nms}{newmoon numbers to be forecast for covariates}

\item{nfcnm}{number of forecast newmoons for covariates}

\item{c_save}{logical if the covariate data should be saved out}

\item{c_filename}{the name of the covariate file for the saving}
}
\value{
control options list for covariates

list of control options lists
}
\description{
Create a list of lists of control options for the data
}
