% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{alternative values}
\alias{alternative values}
\alias{ifnull}
\alias{ifna}
\title{Replace a Value with an Alternative if it is NULL or if it is NA}
\usage{
ifnull(x = NULL, alt = NULL)

ifna(x = NULL, alt = NA)
}
\arguments{
\item{x}{Focal input.}

\item{alt}{Alternative value.}
}
\value{
\code{ifnull}: \code{x} if not \code{NULL}, \code{alt} otherwise. \cr \cr
 \code{ifna}:  \code{x} if not \code{NA}, \code{alt} otherwise.
}
\description{
\code{ifnull} replaces the focal input with the alternative value if it is \code{NULL}. \cr \cr
 \code{ifna} replaces the focal input with the alternative value if it is \code{NA}.
}
\examples{
 ifnull(NULL, 123)
 ifnull(TRUE, 123)
 ifnull(FALSE, 123)
 ifna(NA, 123)
 ifna(FALSE, 123)
 ifna(NA, NA)

}
