% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{alternative_values}
\alias{alternative_values}
\alias{ifnull}
\title{Replace a value with an alternative if it is NULL or if it is NA}
\usage{
ifnull(x = NULL, alt = NULL)
}
\arguments{
\item{x}{Focal input.}

\item{alt}{Alternative value.}
}
\value{
\code{x} if not \code{NULL}, \code{alt} otherwise.
}
\description{
\code{ifnull} replaces the focal input with the alternative value if it
  is \code{NULL}. \cr \cr
}
\examples{
 ifnull(NULL, 123)
 ifnull(TRUE, 123)
 ifnull(FALSE, 123)
 ifna(NA, 123)
 ifna(FALSE, 123)
 ifna(NA, NA)

}
