% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{append_cov_fcast_csv}
\alias{append_cov_fcast_csv}
\title{Append a covariate forecast to existing covariate forecasts}
\usage{
append_cov_fcast_csv(new_forecast_covariates,
  options_covariates = covariates_options())
}
\arguments{
\item{new_forecast_covariates}{Class-\code{covariates} \code{data.frame} of
forecasted covariate data generated by 
\code{\link{forecast_covariates}}.}

\item{options_covariates}{Class-\code{covariates_options} options 
\code{list}. See \code{\link{covariates_options}}.}
}
\value{
\code{new_forecast_covariates} exactly as input.
}
\description{
Add the newest weather and NDVI forecasts to the existing 
  forecasts data file, for use in future hindcasting.
}
