% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{append_covariate_fcast}
\alias{append_covariate_fcast}
\title{Append a Covariate Forecast to Historical Covariate Table}
\usage{
append_covariate_fcast(covariates, metadata, moons = get_moon_data())
}
\arguments{
\item{covariates}{output from \code{get_covariate_data}}

\item{metadata}{model metadata}

\item{moons}{}
}
\value{
no value
}
\description{
combining weather and ndvi forecasts to the existing 
  covariates
}
