% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{append_csv}
\alias{append_csv}
\title{Appending a csv without re-writing the header.}
\usage{
append_csv(df, filename)
}
\arguments{
\item{df}{data table to be written out}

\item{filename}{filename of existing csv to be appended}
}
\value{
Nothing
}
\description{
Appending a csv without re-writing the header.
}
