% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{append_past_moons_to_raw}
\alias{append_past_moons_to_raw}
\title{Append missing past moons to the raw data}
\usage{
append_past_moons_to_raw(data, tree = dirtree(), data_options)
}
\arguments{
\item{data}{moons table with future moons added}

\item{tree}{directory tree}

\item{data_options}{moon data options}
}
\value{
data as came in
}
\description{
Sometimes the raw moon data table is not up-to-date. Because
  the portalr functions \code{weather} and \code{fcast_ndvi} point to the
  raw moons data, that table needs to be updated to produce the correct
  current data table
}
