% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{argument_checking}
\alias{argument_checking}
\alias{check_args}
\alias{check_arg}
\alias{check_arg_list}
\title{Check a function's arguments' values for validity}
\usage{
check_args(arg_checks = TRUE)

check_arg(arg_name, arg_value, fun_name = NULL)

check_arg_list()
}
\arguments{
\item{arg_checks}{\code{logical} value of if the arguments should be
checked.}

\item{arg_name}{\code{character} value of the argument name.}

\item{arg_value}{Input value for the argument.}

\item{fun_name}{\code{character} value of the function name or \code{NULL}.}
}
\value{
\code{check_args}: throws an error or \code{NULL}. \cr \cr
  \code{check_arg}: \code{character} vector of error messages
  associated with the specific situation 
  (\code{<fun_name>(<arg_name> = arg_value)}). \cr \cr
  \code{check_arg_list}: \code{list} of argument checking component 
  \code{list}s.
}
\description{
Provides toggle-able argument checking for nearly all 
 functions in the codebase (save a few utilities), that can be turned off
 to facilitate sandboxing or kept on to support reliability. \cr \cr
 \code{check_args} checks that all of the arguments to a given
 function have valid values within the pipeline to avoid naming collision
 and improper formatting, by wrapping around \code{check_arg} for each
 argument. \cr \cr
 \code{check_arg} returns \code{NULL} if the argument is valid or 
 \code{character} vector of error messages associated 
 with the specific situation (\code{<fun_name>(<arg_name> = arg_value)})
 if the argument is invalid. \cr \cr
 \code{check_arg_list} produces a list of argument details for 
 checking within \code{check_arg}. \cr  \cr 
 \strong{\code{check_args} should only be called within a function.}
 \cr \cr
 See \code{Details} for argument-specific rules.
}
\details{
Argument details used for checking are produced simply via
 \code{check_arg_list}. If you are adding to the codebase and use
 novel arguments, add them to that function's production. \cr
 Each argument should have its own element that is a list with the 
 following elements: \code{class} (\code{character} value of its class), 
 \code{null} (\code{logical} indicator if the argument can be \code{NULL}),
 \code{na} (\code{logical} indicator if the argument can be \code{NA}),
 \code{length} (\code{numeric} value of how many elements the argument can
 be or \code{NULL} if unrestricted), and \code{addl} if needed (an
 unevaluated \code{expression} that can be evaluated and allows for 
 simple evaluation of, e.g., element names of a \code{list}; see the 
 source code). \cr 

 Usage rules for arguments are as follows: \cr

 Must be length-1 \code{logical} values, cannot be \code{NULL} or
 \code{NA}: 
  \code{bline},
  \code{force_model_updates},
  \code{only_if_missing}.

 Must be length-1 \code{logical} values, can be \code{NULL}, but cannot be 
 \code{NA}: 
  \code{add_error},
  \code{add_in_window}, 
  \code{add_lead}, 
  \code{add_obs}, 
  \code{append_cast_csv}, 
  \code{arg_checks}, 
  \code{cast_obs},
  \code{clean},
  \code{cleanup},
  \code{covariatesTF}, 
  \code{effort},
  \code{ensemble},
  \code{fillweight}, 
  \code{interpolate}, 
  \code{include_interp}, 
  \code{na_drop}, 
  \code{nadot}, 
  \code{NULLname},
  \code{overwrite}, 
  \code{quiet}, 
  \code{retain_target_moons},
  \code{return_version},
  \code{save}, 
  \code{silent_jags},
  \code{tail}, 
  \code{total}, 
  \code{unknowns}, 
  \code{verbose},
  \code{with_census},
  \code{zero_drop}. 

 Must be \code{logical} values, can be any length or \code{NULL}, but 
 cannot be \code{NA}: 
  \code{movedTF}, 

 Must be length-1 \code{character} values, can be \code{NULL}, but cannot 
 be \code{NA}:
  \code{colname}, 
  \code{data_name}, 
  \code{data_set}, 
  \code{directory}, 
  \code{extension} (if not \code{NULL}, must have a \code{"."} character),
  \code{filename}, 
  \code{filename_cov}, 
  \code{filename_cov_casts}, 
  \code{filename_config}, 
  \code{filename_meta},
  \code{filename_moons}, 
  \code{freq}, 
  \code{jags_model},  
  \code{level}, 
  \code{main}, 
  \code{method} (if not \code{NULL}, must be \code{"unwtavg"}),
  \code{model} (inputted values are checked via 
   \code{\link{verify_models}}), 
  \code{name}, 
  \code{output} (if not \code{NULL}, must be \code{"abundance"}),
  \code{path},
  \code{plots} (if not \code{NULL}, must be \code{"all"} or 
   \code{"longterm"}),
  \code{raw_data}, 
  \code{sep_char}, 
  \code{set},
  \code{shape},
  \code{source_name},
  \code{source_url}, 
  \code{species_id} (if not \code{NULL}, must be within the species
   produced by \code{rodent_spp("wtotal")}),
  \code{sub} (must be one of the named sub directories: \code{"tmp"},
   \code{"raw"}, \code{"data"}, \code{"models"}, or \code{"casts"}),
  \code{time}, 
  \code{title}, 
  \code{type}, 
  \code{url}, 
  \code{winner} (if not \code{NULL}, must be \code{"hist"} or 
   \code{"cast"}), 
  \code{zip_destin}.

 Must be \code{character} values, can be any length, can be \code{NULL}, 
 but cannot be \code{NA}:
  \code{colnames},
  \code{concept_rec_id}, 
  \code{data}, 
  \code{data_sets}, 
  \code{downloads_versions},
  \code{enquote_args}, 
  \code{eval_args}, 
  \code{factories},
  \code{files}, 
  \code{models} (inputted values are checked via 
   \code{\link{verify_models}}), 
  \code{modules},
  \code{monitor},
  \code{msg}, 
  \code{names},
  \code{paths}, 
  \code{rec_id}, 
  \code{rec_version}, 
  \code{ref_species},
  \code{species} (if not \code{NULL}, must be all within the species
   produced by \code{all_secies(total = TRUE)}),
  \code{subs} (must all be one of the named sub directories: \code{"tmp"},
   \code{"raw"}, \code{"data"}, \code{"models"}, or \code{"casts"}),
  \code{target_cols}, 
  \code{treatment} (if not \code{NULL}, must be \code{"control"})

 Must be \code{data.frame}s, can be any length, can be \code{NULL}, 
 but cannot be \code{NA}:
  \code{casts}, 
  \code{cast_cov}, 
  \code{cast_tab}, 
  \code{casts}, 
  \code{covariates}, 
  \code{df}, 
  \code{hist_cov}, 
  \code{hist_tab}, 
  \code{moons}, 
  \code{obs}, 
  \code{preds}, 
  \code{rodents_tab}.

 Must be a \code{data.frame} or \code{vector} can be any length, can be 
 \code{NULL}, but cannot be \code{NA}:
  \code{dfv}.

 Must be a \code{data.frame} or \code{list} can be any length, can be 
 \code{NULL}, but cannot be \code{NA}:
  \code{dfl}.

 Must be \code{list}s, can be any length, can be \code{NULL}, but cannot 
 be \code{NA}:
  \code{control_climate_dl},
  \code{control_files},
  \code{control_model},
  \code{control_runjags},
  \code{controls_models},
  \code{controls_rodents},
  \code{data_set_controls},
  \code{downloads},
  \code{filenames},
  \code{metadata},
  \code{rodents}.

 Must be \code{list}s, can be any length, can be \code{NULL} or \code{NA}:
  \code{run_status}.

 Must be length-1 \code{Date}-conformable values, can be \code{NULL}, 
 but cannot be \code{NA}:
  \code{cast_date},
  \code{date}
  \code{end},
  \code{from_date},
  \code{start}.

 Must be \code{Date}-conformable values, can be any length, can be
 \code{NULL}, but cannot be \code{NA}:
  \code{cast_dates}.
  \code{dates}.

 Must be length-1 \code{numeric} values, can be \code{NULL}, but
 cannot be \code{NA}:
  \code{confidence_level} (must be between 0 and 1), 
  \code{lat},
  \code{lon}.

 Must be length-1 \code{integer}-conformable values can be \code{NULL},
 but cannot be \code{NA}:
  \code{adapt} (must be positive),
  \code{burnin} (must be positive),
  \code{cast_id} (must be non-negative),
  \code{end_moon} (must be positive),
  \code{lead} (must be positive),
  \code{lead_time} (must be non-negative),
  \code{lev},
  \code{max_E} (must be positive, 1 to 24),
  \code{min_observed} (must be positive),
  \code{min_plots} (must be positive),
  \code{min_traps} (must be positive),
  \code{moon} (must be positive),
  \code{nchains} (must be positive),
  \code{ndates} (must be positive),
  \code{nmoons} (must be non-negative),
  \code{sample} (must be positive),
  \code{start_moon} (must be positive),
  \code{thin} (must be positive),
  \code{topx} (must be positive).

 Must be length-1 \code{integer}-conformable values can be \code{NULL}, 
 and can be \code{NA}:
  \code{lag} (must be positive),
  \code{min_lag} (must be non-negative).

 Must be \code{integer}-conformable values, can be any length, can be 
 \code{NULL}, but cannot be \code{NA}:
  \code{cast_groups} (must be non-negative),
  \code{cast_ids} (must be non-negative),
  \code{end_moons} (must be positive),
  \code{target_moons} (must be positive).

 Must be a \code{list} or \code{function}; can be \code{NULL} or \code{NA}:
  \code{mutate}. 

 Must be a \code{character} vector (any length), \code{list}, or
 \code{function}; can be \code{NULL} but not \code{NA}:
  \code{inits}.
}
\examples{
 check_arg_list()
 check_arg("end_moon", 400)
 check_arg("end_moon", 399.5)

}
