% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoarima.R
\name{autoarima}
\alias{autoarima}
\title{Flexible Auto-Regressive Integrated Moving Average model for Portal 
 Predictions}
\usage{
autoarima(abundances, metadata, level = "All", CL = 0.9)
}
\arguments{
\item{abundances}{table of rodent abundances and time measures}

\item{metadata}{model metadata list}

\item{level}{name of the type of plots included ("All" or "Controls")}

\item{CL}{confidence interval level used for forecast envelope}
}
\value{
list of forecast and aic tables
}
\description{
Model "AutoARIMA" is a flexible auto-regressive integrated 
 moving average model fit to the data using \code{auto.arima} in the
 \code{forecast} package (Hyndman \emph{et al}. 2018).

 Because the seasonality and sampling occurring with different frequencies,
 which \code{auto.arima } cannot accommodate, seasonal models are not 
 included.

 Although the auto.arima model can handle missing data, the other models
 used currently cannot, so we interpolate missing data here for 
 comparison.
}
\references{
Hyndman R., Bergmeir C., Caceres G., Chhay L., O'Hara-Wild M., Petropoulos
 F., Razbash S., Wang E., and Yasmeen F. 2018. forecast: Forecasting 
 functions for time series and linear models. 
 \href{http://pkg.robjhyndman.com/forecast}{R package version 8.3}.
}
