% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{base_path}
\alias{base_path}
\alias{main_path}
\alias{sub_paths}
\title{Determine the path for a specific level of a portalcasting directory}
\usage{
base_path(tree = dirtree())

main_path(tree = dirtree())

sub_paths(tree = dirtree(), specific_subs = NULL)
}
\arguments{
\item{tree}{Directory tree \code{list}. See \code{\link{dirtree}}.}

\item{specific_subs}{\code{character}-value name of the specific 
subdirectory/subdirectories of interest, or \code{NULL} (default)
for all.}
}
\value{
\code{base_path}: The normalized path of the main folder in the 
  directory tree as a character value (see \code{\link{normalizePath}}).

\code{main_path}: The normalized path of the main folder in the 
  directory tree as a character value (see \code{\link{normalizePath}}).

\code{sub_paths}: The normalized path(s) of the 
  \code{specific_subs} folder(s) in the directory tree as \code{character}
   value(s) (see \code{\link{normalizePath}}).
}
\description{
\code{base_path}: Return the path for the \code{base} folder
  in the directory.

\code{main_path}: Return the path for the \code{main} folder
  in the directory.

\code{sub_paths}: Return the path(s) for a specific 
  \code{subs} folder or folders in the directory.
}
\examples{
 create_dir()
 base_path()
 main_path()
 sub_paths()
 sub_paths(specific_subs = "models")

}
