% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{cast}
\alias{cast}
\title{Actually forecast or hindcast}
\usage{
cast(options_cast = cast_options())
}
\arguments{
\item{options_cast}{casting options list}
}
\value{
nothing
}
\description{
Given the data and models are prepared accordingly, run the
  forecast or hindcast(s), (optionally) add an ensemble, combine the 
  results into the predictions folder.
}
