% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_covariates.R
\name{cast_covariates}
\alias{cast_covariates}
\alias{prep_cast_covariates}
\alias{cast_ndvi}
\alias{cast_weather}
\alias{save_cast_cov_csv}
\title{Cast the covariates}
\usage{
cast_covariates(main = ".", moons = prep_moons(main = main),
  hist_cov = NULL, end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), cast_type = "forecast",
  raw_path_archive = "portalPredictions",
  raw_cov_cast_file = "data/covariate_casts.csv",
  raw_path_cov_cast = "cov_casts", source_name = "current_archive",
  control_cdl = list(), quiet = FALSE)

prep_cast_covariates(main = ".", moons = prep_moons(main = main),
  hist_cov = NULL, end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), cast_type = "forecast",
  raw_path_archive = "portalPredictions",
  raw_cov_cast_file = "data/covariate_casts.csv",
  raw_path_cov_cast = "cov_casts", source_name = "current_archive",
  append_cast_csv = TRUE, control_cdl = list(), quiet = FALSE,
  save = TRUE, overwrite = TRUE)

cast_ndvi(main = ".", moons = prep_moons(main = main),
  hist_cov = NULL, lead_time = 12, min_lag = 6)

cast_weather(main = ".", moons = prep_moons(main = main),
  hist_cov = NULL, end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), cast_type = "forecast",
  raw_path_archive = "portalPredictions",
  raw_cov_cast_file = "data/covariate_casts.csv",
  raw_path_cov_cast = "cov_casts", source_name = "current_archive",
  control_cdl = list(), quiet = FALSE)

save_cast_cov_csv(main = ".", cast_cov = NULL,
  cast_type = "forecast", append_cast_csv = TRUE,
  raw_path_archive = "portalPredictions",
  raw_cov_cast_file = "data/covariate_casts.csv",
  source_name = "current_archive", quiet = FALSE, save = TRUE,
  overwrite = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{hist_cov}{\code{data.frame} of historic covariates from 
\code{\link{prep_hist_covariates}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the
number of timesteps forward a cast will cover.}

\item{min_lag}{\code{integer} (or integer \code{numeric}) of the minimum 
covariate lag time used in any model.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}

\item{cast_type}{\code{character} value of the -cast type: 
\code{"forecasts"} or \code{"hindcasts"}.}

\item{raw_path_archive}{\code{character} value of the path to the archive
within the raw sub folder.}

\item{raw_cov_cast_file}{\code{character} value of the path to the
covariate cast file within \code{raw_path_archive}.}

\item{raw_path_cov_cast}{\code{character} value of the path to the folder
that holds any downloaded covariate cast files within the raw sub folder.}

\item{source_name}{\code{character} value for the name to give the 
covariate forecast. Currently is \code{"current_archive"}. Previous to
\code{"current_archive"}, the data were retroactively filled in and are 
given the source name \code{"retroactive"}.}

\item{control_cdl}{\code{list} of specifications for the 
climate download, which are sent to \code{\link{NMME_urls}} to create the 
specific URLs. See \code{\link{climate_dl_control}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{append_cast_csv}{\code{logical} indicator controlling if the new 
cast covariates should be appended to the historical casts for the
purposes of hindcasting later.}

\item{save}{\code{logical} indicator controlling if the output should 
be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing
files should be updated (most users should leave as \code{TRUE}).}

\item{cast_cov}{\code{data.frame} of cast covariates from 
\code{\link{cast_covariates}}.}
}
\value{
\code{cast_covariates}: \code{data.frame} of the covariates cast for the
  moons as defined as needed for saving. \cr \cr
 \code{prep_cast_covariates}: \code{data.frame} of the cast covariates
  after saving, with excess columns removed. \cr \cr
 \code{cast_ndvi}: \code{data.frame} of -casted NDVI values. \cr \cr
 \code{cast_weather}: \code{data.frame} of -casted weather values. \cr \cr
 \code{save_cast_cov_csv}: \code{cast_cov} \code{data.frame} exactly as 
  input. \cr \cr
}
\description{
Forecast or hindcast the covariates for a model run.
 \code{cast_covariates} is the primary function, which produces the 
 coavariates (min, mean, and max temperature; precipitation; and NDVI) 
 required for a model to be forecast or hindcast. \cr \cr
 \code{prep_cast_covariates} provides a wrapper on \code{cast_covariates}
 that includes saving the cast data out via \code{save_cast_cov_csv}
 and tidying the data for use within the model (by removing the vestigial
 columns that are only needed for saving).
}
\details{
If \code{cast_type = "forecast"}, weather is forecast using 
 \code{\link{cast_weather}} and NDVI is forecast using 
 \code{\link{cast_ndvi}}. If \code{cast_type = "hindcast"}, the historical
 forecasted weather and NDVI data are retrieved from the file pointed to
 by \code{raw_cov_cast_file}.
}
\examples{
 \donttest{
  create_dir()
  fill_dir()
  hist_cov <- prep_hist_covariates()
  prep_cast_covariates(hist_cov = hist_cov)
  cast_cov <- cast_covariates(hist_cov = hist_cov)
  save_cast_cov_csv(cast_cov = cast_cov)
  cast_ndvi_data <- cast_ndvi(hist_cov = hist_cov)
  cast_weather_data <- cast_weather(hist_cov = hist_cov)
 }

}
