% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{cast_models}
\alias{cast_models}
\title{Source the selected models to forecast or hindcast with}
\usage{
cast_models(options_cast = cast_options())
}
\arguments{
\item{options_cast}{casting options}
}
\value{
nothing
}
\description{
Source the code of the selected model scripts
}
