% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{castplot_ylab}
\alias{castplot_ylab}
\title{Expand the name with formating for a -cast plot y label}
\usage{
castplot_ylab(tree = dirtree(), species = "total")
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}. Used to access the species list.}

\item{species}{\code{character} value of the species code or \code{"total"}
for the total across species.}
}
\value{
\code{list} of \code{text} and \code{font} elements.
}
\description{
Produce a list with \code{text} and \code{font} elements to
  control the y-axis label in \code{\link{plot_cast}}.
}
