% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{casts}
\alias{casts}
\title{Run potentially multiple forecasts or hindcasts}
\usage{
casts(options_all = all_options())
}
\arguments{
\item{options_all}{\code{all_options}-class \code{list} of options 
controlling the portalcasting directory. See \code{\link{all_options}}.}
}
\description{
Given that the data and models are prepared accordingly, run 
  the forecast or hindcasts, (optionally) add an ensemble, combine the 
  results, and add them to the predictions folder.
}
