% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_args}
\alias{check_args}
\title{Check all of a function's arguments values for validity}
\usage{
check_args()
}
\description{
Check that all of the arguments to a given function have
  valid values, by wrapping around \code{\link{check_arg}}.
}
