% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{check_to_skip}
\alias{check_to_skip}
\title{Check if the newmoon should be skipped during a set of hindcasts}
\usage{
check_to_skip(tree = dirtree(), quiet = FALSE, data_control = list())
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{data_control}{A \code{list} of arguments to control the filling of
the data subdirectory. See \code{\link{data_control}}. 
Values not set assume defaults.}
}
\value{
logical indicating if the cast should be skipped
}
\description{
For hindcasts, skip newmoons with incomplete surveys.
}
