% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_url}
\alias{check_url}
\title{Check if a URL will throw errors}
\usage{
check_url(url, quiet = FALSE)
}
\arguments{
\item{url}{\code{character} value to evaluate.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\value{
Error message or \code{NULL} otherwise.
}
\description{
If the URL throws errors, return them gracefully.
}
