% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{classy}
\alias{classy}
\title{Set the class(es) of an object}
\usage{
classy(x, class = NULL)
}
\arguments{
\item{x}{The object to get the \code{class}(es).}

\item{class}{The \code{class}(es) to apply to \code{x}.}
}
\value{
\code{x} with class(es) of \code{class}.
}
\description{
For inclusion in a simple pipeline, set the class of an object
  and return it.
}
