% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{cleanup_dir}
\alias{cleanup_dir}
\title{Remove the temporary files in PortalData and tmp}
\usage{
cleanup_dir(options_all = all_options())
}
\arguments{
\item{options_all}{list containing all options available for controlling
the set up and population of the forecast directory (see 
\code{\link{all_options}})}
}
\value{
nothing
}
\description{
Remove the PortalData and tmp subdirectories.
}
