% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{column_conformer}
\alias{column_conformer}
\title{Conform column names for use internally}
\usage{
column_conformer(df = NULL)
}
\arguments{
\item{df}{\code{data.frame} to have column names converted if needed.}
}
\value{
\code{df} with column names converted if needed.
}
\description{
Existing cast files may have column names using "forecast"
 rather than the current generalized "cast". This function replaces the
 old with the new for internal use.
}
\examples{
 df <- data.frame(xforecastx = 1:10, fore = 2:11)
 column_conformer(df)

}
