% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{combine_casts}
\alias{combine_casts}
\title{Combine casts and append them to existing files}
\usage{
combine_casts(main = ".", moons = prep_moons(main = main),
  end_moon = NULL, cast_date = Sys.Date(), quiet = FALSE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\value{
\code{list} of [1] \code{"casts"} (the casted abundances)
 and [2] \code{"all_model_aic"} (the model AIC values).
}
\description{
Combine all new casts (from the tmp subdirectory) into
 and append the results to the existing files.
}
\examples{
 \donttest{
  setup_dir()
  end_moon <- 520
  prep_data(end_moon = end_moon)
  models_scripts <- models_to_cast(models = c("ESSS", "AutoArima"))
  sapply(models_scripts, source)
  combine_casts(end_moon = end_moon)
 }

}
