% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{combine_forecasts}
\alias{combine_forecasts}
\title{Combine forecasts and append them to existing files}
\usage{
combine_forecasts(tree = dirtree(), quiet = FALSE,
  data_control = list())
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{data_control}{A \code{list} of arguments to control the filling of
the data subdirectory. See \code{\link{data_control}}. 
Values not set assume defaults.}
}
\value{
\code{list} of [1] \code{"forecasts"} (the forecasted abundances)
  and [2] \code{"all_model_aic"} (the model AIC values).
}
\description{
Combine all new forecasts (from the tmp subdirectory) into
  and append the results to the existing files.
}
