% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{covariates_options}
\alias{covariates_options}
\title{Prepare the data options for covariates data}
\usage{
covariates_options(historical = TRUE, forecasts = TRUE, fdate = today(),
  yr = as.numeric(format(today(), "\%Y")), start = 217, lead_time = 12,
  min_lag = 6, fcast_nms = NULL, nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", save = TRUE,
  filename = "covariates.csv", tree = dirtree(), quiet = FALSE)
}
\arguments{
\item{historical}{logical indicator whether or historical covariates are
to be included}

\item{forecasts}{logical indicator whether or forecasted covariates are
to be included}

\item{fdate}{date from which future is defined, typically today's date.}

\item{yr}{year of today}

\item{start}{newmoon number of the first sample to be included. Default 
value is \code{217}, corresponding to 1995-01-01.}

\item{lead_time}{number of moons into the future the rodents are forecast}

\item{min_lag}{the minimum lag time used in any model}

\item{fcast_nms}{newmoon numbers to be forecast (for the rodents, i.e. 
pre-lag)}

\item{nfcnm}{number of forecast newmoons}

\item{append_fcast_csv}{logical if the new forecast should be appended to
the historical forecasts for the purposes of hindcasting}

\item{hist_fcast_file}{name of the file where the historical 
covariate forecasts are held}

\item{source_name}{character value for the name to give the covariaate
forecast. Currently is "current_archive". Previous to "current_archive",
the data were retroactively filled in and are given the source name
"retroactive"}

\item{save}{logical if the data should be saved out}

\item{filename}{the name of the file for the saving}

\item{tree}{directory tree}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
control options list for covariates
}
\description{
Create a list of control options for the covariates data
}
