% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{create}
\alias{create}
\title{Create a specific folder}
\usage{
create(path = NULL, level = NULL, quiet = FALSE)
}
\arguments{
\item{path}{\code{character} vector of the folder paths.}

\item{level}{\code{character} vector of the folder names.}

\item{quiet}{\code{logical} indicator if messages should be quieted.}
}
\description{
If folders (specified by \code{path} and named by
 \code{level}) do not exist, they are created, with a message or not as
 indicated by \code{quiet}.
}
