% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{create_dir}
\alias{create_dir}
\title{Create the full structure of a portalcasting directory}
\usage{
create_dir(options_dir = dir_options())
}
\arguments{
\item{options_dir}{directory options}
}
\value{
Nothing
}
\description{
Create a full portalcasting directory or any missing 
  components.
}
