% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{create_main_dir}
\alias{create_main_dir}
\title{Create the main directory folder of a portalcasting directory}
\usage{
create_main_dir(options_dir = dir_options())
}
\arguments{
\item{options_dir}{directory options}
}
\value{
Nothing
}
\description{
Create the top-level portalcasting directory folder
}
