% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{create_sub_dir}
\alias{create_sub_dir}
\title{Create a subdirectory folder}
\usage{
create_sub_dir(path = NULL, quiet = FALSE)
}
\arguments{
\item{path}{path of the specific folder to be created}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
Nothing
}
\description{
Create a specific subdirectory folder if it does not already
  exist
}
