% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{create_sub_dirs}
\alias{create_sub_dirs}
\title{Create the sub directory folders of a portalcasting directory}
\usage{
create_sub_dirs(tree = dirtree(), quiet = FALSE)
}
\arguments{
\item{tree}{the name tree of the portalcasting directory}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
Nothing
}
\description{
Create a the portalcasting directory subfolders
}
