% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{create_tmp}
\alias{create_tmp}
\title{Create tmp}
\usage{
create_tmp(tree = dirtree())
}
\arguments{
\item{tree}{directory tree}
}
\value{
nothing
}
\description{
Create the tmp subdirectory specifically
}
