% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dataout}
\alias{dataout}
\title{Save data out to a file and return it}
\usage{
dataout(out, options_out = moons_options())
}
\arguments{
\item{out}{data table to be written out}

\item{options_out}{an options list that includes a save element and 
a filename element}
}
\value{
data (as input)
}
\description{
Save data out to a data file if requested (request is handled
  by the save element in the data_options list and the file name is handled
  by the filename element) and return it to the console
}
