% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{dir_options}
\alias{dir_options}
\title{Prepare the directory options for a portalcasting directory}
\usage{
dir_options(base = "~", main = "forecasting", subs = subdirs(),
  quiet = FALSE)
}
\arguments{
\item{base}{name of the base directory where the forecasting 
directory should exist (will be created if it doesn't)}

\item{main}{name of the portalcasting directory}

\item{subs}{names of the portalcasting subdirectories}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
list of settings controlling the directory structure creeatio
}
\description{
Create a list of control options for the directory set-up
}
