% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{directory configuration file}
\alias{directory configuration file}
\alias{write_directory_configuration}
\alias{read_directory_configuration}
\alias{read_directory_settings}
\alias{update_directory_configuration}
\title{Create, Update, and Read the Directory Configuration File}
\usage{
write_directory_configuration(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

read_directory_configuration(main = ".")

read_directory_settings(main = ".")

update_directory_configuration(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the messages.}
}
\value{
\code{list} of directory configurations, \code{\link[base:invisible]{invisible}}-ly.
}
\description{
The directory configuration file is a special file within the directory setup and has its own set of functions. \cr \cr
\code{write_directory_configuration} creates the YAML metadata configuration file. It is (and should only be) called from within \code{\link{setup_dir}}, as it captures information about the compute environment used to instantiate the directory. \cr \cr
\code{read_directory_configuration} reads the YAML config file into the R session. \cr \cr
\code{read_directory_configuration} reads the YAML config file into the R session and pulls just the directory settings list in.
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "standard")
   setup_dir(main = main1)

   settings1 <- read_directory_settings(main = main1)
   config1   <- read_directory_configuration(main = main1)

   unlink(main1, recursive = TRUE)
}

}
\seealso{
Directory orchestration functions:
\code{\link{directory creation}},
\code{\link{directory paths}},
\code{\link{directory settings}}

File read-write functions:
\code{\link{read write data}}
}
\concept{orchestration}
\concept{read-write}
