% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchestrate.R
\name{directory creation}
\alias{directory creation}
\alias{create_dir}
\alias{setup_dir}
\alias{setup_production}
\alias{setup_sandbox}
\title{Create the Structure of a Directory and Fill with Content}
\usage{
create_dir(main = ".", settings = directory_settings(), quiet = FALSE)

setup_dir(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

setup_production(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  settings = production_settings(),
  quiet = FALSE,
  verbose = TRUE
)

setup_sandbox(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  settings = sandbox_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree. Default value (\code{"."}) puts the directory in the present locations.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{datasets}{\code{character} vector of name(s) of rodent dataset(s) to be created.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or not (and thus just the tidy messages).}
}
\value{
The \code{list} of directory settings \code{\link[base]{invisible}}-ly.
}
\description{
Instantiates the necessary folder structure for a directory, writes the setup configuration file, and fills the directory with core content.
}
\examples{
\donttest{
 create_dir()
 fill_dir()
 setup_dir()
 setup_sandbox()
 setup_production()
}

}
