% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{directory paths}
\alias{directory paths}
\alias{paths}
\alias{models_rmd_path}
\alias{models_html_path}
\alias{about_md_path}
\alias{rodents_profiles_html_path}
\alias{rodents_profiles_csv_path}
\alias{rodents_datasets_paths}
\alias{rodents_datasets_controls_path}
\alias{models_controls_path}
\alias{rodents_dataset_path}
\alias{climate_forecasts_paths}
\alias{forecasts_metadata_path}
\alias{forecasts_evaluations_path}
\alias{forecasts_results_path}
\alias{newmoons_path}
\alias{covariates_path}
\alias{metadata_path}
\alias{app_path}
\alias{data_path}
\alias{forecasts_path}
\alias{fits_path}
\alias{models_path}
\alias{resources_path}
\title{Create a List of Full Directory Paths}
\usage{
paths(main = ".")

models_rmd_path(main = ".")

models_html_path(main = ".")

about_md_path(main = ".")

rodents_profiles_html_path(main = ".")

rodents_profiles_csv_path(main = ".")

rodents_datasets_paths(main = ".")

rodents_datasets_controls_path(main = ".")

models_controls_path(main = ".")

rodents_dataset_path(main = ".", dataset = "all")

climate_forecasts_paths(main = ".")

forecasts_metadata_path(main = ".")

forecasts_evaluations_path(main = ".")

forecasts_results_path(main = ".")

newmoons_path(main = ".")

covariates_path(main = ".")

metadata_path(main = ".")

app_path(main = ".")

data_path(main = ".")

forecasts_path(main = ".")

fits_path(main = ".")

models_path(main = ".")

resources_path(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{dataset}{\code{character} value of name of rodent dataset.}
}
\value{
\code{list} of directory paths or specific \code{character} paths.
}
\description{
Upon creation (or updating) of the directory, all the standard file and subdirectory paths are set based on \code{\link{directory_settings}}. \cr \cr
\code{paths} produces the full path \code{list}, whose contents can then also be accessed with specialized functions, see \code{Details}.
}
\details{
Wrapper functions for specific subdirectories and files include:
\itemize{
\item Files
\itemize{
\item \code{rodents_datasets_paths}
\item \code{rodents_dataset_path}
\item \code{climate_forecasts_paths}
\item \code{forecasts_metadata_path}
\item \code{forecasts_evaluations_path}
\item \code{forecasts_results_path}
\item \code{newmoons_path}
\item \code{covariates_path}
\item \code{metadata_path}
\item \code{models_controls_path}
\item \code{models_rmd_path}
\item \code{rodents_profiles_html_path}
\item \code{rodents_profiles_csv_path}
}
\item Subdirectories
\itemize{
\item \code{app_path}
\item \code{data_path}
\item \code{forecasts_path}
\item \code{fits_path}
\item \code{models_path}
\item \code{resources_path}
}
}
}
\examples{
 \donttest{
   main1 <- file.path(tempdir(), "standard")
   create_dir(main = main1)
   paths(main = main1)
   rodents_datasets_paths(main = main1)
   rodents_dataset_path(main = main1)
   rodents_datasets_controls(main = main1)
   climate_forecasts_paths(main = main1)
   forecasts_metadata_path(main = main1)
   forecasts_evaluations_path(main = main1)
   forecasts_results_path(main = main1)
   newmoons_path(main = main1)
   covariates_path(main = main1)
   metadata_path(main = main1)
   models_controls_path(main = main1)
   models_rmd_path(main = main1)
   models_html_path(main = main1)
   about_md_path(main = main1)
   rodents_profiles_html_path(main = main1)
   rodents_profiles_csv_path(main = main1)
   app_path(main = main1)
   data_path(main = main1)
   forecasts_path(main = main1)
   fits_path(main = main1)
   models_path(main = main1)
   resources_path(main = main1)
 }

}
