% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{directory settings}
\alias{directory settings}
\alias{directory_settings}
\alias{production_settings}
\alias{sandbox_settings}
\title{Create a Directory Settings List}
\usage{
directory_settings(
  files = list(directory_configuration = "directory_configuration.yaml", moons =
    "moon_dates.csv", covariates = "covariates.csv", historical_covariates =
    "historical_covariates.csv", forecast_covariates = "forecast_covariates.csv",
    dataset_controls = "dataset_controls.yaml", model_controls = "model_controls.yaml",
    cast_evaluations = "cast_evaluations.csv", metadata = "metadata.yaml"),
  subdirectories = list(forecasts = "forecasts", fits = "fits", models = "models",
    resources = "resources", data = "data"),
  resources = list(PortalData = list(source = "github", version = "latest"),
    portalPredictions = list(source = "github", version = NULL), climate_forecast =
    list(source = "NMME", version = as.character(Sys.Date()), data = c("tasmin",
    "tasmean", "tasmax", "pr"))),
  save = TRUE,
  overwrite = TRUE,
  unzip_pause = 30,
  download_timeout = getOption("timeout")
)

production_settings()

sandbox_settings()
}
\arguments{
\item{files}{\code{list} of \code{character} names for standard files: 
  \describe{
    \item{directory_configuration}{Directory configuration YAML}
    \item{moons}{Lunar data csv}
    \item{covariates}{Combined historical and forecast covariates csv}
    \item{historical_covariates}{Historical covariates csv}
    \item{forecast_covariates}{Forecast covariates csv}
    \item{metadata}{Forecast metadata YAML}
    \item{dataset_controls}{YAML of dataset control list(s)}
    \item{model_controls}{YAML of model controls list(s)}
    \item{cast_evaluations}{Forecast evaluations csv}
}}

\item{subdirectories}{\code{list} of subdirectories: \code{resources}, \code{data}, \code{models}, \code{fits}, and \code{forecasts}.}

\item{resources}{\code{list} of \code{list}s for standard resources: 
\describe{
  \item{PortalData}{\code{source} and \code{version} elements of \code{character} values for the Portal Data download. Default values retrieve the latest data from github}
  \item{portalPredictions}{\code{source} and \code{version} elements of \code{character} values for the archive download. Default values point to github, but \code{version = NULL} indicates no download}
  \item{climate_forecast}{\code{source}, \code{version}, and \code{data} elements of \code{character} values for the climate forecasts download. Default values retrieve the current day's forecast of min, mean, and max temperature and precipitation from the Northwest Knowledge Network's North American Multi-Model Ensemble (NMME) climate forecasts}
}}

\item{save}{\code{logical} indicator controlling if the output should be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing files should be updated (most users should leave as \code{TRUE}).}

\item{unzip_pause}{Positive \code{integer} or integer \code{numeric} seconds for pausing during steps around unzipping that require time delay.}

\item{download_timeout}{Positive \code{integer} or integer \code{numeric} seconds for timeout on downloads. Temporarily overrides the \code{"timeout"} option in \code{\link[base]{options}}.}
}
\value{
Named \code{list} of settings for the directory.
}
\description{
Most users will not want or need to change the directory folders and file names, but it is helpful to have them be flexible for certain circumstances, and this function gathers them into a list for pipeline functionality.
}
