% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{directory setup}
\alias{directory setup}
\alias{setup_dir}
\alias{setup_production}
\alias{setup_sandbox}
\title{Create and Fill a Forecasting Directory}
\usage{
setup_dir(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

setup_production(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  settings = directory_settings(portalPredictions = list(source = "github", version =
    "latest")),
  quiet = FALSE,
  verbose = FALSE
)

setup_sandbox(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{datasets}{\code{character} vector of name(s) of rodent dataset(s) to be created.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information (and thus just the tidy messages).}

\item{new_dataset_controls}{\code{list} of controls for any new datasets (not in the prefab datasets) listed in \code{datasets}.}

\item{new_model_controls}{\code{list} of controls for any new models (not in the prefab models) listed in \code{models}.}
}
\value{
The \code{list} of directory settings \code{\link[base]{invisible}}-ly.
}
\description{
Combines \code{\link{create_dir}} and \code{\link{fill_dir}} to create a ready-to-run (via \code{\link{portalcast}}) directory where indicated. \cr \cr
             \code{setup_production} creates a standard production directory, which includes downloading the most recent archive of the directory. \cr \cr
             \code{setup_sandbox} creates a sandboxing directory. \cr \cr
}
\examples{
\donttest{
 setup_dir()
 setup_sandbox()
 setup_production()
}

}
