% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{directory_config}
\alias{directory_config}
\alias{write_directory_config}
\alias{update_directory_config}
\alias{read_directory_config}
\title{Create, Update, and Read the Directory Configuration File}
\usage{
write_directory_config(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE
)

update_directory_config(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE
)

read_directory_config(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree. Default value (\code{"."}) puts the forecasting directory in the present locations. Nesting the forecasting directory in a folder can be done by simply adding to the \code{main} input (see \code{Examples}).}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}
}
\value{
\code{list} of directory configurations.
}
\description{
The directory configuration file is a special file within the portalcasting directory setup and has its own set of functions. \cr \cr
             \code{write_directory_config} creates the YAML metadata configuration file. It is (and should only be) called from within \code{\link{create_dir}}, as it captures information about the compute environment used to create the directory. \cr \cr
             \code{update_directory_config} adds key components to the file; currently only adding the versions of the downloaded resources from within \code{\link{fill_raw}} (presently the only place it is and should be called from). \cr \cr
             \code{read_directory_config} reads the YAML config file into the R session.
}
