% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{directory_settings}
\alias{directory_settings}
\title{Create a Directory Settings List}
\usage{
directory_settings(directory_config_file = "dir_config.yaml")
}
\arguments{
\item{directory_config_file}{\code{character} value of the path to the directory config YAML.}
}
\value{
Named \code{list} of settings for the directory.
}
\description{
Most users will not want or need to change the directory folders and file names, but it is helpful to have them be flexible for certain circumstances, and this function gathers them into a list for pipeline functionality.
}
