% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{dirtree}
\alias{dirtree}
\alias{update_tree}
\title{Define the names of the directory tree for a forecasting directory}
\usage{
dirtree(base = ".", main = "", subs = subdirs())

update_tree(tree = dirtree(), base = NULL, main = NULL,
  subs = NULL)
}
\arguments{
\item{base}{\code{character} name of the base folder where the directory 
should or does exist. Default \code{"."} (working directory).}

\item{main}{\code{character} name of the main folder within the base that
contains the subdirectories. Default \code{""} (no main level included).}

\item{subs}{\code{character} vector naming the specific subdirectories
within the portalcasting directory tree. Default \code{subdirs()} sets
the subdirectories as \code{"predictions"}, \code{"models"},
\code{"PortalData"}, \code{"data"}, and \code{"tmp"}. It is generally
not advised to change the subdirectories.}

\item{tree}{Directory tree \code{list}. Consisting of \code{base}, 
\code{main}, and \code{subs} elements. See \code{\link{dirtree}}.}
}
\value{
Named \code{list} (elements: \code{base}, 
  \code{main}, and \code{subs}) representing the directory tree.
}
\description{
A cross-platform, tidy way to pass the full tree structure 
  for a portalcasting directory through the codebase. Default setup (for
  using in Portal Predictions) is with the tree based in the working 
  directory and no additional "main" level (so the subdirectories are 
  located within the working directory).
}
\examples{
 dirtree()
 update_tree(dirtree(), base = "~", main = "forecasting_project")

}
