% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download archive}
\alias{download archive}
\alias{download_archive}
\title{Download the Portal Predictions Repository Archive}
\usage{
download_archive(
  main = ".",
  resources_sub = "resources",
  version = "latest",
  source = "github",
  quiet = FALSE,
  verbose = FALSE,
  force = FALSE,
  pause = 30,
  timeout = getOption("timeout")
)
}
\arguments{
\item{main}{\code{character} value defining the main component of the portalcasting directory tree.}

\item{resources_sub}{\code{character} value defining the resources subdirectory of the portalcasting directory tree.}

\item{version}{\code{character} version of the data to download. Default \code{"latest"} downloads the most recent (by date published). \code{NULL} means no download.}

\item{source}{\code{character} indicator of the source for the download. Either \code{"github"} (default) or \code{"zenodo"}.}

\item{quiet}{\code{logic} indicator if progress messages should be quieted.}

\item{verbose}{\code{logic} indicator if detailed messages should be generated.}

\item{force}{\code{logic} indicator of whether or not existing files or folders (such as the archive) should be over-written if an up-to-date copy exists (most users should leave as \code{FALSE}).}

\item{pause}{Positive \code{integer} or integer \code{numeric} seconds for pausing during steps around unzipping that require time delay.}

\item{timeout}{Positive \code{integer} or integer \code{numeric} seconds for timeout on downloads. Temporarily overrides the \code{"timeout"} option in \code{\link[base:options]{options}}.}
}
\value{
\code{NULL}, \code{\link[base:invisible]{invisible}}-ly.
}
\description{
Downloads a specific \code{version} of the Portal Predictions repository from either GitHub or Zenodo (based on \code{source}) into the \verb{<main>/raw} sub.
}
\note{
There are two calls to \code{\link[base:Sys.sleep]{base::Sys.sleep}} for \code{pause} seconds each to allow for the file unzipping, copying, and such to catch up.
}
