% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\alias{unzip_download}
\alias{unzip_destins}
\alias{download_message}
\alias{download_url}
\alias{download_destin}
\title{Download raw components of a forecasting directory}
\usage{
download(name = NULL, type = NULL, url = NULL,
  concept_rec_id = NULL, rec_version = "latest", rec_id = NULL,
  sep_char = ".", main = ".", specific_sub = "raw", quiet = FALSE,
  cleanup = TRUE, NULLname = FALSE)

unzip_download(name = NULL, zip_destin, main = ".", cleanup = TRUE)

unzip_destins(name = NULL, zip_destin, main = ".")

download_message(name = NULL, type = NULL, url = NULL,
  rec_version = NULL, quiet = FALSE)

download_url(type = NULL, url = NULL, concept_rec_id = NULL,
  rec_version = "latest", rec_id = NULL)

download_destin(name = NULL, source_url, main = ".",
  specific_sub = "raw", sep_char = ".")
}
\arguments{
\item{name}{\code{character} value of the component's name, used to create
the folder within the raw subdirectory. If left as \code{NULL},
\code{\link{record_name_from_url}} tries to obtain a name from the URL,
unless \code{NULLname} is set to \code{TRUE}.}

\item{type}{\code{character} value representing the type of input. 
Allowable types currently include a raw URL (\code{type = "url"}), which
requires a non-\code{NULL} input for \code{url}, and from Zenodo
(\code{type = "zenodo"}), which requires non-\code{NULL} input for either 
the concept record identifier (\code{concept_rec_id}) and 
version (\code{rec_version}) or record identifier (\code{rec_id})  
(\strong{if \code{rec_id} is used, it overrides 
\code{concept_rec_id}.}) \cr \cr}

\item{url}{\code{character} value of the URL to be used if 
\code{type = "url"}.}

\item{concept_rec_id}{Concept record identifier, a \code{character} value
corresponding to the Zenodo concept, used when \code{type = "zenodo"}.}

\item{rec_version}{\code{character} value of the version number or 
\code{"latest"} (default) for the data to be download. 
Used when \code{type = "zenodo"}.}

\item{rec_id}{Optional input record identifier, a \code{character} value
corresponding to the Zenodo record. Used when \code{type = "zenodo"}.}

\item{sep_char}{\code{character} value of the separator that delineates
the extension from the file path. Generally, this will be \code{"."},
but for some API URLs, the extension is actually a query component,
so the separator may sometimes need to be \code{"="}.}

\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{specific_sub}{\code{character} of the name of the subdirectory to
download to.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{cleanup}{\code{logical} indicator if any files put into the tmp
subdirectory should be removed at the end of the process.}

\item{NULLname}{\code{logical} indicator if \code{name} should be kept as
\code{NULL}.}

\item{zip_destin}{\code{character} value of the destination of the 
download of the zip, which is to be unzipped.}

\item{source_url}{\code{character} value of the URL from which the download
should occur}
}
\description{
This suite of functions manages the downloading and (if 
 needed) unzipping of raw files associated with the forecasting directory.
 \code{download} downloads a file from a from a website into the 
 directory, unzipping and cleaning up as needed. \cr \cr
 \code{download_message} creates a customized download message. \cr \cr
 \code{download_url} prepares the URL from the inputs, depending on the
 download type (see \code{Details}). \cr \cr
 \code{download_destin} determines the download destination. \cr \cr
 \code{unzip_download} unzips any compressed downloads. \cr \cr
 \code{unzip_destins} determins the unzipping destinations for any 
  compressed downloads.
}
\details{
If \code{type = NULL}, it is assumed to be a URL (\emph{i.e.}, 
 \code{type = "url"}).
}
\examples{
 \donttest{
  create_dir()
  download("PortalData", "zenodo", concept_rec_id = "1215988")
  source_url <- download_url(type = "zenodo", concept_rec_id = "1215988")
  destin <- download_destin(name = "PortalData", source_url = source_url)
  download_message(type = "zenodo", source_url, rec_version = "latest")
  download.file(source_url, destin, mode = "wb")
  unzip_destins("PortalData", destin)
  unzip_download("PortalData", destin)
 }

}
