% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_covariates.R
\name{download_climate_casts}
\alias{download_climate_casts}
\alias{climate_dl_control}
\alias{read_climate_casts}
\title{Download and read covariate casts (presently only the NMME)}
\usage{
download_climate_casts(main = ".",
  control_climate_dl = climate_dl_control(), quiet = TRUE,
  verbose = FALSE, arg_checks = TRUE)

climate_dl_control(start = Sys.Date(), end = as.Date("2050-01-01"),
  model = "ENSMEAN", lat = 31.9555, lon = -109.0744,
  freq = "daily", data = c(mintemp = "tasmin", meantemp = "tasmean",
  maxtemp = "tasmax", precipitation = "pr"), arg_checks = TRUE)

read_climate_casts(main = ".",
  control_climate_dl = climate_dl_control(), arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{control_climate_dl}{\code{list} of specifications for the download, 
which are sent to \code{\link{NMME_urls}} to create the specific URLs.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be
shown.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}

\item{start, end}{\code{Date} for the start and end of the cast.}

\item{model}{\code{character} value of the model, one of \code{"ENSMEAN"},
(Multi-Model Mean), \code{"CMC1"} (CMC1-CanCM3), \code{"CMC2"}
(CMC2-CanCM4), \code{"CFCSv2"} (NCEP-CFSv2), \code{"GFDL"} (GFDL-CM2.1),
\code{"GFDL-FLOR"} (GFDL-FLOR), or \code{"NCAR"} (NCAR-CCSM4). \cr \cr
Presently can only take one value.}

\item{lat, lon}{\code{numeric} latitude and longitude values used to 
downscale the model.  \cr \cr
Presently can only take one value for each.}

\item{freq}{\code{character} value of the frequency of the data, can 
be \code{"daily"} or \code{"XmonthAverage"}, where \code{"X"} is a
number between \code{1} and \code{7}. \cr \cr
Presently can only take one value.}

\item{data}{\code{character} value of the type of data, one of 
\code{"tasmin"} (minimum temperature),  \code{"tasmean"}
(mean temperature), \code{"tasmax"} (maximum temperature), \code{"pr"}
(precipitation), \code{"dps"} (dew point), \code{"rsds"}
(shortwave radiation; sun intensity), \code{"was"} (wind speed).}
}
\value{
Named \code{character} vector of URLs, or \code{NULL} if
 \code{data}, \code{freq}, or \code{model} is \code{NULL}.

\code{download_climate_casts}: downloads the files and returns a
  \code{character} vector of URLs where the downloads came from,
  as generated by \code{\link{NMME_urls}}. \cr \cr
 \code{read_climate_casts}: \code{data.frame} of the downloaded data
  formatted for casting. \cr \cr
 \code{climate_dl_control}: a \code{list} of control arguments to pass
  to \code{\link{NMME_urls}} to generate the URLs for downloading.
}
\description{
Given the details specified in \code{control_cdl}, download
 climate forecasts from the
 \href{https://bit.ly/2MifqjM}{Northwest Knowledge Network}'s (NKN)
 \href{https://bit.ly/2tCP8NX}{simple API} to the 
 \href{https://bit.ly/2Mdv8gd}{North American Multi-Model Ensemble} (NMME)
 and them ready for analyses. \cr \cr
 \code{download_climate_casts}: downloads the files from the server.
 \cr \cr
 \code{read_climate_casts}: reads the downloaded files into R and does
 some minimal tidying: makes sure \code{-9999}s become \code{NA}s, 
 makes sure the \code{date} column is named as such and is actually a 
 \code{Date}, and converts the temperatures to C and the precipitation
 to mm. \cr \cr
 \code{climate_dl_control}: manages the control list for the URL generation
 via \code{\link{NMME_urls}}.
}
\examples{
 \donttest{
  setup_dir()
  download_climate_casts()
  read_climate_casts()
 }

}
