% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{download_climate_forecasts}
\alias{download_climate_forecasts}
\alias{climate_dl_controls}
\title{Download North American Multi-Model Ensemble climate forecasts}
\usage{
download_climate_forecasts(main = ".",
  raw_path_covfcast = "cov_fcasts", controls = list())

climate_dl_controls(start = Sys.Date(), end = Sys.Date(),
  model = "ENSMEAN", lat = 31.9555, lon = -109.0744,
  freq = "daily", data = c(mintemp = "tasmin", meantemp = "tasmean",
  maxtemp = "tasmax", precipitation = "pr"))
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{raw_path_covfcast}{\code{character} value of the path within the
raw subdirectory to the folder where the downloads are saved.}

\item{controls}{\code{list} of specifications for the download, which
are sent to \code{\link{NMME_urls}} to create the specific URLs.}

\item{start}{\code{Date} for the start of the forecast.}

\item{end}{\code{Date} for the end of the forecast.}

\item{model}{\code{character} value of the model, one of \code{"ENSMEAN"},
(Multi-Model Mean), \code{"CMC1"} (CMC1-CanCM3), \code{"CMC2"}
(CMC2-CanCM4), \code{"CFCSv2"} (NCEP-CFSv2), \code{"GFDL"} (GFDL-CM2.1),
\code{"GFDL-FLOR"} (GFDL-FLOR), or \code{"NCAR"} (NCAR-CCSM4). \cr \cr
Presently can only take one value.}

\item{lat, lon}{\code{numeric} latitude and longitude values used to 
downscale the model.  \cr \cr
Presently can only take one value for each.}

\item{freq}{\code{character} value of the frequency of the data, can 
be \code{"daily"} or \code{"XmonthAverage"}, where \code{"X"} is a
number between \code{1} and \code{7}. \cr \cr
Presently can only take one value.}

\item{data}{\code{character} value of the type of data, one of 
\code{"tasmin"} (minimum temperature),  \code{"tasmean"}
(mean temperature), \code{"tasmax"} (mximum temperature), \code{"pr"}
(precipitation), \code{"dps"} (dew point), \code{"rsds"}
(shortwave radiation; sun intensity), \code{"was"} (wind speed).}
}
\value{
Named \code{character} vector of URLs, or \code{NULL} if
 \code{data}, \code{freq}, or \code{model} is \code{NULL}.

\code{download_climate_forecasts}: downloads the files and returns a
  \code{character} vector of URLs where the downloads came from,
  as generated by \code{\link{NMME_urls}}. \cr \cr
 \code{climate_dl_controls}: a \code{list} of control arguments to pass
  to \code{\link{NMME_urls}} to generate the URLs for downloading.
}
\description{
Given the details specified in \code{controls}, download
 climate forecasts from the
 \href{https://bit.ly/2MifqjM}{Northwest Knowledge Network}'s (NKN) at
 \code{https://bit.ly/2tCP8NX}{simple API} to the 
 \href{https://bit.ly/2Mdv8gd}{North American Multi-Model Ensemble} (NMME).
}
\examples{
 \donttest{
 create_dir()
 download_climate_forecasts()
 }

}
