% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictions.R
\name{download_predictions}
\alias{download_predictions}
\title{Download the predictions files from the portalPredictions repo}
\usage{
download_predictions(tree = dirtree())
}
\arguments{
\item{tree}{the name tree of the forecasting directory}
}
\value{
nothing
}
\description{
Download the files into the predictions subdirectory in the
  the tree. Due to the volume of files, the function \code{download.file}
  is forced quiet.
}
