% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{enforce_covariates_control}
\alias{enforce_covariates_control}
\title{Enforce the specific data options for covariate forecasts based on
  existing data.}
\usage{
enforce_covariates_control(control = covariates_control(),
  hist_cov = prep_hist_covariates(), moons = prep_moons())
}
\arguments{
\item{control}{\code{list} of settings controlling the covariates data 
creation. See \code{\link{covariates_options}}.}

\item{hist_cov}{Historical covariate data table as a code{covariates}-class 
\code{data.frame}, returned from \code{\link{prep_hist_covariates}}.}

\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}
}
\value{
An updated \code{list} of control options for covariates.
}
\description{
Update the covariate data control options based on the 
  historical covariate data and moon data. See 
  \code{\link{covariates_control}}.
}
