% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{enforce_rodents_control}
\alias{enforce_rodents_control}
\title{Enforce the specific options for a given rodent data type}
\usage{
enforce_rodents_control(control = rodents_control(), tmnt_type = NULL)
}
\arguments{
\item{control}{\code{list} of settings controlling the rodents data 
creation. See \code{\link{rodents_control}}.}

\item{tmnt_type}{\code{character}-value representation of treatment type 
(\code{"all"} or \code{"controls"}) used to enforce certain elements of 
the \code{control} \code{list}. Alternatively, use \code{NULL}
to not overwrite any options.}
}
\value{
An updated \code{list} of control options for rodents.
}
\description{
Enforce a list of control options for the rodent data 
  (\code{rodents_control}, see \code{\link{rodents_control}}) based on
  the requested data type. Specifically, the \code{tmnt_type}, 
  \code{level}, \code{length}, \code{plots}, \code{output}, 
  \code{treatment}, and \code{filename} elements are set by the selection.
}
\details{
If \code{tmnt_type = "all"}, the elements are \code{tmnt_type} as
  \code{"all"}, \code{level} as \code{"Site"}, \code{length} as 
  \code{"all"}, \code{plots} as \code{"all"}, \code{output} as 
  \code{"abundance"}, \code{treatment} as \code{NULL}, 
  and \code{filename} as \code{"all.csv"}. \cr \cr
  If \code{tmnt_type = "controls"}, the elements are \code{tmnt_type} as
  \code{"controls"}, \code{level} as \code{"Treatment"}, \code{length} as 
  \code{"Longterm"}, \code{plots} as \code{"Longterm"}, \code{output} as 
  \code{"abundance"}, \code{treatment} as \code{controls}, 
  and \code{filename} as \code{"controls.csv"}. \cr \cr
}
