% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{enforce_rodents_options}
\alias{enforce_rodents_options}
\title{Enforce the specific data options for the two rodent data types}
\usage{
enforce_rodents_options(options_rodents = rodents_options(),
  tmnt_type = NULL)
}
\arguments{
\item{options_rodents}{existing control options list for rodents}

\item{tmnt_type}{treatment type ("all" or "controls") used to enfore the 
options list}
}
\value{
updated control options list for rodents
}
\description{
Force a list of control options for the rodent data based on
  the rodent data type
}
