% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembling.R
\name{ensemble_forecasts}
\alias{ensemble_forecasts}
\title{Combine (Ensemble) Casts}
\usage{
ensemble_forecasts(
  main = ".",
  method = "unwtavg",
  forecast_groups = NULL,
  forecast_ids = NULL,
  forecast_table = NULL,
  historic_end_newmoonnumber = NULL,
  models = NULL,
  dataset = NULL,
  species = NULL
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{method}{\code{character} value of the name of the ensemble method to use. Presently, only \code{"unwtavg"} (unweighted average) is allowed.}

\item{forecast_groups}{\code{integer} (or integer \code{numeric}) value of the forecast group to combine with an ensemble. If \code{NULL} (default), the most recent forecast group is ensembled.}

\item{forecast_ids}{\code{integer} (or integer \code{numeric}) values representing the forecasts of interest for restricting ensembling, as indexed within the directory in the \code{casts} sub folder. See the forecasts metadata file (\code{forecasts_metadata.csv}) for summary information.}

\item{forecast_table}{Optional \code{data.frame} of forecast table outputs. If not input, will be loaded.}

\item{historic_end_newmoonnumber}{\code{integer} (or integer \code{numeric}) newmoon number of the forecast origin. Default value is \code{NULL}, which equates to no selection.}

\item{models}{\code{character} value(s) of the name of the model to include. Default value is \code{NULL}, which equates to no selection with respect to \code{model}. \code{NULL} translates to all \code{models} in the table.}

\item{dataset}{\code{character} value of the rodent data set to include Default value is \code{NULL}, which equates to the first data set encountered.}

\item{species}{\code{character} vector of the species code(s) or \code{"total"} for the total across species) to be plotted \code{NULL} translates to the species defined by \code{\link[portalr:rodent_species]{forecasting_species}}.}
}
\value{
\code{data.frame} of ensembled forecasts.
}
\description{
Combine multiple forecasts' output into a single ensemble. Presently, only a general average ensemble is available.
}
\details{
A pre-loaded table of forecasts can be input, but if not (default), the table will be efficiently (as defined by the inputs) loaded and trimmed. \cr
The forecasts can be trimmed specifically using the \code{forecast_ids} input, otherwise, all relevant forecasts from the stated \code{forecast_groups} will be included.
}
