% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESSS.R
\name{ESSS}
\alias{ESSS}
\title{Exponential Smoothing State Space model for Portal Predictions}
\usage{
ESSS(tree = dirtree(), level = "All", quiet = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{level}{\code{character} value name of the type of plots included 
(\code{"All"} or \code{"Controls"}).}

\item{quiet}{\code{logical} value indicating if the function should be 
quiet.}
}
\value{
\code{list} of [1] \code{"forecast"} (the forecasted abundances)
  and [2] \code{"all_model_aic"} (the model AIC values).
}
\description{
Fit an ESSS model in the portalcasting pipeline.
}
\details{
Model "ESSS" is a flexible exponential smoothing state space 
 model with the possibility of multiplicative trends fit using 
 \code{\link[forecast]{ets}} in the
 \href{http://pkg.robjhyndman.com/forecast/}{\code{forecast} package}
 (Hyndman \emph{et al}. 2018).
}
\examples{
\dontrun{

setup_dir()
ESSS()
}

}
\references{
Hyndman R., Bergmeir C., Caceres G., Chhay L., O'Hara-Wild M., Petropoulos
 F., Razbash S., Wang E., and Yasmeen F. 2018. forecast: Forecasting 
 functions for time series and linear models. 
 \href{http://pkg.robjhyndman.com/forecast}{R package version 8.3}.
}
