% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fcast0}
\alias{fcast0}
\title{Create a 0-abundance forecast}
\usage{
fcast0(nfcnm)
}
\arguments{
\item{nfcnm}{number of forecast newmoons}
}
\value{
list of mean and interval values for a 0-abundance filler forecast
}
\description{
Create a 0-abundance forecast
}
