% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{fcast_covariates}
\alias{fcast_covariates}
\title{Build a covariate forecast for model predictions}
\usage{
fcast_covariates(covariate_data, moons, covariate_forecast_newmoons,
  forecast_date = Sys.Date(), min_lag = 6)
}
\arguments{
\item{covariate_data}{historical covariate data table}

\item{moons}{moon data}

\item{covariate_forecast_newmoons}{newmoon numbers for the covariate
forecast}

\item{forecast_date}{date of the forecast (i.e., today)}

\item{min_lag}{minimum (of non-0) lag times that need to be accommodate}
}
\value{
a data.frame with 12 new moons of weather values
}
\description{
combining weather and ndvi forecasts
}
