% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{fcast_weather}
\alias{fcast_weather}
\title{Build a Weather Forecast for Model Predictions}
\usage{
fcast_weather(start = as.numeric(format(Sys.Date(), "\%Y")), moons, lag = 6,
  lead_time = 6)
}
\arguments{
\item{start}{end year of data and beginning of forecast}

\item{moons}{moon data}

\item{lag}{newmoons by which weather data is lagged}

\item{lead_time}{the number of newmoons into the future to obtain forecasts
Max of 7. lag + lead_time should equal 12}
}
\value{
a data.frame with 12 new moons of weather values
}
\description{
created using downscaled climate forecasts
}
