% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{file_path}
\alias{file_path}
\title{Determine the path for a file in the forecasting directory}
\usage{
file_path(tree = dirtree(), local_path = NULL)
}
\arguments{
\item{tree}{the name tree of the forecasting directory}

\item{local_path}{the file path within the forecasting directory}
}
\value{
the full path of the file
}
\description{
Return the path for a specific file
}
