% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{file_paths}
\alias{file_paths}
\title{Determine the path for a file or files in the forecasting directory}
\usage{
file_paths(tree = dirtree(), local_paths = NULL)
}
\arguments{
\item{tree}{\code{dirtree}-class list.}

\item{local_paths}{\code{character} file path(s) within the \code{main}
level of the portalcasting directory.}
}
\value{
The normalized path(s) of the specified file(s) (see 
  \code{\link{normalizePath}}).
}
\description{
Return the normalized path for a specific file or files 
  within the directory.
}
\examples{
\dontrun{

setup_dir()
file_paths(local_paths = "PortalData/Rodents/Portal_rodent_species.csv")
}

}
