% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{files_control}
\alias{files_control}
\title{Create a control list for files}
\usage{
files_control(
  directory = "portalPredictions",
  raw_data = "PortalData",
  filename_moons = "moon_dates.csv",
  filename_config = "dir_config.yaml",
  filename_cov = "covariates.csv",
  filename_cov_casts = "covariate_casts.csv",
  filename_meta = "metadata.yaml",
  save = TRUE,
  overwrite = TRUE,
  cleanup = TRUE,
  source_name = "current_archive",
  append_cast_csv = TRUE
)
}
\arguments{
\item{directory}{\code{character} value of the directory name.}

\item{raw_data}{\code{character} value indicating the name of the raw
data directory. A standard portalcasting directory downloads the raw data
files into from the PortalData repository, so 
\code{raw_data = "PortalData"}.}

\item{filename_moons}{\code{character} name of the file for saving the 
moons data.}

\item{filename_config}{\code{character} filename of the directory
configuration YAML.}

\item{filename_cov}{\code{character} filename for saving the output.}

\item{filename_cov_casts}{\code{character} filename for saving the 
covariate casts output.}

\item{filename_meta}{\code{character} filename for saving the metadata.}

\item{save}{\code{logical} indicator controlling if the output should 
be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing
files should be updated (most users should leave as \code{TRUE}).}

\item{cleanup}{\code{logical} indicator of whether or not the tmp files
should be cleaned up.}

\item{source_name}{\code{character} value for the name to give the 
covariate forecast. Currently is \code{"current_archive"}. Previous to
\code{"current_archive"}, the data were retroactively filled in and are 
given the source name \code{"retroactive"}.}

\item{append_cast_csv}{\code{logical} indicator controlling if the new 
cast covariates should be appended to the historical casts later use.}
}
\value{
Named \code{list} of names of the folders and files within the
 directory structure as well as saving strategies (directly as input).
}
\description{
Most users will not want or need to change data file and 
 folder names or saving conditions, but it is helpful to have them be 
 flexible for certain circumstances, and this function gathers them into a
 list for higher-in-the-pipeline functions.
}
