% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_data}
\alias{fill_data}
\title{Populate the for-use data of a portalcasting directory}
\usage{
fill_data(options_all = all_options())
}
\arguments{
\item{options_all}{list containing all options available for controlling
the set up and population of the forecast directory}
}
\value{
Nothing
}
\description{
Populate the for-use data folder
}
