% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_dir}
\alias{fill_dir}
\title{Populate the base components of a portalcasting directory}
\usage{
fill_dir(options_all = all_options())
}
\arguments{
\item{options_all}{list containing all options available for controlling
the set up and population of the forecast directory}
}
\value{
Nothing
}
\description{
Populate an existing portalcasting directory structure in full
   or any missing components.
}
