% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dir.R
\name{fill_directory}
\alias{fill_directory}
\alias{fill_dir}
\alias{fill_data}
\alias{fill_resources}
\alias{fill_casts}
\alias{fill_fits}
\alias{fill_models}
\title{Fill a Portalcasting Directory with Basic Components}
\usage{
fill_dir(
  main = ".",
  PortalData_source = "gitub",
  PortalData_version = "latest",
  portalPredictions_source = "github",
  portalPredictions_version = NULL,
  climate_forecast_source = "NMME",
  climate_forecast_version = Sys.Date(),
  models = prefab_models(),
  datasets = prefab_datasets(),
  quiet = FALSE,
  verbose = TRUE
)

fill_data(
  main = ".",
  datasets = prefab_datasets(),
  quiet = FALSE,
  verbose = TRUE
)

fill_resources(
  main = ".",
  PortalData_version = "latest",
  PortalData_source = "gitub",
  portalPredictions_version = NULL,
  portalPredictions_source = "github",
  climate_forecast_version = Sys.Date(),
  climate_forecast_source = "NMME",
  quiet = FALSE,
  verbose = FALSE
)

fill_casts(main = ".", quiet = FALSE, verbose = FALSE)

fill_fits(main = ".", quiet = FALSE, verbose = FALSE)

fill_models(
  main = ".",
  models = prefab_models(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value defining the main component of the 
portalcasting directory tree. Default value (\code{"."}) 
puts the directory in the present location.}

\item{PortalData_source, PortalData_version}{\code{character} values for the
source and version of the Portal Data to download. 
\cr \cr 
Default values retrieve the latest data from github.
\cr \cr 
See \code{\link[portalr]{download_observations}}.}

\item{portalPredictions_source, portalPredictions_version}{\code{character} 
values for the source and version of the archive to download.
\cr \cr 
Default values point to github, but \code{version = NULL} indicates
no download.
\cr \cr 
See \code{\link{download_archive}}.}

\item{models}{\code{vector} of models to be written via 
\code{\link{write_model}}.}

\item{datasets}{\code{list} of datasets to be created using 
\code{\link{do.call}} on the defined functions. 
\cr \cr
For example, \code{\link{prefab_datasets}} use 
\code{\link{prepare_rodent_data}}.}

\item{quiet}{\code{logical} indicator if messages should be quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be
printed.}

\item{climate_source, climate_version}{\code{character} values for the
source and version of the climate forecasts to download.
\cr \cr 
Default values retrieve the current day's forecast from the
Northwest Knowledge Network's North American Multi-Model Ensemble 
(NMME) climate forecasts.
\cr \cr 
See \code{\link{download_climate_forecasts}}.}
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Fill the directory with components including (optionally) 
             the raw data (\code{\link[portalr]{download_observations}}), 
             portalPredictions archive (\code{\link{download_archive}}), 
             climate forecasts (\code{\link{download_climate_forecasts}}), 
             models (\code{\link{write_model}}), 
             and model-ready data (\code{\link{prepare_rodent_data}}).
}
\examples{
 \donttest{

  create_dir("./portalcasting")
  fill_dir("./portalcasting")

  create_dir("./pc")
  fill_resources("./pc")
  fill_casts("./pc")
  fill_fits("./pc")
  fill_models("./pc")
 }

}
