% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_models}
\alias{fill_models}
\title{Populate the model scripts of a portalcasting directory}
\usage{
fill_models(options_models = models_options())
}
\arguments{
\item{options_models}{model option list}
}
\value{
Nothing
}
\description{
Populate the model folder with specified scripts to be run
}
