% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_PortalData}
\alias{fill_PortalData}
\title{Populate the raw data of a portalcasting directory}
\usage{
fill_PortalData(tree = dirtree(), quiet = FALSE)
}
\arguments{
\item{tree}{the name tree of the portalcasting directory}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
Nothing
}
\description{
Populate the raw data folder
}
