% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_predictions}
\alias{fill_predictions}
\title{Populate the predictions subdirectory of a portalcasting directory}
\usage{
fill_predictions(options_predictions = predictions_options())
}
\arguments{
\item{options_predictions}{predictions options list}
}
\value{
Nothing
}
\description{
Populate the predictions folder with existing predictions
  housed on the main portalPredictions repo
}
