% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_jags.R
\name{fit_runjags}
\alias{fit_runjags}
\title{Create and Run a runjags Model Run}
\usage{
fit_runjags(abundance, metadata, covariates, model_controls, control_runjags)
}
\arguments{
\item{abundance}{Non-negative \code{integer}-conformable vector of rodent abundances to use in forecasting. See \code{\link{prepare_abundance}}.}

\item{metadata}{\code{list} of model control elements. See \code{\link{prepare_metadata}}.}

\item{covariates}{\code{data.frame} of covariates used in modeling. See \code{\link{prepare_covariates}}.}

\item{model_controls}{\code{list} of model controls for the model of interest. See \code{\link{prefab_model_controls}}.}

\item{control_runjags}{\code{list} of controls for running runjags models. See \code{\link{runjags_control}}. Optional. If not provided here, will be taken from the model controls list.}
}
\value{
An object of class \code{"runjags"} of model components. See \code{\link[runjags]{run.jags}}.
}
\description{
Wraps up the runjags model object preparation functions with the model running (\code{\link[runjags]{run.jags}} function in the runjags (Denwood 2016) package) we use to run JAGS (Plummer 2003) models in portalcasting.
}
\references{
Denwood, M. J. 2016. runjags: an R package providing interface utilities, model templates, parallel computing methods and additional distributions for MCMC models in JAGS. Journal of Statistical Software, 71:9. \href{https://www.jstatsoft.org/article/view/v071i09}{URL}. 
 
 Plummer, M. 2003. JAGS: A program for analysis of Bayesian graphical models using Gibbs Sampling. Proceedings of the 3rd International Workshop on Distributed Statistical Computing (DSC 2003). ISSN 1609-395X. \href{https://bit.ly/33aQ37Y}{URL}.
}
