% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_jags.R
\name{forecast.runjags}
\alias{forecast.runjags}
\title{Forecast a runjags Object}
\usage{
\method{forecast}{runjags}(object, h, level)
}
\arguments{
\item{object}{A \code{runjags}-class object with columns of \code{"X"} values (state variables) in the the \code{mcmc} element.}

\item{h}{\code{integer}-conformable number of steps forward to include in the forecast.}

\item{level}{\code{numeric} of the confidence level to use in summarizing the predictions.}
}
\value{
\code{list} with \code{"forecast"}-class with named elements including \code{"mean"}, \code{"lower"}, \code{"upper"}, and \code{"level"}.
}
\description{
A convenience function for extracting existing forecasts from runjags objects and summarizing them into a \code{"forecast"}-class object.
}
