% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tsglm.R
\name{forecast.tsglm}
\alias{forecast.tsglm}
\title{Forecast a TSGLM Object}
\usage{
\method{forecast}{tsglm}(object, h, level, ...)
}
\arguments{
\item{object}{A \code{tsglm}-class object.}

\item{h}{\code{integer}-conformable number of steps forward to forecast. Passed into \code{predict} as \code{n.ahead}.}

\item{level}{\code{numeric} of the confidence level to use in summarizing the predictions.}

\item{...}{Additional parameters passed into \code{predict}.}
}
\value{
\code{list} with \code{"forecast"}-class with named elements including \code{"mean"}, \code{"lower"}, \code{"upper"}, and \code{"newxreg"} (if provided for prediction) as well as the other elements returned by \code{predict}.
}
\description{
A wrapper around the \code{predict} function for tsglm objects that produces a \code{"forecast"}-class object.
}
