% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoarima.R
\name{forecast_autoarima}
\alias{forecast_autoarima}
\title{Flexible Auto-Regressive Integrated Moving Average model for Portal 
 Predictions}
\usage{
forecast_autoarima(abundances, forecast_date, forecast_months, forecast_years,
  forecast_newmoons, level, num_forecast_newmoons, CI_level = 0.9)
}
\arguments{
\item{abundances}{table of rodent abundances and time measures}

\item{forecast_date}{the dates to be forecast from}

\item{forecast_months}{the months of the dates to be forecast}

\item{forecast_years}{the years of the dates to be forecast}

\item{forecast_newmoons}{the numbers of the new moons to be forecast}

\item{level}{name of the type of plots included ("All" or "Controls")}

\item{num_forecast_newmoons}{number of new moons to forecast}

\item{CI_level}{confidence interval level used for forecast envelope}
}
\value{
list of forecast and aic tables
}
\description{
Model "AutoARIMA" is a flexible auto-regressive integrated 
 moving average model fit to the data using the auto.arima and forecast 
 functions. Unfortunately since the seasonality and sampling occurring with 
 different frequencies, which the auto.arima function cannot accommodate, 
 seasonal models are not included.
 Also, while the auto.arima model can handle missing data, the other models
 used currently cannot, so to provide an appropriate comparison, we 
 interpolate missing data here as well.
}
