% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbGARCH.R
\name{forecast_nbgarch}
\alias{forecast_nbgarch}
\title{negative binomial Generalized AutoRegressive Conditional
  Heteroscedasticity model for Portal Predictions}
\usage{
forecast_nbgarch(abundances, forecast_date, forecast_months, forecast_years,
  forecast_newmoons, level, num_forecast_newmoons, CI_level = 0.9)
}
\arguments{
\item{abundances}{table of rodent abundances and time measures}

\item{forecast_date}{the dates to be forecast from}

\item{forecast_months}{the months of the dates to be forecast}

\item{forecast_years}{the years of the dates to be forecast}

\item{forecast_newmoons}{the numbers of the new moons to be forecast}

\item{level}{name of the type of plots included ("All" or "Controls")}

\item{num_forecast_newmoons}{number of new moons to forecast}

\item{CI_level}{confidence interval level used for forecast envelope}
}
\value{
list of forecast and aic tables
}
\description{
Model "nbGARCH" is a generalized autoregresive conditional 
  heteroscedasticity model with overdispersion (i.e. a negative binomial 
  response variable)
}
