% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{forecast_ndvi}
\alias{forecast_ndvi}
\title{Build an ndvi forecast for model predictions}
\usage{
forecast_ndvi(covariate_data, moons, data_options)
}
\arguments{
\item{covariate_data}{historical covariate data table}

\item{moons}{moon data}

\item{data_options}{covariate data options}
}
\value{
a data.frame with needed foecasted ndvi values
}
\description{
creating forecasts
}
