% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pevGARCH.R
\name{forecast_pevgarch}
\alias{forecast_pevgarch}
\title{poisson environmental variable Generalized Auto-Regressive 
  Conditional Heteroscedasticity model for Portal Predictions}
\usage{
forecast_pevgarch(abundances, forecast_date, forecast_months, forecast_years,
  forecast_newmoons, level, num_forecast_newmoons, CI_level = 0.9,
  weather_data, weather_fcast, ndvi_data, ndvi_fcast)
}
\arguments{
\item{abundances}{table of rodent abundances and time measures}

\item{forecast_date}{the dates to be forecast from}

\item{forecast_months}{the months of the dates to be forecast}

\item{forecast_years}{the years of the dates to be forecast}

\item{forecast_newmoons}{the numbers of the new moons to be forecast}

\item{level}{name of the type of plots included ("All" or "Controls")}

\item{num_forecast_newmoons}{number of new moons to forecast}

\item{CI_level}{confidence interval level used for forecast envelope}

\item{weather_data}{table of already lagged weather data}

\item{weather_fcast}{table of forecasted weather}

\item{ndvi_data}{table of already lagged NDVI data}

\item{ndvi_fcast}{table of forecasted NDVI}
}
\value{
list of forecast and aic tables
}
\description{
Model "pevGARCH" is a generalized autoregresive conditional 
  heteroscedasticity model with a Poisson response variable
  For this model, environmental data are included as predictors of ,
  abundance but with a 6 month lag between the covariate values and the 
  abundances.
}
