% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{forecast_weather}
\alias{forecast_weather}
\title{Build a weather forecast for model predictions}
\usage{
forecast_weather(tree = dirtree(), moons, data_options)
}
\arguments{
\item{tree}{directory tree}

\item{moons}{moon data}

\item{data_options}{covariate data options}
}
\value{
a data.frame with needed foecasted weather values
}
\description{
creating forecasts
}
