% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_tools.R
\name{forecastall}
\alias{forecastall}
\title{Combine Forecasts and Add Ensembles}
\usage{
forecastall(model_metadata)
}
\arguments{
\item{model_metadata}{model metadata list}
}
\value{
list of [1] the forecasts and [2] the model AIC values
}
\description{
Combine all new forecasts (from the tmp directory), add 
  ensembles
}
