% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{format_moons}
\alias{format_moons}
\title{Format the moon data}
\usage{
format_moons(moons = prep_moons())
}
\arguments{
\item{moons}{newmoon table}
}
\value{
appended moon table
}
\description{
Add year and month columns and encode newmoon date as a Date
}
