% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{full_path}
\alias{full_path}
\title{Normalize Path}
\usage{
full_path(reference, base = getwd())
}
\arguments{
\item{reference}{a path to join with current working directory}

\item{base}{Current working directory else path given}
}
\value{
normalized path
}
\description{
Return normalized path for all operating systems
}
\examples{
full_path("PortalData/Rodents/Portal_rodent.csv")
full_path("PortalData/Rodents/Portal_rodent.csv", "~")

}
