% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{get_climate_forecasts}
\alias{get_climate_forecasts}
\title{Download downscaled weather data}
\usage{
get_climate_forecasts(moons = prep_moons(),
  options_covariates = covariates_options())
}
\arguments{
\item{moons}{newmoon data table}

\item{options_covariates}{covariate options control list}
}
\value{
a data.frame with precipitation(mm), temperature(C), year, and
  month. Temperature is the mean temperature for the month, while 
  precipitation is the total forecasted precip.
}
\description{
ENSMEAN (ensemble mean) obtained from 
  https://climate.northwestknowledge.net/RangelandForecast/download.php
  and downscaled to Portal, AZ (31.9555, -109.0744)
}
