% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{get_covariate_data}
\alias{get_covariate_data}
\title{Collect Historic Weather and NDVI Data}
\usage{
get_covariate_data()
}
\value{
covariate data table
}
\description{
Gather the downloaded historical weather and ndvi data
}
