% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{get_moon_data}
\alias{get_moon_data}
\title{Collect Moon Data}
\usage{
get_moon_data(future = FALSE, forecast_date = Sys.Date())
}
\arguments{
\item{future}{logical value indicating if future moons (for forecasts) are
to be added to the table or not}

\item{forecast_date}{date from which forecasting is happening, typically
today's date. Required if \code{future = TRUE}}
}
\value{
newmoons table
}
\description{
Get dates and data for newmoonnumbers and related trapping 
  periods
}
