% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictions.R
\name{get_releases}
\alias{get_releases}
\alias{get_github_releases}
\alias{get_latest_zenodo_release}
\title{Get release information about Portal Predictions versions}
\usage{
get_releases(from_zenodo = TRUE)

get_github_releases()

get_latest_zenodo_release()
}
\arguments{
\item{from_zenodo}{\code{logical} indicator of whether or not the 
predictions should come from Zenodo (they will come from GitHub if not).}
}
\value{
\code{data.frame} of releases with columns named
  \code{version} and \code{zipball_url}.
}
\description{
\code{get_releases}: find and return releases from either
  \href{https://zenodo.org/record/3333770}{Zenodo} (default) or from the 
  \href{https://bit.ly/2Lo2xnQ}{portalPredictions repo}. \cr \cr
  \code{get_latest_zenodo_release}: retrieve latest Zenodo release \cr \cr  
  \code{get_github_releases}: retrieve all GitHub releases \cr \cr
}
