% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{get_rodent_data}
\alias{get_rodent_data}
\title{Collect Historic Rodent Data}
\usage{
get_rodent_data(moons, forecast_date)
}
\arguments{
\item{moons}{current newmoonnumber table}

\item{forecast_date}{date the forecast is run}
}
\value{
a list of two dataframes, all plots and control plots
}
\description{
Get rodent data, tailored for forecasting (all plots and 
  controls only)
}
\examples{
get_rodent_data(moons, forecast_date)
}
