% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{get_weather_data}
\alias{get_weather_data}
\title{Collect Historic Weather Data}
\usage{
get_weather_data()
}
\value{
weather data table
}
\description{
Gather the downloaded historical weather data
}
