% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ifna}
\alias{ifna}
\title{Replace if NA}
\usage{
ifna(x = NULL, alt = NA)
}
\arguments{
\item{x}{Focal input.}

\item{alt}{Alternative value.}
}
\value{
\code{x} if not \code{NA}, \code{alt} otherwise.
}
\description{
If the focal input is \code{NA}, replace it with 
  alternative.
}
\examples{
 ifna(NA, 123)
 ifna(FALSE, 123)
 ifna(NA, NA)

}
