% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ifnull}
\alias{ifnull}
\title{Replace a Value with an Alternative if it is NULL}
\usage{
ifnull(x = NULL, alt = NULL)
}
\arguments{
\item{x}{Focal input.}

\item{alt}{Alternative value.}
}
\value{
\code{x} if not \code{NULL}, \code{alt} otherwise.
}
\description{
Replaces the focal input with the alternative value if it is \code{NULL}.
}
\examples{
   ifnull(NULL, 123)
   ifnull(TRUE, 123)
   ifnull(FALSE, 123)

}
\seealso{
Other utilities: 
\code{\link{file_ext}()},
\code{\link{foy}()},
\code{\link{messages}},
\code{\link{named_null_list}()},
\code{\link{package_version_finder}()},
\code{\link{round_na.interp}()},
\code{\link{update_list}()}
}
\concept{utilities}
