% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{is.spcol}
\alias{is.spcol}
\title{Which columns are species columns}
\usage{
is.spcol(x, spp_names = rodent_spp())
}
\arguments{
\item{x}{data table of interest}

\item{spp_names}{\code{character} vector of species name abbreviations.}
}
\value{
\code{logical} vector corresponding to the columns in \code{x}.
}
\description{
Check column names to verify if they are species names.
}
