% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{is_sp_col}
\alias{is_sp_col}
\title{Determine if columns in a table are species columns}
\usage{
is_sp_col(rodents_tab, species = NULL, total = FALSE, nadot = FALSE,
  arg_checks = TRUE)
}
\arguments{
\item{rodents_tab}{\code{data.frame} of columns to be checked.}

\item{species}{\code{character} vector indicating species names to 
use in determining if columns are species columns. Defaults to
\code{\link{all_species}}.}

\item{total}{\code{logical} indicator if the total should be included.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{logical} vector indicating if each column is a species' 
 column or not.
}
\description{
Given a table, returns a \code{logical} vector indicating
 if each column is a species' column or not.
}
\examples{
 \donttest{
  create_dir()
  fill_raw()
  rodents_tab <- prep_rodents_table()
  is_sp_col(rodents_tab)
 }

}
