% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_jags.R
\name{jags_ss}
\alias{jags_ss}
\title{Run a single-species JAGS model}
\usage{
jags_ss(
  main = ".",
  model_name = NULL,
  dataset = "all",
  settings = directory_settings(),
  control_runjags = runjags_control(),
  jags_model = NULL,
  monitor = NULL,
  inits = NULL,
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{model_name}{\code{character} value name of the model name to save in the cast tab.}

\item{dataset}{\code{character} value name of the rodent data set, such as (\code{"all"} or \code{"controls"}).}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{control_runjags}{\code{list} of arguments passed to \code{\link[runjags]{run.jags}} via \code{\link{runjags_control}}}

\item{jags_model}{\code{character} value of the model. See \code{Examples}.}

\item{monitor}{\code{character} vector of the names of the variables to monitor. See \code{Examples} and \code{\link[runjags]{run.jags}}.}

\item{inits}{\code{function}, \code{character} vector, or \code{list} of values (including random number generation components) for initializing runs of the model. See \code{Examples} and \code{\link[runjags]{run.jags}}.}

\item{quiet}{\code{logical} value indicating if the function should be quiet.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or not (and thus just the tidy messages).}
}
\value{
\code{list} of [1] model metadata \code{list} (\code{"metadata"}), [2] cast summary \code{data.frame} (\code{"cast_tab"}), [3] \code{list} of model fit objects (\code{"model_fits"}), and [4] \code{list} of model cast objects (\code{"model_casts"}).
}
\description{
Provides an API to \code{\link[runjags]{run.jags}} for single-species portalcasting models that requires the user only provide the JAGS model, initializer, monitor, and control list.
}
\details{
The data structure constructed internally, which is available to the \code{jags_model} and \code{inits} components, contains the following elements: \code{count}, \code{moon}, and \code{ntraps} (all three include the forecast horizon as well, with \code{NA} for \code{count}, \code{moon} extended as expected, and \code{ntraps} using the max number of traps); \code{N} (the length of the \code{count} time series); \code{past_count}, \code{past_moon}, and \code{past_ntraps}, which provide all of the historic data that preceded \code{start_moon} (only including samples that are present; all \code{NA} counts are removed as are associated trap count sand moons); \code{past_N} (the length of the \code{past_count} time series); and if desired, \code{covariates} (table of all covariates for the span of \code{moon}). and \code{past_covariates} (table of all covariates for the span of \code{past_moon}).
}
\references{
Denwood, M. J. 2016. runjags: an R package providing interface utilities, model templates, parallel computing methods and additional distributions for MCMC models in JAGS. Journal of Statistical Software, 71:9. \href{https://www.jstatsoft.org/article/view/v071i09}{URL}. 
 
 Plummer, M. 2003. JAGS: A program for analysis of Bayesian graphical models using Gibbs Sampling. Proceedings of the 3rd International Workshop on Distributed Statistical Computing (DSC 2003). ISSN 1609-395X. \href{https://bit.ly/33aQ37Y}{URL}.
}
