% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{lag_covariates}
\alias{lag_covariates}
\title{Lag covariate data}
\usage{
lag_covariates(covariates, lag, tail = FALSE, arg_checks = TRUE)
}
\arguments{
\item{covariates}{\code{data.frame} of covariate data to be lagged.}

\item{lag}{\code{integer} lag between rodent census and covariate data, in
new moons.}

\item{tail}{\code{logical} indicator if the data lagged to the tail end 
should be retained.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{data.frame} with a \code{newmoonnumber} column reflecting
 the lag.
}
\description{
Lag the covariate data together based on the new moons
}
\examples{
 \donttest{
  setup_dir()
  covariate_casts <- read_covariate_casts()
  covar_casts_lag <- lag_covariates(covariate_casts, lag = 2, tail = TRUE)
 }

}
