% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{last_census}
\alias{last_census}
\title{Determine the most recent data collection}
\usage{
last_census(main = ".", settings = directory_settings())
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}
}
\value{
\code{Date} of the last census.
}
\description{
Determine the most recent census.
}
\examples{
 \donttest{
  setup_dir()
  last_census()
 }

}
