% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_depth}
\alias{list_depth}
\title{Determine the depth of a list}
\usage{
list_depth(x)
}
\arguments{
\item{x}{Focal input.}
}
\value{
\code{integer} value of the depth of the list.
}
\description{
Evaluate an input for the depth of its nesting.
}
\examples{
 list_depth("a")
 list_depth(list("a"))
 list_depth(list(list("a")))

}
