% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match.call.defaults}
\alias{match.call.defaults}
\title{Argument matching with defaults}
\usage{
match.call.defaults(definition = sys.function(sys.parent()),
  call = sys.call(sys.parent()), expand.dots = TRUE,
  envir = parent.frame(2L))
}
\arguments{
\item{definition}{A \code{function}, by default the function from which 
\code{match.call.defaults} is called.}

\item{call}{An unevaluated \code{call} to the function specified by 
\code{definition}, as generated by \code{\link[base]{call}}.}

\item{expand.dots}{\code{logical} defining if arguments matching 
\code{...} in the call be included or left as a \code{...} argument}

\item{envir}{An \code{environment}, from which the \code{...} in 
\code{call} are retrieved, if any.}
}
\description{
Expansion of \code{\link[base]{match.call}} to include
 default formal values.
}
\examples{
 fun <- function(arg1 = "ok", ...) {
   match.call.defaults()
 }
 fun()
 fun(arg2 = "hi")

}
\references{
DesignLibrary's \code{match.call.defaults} function. \cr
 Stack overflow post reply by Roland. \href{https://bit.ly/2PtEgy1}{URL}
}
