% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{measure_cast_level_error}
\alias{measure_cast_level_error}
\title{Measure Error and Fit Metrics for Forecasts}
\usage{
measure_cast_level_error(cast_tab = NULL)
}
\arguments{
\item{cast_tab}{A \code{data.frame} of a cast's output. See \code{\link{read_cast_tab}}.}
}
\value{
\code{data.frame} of metrics for each cast for each species.
}
\description{
Summarize the cast-level errors or fits to the observations. \cr 
 Presently included are root mean square error and coverage.
}
