% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{message_break}
\alias{message_break}
\title{Produce a Horizontal Break Line for Messaging}
\usage{
message_break(char = "-", reps = 60)
}
\arguments{
\item{char}{\code{character} value to repeated \code{reps} times to form the break.}

\item{reps}{\code{integer}-conformable value for number of times \code{char} is replicated.}
}
\value{
\code{NULL} (message is put out to console).
}
\description{
Creates a horizontal line of characters for messages.
}
\examples{
 message_break()

}
