% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{messageq_break}
\alias{messageq_break}
\title{Produce a horizontal break line for messaging}
\usage{
messageq_break(bline = TRUE, quiet = FALSE, arg_checks = TRUE)
}
\arguments{
\item{bline}{\code{logical} indicator if horizontal break lines should be
made or not. For toggling separately from the more general \code{quiet}
argument.}

\item{quiet}{\code{logical} indicator if messages should be quieted.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
\code{NULL} (message is put out to console).
}
\description{
Optionally creates and returns a message that is a
 horizontal break line of dashes.
}
\examples{
 messageq_break()

}
