% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tsglm.R
\name{meta_tsglm}
\alias{meta_tsglm}
\title{Create and Run a runjags Model Run}
\usage{
meta_tsglm(
  ts,
  model,
  distr,
  link,
  lag,
  submodels,
  covariates,
  metadata,
  quiet = FALSE
)
}
\arguments{
\item{ts}{Non-negative \code{integer}-conformable vector of rodent abundances to use in forecasting. See \code{\link{prepare_abundance}}.}

\item{model}{A named \code{list} of model linear predictors. See \code{\link[tscount]{tsglm}}.}

\item{distr}{\code{character} of the response distribution. See \code{\link[tscount]{tsglm}}.}

\item{link}{\code{character} of the link funciton. See \code{\link[tscount]{tsglm}}.}

\item{lag}{\code{integer}-conformable value of the number of timesteps used in a bulk lagging for all covariates in all submodels.}

\item{submodels}{\code{list} of \code{character} vectors definind the covariates to include in each of the submodels.}

\item{covariates}{\code{data.frame} of covariates used in modeling. See \code{\link{prepare_covariates}}.}

\item{metadata}{\code{list} of model control elements. See \code{\link{prepare_metadata}}.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}
}
\value{
An object of class \code{"tsglm"} with additional elements defining the submodel and lag.
}
\description{
Wraps up the runjags model object preparation functions with the model running (\code{\link[runjags]{run.jags}} function in the runjags (Denwood 2016) package) we use to run JAGS (Plummer 2003) models in portalcasting.
}
\references{
Denwood, M. J. 2016. runjags: an R package providing interface utilities, model templates, parallel computing methods and additional distributions for MCMC models in JAGS. Journal of Statistical Software, 71:9. \href{https://www.jstatsoft.org/article/view/v071i09}{URL}. 
 
 Plummer, M. 2003. JAGS: A program for analysis of Bayesian graphical models using Gibbs Sampling. Proceedings of the 3rd International Workshop on Distributed Statistical Computing (DSC 2003). ISSN 1609-395X. \href{https://bit.ly/33aQ37Y}{URL}.
}
