% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{metadata_options}
\alias{metadata_options}
\title{Prepare model metadata options}
\usage{
metadata_options(fdate = today(), filename_suffix = "forecasts",
  confidence_level = 0.9, lead_time = 12, save = TRUE,
  filename = "metadata.yaml", quiet = FALSE, tree = dirtree())
}
\arguments{
\item{fdate}{date of the forecast (i.e., today)}

\item{filename_suffix}{"forecasts" or "hindcasts"}

\item{confidence_level}{confidence level used in summarizing model output}

\item{lead_time}{number of newmoons forward to forecast abundances}

\item{save}{logical if the data should be saved out}

\item{filename}{the name of the file for the saving}

\item{quiet}{logical indicator if progress messages should be quieted}

\item{tree}{directory tree}
}
\value{
metadata options list
}
\description{
Create a list of control options for the metadata
}
