% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefabricated_models.R
\docType{data}
\name{model_controls}
\alias{model_controls}
\title{Model Controls List}
\format{
\code{list} of control \code{list}s, each corresponding to a model.
}
\usage{
model_controls
}
\description{
\code{list} of the models available for use via \code{\link{prepare_models}}.
}
\details{
The currently prefabricated models include \code{"AutoArima"}
}
\keyword{datasets}
