% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_options.R
\name{model_options}
\alias{model_options}
\alias{AutoArima_options}
\alias{ESSS_options}
\alias{nbGARCH_options}
\alias{pevGARCH_options}
\title{Prepare options for a portalcasting model}
\usage{
model_options(tree = dirtree(), model = "AutoArima",
  mod_covariates = FALSE, lag = NULL, quiet = FALSE)

AutoArima_options(tree = dirtree(), model = "AutoArima",
  mod_covariates = FALSE, lag = NULL, quiet = FALSE)

ESSS_options(tree = dirtree(), model = "ESSS",
  mod_covariates = FALSE, lag = NULL, quiet = FALSE)

nbGARCH_options(tree = dirtree(), model = "nbGARCH",
  mod_covariates = FALSE, lag = NULL, quiet = FALSE)

pevGARCH_options(tree = dirtree(), model = "pevGARCH",
  mod_covariates = TRUE, lag = 6, quiet = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{model}{\code{character} value of the name of the model.}

\item{mod_covariates}{\code{logical} indicator for if the model requires 
covariates.}

\item{lag}{\code{integer} (or integer \code{numeric}) lag time used for the
covariates or \code{NULL} if \code{covariates} is \code{FALSE}.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}
}
\value{
\code{model_options}: a class-\code{model_options} \code{list} of 
  options used to set up a general model script.

\code{AutoArima_options}: a \code{model_options} \code{list} of 
  settings controlling the \code{\link{AutoArima}} model creation.

\code{ESSS_options}: a \code{model_options} \code{list} of 
  settings controlling the \code{\link{ESSS}} model creation.

\code{nbGARCH_options}: a \code{model_options} \code{list} of 
  settings controlling the \code{\link{nbGARCH}} model creation.

\code{pevGARCH_options}: a \code{model_options} \code{list} of 
  settings controlling the \code{\link{pevGARCH}} model creation.
}
\description{
Suite of functions used to generate model-writing options 
  lists: \cr \cr \code{model_options} is a basic template that provides
  validation of all of the inputs (specific models' functions can thus just 
  wrap around this generalized function). Not to be confused with 
  \code{\link{models_options}}, which creates the options list for the 
  \code{models} subdirectory.

\code{AutoArima_options} creates a \code{model_options} 
  \code{list} of control options for the model script controlling the 
  \code{\link{AutoArima}} model.

\code{ESSS_options} creates a \code{model_options} 
  \code{list} of control options for the model script controlling the 
  \code{\link{ESSS}} model.

\code{nbGARCH_options} creates a \code{model_options} 
  \code{list} of control options for the model script controlling the 
  \code{\link{nbGARCH}} model.

\code{pevGARCH_options} creates a \code{model_options} 
  \code{list} of control options for the model script controlling the 
  \code{\link{pevGARCH}} model.
}
