% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{model_path}
\alias{model_path}
\title{Determine the path for a model in the model sub directory}
\usage{
model_path(tree = dirtree(), model = NULL, extension = ".R")
}
\arguments{
\item{tree}{\code{dirtree}-class list.}

\item{model}{\code{character} name of the specific model.}

\item{extension}{\code{character} file extension (including the period).}
}
\value{
The normalized path of the specified model script (see 
  \code{\link{normalizePath}}).
}
\description{
Return the normalized path for a specific model.
}
