% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{model_paths}
\alias{model_paths}
\title{Determine the path for one or more models in the model sub directory}
\usage{
model_paths(tree = dirtree(), models = NULL, extension = ".R")
}
\arguments{
\item{tree}{Directory tree \code{list}. See \code{\link{dirtree}}.}

\item{models}{\code{character} name of the specific model(s).}

\item{extension}{\code{character} file extension (including the period).}
}
\value{
The normalized path of the specified model script (see 
  \code{\link{normalizePath}}) or \code{NULL} if \code{models = NULL}.
}
\description{
Return the normalized path(s) for a specific model or models.
}
\examples{
 model_paths(models = "AutoArima")

}
