% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{model_scripts}
\alias{model_scripts}
\title{Provide the names of model scripts in the models subdirectory}
\usage{
model_scripts(tree = dirtree())
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}
}
\value{
\code{character} vector of model names.
}
\description{
Lets the user know what model scripts are available.
}
