% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_script_generation.R
\name{model_template}
\alias{model_template}
\title{Create the text to be written out to a script for a model}
\usage{
model_template(tree = dirtree(), model = "autoarima")
}
\arguments{
\item{tree}{directory tree}

\item{model}{names of model script text to write}
}
\value{
Nothing
}
\description{
Create the text for a model script to be housed in the model 
  directory
}
