% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{models}
\alias{models}
\title{Provide the names of models}
\usage{
models(set = "prefab", add = NULL)
}
\arguments{
\item{set}{type of model options (currently only support for "prefab")}

\item{add}{character name of model to add to the list to run}
}
\value{
vector of model names
}
\description{
Based on a \code{set}, returns a character vectory of model
  names to be included. Currently only support for 
  \code{set = "portalcasting"}.
}
