% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{models_control}
\alias{models_control}
\title{Create a models control list}
\usage{
models_control(AutoArima = model_control("AutoArima"),
  ESSS = model_control("ESSS"), nbGARCH = model_control("nbGARCH"),
  nbsGARCH = model_control("nbsGARCH"),
  pevGARCH = model_control("pevGARCH", TRUE, 6), ...)
}
\arguments{
\item{AutoArima, ESSS, nbGARCH, nbsGARCH, pevGARCH}{Control
\code{lists} for each model, each containing \code{name}
(\code{character} value of the model name),
\code{covariates} (\code{logical} indicator of if the model needs
covariates), \code{lag} (\code{integer}-conformable value of the
lag to use with the covariates), and \code{update} (\code{logical}
indicator of if the model script should be overwritten if it's present)
elements.}

\item{...}{Additional \code{list}s or \code{model_control} \code{list}s
defining control options for models to add to the prefab set.}
}
\value{
\code{list} of models metadata, where each element is a model's
  list of model-creation variables (presently: name, covariates, lag, and
  update). The returned \code{list} will contain all models in the prefab 
  set and the models input via \code{control}.
}
\description{
Creates a \code{list} of options for filling the models  
  subdirectory and writing the specific scripts.
}
\details{
Genetates a \code{list} of controls used in setting up the
  model scripts. Presently used elements are name, covariates and
  lag. Any model not in \code{\link{model_names}()} should be provided
  an entry, otherwise a name is autogenerated, and covariate is set
  to FALSE so lag = 0. \code{models_meta_prefab} is an available 
  data structure within the package that allows scaffolding, but input
  values override the existing values.
}
\examples{
  models_control()

}
