% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{models_options}
\alias{models_options}
\title{Prepare the model options for a portalcasting directory}
\usage{
models_options(base = "~", main = "forecasting", subs = subdirs(),
  quiet = FALSE, model = models())
}
\arguments{
\item{base}{name of the base directory where the forecasting 
directory should exist (will be created if it doesn't)}

\item{main}{name of the portalcasting directory}

\item{subs}{names of the portalcasting subdirectories}

\item{quiet}{logical indicator if progress messages should be quieted}

\item{model}{names of model scripts to include}
}
\value{
list of model control options
}
\description{
Create a list of control options for the model set-up
}
