% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{models_to_cast}
\alias{models_to_cast}
\title{Select models to forecast or hindcast with}
\usage{
models_to_cast(options_cast = cast_options())
}
\arguments{
\item{options_models}{models options}
}
\value{
names of files to be run
}
\description{
Verify that models requested are available and select them
}
