% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{moons_options}
\alias{moons_options}
\title{Prepare the data options for moon data}
\usage{
moons_options(n_future_moons = 0, fdate = today(), save = TRUE,
  filename = "moons.csv")
}
\arguments{
\item{n_future_moons}{integer value for the number of future moons to add}

\item{fdate}{date from which future is defined, typically today's date.
Required if \code{n_future_moons > 0}}

\item{save}{logical if the data should be saved out}

\item{filename}{the name of the file for the saving.}
}
\value{
control options list for moons
}
\description{
Create a list of control options for the moon data
}
