% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{most_recent_census}
\alias{most_recent_census}
\title{Determine the most recent data collection}
\usage{
most_recent_census(tree = dirtree())
}
\arguments{
\item{tree}{\code{dirtree}-class list. See \code{\link{dirtree}}.}
}
\value{
\code{Date} of the most recent census.
}
\description{
Determine the most recent census.
}
