% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveArima.R
\name{naiveArima}
\alias{naiveArima}
\title{Naive Arima model for Portal Predictions}
\usage{
naiveArima(tree = dirtree(), level = "All", quiet = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{level}{\code{character} value name of the type of plots included 
(\code{"All"} or \code{"Controls"}).}

\item{quiet}{\code{logical} value indicating if the function should be 
quiet.}
}
\value{
\code{list} of [1] \code{"forecast"} (the forecasted abundances)
  and [2] \code{"all_model_aic"} (the model AIC values).
}
\description{
Fit a Naive Arima model in the portalcasting pipeline.
}
\details{
Model "naiveArima" is a naive baseline model commonly used for  
forecasting. It is sometimes referred to as a random walk forecast with 'no  
drift'. Operationally, it is implemented as an ARIMA(0,1,0) using 
\code{\link[forecast]{Arima}} in the \href{http://pkg.robjhyndman.com/forecast/}{\code{forecast} package}
 (Hyndman \emph{et al}. 2018).
}
\examples{
\dontrun{

setup_dir()
naiveArima()
}

}
\references{
Hyndman R., Bergmeir C., Caceres G., Chhay L., O'Hara-Wild M., Petropoulos
 F., Razbash S., Wang E., and Yasmeen F. 2018. forecast: Forecasting 
 functions for time series and linear models. 
 \href{http://pkg.robjhyndman.com/forecast}{R package version 8.3}.
}
