% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbgarch.R
\name{nbgarch}
\alias{nbgarch}
\title{negative binomial Generalized AutoRegressive Conditional
  Heteroscedasticity model for Portal Predictions}
\usage{
nbgarch(abundances, metadata, level = "All")
}
\arguments{
\item{abundances}{table of rodent abundances and time measures}

\item{metadata}{model metadata list}

\item{level}{name of the type of plots included ("All" or "Controls")}
}
\value{
list of forecast and aic tables
}
\description{
Model "nbGARCH" is a generalized autoregresive conditional 
  heteroscedasticity model with overdispersion (\emph{i.e.}, a negative 
  binomial response variable) fit to the data using \code{tsglm} in the
  \code{tscount} package (Liboschik \emph{et al}. 2017).
}
\references{
Liboschik T., Fokianos K., and Fried R. 2017. tscount: An R Package for 
  Analysis of Count Time Series Following Generalized Linear Models. 
  \emph{Journal of Statistical Software} \strong{82}:5, 1-51. 
  \href{http://doi.org/10.18637/jss.v082.i05}{URL}.
}
