% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchestrate.R
\name{package_version_finder}
\alias{package_version_finder}
\title{Find an Object's Host Package and Version Information}
\usage{
package_version_finder(what)
}
\arguments{
\item{what}{An R object.}
}
\value{
\code{list} of the object, its class, the packages it is sourced from / through, and the versions of those packages.
}
\description{
Locate basic package information of an R object. If nothing is input, it operates on itself. If the object is sourced through multiple packages, each and its version are included.
}
