% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pc_path}
\alias{pc_path}
\title{Portalcasting full path}
\usage{
pc_path(reference, base = getwd())
}
\value{
normalized path to portalcasting
}
\description{
Simplification wrapper to use \code{full_path} to the 
  \code{portalcasting} folder within the base folder
}
