% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_cast_point}
\alias{plot_cast_point}
\alias{plot_cast_point_yaxis}
\title{Plot forecast or hindcast predictions for a given point in time 
 across multiple species.}
\usage{
plot_cast_point(main = ".", species = base_species(total = TRUE),
  level = "Controls", cast_type = "forecast", cast_date = NULL,
  model = "Ensemble", lead = 1, from_date = NULL,
  with_census = FALSE)

plot_cast_point_yaxis(main = ".", species = "total")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{species}{\code{character} vector of the species codes (or 
\code{"total"} for the total across species) to be plotted.}

\item{level}{\code{character} value of the level of interest (\code{"All"} 
or \code{"Controls"}).}

\item{cast_type}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"forecast"}.}

\item{cast_date}{\code{Date} the predictions were made. Used to select the
file in the predictions subdirectory. If \code{NULL} (default), selects
the most recent -casts (if \code{with_census = FALSE}) or the most 
recent -casts with a subsequent census (if \code{with_census = TRUE}).}

\item{model}{\code{character} value of the name (or \code{"Ensemble"}) of
the model to be plotted.}

\item{lead}{\code{integer}-conformable lead of the newmoon number used to
select the data plotted.}

\item{from_date}{\code{Date} to be used as a reference of when to count
the \code{lead} from. If \code{NULL} (default), for 
\code{cast_type = "forecast"}, \code{from_date = cast_date} and 
\code{plot_cast_point} is not yet reliable for 
\code{cast_type = "hindcast"}.}

\item{with_census}{\code{logical} toggle if the plot should include the
observed data collected during the predicted census.}
}
\value{
\code{plot_cast_point}: creates the plot and returns \code{NULL}.
 \code{plot_cast_point_yaxis}: adds the y axis and returns \code{NULL}.
}
\description{
Plot the point value with confidence interval for a step in
 a forecast or hindcast across multiple species. Currently only reliable
 for forecasts. \cr \cr 
 \code{plot_cast_point_yaxis} expands the names with formatting for a 
 multi-species -cast plot y axis.
}
\examples{
 \donttest{
  setup_dir()
  plot_cast_point()
  plot_cast_point(with_census = TRUE)
 }

}
