% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_cast_point_yaxis}
\alias{plot_cast_point_yaxis}
\title{Expand the names with formating for a multi-species -cast plot y 
  axis}
\usage{
plot_cast_point_yaxis(tree = dirtree(), species = "total")
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}. Used to access the species list.}

\item{species}{\code{character} vector of the species codes (or 
\code{"total"} for the total across species) being plotted.}
}
\value{
\code{list} of \code{text} and \code{font} elements.
}
\description{
Add the y-axis of formatted species names to a 
  \code{\link{plot_cast_point}}.
}
