% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_cast_ts}
\alias{plot_cast_ts}
\alias{plot_cast_ts_xaxis}
\alias{plot_cast_ts_ylab}
\title{Visualize a forecast or hindcast}
\usage{
plot_cast_ts(main = ".", species_id = "total",
  tmnt_type = "Controls", cast_type = "forecast", cast_date = NULL,
  model = "Ensemble", start_moon = 300, add_obs = TRUE,
  arg_checks = TRUE)

plot_cast_ts_xaxis(main = ".", rangex, arg_checks = TRUE)

plot_cast_ts_ylab(main = ".", tmnt_type = "Controls",
  species_id = "total", model = "Ensemble", arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{species_id}{\code{character} value of the species code or 
\code{"total"} for the total across species.}

\item{tmnt_type}{\code{character} value of the level of interest 
(\code{"All"} or \code{"Controls"}).}

\item{cast_type}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"forecast"}.}

\item{cast_date}{\code{Date} the predictions were made. Used to select the
file in the predictions subdirectory.}

\item{model}{\code{character} value of the name (or \code{"Ensemble"}) of
the model to be plotted.}

\item{start_moon}{\code{integer}-conformable newmoon number used as the
the minimum x value for the plot.}

\item{add_obs}{\code{logical} indicator if values observed during the 
-cast time should be added (default is \code{TRUE}).}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}

\item{rangex}{\code{integer}-conformable vector of two values corresponding
to the minimum and maximum newmoonnumbers plotted.}
}
\value{
\code{plot_cast_ts}: creates the plot and returns \code{NULL}. \cr \cr
 \code{plot_cast_ts_xaxis}: adds the x axis and returns \code{NULL}.\cr \cr
 \code{plot_cast_ts_ylab}: \code{list} of \code{text} and \code{font} 
  elements.
}
\description{
Plot an observed timeseries and forecast or hindcast 
 timeseries with a prediction interval. Currently only reliable for 
 forecasts. \cr \cr 
 \code{plot_cast_point_xaxis} adds the x-axis to the plot. \cr \cr
 \code{plot_cast_point_ylab} produces a list with \code{text} and 
  \code{font} elements to control the y-axis label
}
\examples{
 \donttest{
  setup_dir()
  plot_cast_ts()
 }

}
