% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_cov_RMSE_mod_spp}
\alias{plot_cov_RMSE_mod_spp}
\title{Plot RMSE and coverage as a function of model for each species}
\usage{
plot_cov_RMSE_mod_spp(tree = dirtree(), cast_type = "hindcasts",
  species = rodent_spp(species_set = "evalplot"), level = "Controls",
  cast_dates = NULL, min_observed = 1,
  models = model_names("wEnsemble"))
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{cast_type}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"hindcasts"}.}

\item{species}{\code{character} vector of the species codes (or 
\code{"total"} for the total across species) to be selected from or 
\code{NULL} to include all species and the total. Default set to 
\code{rodent_spp(set = "evalplot")}, the only setting for which the
plot is reliably coded/}

\item{level}{\code{character} value of the treatment type of interest 
(\code{"All"} or \code{"Controls"}).}

\item{cast_dates}{\code{Date}s the predictions were made. Used to select 
the files in the predictions subdirectory. Can be length 1 or more and if 
\code{NULL} (default), selects all available -casts.}

\item{min_observed}{\code{integer} value for the minimum number of observed
values needed for a -cast to be retained in the output table. Default is
\code{1}, which returns all -casts with any observations. To include all
-casts (even those without any evaluations), set to \code{0}.}

\item{models}{\code{character} value(s) of the name(s) (or 
\code{"Ensemble"}) of the model(s) of interest. If \code{NULL}, all
models are returned.}
}
\description{
Plot the RMSE and coverage by species matrix of cast error 
  as a function of model.
}
\examples{
\dontrun{

setup_dir(all_options(download_existing_predictions = TRUE))
plot_cov_RMSE_mod_spp()
}

}
