% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_covariates}
\alias{plot_covariates}
\title{Visualize a Time Series of Covariates}
\usage{
plot_covariates(main = ".", to_plot = "ndvi")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{to_plot}{\code{character} of the covariate to plot, restricted to column names in the covariates table (see \code{\link{read_covariates}}).}
}
\value{
\code{NULL}. Plot is generated.
}
\description{
Plot an observed timeseries and cast timeseries of the covariates used.
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "standard")
   setup_dir(main = main1)

   plot_covariates(main = main1)

   unlink(main1, recursive = TRUE)
}

}
