% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_err_lead_spp_mods}
\alias{plot_err_lead_spp_mods}
\title{Plot error as a function of lead time for each of a combination of
  species and model}
\usage{
plot_err_lead_spp_mods(tree = dirtree(), cast_type = "forecasts",
  species = rodent_spp(species_set = "evalplot"), level = "Controls",
  ndates = 3, models = model_names("wEnsemble"))
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{cast_type}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"forecasts"}.}

\item{species}{\code{character} vector of the species codes (or 
\code{"total"} for the total across species) to be selected from or 
\code{NULL} to include all species and the total. Default set to 
\code{rodent_spp(set = "evalplot")}, the only setting for which the
plot is reliably coded/}

\item{level}{\code{character} value of the level of interest (\code{"All"} 
or \code{"Controls"}).}

\item{ndates}{\code{integer} number of -cast issue dates to include.}

\item{models}{\code{character} value(s) of the name(s) (or 
\code{"Ensemble"}) of the model(s) of interest. If \code{NULL}, all
models with available -casts are returned.}
}
\description{
Plot the model-species matrix of error as a function of
  lead time.
}
\examples{
\dontrun{

setup_dir(all_options(download_existing_predictions = TRUE))
plot_err_lead_spp_mods()
}

}
