% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_species_casts}
\alias{plot_species_casts}
\title{Plot forecast or hindcast predictions for a given point in time 
  across multiple species.}
\usage{
plot_species_casts(tree = dirtree(), species = NULL,
  level = "Controls", casttype = "forecasts", castdate = today(),
  model = "Ensemble", lead = 1)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{species}{\code{character} vector of the species codes (or 
\code{"total"} for the total across species) to be plotted or 
\code{NULL} (default) to plot all species and the total.}

\item{level}{\code{character} value of the level of interest (\code{"All"} 
or \code{"Controls"}).}

\item{casttype}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"forecasts"}.}

\item{castdate}{\code{Date} the predictions were made. Used to select the
file in the predictions subdirectory.}

\item{model}{\code{character} value of the name (or \code{"Ensemble"}) of
the model to be plotted.}

\item{start_newmoon}{\code{integer}-conformable newmoon number used as the
the minimum x value for the plot.}
}
\description{
Plot the point value with confidence interval for a step in
  a forecast or hindcast across multiple species.
}
