% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plotcastts_xaxis}
\alias{plotcastts_xaxis}
\title{Add a -cast plot x-axis}
\usage{
plotcastts_xaxis(tree, rangex)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}. Used to access the species list.}

\item{rangex}{\code{integer}-conformable vector of two values corresponding
to the minimum and maximum newmoonnumbers plotted.}
}
\description{
Add the x-axis in \code{\link{plot_cast_ts}}.
}
