% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plots}
\alias{plots}
\alias{plot_forecasts_error_lead}
\alias{plot_forecasts_cov_RMSE}
\alias{plot_forecast_point}
\alias{plot_forecast_ts}
\alias{plot_covariates}
\title{Visualize Portalcasting Data and Forecasts}
\usage{
plot_forecasts_error_lead(
  main = ".",
  forecast_ids = NULL,
  historic_end_newmoonnumbers = NULL,
  models = NULL,
  datasets = NULL,
  species = NULL
)

plot_forecasts_cov_RMSE(
  main = ".",
  forecast_ids = NULL,
  historic_end_newmoonnumbers = NULL,
  models = NULL,
  datasets = NULL,
  species = NULL
)

plot_forecast_point(
  main = ".",
  forecast_id = NULL,
  dataset = NULL,
  model = NULL,
  historic_end_newmoonnumber = NULL,
  species = NULL,
  highlight_sp = NULL,
  newmoonnumber = NULL,
  with_census = FALSE
)

plot_forecast_ts(
  main = ".",
  forecast_id = NULL,
  forecast_group = NULL,
  dataset = NULL,
  model = NULL,
  historic_start_newmoonnumber = NULL,
  historic_end_newmoonnumber = NULL,
  species = NULL
)

plot_covariates(main = ".", to_plot = "ndvi")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{species}{\code{character} vector of the species code(s) or \code{"total"} for the total across species) to be plotted \code{NULL} translates to the species defined by \code{\link[portalr:rodent_species]{portalr::forecasting_species}}.}

\item{forecast_id, forecast_ids}{\code{integer} (or integer \code{numeric}) values representing the forecasts of interest for restricting plotting, as indexed within the directory in the \code{casts} sub folder. See the forecasts metadata file (\code{forecasts_metadata.csv}) for summary information. \code{forecast_id} can only be length-1 or \code{NULL}, whereas \code{forecast_ids} is not length-restricted.}

\item{dataset, datasets}{\code{character} value of the rodent data set(s) to include. \code{dataset} can only be length-1 or \code{NULL}, whereas \code{datasets} is not length-restricted.}

\item{model, models}{\code{character} value(s) of the name of the model to include. Default value is \code{NULL}, which equates to no selection with respect to \code{model} or \code{models}. \code{model} can only be length-1 or \code{NULL}, whereas \code{models} is not length-restricted.}

\item{historic_end_newmoonnumber, historic_end_newmoonnumbers}{\code{integer} (or integer \code{numeric}) newmoon number(s) of the forecast origin. Default value is \code{NULL}, which equates to no selection. \code{historic_end_newmoonnumber} can only be length-1 or \code{NULL}, whereas \code{historic_end_newmoonnumbers} is not length-restricted.}

\item{highlight_sp}{\code{character} vector of the species codes (or \code{"total"} for the total across species) to be highlighted or \code{NULL} (default) to not highlight anything.}

\item{newmoonnumber}{\code{integer} (or integer \code{numeric}) newmoon number for the plot.}

\item{with_census}{\code{logical} toggle if the plot should include the observed data collected during the predicted census.}

\item{forecast_group}{\code{integer} (or integer \code{numeric}) value of the forecast group to include. Default value is \code{NULL}, which equates to no selection with respect to \code{forecast_group}.}

\item{historic_start_newmoonnumber}{\code{integer} (or integer \code{numeric}) newmoon number for the beginning of the x-axis of the plot. \cr
Does not influence the fit of the models, just the presentation.}

\item{to_plot}{\code{character} of the covariate to plot, restricted to column names in the covariates table (see \code{\link{read_covariates}}).}
}
\value{
\code{NULL}. Plot is generated.
}
\description{
\code{plot_forecasts_error_lead}: lots the raw error (estimate - observation) as a function of lead time across model runs from different forecast origins for multiple models and multiple species (or total) within a data set. \cr
\code{plot_covariates}: plots an observed timeseries and forecast timeseries of the covariates used. \cr
\code{plot_forecast_ts}: plots an observed timeseries and forecast timeseries with a prediction interval. Observations that occurred after the forecast are shown connected directly to the pre-cast observation data (as the black solid line with points).\cr
\code{plot_forecast_point}: plots the point value with confidence interval for a time point across multiple species. Casts can be selected either by supplying a \code{forecast_id} number or any combination of \code{dataset}, \code{model}, and \code{historic_end_newmoonnumber}, which filter the available forecasts in unison. This plot type can only handle output from a single forecast, so if multiple forecasts still remain, the one with the highest number is selected. To be more certain about forecast selection, use the \code{forecast_id} input.  \cr
\code{plot_forecasts_cov_RMSE}: plots the coverage (fraction of predictions within the CI) and RMSE (root mean squared error) of each model among multiple species.
}
\details{
Casts can be selected either by supplying a \code{forecast_id} number or any combination of \code{dataset}, \code{model}, and \code{historic_end_newmoonnumber}, which filter the available forecasts in unison. This plot type can only handle output from a single forecast, so if multiple forecasts still remain, the one with the highest number is selected. To be more certain about forecast selection, use the \code{forecast_id} input. \cr
As of \verb{portalcasting v0.9.0}, the line and bands in \code{plot_forecast_ts} and point and bars in \code{plot_forecast_point} represent the mean and the 95 percent prediction interval. \cr
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "figures")
   setup_production(main = main1)

   plot_covariates(main = main1)

   portalcast(main = main1, models = "AutoArima")

   ids <- select_forecasts(main     = main3, 
                           species  = c("DM", "PP", "total"),
                           models   = c("AutoArima", "ESSS", "pevGARCH", "nbGARCH", "jags_RW"),
                           datasets = c("all", "controls"))$forecast_id
   nids         <- length(ids)
   nsample_ids  <- 1000
   forecast_ids <- ids[round(seq(1, nids, length.out = nsample_ids))]
   evaluate_forecasts(main         = main3, 
                      forecast_ids = forecast_ids) 

   plot_forecast_ts(main = main1)
   plot_forecast_point(main = main1)
   plot_forecasts_error_lead(main = main1)
   plot_forecasts_cov_RMSE(main    = main1, 
                           models  = "AutoArima", 
                           species = "DM")

   unlink(main1, recursive = TRUE)
}

}
