% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{populate_models}
\alias{populate_models}
\title{Populate a portalcasting models subdirectory}
\usage{
populate_models(model_dir = pc_path("models", "~"), models = c("autoarima",
  "esss", "nbgarch", "pevgarch"), data_dir = pc_path("data", "~"))
}
\arguments{
\item{model_dir}{directory name where the model files will reside}

\item{models}{names of model scripts to include}

\item{data_dir}{directory name where the data files will reside}
}
\value{
Nothing
}
\description{
Create and populate a portalcasting models subdirectory or add
  models to an existing one
}
