% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{portalcast}
\alias{portalcast}
\title{Forecast Portal rodents}
\usage{
portalcast(options_cast = cast_options())
}
\arguments{
\item{options_cast}{options controlling the (fore- or hind-)casting}
}
\value{
Nothing
}
\description{
Main function for controlling the running of potentially 
  multiple models for either a forecast or a hindcast. \cr \cr 
  Note: currently \code{portalcast} can only run "forecasts".
}
