% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{portalcast_welcome}
\alias{portalcast_welcome}
\alias{portalcast_goodbye}
\title{Directory welcome and goodbye messages}
\usage{
portalcast_welcome(quiet = FALSE, arg_checks = TRUE)

portalcast_goodbye(quiet = FALSE, arg_checks = TRUE)
}
\arguments{
\item{quiet}{\code{logical} indicator if message should be quieted.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{NULL}, as message is printed.
}
\description{
Create a welcome message for the directory based on the 
 package version or a goodbye message for when models are done.
}
\examples{
 portalcast_welcome()
 portalcast_goodbye()

}
