% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp.R
\name{portalcasting app helpers}
\alias{portalcasting app helpers}
\alias{available_newmoonnumbers}
\alias{selected_newmoonnumber}
\alias{available_historic_end_newmoonnumbers}
\alias{selected_historic_end_newmoonnumber}
\alias{available_species}
\alias{selected_species}
\alias{available_datasets}
\alias{selected_dataset}
\alias{available_models}
\alias{selected_model}
\title{Helper Functions for the Portalcasting App}
\usage{
available_newmoonnumbers(global = global_list(), event_name, rv = NULL)

selected_newmoonnumber(global = global_list(), event_name, rv = NULL)

available_historic_end_newmoonnumbers(
  global = global_list(),
  event_name,
  rv = NULL
)

selected_historic_end_newmoonnumber(
  global = global_list(),
  event_name,
  rv = NULL
)

available_species(global = global_list(), event_name, rv = NULL)

selected_species(global = global_list(), event_name, rv = NULL)

available_datasets(global = global_list(), event_name, rv = NULL)

selected_dataset(global = global_list(), event_name, rv = NULL)

available_models(global = global_list(), event_name, rv = NULL)

selected_model(global = global_list(), event_name, rv = NULL)
}
\arguments{
\item{global}{A \code{list} of global values for the app.}

\item{event_name}{\code{character} value of the server event. \cr
Options include \code{"initial_forecast_tab"}, \code{"initial_evaluation_tab"}, \code{"forecast_tab_species"}, \code{"forecast_tab_dataset"}, \code{"forecast_tab_model"}, \code{"forecast_tab_historic_end_newmoonnumber"}, \code{"evaluation_tab_species"}, \code{"evaluation_tab_dataset"}, \code{"evaluation_tab_model"}, \code{"evaluation_tab_historic_end_newmoonnumber"}, and \code{"evaluation_tab_newmoonnumber"})}

\item{rv}{\code{\link[shiny:reactiveValues]{reactiveValues}} \code{list} for the UI.}
}
\value{
\code{character} or \code{integer}, depending upon the function.
}
\description{
Construct vectors of available choices and make selections.
}
