% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp_server.R
\name{portalcasting app server}
\alias{portalcasting app server}
\alias{web-app-server}
\alias{app-server}
\alias{server}
\alias{app_server}
\alias{initial_reactive_values}
\alias{initial_output}
\alias{event_reaction}
\alias{update_reactive_values}
\alias{update_output}
\alias{update_input}
\title{Generate the Server Code for the Web App}
\usage{
app_server(input, output, session)

initial_reactive_values(global = global_list())

initial_output(main = ".", global, rv, output)

event_reaction(main, global, event, rv, input, output, session)

update_reactive_values(event, rv, input)

update_output(main, global, event, rv, input, output)

update_input(global, event, rv, input, session)
}
\arguments{
\item{input}{\code{input} \code{list} for the UI.}

\item{output}{\code{output} \code{list} for the UI.}

\item{session}{Environment for the UI.}

\item{global}{A \code{list} of global values for the app.}

\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{rv}{\code{\link[shiny:reactiveValues]{reactiveValues}} \code{list} for the UI.}

\item{event}{\code{character} value of the server event. \cr
Options include \code{"forecast_tab_species"}, \code{"forecast_tab_dataset"}, \code{"forecast_tab_model"}, \code{"forecast_tab_historic_end_newmoonnumber"}, \code{"evaluation_tab_species"}, \code{"evaluation_tab_dataset"}, \code{"evaluation_tab_model"}, and \code{"evaluation_tab_historic_end_newmoonnumber"})}
}
\value{
\code{app_server}: an observer reference class object (see \code{\link[shiny:observeEvent]{observeEvent}} and \code{\link[shiny:observe]{observe}}). \cr
\code{initial_reactive_values}: a \code{\link[shiny:reactiveValues]{reactiveValues}} \code{list}. \cr
\code{initial_output}: an \code{output} \code{list}. \cr
\code{event_reaction}: updates the \code{rv}, \code{output}, and \code{input} \code{list}s, but does not return them, per se. \cr
\code{update_reactive_values}: a \code{\link[shiny:reactiveValues]{reactiveValues}} \code{list}. \cr
\code{update_output}: an \code{output} \code{list}. \cr
\code{update_input}: updates the \code{input} \code{list}, but does not return it.
}
\description{
Generate the Server Code for the Web App
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "app")
   setup_dir(main = main1)

   global <- global_list(main = main1)
 
   rv     <- initial_reavtive_values(global = global)

   output <- initial_output(main   = main1,
                            global = global,
                            rv     = rv,
                            output = list())

   unlink(main1, recursive = TRUE)
 }

}
\seealso{
Other shinyapp: 
\code{\link{portalcasting app selection helpers}},
\code{\link{portalcasting app ui}},
\code{\link{portalcasting app}}
}
\concept{shinyapp}
