% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp_ui.R
\name{portalcasting app ui}
\alias{portalcasting app ui}
\alias{app_ui}
\alias{app_theme}
\alias{main_panel}
\alias{title_panel}
\alias{subtitle_panel}
\alias{about_tab}
\alias{models_tab}
\alias{rodents_profiles_tab}
\alias{forecast_tab}
\alias{forecast_tab_input_selection_checks_row}
\alias{forecast_tab_input_selection_row}
\alias{forecast_tab_input_selection_row_species}
\alias{forecast_tab_input_selection_row_dataset}
\alias{forecast_tab_input_selection_row_model}
\alias{forecast_tab_input_selection_row_historic_end_newmoonnumber}
\alias{evaluation_tab}
\alias{evaluation_tab_input_selection_checks_row}
\alias{evaluation_tab_input_selection_row}
\alias{evaluation_tab_input_selection_row_species}
\alias{evaluation_tab_input_selection_row_dataset}
\alias{evaluation_tab_input_selection_row_model}
\alias{evaluation_tab_input_selection_row_historic_end_newmoonnumber}
\alias{evaluation_tab_input_selection_row_newmoonnumber}
\alias{covariates_tab}
\alias{data_sources_section}
\alias{portal_project_href}
\alias{portal_data_href}
\alias{portal_weather_href}
\alias{portal_ndvi_href}
\alias{write_rodents_profiles_tab_html}
\alias{write_models_tab_html}
\title{Generate the User Interface for the Web App}
\usage{
app_ui(global = global_list())

app_theme()

main_panel(global = global_list())

title_panel()

subtitle_panel()

about_tab(global = global_list())

models_tab(global = global_list())

rodents_profiles_tab(global = global_list())

forecast_tab(global = global_list())

forecast_tab_input_selection_checks_row()

forecast_tab_input_selection_row(global = global_list())

forecast_tab_input_selection_row_species(global = global_list())

forecast_tab_input_selection_row_dataset(global = global_list())

forecast_tab_input_selection_row_model(global = global_list())

forecast_tab_input_selection_row_historic_end_newmoonnumber(
  global = global_list()
)

evaluation_tab(global = global_list())

evaluation_tab_input_selection_checks_row()

evaluation_tab_input_selection_row(global = global_list())

evaluation_tab_input_selection_row_species(global = global_list())

evaluation_tab_input_selection_row_dataset(global = global_list())

evaluation_tab_input_selection_row_model(global = global_list())

evaluation_tab_input_selection_row_historic_end_newmoonnumber(
  global = global_list()
)

evaluation_tab_input_selection_row_newmoonnumber(global = global_list())

covariates_tab(global = global_list())

data_sources_section()

portal_project_href(text = "The Portal Project")

portal_data_href(text = "The Portal Data Repository")

portal_weather_href(text = "weather data")

portal_ndvi_href(text = "NDVI data")

write_rodents_profiles_tab_html(main = ".")

write_models_tab_html(main = ".")
}
\arguments{
\item{global}{A \code{list} of global values for the app.}

\item{text}{\code{character} value of the text used in \code{\link[htmltools:builder]{htmltools::a}}.}

\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
A UI definition, component shiny tags, or bootswatch theme.
}
\description{
\code{app_ui} constructs the user interface (UI) for the web application by updating the static pages (models and rodent profiles) then running \code{\link[shiny:fluidPage]{fluidPage}} on the UI components. \cr
\code{write_rodent_profiles_tab_html} and \code{write_model_tab_html} build and write-out static html files for the rodent profiles and models tabs during \code{\link{fill_app}}. \cr
The \verb{<_>_href} functions provide simplified calls to hyperlinked texts that are repeatedly used. \cr
See \code{Details} for hierarchy of functions.
}
\details{
The UI is hierarchical built as:
\itemize{
\item \code{app_ui}
\itemize{
\item \code{title_panel}
\item \code{subtitle_panel}
\item \code{main_panel}
\itemize{
\item \code{forecast_tab}
\itemize{
\item \code{forecast_tab_input_selection_row}
\itemize{
\item \code{forecast_tab_input_selection_row_species}
\item \code{forecast_tab_input_selection_row_dataset}
\item \code{forecast_tab_input_selection_row_model}
\item \code{forecast_tab_input_selection_row_historic_end_newmoonnumber}
}
\item \code{forecast_tab_input_selection_checks_row}  # commented out, but available for checking reactive inputs in dev
\item \code{plot_forecast_ts}
\item \code{plot_forecast_point}
}
\item \code{evaluation_tab}
\itemize{
\item \code{evaluation_tab_input_selection_row}
\itemize{
\item \code{evaluation_tab_input_selection_row_species}
\item \code{evaluation_tab_input_selection_row_dataset}
\item \code{evaluation_tab_input_selection_row_model}
\item \code{evaluation_tab_input_selection_row_historic_end_newmoonnumber}
\item \code{evaluation_tab_input_selection_row_newmoonnumber}
}
\item \code{evaluation_tab_input_selection_checks_row}  # commented out, but available for checking reactive inputs in dev
\item \code{plot_forecast_point}
\item \code{plot_forecasts_cov_RMSE}
}
\item \code{about_tab}
\itemize{
\item \code{htmltools::includeMarkdown}
}
\item \code{models_tab}
\itemize{
\item \code{htmltools::includeHTML}
}
\item \code{rodents_profiles_tab}
\itemize{
\item \code{htmltools::includeHTML}
}
\item \code{covariates_tab}
\itemize{
\item \code{data_sources_section}
}
}
}
}
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "app_ui")
   setup_dir(main = main1)

   global <- global_list(main = main1)
 
   app_ui(global = global)
   title_panel( )
   subtitle_panel( )
   main_panel(global = global)
   forecast_tab(global = global)
   forecast_tab_input_selection_row(global = global)
   forecast_tab_input_selection_row_species(global = global)
   forecast_tab_input_selection_row_dataset(global = global)
   forecast_tab_input_selection_row_model(global = global)
   forecast_tab_input_selection_row_historic_end_newmoonnumber(global = global)
   forecast_tab_input_selection_checks_row( )
   evaluation_tab(global = global)
   evaluation_tab_input_selection_row(global = global)
   evaluation_tab_input_selection_row_species(global = global)
   evaluation_tab_input_selection_row_dataset(global = global)
   evaluation_tab_input_selection_row_model(global = global)
   evaluation_tab_input_selection_row_historic_end_newmoonnumber(global = global)
   evaluation_tab_input_selection_row_newmoonnumber(global = global)
   evaluation_tab_input_selection_checks_row( )
   about_tab( )
   models_tab(global = global)
   rodents_profiles_tab(global = global)
   covariates_tab(global = global)
   data_sources_section( )

   unlink(main1, recursive = TRUE)
 }

}
\seealso{
Other shinyapp: 
\code{\link{portalcasting app selection helpers}},
\code{\link{portalcasting app server}},
\code{\link{portalcasting app}}
}
\concept{shinyapp}
