% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp.R
\name{portalcasting app}
\alias{portalcasting app}
\alias{run_app}
\alias{global_list}
\title{Build and Launch the Portal Forecast Web Application}
\usage{
run_app(main = ".")

global_list(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
\code{global_list}: a \code{list} of values that will be globally available in the app. \cr
\code{run_app} runs the app itself.
}
\description{
\code{run_app} constructs and launches a local version of the web application by running \code{\link[shiny:runApp]{shiny::runApp}} pointed to the \code{app} subdirectory in the local \code{portalcasting} package folder. \cr \cr
\code{global_list} creates a list of values that are globally available in an app run.
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "app")
   setup_dir(main = main1)

   global_list(main = main1)
 

   if (getShinyOption("shiny.launch.browser")) {
     run_app(main = main1)
   }

   unlink(main1, recursive = TRUE)
 }

}
\seealso{
Other shinyapp: 
\code{\link{portalcasting app selection helpers}},
\code{\link{portalcasting app server}},
\code{\link{portalcasting app ui}}
}
\concept{shinyapp}
