% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcasting.R
\docType{package}
\name{portalcasting-package}
\alias{portalcasting}
\alias{portalcasting-package}
\title{portalcasting: Functions Used in Predicting Portal Rodent Dynamics}
\description{
Create a directory, populate it with data, build models, forecast rodent populations, and visualize the results. This package provides the functionality underlying the Portal Predictions Project continuous integration system \url{https://github.com/weecology/portalpredictions} and website \url{https://portal.naturecast.org} as well as support for exploring and developing new models in sandbox environments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://weecology.github.io/portalcasting/}
  \item \url{https://github.com/weecology/portalcasting}
  \item Report bugs at \url{https://github.com/weecology/portalcasting/issues}
}

}
\author{
\strong{Maintainer}: Juniper L. Simonis \email{juniper.simonis@weecology.org} (\href{https://orcid.org/0000-0001-9798-0460}{ORCID})

Authors:
\itemize{
  \item Glenda M. Yenni (\href{https://orcid.org/0000-0001-6969-1848}{ORCID})
  \item Ellen K. Bledsoe (\href{https://orcid.org/0000-0002-3629-7235}{ORCID})
  \item Erica M. Christensen (\href{https://orcid.org/0000-0002-5635-2502}{ORCID})
  \item Henry Senyondo (\href{https://orcid.org/0000-0001-7105-5808}{ORCID})
  \item Hao Ye (\href{https://orcid.org/0000-0002-8630-1458}{ORCID})
  \item Ethan P. White (\href{https://orcid.org/0000-0001-6728-7745}{ORCID})
  \item S. K. Morgan Ernest (\href{https://orcid.org/0000-0002-6026-8530}{ORCID})
}

}
\keyword{internal}
