% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcasting.R
\docType{package}
\name{portalcasting}
\alias{portalcasting}
\alias{portalcasting-package}
\title{Functions for Portal Forecasting}
\description{
This package contains the functions used for continuous
  analysis and forecasting of 
  \href{https://portal.weecology.org/}{Portal rodent populations},
  as implemented in the 
  \href{https://github.com/weecology/portalPredictions}{Portal Predictions repository}.
  Functions in \code{portalcasting} are designed to be portable, allowing
  users to set up a fully-functional replica repository for development
  and testing of new models.
}
\keyword{package}
