% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{PortalData_control}
\alias{PortalData_control}
\title{Create a PortalData control list}
\usage{
PortalData_control(version = "latest", from_zenodo = TRUE,
  update = TRUE)
}
\arguments{
\item{version}{\code{character} values of the version number or 
\code{"latest"} (default) for the Portal Data to be download.}

\item{from_zenodo}{\code{logical} indicator of whether or not the Portal 
Data should come from Zenodo (come from GitHub if not).}

\item{update}{\code{logical} indicator of whether or not the Portal Data
should be updated.}
}
\value{
A \code{list} of settings controlling the population of the
  PortalData subdirectory.
}
\description{
Creates a \code{list} of options for filling the PortalData  
  subdirectory, setting any non-set options to default values.
}
