% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PortalPredictions}
\alias{PortalPredictions}
\title{Portal Predictions repository setup list}
\format{A \code{list} of elements used as inputs to the functions to 
 set up the directory. Specific elements include:}
\source{
\href{https://bit.ly/2Lo2xnQ}{Portal Predictions repository}
}
\usage{
PortalPredictions
}
\description{
The list of inputs used to set up the continuous analysis 
 and forecasting of
 \href{https://portal.weecology.org/}{Portal rodent populations}, as
 implemented in the 
 \href{https://bit.ly/2Lo2xnQ}{Portal Predictions repository}.
}
\keyword{datasets}
