% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dir.R
\name{post_process}
\alias{post_process}
\title{After portalcast and evaluate_forecast run post_process to zip files}
\usage{
post_process(main = ".")
}
\arguments{
\item{main}{location of main directory with a forecast dir}
}
\description{
Zips both main/forecasts and resources/portal-forecasts/forecasts (if it exists).
All computation (portalcast, evaluate_forecasts) writes to main/forecasts.
}
