% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{predictions_control}
\alias{predictions_control}
\title{Create a predictions control list}
\usage{
predictions_control(download = FALSE, version = "latest",
  from_zenodo = TRUE, update = TRUE, verbose = TRUE,
  use_existing_zip = TRUE)
}
\arguments{
\item{download}{\code{logical} indicator for if the existing predictions 
files should be retrieved.}

\item{version}{\code{character} values of the version number or 
\code{"latest"} (default) for the predictions to be download. 
Versioning of the predictions on Zenodo is done by date.}

\item{from_zenodo}{\code{logical} indicator of whether or not the 
predictions should come from Zenodo (they will come from GitHub if not).}

\item{update}{\code{logical} indicator of whether or not the predictions
should be updated.}

\item{verbose}{\code{logical} indicator of whether or not to print out
all of the details.}

\item{use_existing_zip}{\code{logical} indicator of whether an existing
zip in the temp directory should be used or overwritten when downloading 
files.}
}
\value{
A \code{list} of settings controlling the population of the
  PortalData subdirectory.
}
\description{
Creates a \code{list} of options for filling the predictions  
  subdirectory, setting any non-set options to default values.
}
\details{
Predictions can be downloaded from 
  \href{https://zenodo.org/record/3333770}{Zenodo} (default) or from the 
  \href{https://bit.ly/2Lo2xnQ}{portalPredictions repo}.
}
