% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{predictions_options}
\alias{predictions_options}
\title{Prepare the predictions options for a portalcasting directory}
\usage{
predictions_options(base = "~", main = "forecasting", subs = subdirs(),
  download_existing_predictions = TRUE, quiet = FALSE)
}
\arguments{
\item{base}{name of the base directory where the forecasting 
directory should exist (will be created if it doesn't)}

\item{main}{name of the portalcasting directory}

\item{subs}{names of the portalcasting subdirectories}

\item{download_existing_predictions}{logical indicator is the existing
predictions files should be retrieved from the portalPredictions repo}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
list of settings controlling the directory structure creeatio
}
\description{
Create a list of control options for populating the 
  predictions subdirectory
}
