% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{prefab_rodent_datasets}
\alias{prefab_rodent_datasets}
\title{Provide the Names of the Prefabricated Rodent Datasets}
\usage{
prefab_rodent_datasets()
}
\value{
\code{character} vector of model names.
}
\description{
Create a \code{character} vector of the names of the datasets included including the prefabricated (prefab) set. \cr \cr
             The currently prefabricated datasets include \code{"all"}, \code{"all_interp"}, \code{"controls"}, \code{"controls_interp"}, \code{"exclosures"}, \code{"exclosures_interp"}, \code{"dm_controls"}, and \code{"dm_controls_interp"}.
}
\examples{
 prefab_models()

}
