% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab_models.R
\name{prefabricated models}
\alias{prefabricated models}
\alias{prefab_models_controls}
\alias{prefab_models}
\title{Provide the Names or Controls for the Prefabricated Models}
\usage{
prefab_models_controls()

prefab_models()
}
\value{
\code{prefab_models}: \code{character} vector of model names. \cr
\code{prefab_models_controls}: \code{list} of model controls.
}
\description{
Create a \code{character} vector of the names of the pre-fabricated (prefab) models or a \code{list} of their controls.
}
\examples{
  prefab_models_controls( )
  prefab_models( )

}
\seealso{
Helper functions for prefab models:
\code{\link{runjags models}},
\code{\link{tsglm models}}
}
\concept{prefabmodels}
