% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab_rodents_datasets.R
\name{prefabricated rodents datasets}
\alias{prefabricated rodents datasets}
\alias{prefab-rodents}
\alias{prefab-datasets}
\alias{prefab-rodents-datasets}
\alias{prefab_datasets}
\alias{prefab_datasets_controls}
\alias{prefab_species}
\title{Provide the Names or Controls for the Prefabricated Rodent Datasets}
\usage{
prefab_datasets()

prefab_datasets_controls()

prefab_species(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
\code{prefab_datasets}: \code{character} vector of dataset names. \cr
\code{prefab_datasets_controls}: \code{list} of dataset controls. \cr
\code{prefab_species}: \code{character} vector of species abbreviations.
}
\description{
Create a \code{character} vector of the names of the pre-fabricated (prefab) rodent datasets or species or a \code{list} of their controls
}
\examples{
  prefab_datasets_controls( )
  prefab_datasets( )

}
