% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab_rodents_datasets.R
\name{prefabricated_datasets}
\alias{prefabricated_datasets}
\alias{prefab_datasets}
\alias{prefab_dataset_controls}
\title{Provide the Names or Controls for the Prefab Rodent Datasets}
\usage{
prefab_datasets(interpolate = NULL)

prefab_dataset_controls(interpolate = NULL)
}
\arguments{
\item{interpolate}{\code{logical} value indicating if the interpolated data only should be listed (\code{interpolate = TRUE}), if the non-interpolated data only should be listed (\code{interpolate = FALSE}), or if both should be listed (\code{interpolate = NULL}, default).}
}
\value{
\code{prefab_datasets}: \code{character} vector of dataset names. \cr
        \code{prefab_dataset_controls}: \code{list} vector of dataset controls. \cr
}
\description{
Create a \code{character} vector of the names of the pre-fabricated (prefab) rodent datasets or a \code{list} of their controls
}
\examples{
 prefab_datasets()
 prefab_dataset_controls()

}
