% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab_models.R
\name{prefabricated_models}
\alias{prefabricated_models}
\alias{prefab_model_controls}
\alias{prefab_models}
\title{Provide the Names or Controls for the Prefab Models}
\usage{
prefab_model_controls()

prefab_models()
}
\value{
\code{prefab_models}: \code{character} vector of model names. \cr
        \code{prefab_model_controls}: \code{list} vector of model controls. \cr
}
\description{
Create a \code{character} vector of the names of the pre-fabricated (prefab) models or a \code{list} of their controls
}
\examples{
 prefab_models( )
 prefab_model_controls( )

}
