% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_covariate_data}
\alias{prep_covariate_data}
\title{Prepare covariate data table}
\usage{
prep_covariate_data(moons = prep_moon_data(future = TRUE),
  forecast_date = Sys.Date(), historical = TRUE, forecasts = TRUE,
  lead_time = 12, min_lag = 6, append_fcast_csv = FALSE,
  covariate_file = NULL, source_name = "current_archive",
  data_dir = pc_path("data", "~"))
}
\arguments{
\item{moons}{current newmoonnumber table}

\item{forecast_date}{date from which forecasting is happening, typically
today's date.}

\item{historical}{logical input dictating if historical covariates should
be included in the output}

\item{forecasts}{logical input dictating if covariate forecasts should be
made and included in the output}

\item{lead_time}{number of newmoons forward to forecast abundances}

\item{min_lag}{minimum (of non-0) lag times that need to be accommodate}

\item{append_fcast_csv}{logical input dictating if covariate forecasts
should be added to the existing covariate forecast data file}

\item{covariate_file}{file name for the existing historical covariate data
file}

\item{source_name}{character value for the name to give the covariaate
forecast. Currently is "current_archive". Previous to "current_archive",
the data were retroactively filled in and are given the source name
"retroactive"}

\item{data_dir}{directory name where the data files will reside}
}
\value{
covariate data table including both historical and covariate data
  with an additional column (\code{source}) distinguishing the source of
  the data. In addition, if \code{append_fcast_csv} is \code{TRUE}, the
  forecasts will be appended to \code{covariate_file} tagged with  
  \code{source_name} dictating the source name tag in the forecast file.
}
\description{
Create a covariate (weather and NDVI) data table, including
  potentially both historical and forecast data
}
