% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_covariates}
\alias{prep_covariates}
\alias{download_hist_covariate_forecasts}
\alias{prep_hist_covariates}
\alias{prep_fcast_covariates}
\alias{prep_weather_data}
\alias{append_cov_fcast_csv}
\title{Prepare covariate data table}
\usage{
prep_covariates(moons = prep_moons(), tree = dirtree(),
  quiet = FALSE, control = list())

download_hist_covariate_forecasts(tree = dirtree(), quiet = FALSE,
  control = list())

prep_hist_covariates(tree = dirtree(), control = list())

prep_fcast_covariates(hist_cov = prep_hist_covariates(),
  moons = prep_moons(), tree = dirtree(), quiet = FALSE,
  control = list())

prep_weather_data(tree = dirtree())

append_cov_fcast_csv(new_forecast_covariates, tree = dirtree(),
  control = list)
}
\arguments{
\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}

\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{control}{\code{list} of parameters to control the preparation of 
moons. See \code{\link{covariates_control}}.}

\item{hist_cov}{Historical covariate data table as a code{covariates}-class 
\code{data.frame}, returned from \code{\link{prep_hist_covariates}}.}

\item{new_forecast_covariates}{Class-\code{covariates} \code{data.frame} of
forecasted covariate data generated by 
\code{\link{forecast_covariates}}.}
}
\value{
\code{prep_covariates}: Covariate data table as a 
  code{covariates}-class \code{data.frame}. \cr \cr
  \code{prep_hist_covariates}: Historical covariate data table as a 
  code{covariates}-class \code{data.frame}. \cr \cr
  \code{prep_weather_data}: \code{data.frame} of historical weather data.
  \cr \cr
  \code{append_cov_fcast_csv}: appended \code{new_forecast_covariates}.
}
\description{
Set of functions to prepare a covariate data table used in
  forecasting. \cr \cr
  \code{prep_covariates}: interface-level function
  code{download_hist_covariate_forecasts}: download the historical 
  covariate data forecasts to the data folder \cr \cr
  \code{prep_hist_covariates}: prepares the historical covariates data
  \cr \cr
  \code{prep_weather_data}: prepares historical weather data \cr \cr
  \code{append_cov_fcast_csv}: appends a covariate forecast to 
  existing covariate forecasts
}
\details{
\code{prep_covariates} is the overall function that prepares the 
  covariate data for forecasting or hindcasting, 
  including saving out the data and appending the new forecasts of 
  covariates to the existing covariate forecast table. \cr \cr
  \code{download_hist_covariate_forecasts}: Download the
  historical covariates predictions file into the data subdirectory in the
  the portalcasting directory tree from 
  \href{https://zenodo.org/record/3333770}{Zenodo} (default), 
  \href{https://bit.ly/2Lo2xnQ}{portalPredictions repo}, or an old local
  version from the package's data).
  \code{prep_hist_covariates}: creates a data table of historical 
  weather and NDVI data. \cr \cr
  \code{prep_weather_data}: creates a data table of historical weather 
  data using \code{\link[portalr]{weather}}. \cr \cr
  \code{append_cov_fcast_csv}: Add the newest weather and NDVI forecasts to
  the existing forecasts data file, for use in future hindcasting.
}
\examples{
\dontrun{

setup_dir()
prep_covariates()
}

}
