% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data Subdirectory for a Forecast Run or Hindcast Runs}
\usage{
prep_data(options_data = data_options())
}
\arguments{
\item{options_data}{the data options list}
}
\value{
nothing
}
\description{
Prepare the data directory for a forecasting run or a set
  of hindcasting runs
}
