% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_fcast_covariates}
\alias{prep_fcast_covariates}
\title{Prepare forecast covariate data table}
\usage{
prep_fcast_covariates(hist_cov = prep_hist_covariates(),
  moons = prep_moons(), options_covariates = covariates_options())
}
\arguments{
\item{hist_cov}{Historical covariate data table as a code{covariates}-class 
\code{data.frame}, returned from \code{\link{prep_hist_covariates}}.}

\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}

\item{options_covariates}{A class-\code{covariates_options} \code{list} of 
settings controlling the covariates data creation.}
}
\value{
Covariate data table as a code{covariates}-class \code{data.frame}
  ready for casting.
}
\description{
Prepare forecasts of covariate data as needed for rodent
  forecasting, including appending the new forecasts of covariates to the
  existing covariate forecast table
}
