% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prep_hist_covariate_data}
\alias{prep_hist_covariate_data}
\title{Collect Historic Weather and NDVI Data}
\usage{
prep_hist_covariate_data()
}
\value{
covariate data table
}
\description{
Gather the downloaded historical weather and ndvi data and
  join them into a single data table
}
