% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_hist_covariates}
\alias{prep_hist_covariates}
\title{Prepare historical covariates data}
\usage{
prep_hist_covariates(options_covariates = covariates_options())
}
\arguments{
\item{options_covariates}{A class-\code{covariates_options} \code{list} of 
settings controlling the covariates data creation.}
}
\value{
Historical covariate data table as a code{covariates}-class 
  \code{data.frame}.
}
\description{
Create a data table of historical weather and NDVI data.
}
