% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prep_metadata}
\alias{prep_metadata}
\title{Prepare model metadata}
\usage{
prep_metadata(moons = prep_moons(), rodents_list = prep_rodents_list(),
  covariates = prep_covariates(), tree = dirtree(), quiet = FALSE,
  control = list())
}
\arguments{
\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}

\item{rodents_list}{A class-\code{rodents_list} \code{list} of two class-
\code{rodents} \code{data.frame}s, \code{all} (abundances on all plots)
and \code{controls} (abundances on control plots only).}

\item{covariates}{Covariate data table as a code{covariates}-class 
\code{data.frame} generated by \code{\link{prep_covariates}}}

\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{control}{\code{list} of parameters to control the preparation of 
metadata. See \code{\link{metadata_control}}.}
}
\value{
Model metadata as a \code{metadata}-class \code{list}.
}
\description{
Set up the metadata list used for -casting, primarily the 
  time periods.
}
\examples{
\dontrun{

setup_dir()
prep_metadata()
}

}
