% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{prep_models}
\alias{prep_models}
\title{Select Models to Forecast or Hindcast With}
\usage{
prep_models(options_models = models_options())
}
\arguments{
\item{options_models}{models options}
}
\value{
names of files to be run
}
\description{
Verify that models requested are available and select them
}
