% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{prep_moons}
\alias{prep_moons}
\alias{append_past_moons_to_raw}
\alias{add_future_moons}
\alias{add_addl_future_moons}
\title{Prepare temporal (lunar) data}
\usage{
prep_moons(tree = dirtree(), quiet = FALSE, control = list())

append_past_moons_to_raw(moons, tree = dirtree(), control = list())

add_future_moons(moons = prep_moons(), control = list())

add_addl_future_moons(future_moons, control = list())
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{control}{\code{list} of parameters to control the preparation of 
moons. See \code{\link{moons_control}}.}

\item{moons}{Moons data table as a code{moons}-class \code{data.frame}.}

\item{future_moons}{A \code{moons}-class \code{data.frame} forward from 
the historic moons table (produced by \code{\link{prep_moons}}), as
generated by \code{\link[portalr]{get_future_moons}}.}
}
\value{
\code{prep_moons}: Fully appended \code{moons}data table as a 
   code{moons}-class \code{data.frame}.
 \code{append_past_moons_to_raw}, \code{add_future_moons}, 
   \code{add_addl_future_moons}: Appropropriately appended \code{moons} 
   data table as a \code{moons}-class \code{data.frame}.
}
\description{
A set of functions that get time information (calendar dates, 
  census periods, and newmoon numbers) associated with trapping events 
  (achieved and missed) based on a lunar survey schedule. If needed, 
  additional moons will be added to both the in-use and raw versions of 
  the data table. \cr \cr
  \code{append_past_moons_to_raw} Appends missing past moon dates to the
  raw data. \cr \cr
  \code{add_future_moons} Adds future moon dates to the moon table. \cr \cr
  \code{add_addl_future_moons} Add more more moons to accomplish requested 
   forecast length. \cr \cr
}
\details{
\code{append_past_moons_to_raw}:
  Sometimes the raw moon data table is not up-to-date. Because
  the \code{portalr} functions \code{\link[portalr]{weather}} and 
  \code{\link[portalr]{fcast_ndvi}} point to the raw moons data, that table
  needs to be updated to produce the correct current data table for 
  casting. \cr \cr
  \code{add_future_moons}:
  Add future moons to the moon table, counting forward from
  the forecast date (\code{options_moons$cast_date}). Because the 
  \code{moons} table might not have the most recent moons, more rows than 
  anticipated (\code{options_moons$n_future_moons}) may need to be added to 
  the table. \cr \cr
  \code{add_addl_future_moons}: 
  Add extra future moons to the moon table to achieve requested
  forecast length. Because the moon table might not have the most recent 
  moons, more rows than initially requested may need to be added to the 
  table for it to be of the required coverage.
}
\examples{
\dontrun{

setup_dir()
prep_moons()
}

}
