% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{prep_rodent_data}
\alias{prep_rodent_data}
\title{Prepare historic rodent data}
\usage{
prep_rodent_data(moons = prep_moon_data(), start_newmoonnumber = 217)
}
\arguments{
\item{moons}{current newmoonnumber table}

\item{start_newmoonnumber}{newmoon number of the first sample to be 
included. Default value is \code{217}, corresponding to 1995-01-01.}
}
\value{
a list of two dataframes, all plots and control plots
}
\description{
Prepare the rodent dataset, tailored for forecasting. 
  Specifically, return data for all plots and control plots only, for each 
  species except PI and the total
}
