% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{prep_rodents}
\alias{prep_rodents}
\title{Prepare historic rodent data}
\usage{
prep_rodents(tree = dirtree(), moons = prep_moons(), start = 217)
}
\arguments{
\item{tree}{the name tree of the portalcasting directory}

\item{moons}{current newmoon table}

\item{start}{newmoon number of the first sample to be included. Default 
value is \code{217}, corresponding to 1995-01-01.}
}
\value{
a list of two dataframes, all plots and control plots
}
\description{
Prepare the rodent dataset, tailored for forecasting. 
  Specifically, return data for all plots and control plots only, for each 
  species except PI and the total
}
