% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{prep_rodents_list}
\alias{prep_rodents_list}
\alias{prep_rodents}
\title{Prepare historic rodent data}
\usage{
prep_rodents_list(moons = prep_moons(), tree = dirtree(),
  quiet = FALSE, control = list())

prep_rodents(moons = prep_moons(), tree = dirtree(),
  control = list(), tmnt_type = NULL)
}
\arguments{
\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}

\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{control}{A \code{list} of arguments to 
\code{\link[portalr]{summarize_rodent_data}}. Preppared via 
\code{\link{enforce_rodents_control}}.}

\item{tmnt_type}{\code{character}-value representation of treatment type 
(\code{"all"} or \code{"controls"}) used to enforce certain elements of 
the \code{control} \code{list} in \code{prep_rodents}.}
}
\value{
\code{prep_rodents_list}: 
  A class-\code{rodents_list} \code{list} of two class-
  \code{rodents} \code{data.frame}s, \code{all} (abundances on all plots)
  and \code{controls} (abundances on control plots only) (only returned if
  \code{update = TRUE}). \cr \cr
  \code{prep_rodents} 
  One of two class-\code{rodents} \code{data.frame}s, as defined by
  \code{tmnt_type}: \code{"all"} (abundances on all plots) or 
  \code{"controls"} (abundances on control plots only).
}
\description{
Prepare the rodent dataset, tailored for forecasting. 
  Specifically, \code{prep_rodents_list} returns data for all plots and 
  control plots only, for each species except PI and the total.
  \code{prep_rodents} prepares a particular version of the rodent dataset,
  tailored for forecasting or hindcasting. Available versions of the data 
  currently include \code{"all"} (abundances on all plots) and 
  \code{"controls"} (abundances on control plots only).
}
\examples{
\dontrun{

setup_dir()
prep_rodents()
}

}
