% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{prep_rodents_table}
\alias{prep_rodents_table}
\alias{trim_time}
\alias{add_moons}
\alias{add_total}
\alias{trim_species}
\alias{trim_treatment}
\title{Prepare a rodents data table for forecasting}
\usage{
prep_rodents_table(main = ".", moons = prep_moons(main = main),
  end_moon = NULL, start_moon = 217, species = base_species(),
  total = TRUE, level = "Site", treatment = NULL, plots = "all",
  min_plots = 24, min_traps = 1, output = "abundance",
  ref_species = all_species(), quiet = TRUE, save = TRUE,
  overwrite = TRUE, filename = "rodents_all.csv")

trim_time(rodents_tab, start_moon = NULL, end_moon = NULL)

add_moons(rodents_tab, moons = prep_moons())

add_total(rodents_tab, total = TRUE)

trim_species(rodents_tab, species = base_species(),
  ref_species = all_species())

trim_treatment(rodents_tab, level = "Site", treatment = NULL)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{start_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the first sample to be included. Default value is \code{217}, 
corresponding to \code{1995-01-01}.}

\item{species}{\code{character}-valued vector of species names to include 
in the forecasting data table. Defaults to all species but \code{"PI"}
via \code{\link{base_species}}}

\item{total}{\code{logical} value indicating if a total (sum across 
species) should be added or not. Only available if more than one species
is included.}

\item{level}{\code{character} indicating the type of summary:
\code{"Plot"}, \code{"Treatment"}, or \code{"Site"}. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}

\item{treatment}{\code{character} indicating the specific treatment(s) to
trim to if \code{level = "Treatment"}: \code{"control"},
\code{"exclosure"}, \code{"removal"}, or \code{"spectabs"} .}

\item{plots}{Specification of subset of plots. Can be a vector of 
\code{numeric} plots indicators or specific sets indicated by
\code{character} values: \code{"all"} plots or \code{"Longterm"} plots
(plots that have had the same treatment for the entire time series).}

\item{min_plots}{\code{integer} (or integer \code{numeric}) of the minimum 
number of plots surveyed for a survey to be used.Pipes directly to 
\code{\link[portalr]{summarize_rodent_data}}.}

\item{min_traps}{\code{integer} (or integer \code{numeric}) of the minimum 
number of traps collected for a plot to be used. Pipes directly to 
\code{\link[portalr]{summarize_rodent_data}}.}

\item{output}{\code{character} indicating the type of data:
\code{"abundance"}, \code{"biomass"}, or \code{"energy"}. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}

\item{ref_species}{\code{character}-valued vector of all possible species 
names that could be used should be set up via \code{\link{all_species}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{save}{\code{logical} indicator controlling if the output should 
be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing
files should be updated (most users should leave as \code{TRUE}).}

\item{filename}{\code{character} name of the file for saving the output.}

\item{rodents_tab}{\code{data.frame} rodents table, with varying levels of 
editing.}
}
\value{
\code{data.frame} rodents table, with varying levels of editing
 for use in forecasting. \cr \cr
 \code{prep_rodents_table}, \code{trim_time}: a fully appended and 
 formatted \code{data.frame} (also saved out if \code{save = TRUE}). 
 \cr \cr
 \code{trim_species}, \code{add_total}, \code{trim_treatment}, 
 \code{add_moons}: appropriately appended rodents \code{data.frame}.
}
\description{
This set of functions provides a convenient wrapper on
 \code{\link[portalr]{summarize_rodent_data}} to generate a specific
 data table for use in the forecasting pipeline. Many arguments pipe
 directly to \code{\link[portalr]{summarize_rodent_data}}. \cr \cr
 \code{prep_rodents_table} generates the ready-to-model table. \cr \cr
 \code{add_total} adds a total (sum-across-species) column. \cr \cr
 \code{trim_treatment} removes not-requested plots according to treatment
 levels and then removes the treatment column. \cr \cr
 \code{add_moons} adds the lunar data associated with the samples. \cr \cr
 \code{trim_time} reduces the table to the samples within the time window
 and then removes extra time columns. \cr \cr
}
\examples{
 \donttest{
  create_dir()
  fill_raw()
  prep_rodents_table()
  raw_path <- sub_paths(".", "raw")
  dat1 <- portalr::summarize_rodent_data(path = raw_path, clean = FALSE)
  dat2 <- trim_species(dat1)
  dat3 <- add_total(dat2)
  dat4 <- trim_treatment(dat3)
  dat5 <- add_moons(dat4)
  dat6 <- trim_time(dat5, 217)
 }

}
