% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prepare covariates}
\alias{prepare covariates}
\alias{prepare_covariates}
\title{Prepare Covariate Data for Casting}
\usage{
prepare_covariates(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
\code{data.frame} of historical and forecasted covariates that is also saved out to \code{settings$files$covariates} if indicated by \code{settings$save}.
}
\description{
Prepare and combine the historical and forecast covariate data for a model run, according to the \code{\link{directory_settings}}.
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "covariates")

   create_dir(main = main1)
   fill_resources(main = main1)
   fill_forecasts(main = main1)
   fill_fits(main = main1)
   fill_models(main = main1)

   prepare_newmoons(main   = main1)
   prepare_rodents(main    = main1) 
   prepare_covariates(main = main1)

   unlink(main1, recursive = TRUE)
}

}
\seealso{
Content preparation functions:
\code{\link{directory filling}},
\code{\link{prepare metadata}},
\code{\link{prepare models}},
\code{\link{prepare newmoons}},
\code{\link{prepare rodents}}
}
\concept{content-prep}
