% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{prepare models}
\alias{prepare models}
\alias{write_models_controls}
\alias{write_models_scripts}
\alias{prepare-models}
\alias{read_models_controls}
\alias{models_controls}
\title{Prepare Portalcasting Models}
\usage{
write_models_controls(
  main = ".",
  new_models_controls = NULL,
  models = prefab_models()
)

write_models_scripts(main = ".", controls = prefab_models_controls())

read_models_controls(main = ".")

models_controls(main = ".", models = prefab_models())
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{new_models_controls}{\code{list} of controls for any new models (not in the prefab models) listed in \code{models} that are to be added to the control list and file.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{controls}{\code{list} of controls for the models.}
}
\value{
\code{model_controls}: \code{list} of \code{models}' control \code{list}s, \code{\link[base:invisible]{invisible}}-ly. \cr \cr
\code{read_models_controls}: \code{list} of all \code{models}' control \code{list}s, from the file defined in \code{\link{directory_settings}}, \code{\link[base:invisible]{invisible}}-ly. \cr \cr
\code{write_models_controls}: \code{list} of \code{models}' control \code{list}s, \code{\link[base:invisible]{invisible}}-ly. \cr \cr
\code{write_models_scripts}: \code{NULL}, \code{\link[base:invisible]{invisible}}-ly.
}
\description{
Add model controls and scripts to the portalcasting directory and read in controls and model names.
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "models")

   create_dir(main = main1)
   fill_resources(main = main1)
   fill_forecasts(main = main1)
   fill_fits(main = main1)

   controls <- write_models_controls(main = main1)
   write_models_scripts(main     = main1, 
                      controls = controls)

   unlink(main1, recursive = TRUE)
}

}
