% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prepare_historical_covariates}
\alias{prepare_historical_covariates}
\alias{prep_hist_covariates}
\title{Prepare historical covariates data}
\usage{
prep_hist_covariates(
  main = ".",
  moons = NULL,
  control_files = files_control(),
  quiet = TRUE,
  arg_checks = TRUE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Lunar data \code{data.frame}. See \code{\link{prep_moons}}.}

\item{control_files}{\code{list} of names of the folders and files within
the sub directories and saving strategies (save, overwrite, append, etc.).
Generally shouldn't need to be edited. See \code{\link{files_control}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{prep_hist_covariates}: daily historical covariate data table as a 
  \code{data.frame}.
}
\description{
\code{prep_hist_covariates} creates a \code{data.frame} of historical 
 weather and NDVI data. Automatically goes all the way back to the 
 beginning of the available data, as the table is automatically trimmed 
 later. \cr \cr
 NDVI data are generated by \code{\link[portalr]{ndvi}}.
}
\examples{
 \donttest{   
  create_dir()
  fill_raw()
  prep_hist_covariates()
 }

}
