% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{prepare_rodents_table}
\alias{prepare_rodents_table}
\alias{prep_rodents_table}
\alias{process_rodent_data}
\title{Prepare a rodents data tables for forecasting}
\usage{
prep_rodents_table(
  main = ".",
  moons = NULL,
  end_moon = NULL,
  species = base_species(),
  total = TRUE,
  interpolate = TRUE,
  clean = FALSE,
  type = "Rodents",
  level = "Site",
  plots = "all",
  treatment = NULL,
  min_plots = 24,
  min_traps = 1,
  output = "abundance",
  shape = "crosstab",
  fillweight = (output != "abundance"),
  unknowns = FALSE,
  time = "newmoon",
  na_drop = FALSE,
  zero_drop = switch(tolower(level), plot = FALSE, treatment = TRUE, site = TRUE),
  ref_species = all_species(),
  effort = FALSE,
  quiet = TRUE,
  verbose = FALSE,
  control_files = files_control(),
  save = TRUE,
  overwrite = TRUE,
  filename = "rodents_all.csv",
  arg_checks = TRUE
)

process_rodent_data(
  rodents_tab,
  main = ".",
  moons = NULL,
  end_moon = NULL,
  species = base_species(),
  total = TRUE,
  interpolate = FALSE,
  clean = FALSE,
  level = "Site",
  treatment = NULL,
  na_drop = FALSE,
  time = "newmoon",
  ref_species = all_species(),
  quiet = TRUE,
  verbose = FALSE,
  arg_checks = TRUE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{species}{\code{character}-valued vector of species names 
to include.}

\item{total}{\code{logical} value indicating if a total 
(sum across species) should be added or not. Only available if more than 
one species is included.}

\item{interpolate}{\code{logical} value indicating if the data should be
interpolated to fill in \code{NA} values (using 
\code{\link[forecast]{na.interp}}).}

\item{clean}{\code{logical} indicator of if only the rodent data
that passed QA/QC (\code{clean = TRUE}) or if all data 
(\code{clean = FALSE}) should be loaded.}

\item{type}{\code{character} value of the rodent data set type, according 
to pre-existing definitions. An alternative toggle to \code{species}. \cr
Either all species 
(\code{type = "Rodents"}) or only granivoes (\code{type = "Granivores"}).}

\item{level}{\code{character} indicating the type of summary:
\code{"Plot"}, \code{"Treatment"}, or \code{"Site"}. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}

\item{plots}{Specification of subset of plots. Can be a vector of 
\code{numeric} plots indicators or specific sets indicated by
\code{character} values: \code{"all"} plots or \code{"Longterm"} plots
(plots that have had the same treatment for the entire time series).}

\item{treatment}{\code{character} indicating the specific 
treatment(s) to trim to if \code{level = "Treatment"}: \code{"control"},
\code{"exclosure"}, \code{"removal"}, or \code{"spectabs"}}

\item{min_plots}{\code{integer} (or integer \code{numeric}) of the 
minimum number of plots surveyed for a survey to be used. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}

\item{min_traps}{\code{integer} (or integer \code{numeric}) of the 
minimum number of traps collected for a plot to be used. Pipes directly
to \code{\link[portalr]{summarize_rodent_data}}.}

\item{output}{\code{character} indicating the type of data:
\code{"abundance"}, \code{"biomass"}, or \code{"energy"}. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}

\item{shape}{\code{character} value indicating a "crosstab" or "flat" 
output.}

\item{fillweight}{\code{logical} specifier of whether to fill in unknown 
weights with other records from that individual or species, where 
possible.}

\item{unknowns}{\code{logical} indicator to either remove all 
individuals not identified to species (\code{unknowns = FALSE}) or 
sum them in an additional column (\code{unknowns = TRUE}.}

\item{time}{\code{character} value specifying the format of the time index 
in the output. Options are \code{"period"} (sequential Portal surveys), 
\code{"newmoon"} (lunar cycle numbering), and \code{"date"} (calendar 
date). \cr
The default \code{time = "newmoon"} produces an equispaced observation 
timestep, a common format format for discrete-time modeling.}

\item{na_drop}{\code{logical} indicator of if \code{NA} values 
(representing insufficient sampling) should be dropped.}

\item{zero_drop}{\code{logical} indicator of if \code{0} values 
(representing sufficient sampling but no detection) should be dropped.}

\item{ref_species}{\code{character} vector of the reference species. Sets
a boundary outside of the species included in the model.}

\item{effort}{\code{logical} indicator of if the effort columns should
be included in the output.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be
shown.}

\item{control_files}{\code{list} of names of the folders and files within
the sub directories and saving strategies (save, overwrite, append, etc.).
Generally shouldn't need to be edited. See \code{\link{files_control}}. 
Only used here to point to the moons data.}

\item{save}{\code{logical} indicator controlling if the output should 
be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing
files should be updated (most users should leave as \code{TRUE}).}

\item{filename}{\code{character} name of the file for saving the 
output. The extension of the filename may be used for identifying
object requirements (for example, if the extension of \code{filename} is
\code{".csv"}, it will trigger use of \code{\link[utils]{write.csv}}).}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}

\item{rodents_tab}{Rodents \code{data.frame} data table.}
}
\value{
\code{data.frame} for the specified data set.
}
\description{
This function wraps around 
 \code{\link[portalr]{summarize_rodent_data}} to 
 produce a \code{data.frame}s associated with a set of data 
 specifications.
}
\examples{
 \donttest{
   create_dir()
   fill_raw()
   prep_rodents_table()
 }

}
