% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{read and write rodent dataset controls}
\alias{read and write rodent dataset controls}
\alias{read_datasets_controls}
\alias{datasets_controls}
\alias{write_datasets_controls}
\title{Read and Write Rodent Datasets Control Lists}
\usage{
read_datasets_controls(main = ".")

datasets_controls(main = ".", datasets = prefab_datasets())

write_datasets_controls(
  main = ".",
  new_datasets_controls = NULL,
  datasets = prefab_datasets()
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{datasets}{\code{character} vector of name(s) of rodent dataset(s) to include.}

\item{new_datasets_controls}{\code{list} of controls for any new datasets (not in the prefab datasets) listed in \code{datasets} that are to be added to the control list and file.}
}
\value{
\code{list} of \code{datasets}' control \code{list}s, \code{\link[base:invisible]{invisible}}-ly for \code{write_datasets_controls}.
}
\description{
Input/Output functions for dataset control lists.
}
