% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{read cast output}
\alias{read cast output}
\alias{read_cast_tab}
\alias{read_cast_tabs}
\alias{read_cast_metadata}
\alias{read_model_fits}
\alias{read_model_casts}
\title{Read in Cast Output From a Given Cast}
\usage{
read_cast_tab(main = ".", settings = directory_settings(), cast_id = NULL)

read_cast_tabs(main = ".", settings = directory_settings(), cast_ids = NULL)

read_cast_metadata(main = ".", settings = directory_settings(), cast_id = NULL)

read_model_fits(main = ".", settings = directory_settings(), cast_id = NULL)

read_model_casts(main = ".", settings = directory_settings(), cast_id = NULL)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{cast_ids, cast_id}{\code{integer} (or integer \code{numeric}) value(s) representing the cast(s) of interest, as indexed within the directory in the \code{casts} sub folder. See the casts metadata file (\code{casts_metadata.csv}) for summary information. If \code{NULL} (the default), the most recently generated cast's output is read in. \cr 
\code{cast_ids} can be NULL, one value, or more than one values, \code{cast_id} can only be NULL or one value.}
}
\value{
\code{read_cast_tab}: \code{data.frame} of the \code{cast_tab}. \cr \cr
 \code{read_cast_tabs}: \code{data.frame} of the \code{cast_tab}s with a \code{cast_id} column added to distinguish among casts. \cr \cr
 \code{read_cast_metadata}: \code{list} of \code{cast_metadata}. \cr \cr
 \code{read_model_fits}: \code{list} of \code{model_fits}. \cr \cr
 \code{read_model_casts}: \code{list} of \code{model_casts}.
}
\description{
Read in the various output files of a cast or casts in the casts sub directory.
}
