% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_output.R
\name{read write data}
\alias{read write data}
\alias{write_data}
\alias{read_data}
\alias{read_rodents_dataset}
\alias{read_rodents}
\alias{read_newmoons}
\alias{read_covariates}
\alias{read_metadata}
\alias{read_climate_forecasts}
\title{Read from and Write to a Data File}
\usage{
write_data(
  x = NULL,
  main = ".",
  subdirectory = "data",
  save = TRUE,
  overwrite = TRUE,
  filename = NULL,
  quiet = FALSE
)

read_data(
  main = ".",
  data_name = NULL,
  dataset = "all",
  datasets = prefab_datasets()
)

read_rodents_dataset(main = ".", dataset = "all")

read_rodents(main = ".", datasets = prefab_datasets())

read_newmoons(main = ".")

read_covariates(main = ".")

read_metadata(main = ".")

read_climate_forecasts(main = ".")
}
\arguments{
\item{x}{Data, such as a \code{data.frame} or \code{list}, to be written out.}

\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{subdirectory}{\code{character} value defining the data subdirectory of the portalcasting directory tree.}

\item{save}{\code{logic} indicator controlling if \code{x} should be saved out.}

\item{overwrite}{\code{logic} indicator of whether or not file writing sould occur even if a local copy already exists.}

\item{filename}{\code{character} name of the file for saving \code{x}.}

\item{quiet}{\code{logic} indicator if messages should be quieted.}

\item{data_name}{\code{character} representation of the data needed. Current options include \code{"rodents"}, \code{"rodents_table"}, \code{"covariates"}, \code{"climate_forecasts"}, \code{"newmoons"}, and \code{"metadata"}.}

\item{dataset, datasets}{\code{character} representation of the grouping name(s) used to define the rodents. Standard options are \code{"all"}, \code{"controls"}, and \code{"exclosures"}. \code{dataset} can only be length 1, \code{datasets} is not restricted in length.}
}
\value{
\code{write_data} \code{x} as input, \code{\link[base:invisible]{invisible}}-ly.
}
\description{
Generalized data input-output functionality with specifics for common files. \cr \cr
\code{write_data} saves inputted data out to a data file if requested and returns it to the console, \code{\link[base:invisible]{invisible}}-ly. Currently available for \code{yaml}, \code{csv}, and \code{json} file extensions. \cr \cr
\code{read_data} reads in a specified data file. Specific functions available include \code{read_rodents}, \code{read_rodents_dataset}, \code{read_covariates}, \code{read_climate_forecasts}, \code{read_newmoons}, and \code{read_metadata}.
}
\examples{
 
 \donttest{
   main1 <- file.path(tempdir(), "standard")
   setup_dir(main = main1)
   write_data(main = main1, x = data.frame(rnorm(10)), filename = "xx.csv")
   read_data(main = main1)
   read_rodents(main = main1)
   read_rodents_dataset(main = main1)
   read_covariates(main = main1)
   read_climate_forecasts(main = main1)
   read_newmoons(main = main1)
   read_metadata(main = main1)
   unlink(main1, force = TRUE, recursive = TRUE)
 }

}
