% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{read_cast}
\alias{read_cast}
\alias{read_casts}
\title{Read in a cast file or multiple cast files and set the class}
\usage{
read_cast(tree = dirtree(), cast_type = "forecasts",
  cast_date = NULL)

read_casts(tree = dirtree(), cast_type = "forecasts",
  cast_dates = NULL)
}
\arguments{
\item{tree}{\code{dirtree}-class list. See \code{\link{dirtree}}.}

\item{cast_type}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"forecasts"}.}

\item{cast_date}{\code{Date} the predictions were made. Used to select the
file in the predictions subdirectory. Can only be length 1 and if 
\code{NULL} (default), selects the most recent -casts.}

\item{cast_dates}{\code{Date}s the predictions were made. Used to select 
the files in the predictions subdirectory. Can be length 1 or more and if 
\code{NULL} (default), selects all available -casts.}
}
\value{
Class \code{casts} \code{data.frame} of requested fore- or 
  hindcasts.
}
\description{
Read in a specified forecast or hindcast data file and ensure 
  its class attribute is appropriate for usage within the portalcasting 
  pipeline. Current not reliably coded for hindcasts.
  \cr \cr \code{read_cast}: Only allows for a single date and defaults to
  the most recent.
  \cr \cr \code{read_casts}: Allows for multiple dates and defaults to all.
}
