% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{read_cast_tab}
\alias{read_cast_tab}
\title{Read in the cast tab output from a given cast}
\usage{
read_cast_tab(main = ".", cast_id = NULL, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{cast_id}{\code{integer} (or integer \code{numeric}) value 
representing the cast of interest, as indexed within the directory in
the \code{casts} sub folder. See the casts metadata file 
(\code{casts_metadata.csv}) for summary information. If \code{NULL} (tjhe
default), no \code{cast_tab} is read in.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
\code{data.frame} of the \code{cast_tab}.
}
\description{
Retrieve the \code{cast_tab} of a cast in the casts 
 sub directory.
}
\examples{
 \donttest{
  setup_dir()
  portalcast(models = c("AutoArima", "NaiveArima"), end_moons = 515:520)
  read_cast_tab(cast_id = 1)
}

}
