% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{read_casts}
\alias{read_casts}
\title{Read in a cast file and format it for specific class}
\usage{
read_casts(tree, casttype = "forecasts", castdate = today())
}
\arguments{
\item{tree}{\code{dirtree}-class list. See \code{\link{dirtree}}.}

\item{casttype}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"forecasts"}.}

\item{castdate}{\code{Date} the predictions were made. Used to select the
file in the predictions subdirectory.}
}
\value{
Class \code{casts} \code{data.frame} of requested fore- or 
  hindcasts.
}
\description{
Read in a specified forecast or hindcast data file and ensure 
  its class attribute is appropriate for usage within the portalcasting 
  pipeline. Current
}
