% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{read_casts_evaluations}
\alias{read_casts_evaluations}
\title{Read in the Casts Evaluations File}
\usage{
read_casts_evaluations(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
Evaluations requested.
}
\description{
Read in the casts evaluations file.
}
