% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_output.R
\name{read_data}
\alias{read_data}
\alias{read_rodents_table}
\alias{read_rodents}
\alias{read_newmoons}
\alias{read_covariates}
\alias{read_climate_forecasts}
\alias{read_metadata}
\title{Read in and Format a Portalcasting Data File}
\usage{
read_data(
  main = ".",
  data_name = NULL,
  dataset = "all",
  datasets = prefab_datasets()
)

read_rodents_table(main = ".", dataset = "all")

read_rodents(main = ".", datasets = prefab_datasets())

read_newmoons(main = ".")

read_covariates(main = ".")

read_climate_forecasts(main = ".")

read_metadata(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{data_name}{\code{character} representation of the data needed. Current options include \code{"rodents"}, \code{"rodents_table"}, \code{"covariates"}, \code{"climate_forecasts"}, \code{"newmoons"}, and \code{"metadata"}.}

\item{dataset, datasets}{\code{character} representation of the grouping name(s) used to define the rodents. Standard options are \code{"all"}, \code{"controls"}, and \code{"exclosures"}. \code{dataset} can only be length 1, \code{datasets} is not restricted in length.}
}
\value{
Data requested.
}
\description{
Read in a specified data file.
}
\examples{
   \donttest{
     main1 <- file.path(tempdir(), "standard")
     setup_dir(main = main1)
     read_data(main = main1)
     read_rodents(main = main1)
     read_rodents_table(main = main1)
     read_covariates(main = main1)
     read_climate_forecast(main = main1)
     read_newmoons(main = main1)
     read_metadata(main = main1)
     unlink(main1, force = TRUE, recursive = TRUE)
   }

}
