% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{read_foreforecasts_metadata}
\alias{read_foreforecasts_metadata}
\title{Read in the Casts Metadata File}
\usage{
read_foreforecasts_metadata(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
Data requested.
}
\description{
Read in the casts metadata file. If the data file does not exist, an effort is made to create the file.
}
