% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{record_name_from_url}
\alias{record_name_from_url}
\title{Attempt to extract the record name from a URL}
\usage{
record_name_from_url(url, NULLname = FALSE)
}
\arguments{
\item{url}{\code{character} value of the URL.}

\item{NULLname}{\code{logical} indicator of if the name should be kept as
\code{NULL}, rather than given a name based on the URL.}
}
\value{
\code{character} value of the name or \code{NULL}.
}
\description{
The record name is often encoded in a URL, so in the case that
 a name is not given for the record, this function attempts to extract it,
 unless told to keep the name as NULL (via \code{NULLname}).
}
\examples{
 source_url <- zenodo_url(concept_rec_id = "1215988")
 record_name_from_url(source_url)

}
