% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_incompletes}
\alias{remove_incompletes}
\title{Remove any Specific Incomplete Entries as Noted by an NA}
\usage{
remove_incompletes(df, colname)
}
\arguments{
\item{df}{\code{data.frame} table to be written out.}

\item{colname}{A single \code{character} value of the column to use to remove incomplete entries.}
}
\value{
\code{df} without any incomplete entries.
}
\description{
Remove any incomplete entries in a table, as determined by the presence of an \code{NA} entry in a specific column (\code{colname}).
}
\examples{
 df <- data.frame(c1 = c(1:9, NA), c2 = 11:20)
 remove_incompletes(df, "c1")

}
