% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{remove_spp}
\alias{remove_spp}
\title{Remove species from the rodent data table}
\usage{
remove_spp(data, drop_spp = rodent_options()$drop_spp)
}
\arguments{
\item{data}{rodent data table}

\item{drop_spp}{species name (column header) to drop from data}
}
\value{
trimmed data
}
\description{
Remove species from the rodent data table
}
