% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefabricated_rodent_datasets.R
\docType{data}
\name{rodent_dataset_controls}
\alias{rodent_dataset_controls}
\title{Rodent Dataset Controls List}
\format{
\code{list} of control \code{list}s, each corresponding to a dataset.
}
\usage{
rodent_dataset_controls
}
\description{
\code{list} of the rodent datasets available for use via \code{\link{prepare_rodents}}.
}
\details{
The currently prefabricated datasets include \code{"all"}, \code{"all_interp"}, \code{"controls"}, \code{"controls_interp"}, \code{"exclosures"}, \code{"exclosures_interp"}, \code{"dm_controls"}, and \code{"dm_controls_interp"}.
}
\keyword{datasets}
