% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodent_species.R
\name{rodent_species}
\alias{rodent_species}
\alias{all_species}
\alias{base_species}
\title{Rodent species abbreviations}
\usage{
rodent_species(
  species = NULL,
  set = NULL,
  nadot = FALSE,
  total = FALSE,
  type = "abbreviation"
)

all_species(
  species = NULL,
  nadot = FALSE,
  total = FALSE,
  type = "abbreviation"
)

base_species(
  species = NULL,
  nadot = FALSE,
  total = FALSE,
  type = "abbreviation"
)
}
\arguments{
\item{species}{\code{character}-valued vector of species names to include.}

\item{set}{\code{character} input of a specified set of species. Options include \code{all} (all species included) and \code{base} (all but PI).}

\item{nadot}{\code{logical} indicator if the dot should be added to the \code{"NA"} species name. Defaults to \code{FALSE}.}

\item{total}{\code{logical} value indicating if \code{"total"} should be added or not.}

\item{type}{\code{character} value indicating the output type. Current options include \code{abbreviation} (default, two-letter abbreviation) and \code{Latin} (full scientific names).}
}
\value{
\code{character} vector of species abbreviations.
}
\description{
Creates a simple \code{character} vector of abbreviations for the Portal \href{https://portal.naturecast.org/profiles.html}{Rodents}.
}
\examples{
 rodent_species()
 rodent_species(set = "all")
 rodent_species("BA")
 rodent_species("BA", set = "all")
 all_species()
 all_species(nadot = TRUE)
 all_species(total = TRUE)
 base_species()

}
