% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{rodent_spp}
\alias{rodent_spp}
\title{Rodent species abbreviations}
\usage{
rodent_spp(nadot = FALSE)
}
\arguments{
\item{nadot}{\code{logical} indicator if the dot should be added to the 
\code{"NA"} species name. Defaults to \code{FALSE}.}
}
\value{
\code{character} vector of species abbreviations.
}
\description{
Creates a simple \code{character} vector of abbreviations for
  the \href{https://portal.naturecast.org/profiles.html}{Portal Rodents}.
}
