% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{rodents_control}
\alias{rodents_control}
\title{Create a control lists for generating a rodents data table}
\usage{
rodents_control(name = "name", species = base_species(),
  total = TRUE, interpolate = FALSE, clean = FALSE,
  type = "Rodents", level = "Site", plots = "all",
  treatment = NULL, min_plots = 24, min_traps = 1,
  output = "abundance", fillweight = (output != "abundance"),
  unknowns = FALSE, time = "newmoon", na_drop = FALSE,
  zero_drop = switch(tolower(level), plot = FALSE, treatment = TRUE, site
  = TRUE), effort = FALSE, filename = paste0("rodents_", name, ".csv"),
  arg_checks = TRUE)
}
\arguments{
\item{name}{\code{character} value of the rodent data set name
used to enforce certain arguments. Currently used prefab data
sets are \code{"all"}, \code{"all_interp"}, \code{"controls"}, 
and \code{"controls_interp"}.}

\item{species}{\code{character}-valued vector of species names 
to include.}

\item{total}{\code{logical} value indicating if a total 
(sum across species) should be added or not. Only available if more than 
one species is included.}

\item{interpolate}{\code{logical} value indicating if the data should be
interpolated to fill in \code{NA} values (using 
\code{\link[forecast]{na.interp}}).}

\item{clean}{\code{logical} indicator of if only the rodent data
that passed QA/QC (\code{clean = TRUE}) or if all data 
(\code{clean = FALSE}) should be loaded.}

\item{type}{\code{character} value of the rodent data set type, according 
to pre-existing definitions. An alternative toggle to \code{species}. \cr
Either all species 
(\code{type = "Rodents"}) or only granivoes (\code{type = "Granivores"}).}

\item{level}{\code{character} indicating the type of summary:
\code{"Plot"}, \code{"Treatment"}, or \code{"Site"}. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}

\item{plots}{Specification of subset of plots. Can be a vector of 
\code{numeric} plots indicators or specific sets indicated by
\code{character} values: \code{"all"} plots or \code{"Longterm"} plots
(plots that have had the same treatment for the entire time series).}

\item{treatment}{\code{character} indicating the specific 
treatment(s) to trim to if \code{level = "Treatment"}: \code{"control"},
\code{"exclosure"}, \code{"removal"}, or \code{"spectabs"}}

\item{min_plots}{\code{integer} (or integer \code{numeric}) of the 
minimum number of plots surveyed for a survey to be used. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}

\item{min_traps}{\code{integer} (or integer \code{numeric}) of the 
minimum number of traps collected for a plot to be used. Pipes directly
to \code{\link[portalr]{summarize_rodent_data}}.}

\item{output}{\code{character} indicating the type of data:
\code{"abundance"}, \code{"biomass"}, or \code{"energy"}. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}

\item{fillweight}{\code{logical} specifier of whether to fill in unknown 
weights with other records from that individual or species, where 
possible.}

\item{unknowns}{\code{logical} indicator to either remove all 
individuals not identified to species (\code{unknowns = FALSE}) or 
sum them in an additional column (\code{unknowns = TRUE}.}

\item{time}{\code{character} value specifying the format of the time index 
in the output. Options are \code{"period"} (sequential Portal surveys), 
\code{"newmoon"} (lunar cycle numbering), and \code{"date"} (calendar 
date). \cr
The default \code{time = "newmoon"} produces an equispaced observation 
timestep, a common format format for discrete-time modeling.}

\item{na_drop}{\code{logical} indicator of if \code{NA} values 
(representing insufficient sampling) should be dropped.}

\item{zero_drop}{\code{logical} indicator of if \code{0} values 
(representing sufficient sampling but no detection) should be dropped.}

\item{effort}{\code{logical} indicator of if the effort columns should
be included in the output.}

\item{filename}{\code{character} name of the file for saving the 
output. The extension of the filename may be used for identifying
object requirements (for example, if the extension of \code{filename} is
\code{".csv"}, it will trigger use of \code{\link[utils]{write.csv}}).}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
Named \code{list} of \code{data_set}'s data-generating controls, 
 for input as part of \code{controls_rodents} in 
 \code{\link{prep_rodents}}.
}
\description{
Given the number of arguments into 
 \code{\link[portalr]{summarize_rodent_data}} via 
 \code{\link{prep_rodents_table}}, it helps to have a control \code{list}  
 to organize them. \code{rodents_control} provides a template for the 
 controls \code{list} used to define the specific arguments for a user's
 novel data set, setting the formal arguments to the basic default 
 values
}
\examples{
 rodents_controls(c("all", "controls"))
 all_PPonly <- rodents_control("all_PPonly", species = "PP")
 rodents_controls("all_PPonly", all_PPonly, arg_checks = FALSE)
 rodents_controls(c("all", "all_PPonly"), all_PPonly, arg_checks = FALSE)

}
