% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{rodents_options}
\alias{rodents_options}
\title{Prepare the data options for rodent data}
\usage{
rodents_options(type = NULL, start = 217, drop_spp = "PI",
  min_plots = 24, level = "Site", treatment = NULL, length = "all",
  output = "abundance", save = TRUE, filename = "all.csv")
}
\arguments{
\item{type}{"all" or "controls"}

\item{start}{newmoon number of the first sample to be included. Default 
value is \code{217}, corresponding to 1995-01-01.}

\item{drop_spp}{species names to drop from the table}

\item{min_plots}{minimum number of plots surveyed for a survey to be used}

\item{level}{for get_rodent_data, automatically set by type}

\item{treatment}{for get_rodent_data, automatically set by type}

\item{length}{for get_rodent_data, automatically set by type}

\item{output}{for get_rodent_data, automatically set by type}

\item{save}{logical if the data should be saved out}

\item{filename}{the name of the file for the saving, automatically set by
type}
}
\value{
control options list for rodents
}
\description{
Create a list of control options for the rodent data
}
