% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round_na.interp}
\alias{round_na.interp}
\title{Round an Interpolated Series}
\usage{
round_na.interp(
  x,
  lambda = NULL,
  linear = (frequency(x) <= 1 | sum(!is.na(x)) <= 2 * frequency(x)),
  digits = 0,
  min_val = 0
)
}
\arguments{
\item{x}{A time series passed directly to \code{\link[forecast:na.interp]{forecast::na.interp}}.}

\item{lambda}{Box-Cox transformation parameter passed directly to \code{\link[forecast:na.interp]{forecast::na.interp}}.}

\item{linear}{\code{logical} indicator of if linear interpolation should be used. Passed directly to \code{\link[forecast:na.interp]{forecast::na.interp}}.}

\item{digits}{\code{integer} or \code{numeric} integer of how many digits to round to. Passed directly to \code{\link{round}}.}

\item{min_val}{\code{integer} or \code{numeric} integer of minimum value allowable in the series.}
}
\value{
\code{numeric} series.
}
\description{
Wraps \code{\link{round}} around \code{\link[forecast:na.interp]{forecast::na.interp}} to provide a rounded interpolated series, which is then enforced to be greater than or equal to a minimum value (default \code{min_val = 0}) via \code{\link{pmax}}.
}
\examples{
   round_na.interp(x = c(1, 2, 3, NA, NA, 170))
   round_na.interp(x = c(-1, 2, 3, NA, NA, 170), min_val = 1)

}
\seealso{
Other utilities: 
\code{\link{file_ext}()},
\code{\link{foy}()},
\code{\link{ifnull}()},
\code{\link{messages}},
\code{\link{named_null_list}()},
\code{\link{package_version_finder}()},
\code{\link{update_list}()}
}
\concept{utilities}
