% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp.R
\name{run_web_app}
\alias{run_web_app}
\title{Launch the Portal Forecast Web Application}
\usage{
run_web_app()
}
\description{
Launches a local version of the web application by running \code{\link[shiny]{runApp}} pointed to the \code{app} subdirectory in the local \code{portalcasting} package folder.
}
