% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{sandbox_welcome}
\alias{sandbox_welcome}
\title{Sandbox welcome}
\usage{
sandbox_welcome(main = ".", quiet = FALSE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{quiet}{\code{logical} indicator if progress message should be
quieted.}
}
\value{
\code{NULL}, as message is printed.
}
\description{
Create a welcome message for the sandbox
}
\examples{
 sandbox_welcome()

}
