% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{save_forecast_output}
\alias{save_forecast_output}
\title{Save forecast output to files}
\usage{
save_forecast_output(all, controls, name, tree)
}
\arguments{
\item{all}{Ouput from a forecasting model function (e.g., 
\code{\link{AutoArima}}) run on the "all" data. Required to be a 
\code{list} with two elements named \code{forecast} and \code{aic}.}

\item{controls}{Ouput from a forecasting model function (e.g., 
\code{\link{AutoArima}}) run on the "controls" data. Required to be a 
\code{list} with two elements named \code{forecast} and \code{aic}.}

\item{name}{\code{character} model name for saving in files.}

\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}
}
\description{
Save out the forecast abundances and AIC values from a given 
  set of forecasts on all and control-only rodent data.
}
