% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{select_forecasts}
\alias{select_forecasts}
\title{Find Casts that Fit Specifications}
\usage{
select_forecasts(
  main = ".",
  forecast_ids = NULL,
  forecast_groups = NULL,
  models = NULL,
  datasets = NULL,
  species = NULL,
  historic_end_newmoonnumbers = NULL
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{forecast_ids}{\code{integer} (or integer \code{numeric}) values representing the casts of interest, as indexed within the directory in the \code{casts} sub folder. See the casts metadata file (\code{forecasts_metadata.csv}) for summary information.}

\item{forecast_groups}{\code{integer} (or integer \code{numeric}) value of the cast groups to include. Default value is \code{NULL}, which equates to no selection with respect to \code{forecast_group}.}

\item{models}{\code{character} values of the names of the models to include. Default value is \code{NULL}, which equates to no selection with respect to \code{model}.}

\item{datasets}{\code{character} values of the rodent data sets to include. Default value is \code{NULL}, which equates to no selection with respect to \code{dataset}.}

\item{species}{\code{character} value of the species codes (or \code{"total"} for the total across species) to include. Default value is \code{NULL}, which equates to no selection with respect to \code{species}.}

\item{historic_end_newmoonnumbers}{\code{integer} (or integer \code{numeric}) newmoon numbers of the forecast origin. Default value is \code{NULL}, which equates to no selection.}
}
\value{
\code{data.frame} of the \code{cast_tab}.
}
\description{
Determines the casts that match user specifications. \cr
Functionally, a wrapper on \code{\link{read_foreforecasts_metadata}} with filtering for specifications that provides a simple user interface to the large set of available casts via the metadata.
}
