% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{select_models}
\alias{select_models}
\title{Select Models to Forecast or Hindcast With}
\usage{
select_models(model_dir, models)
}
\arguments{
\item{model_dir}{directory name where the model files reside}

\item{models}{names of models to run or "all" to run them all}
}
\value{
names of files to be run
}
\description{
Verify that models requested are available and select them
}
