% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{select_most_ab_spp}
\alias{select_most_ab_spp}
\title{Select the most abundant species from a forecast or hindcast}
\usage{
select_most_ab_spp(topx = 3, tree = dirtree(),
  species = rodent_spp(), level = "Controls",
  cast_type = "forecasts", cast_date = NULL, model = "Ensemble",
  lead = 1, from_date = NULL)
}
\arguments{
\item{topx}{\code{integer}-conformable numeric value for the top number of
species to select.}

\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{species}{\code{character} vector of the species codes (or 
\code{"total"} for the total across species) to be selected from or 
\code{NULL} to include all species and the total.}

\item{level}{\code{character} value of the level of interest (\code{"All"} 
or \code{"Controls"}).}

\item{cast_type}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"forecasts"}.}

\item{cast_date}{\code{Date} the predictions were made. Used to select the
file in the predictions subdirectory. If \code{NULL} (default), the
most recently made -cast is selected.}

\item{model}{\code{character} value of the name (or \code{"Ensemble"}) of
the model to be plotted.}

\item{lead}{\code{integer}-conformable lead of the newmoon number used to
select the data plotted.}

\item{from_date}{\code{Date} to be used as a reference of when to count
the \code{lead} from. If \code{NULL} (default), for 
\code{cast_type = "forecasts"}, \code{from_date = cast_date} and 
\code{plot_species_casts} is not yet reliable for 
\code{cast_type = "hindcasts"}.}
}
\value{
\code{character} vector of length \code{topx} of the species codes
  (see \code{\link{rodent_spp}}) of the selected species.
}
\description{
Given a forecast or hindcast, determine the most abundant
  species predicted. Currently only reliable for forecasts.
}
