% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{setup_portalcast_dir}
\alias{setup_portalcast_dir}
\title{Create a portalcasting directory}
\usage{
setup_portalcast_dir(main_path = pc_path("", "~"), models = c("autoarima",
  "esss", "nbgarch", "pevgarch"))
}
\arguments{
\item{main_path}{directory name for the location of the portalcasting 
folder}

\item{models}{names of model scripts to include}
}
\value{
Nothing
}
\description{
Create and populate a full portalcasting directory or any 
  missing components.
}
