% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{step_casts}
\alias{step_casts}
\title{Step through the subsequent casts}
\usage{
step_casts(options_all = all_options())
}
\arguments{
\item{options_all}{full options list}
}
\value{
nothing
}
\description{
If there are multiple steps in the hindcast, update the data
  and continue casting
}
