% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{step_hind_forward}
\alias{step_hind_forward}
\title{Step the hind_step values forward}
\usage{
step_hind_forward(data_control = list())
}
\arguments{
\item{data_control}{A \code{list} of arguments to control the filling of
the data subdirectory. See \code{\link{data_control}}. 
Values not set assume defaults.}
}
\value{
Updated \code{options_all} \code{list} for the next -cast.
}
\description{
Iterate the \code{hind_step} elements of options \code{list}s
  through subsequent-to-the-first -cast(s) within \code{\link{casts}}.
}
