% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{sub_path}
\alias{sub_path}
\title{Determine the path for a specific sub directory}
\usage{
sub_path(tree = dirtree(), specific_sub)
}
\arguments{
\item{tree}{the name tree of the forecasting directory}

\item{specific_sub}{the name of the specific subdirectory of interest}
}
\value{
the paths of the sub folders in the directory tree
}
\description{
Return the paths for the sub directories
}
