% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{sub_paths}
\alias{sub_paths}
\title{Determine the paths for the sub directories}
\usage{
sub_paths(tree = dirtree())
}
\arguments{
\item{tree}{the name tree of the forecasting directory}
}
\value{
the paths of the sub folders in the directory tree
}
\description{
Return the paths for the sub directories
}
