% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{subdirs}
\alias{subdirs}
\title{Define the names of the subdirectories in a forecasting directory}
\usage{
subdirs(type = "portalcasting", subs = NULL)
}
\arguments{
\item{type}{name for quick generation of subdirectory names}

\item{subs}{names of additional subdirectories to add}
}
\value{
character vectory of subdirectory names
}
\description{
Convenience function for producing a vector of subdirectory
  names
}
