% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{summarize_daily_weather_by_moon}
\alias{summarize_daily_weather_by_moon}
\title{Summarize a daily weather table by newmoons}
\usage{
summarize_daily_weather_by_moon(x)
}
\arguments{
\item{x}{\code{data.frame} of daily weather, with columns named
\code{"date"}, \code{"mintemp"}, \code{"maxtemp"}, \code{"meantemp"}, 
\code{"precipitation"}, and \code{"moon"}.}
}
\value{
\code{data.frame} of \code{x} summarized and arranged by
 \code{x$moon}.
}
\description{
Summarizes a daily weather table by newmoons. Taking the 
 max date, min of the min temperatures, max of the max temperatures,
 mean of the mean temperatures, and sum of the precipitation.
}
\examples{
 \donttest{
  create_dir()
  fill_raw()
  raw_path <- raw_path()
  moons <- prep_moons()
  weather <- portalr::weather("daily", TRUE, raw_path)
  weather <- add_date_from_components(weather)
  weather <- add_moons_from_date(weather, moons)
  summarize_daily_weather_by_moon(weather)
 }

}
