% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{summarize_daily_weather_by_newmoon}
\alias{summarize_daily_weather_by_newmoon}
\title{Summarize a daily weather table by newmoons}
\usage{
summarize_daily_weather_by_newmoon(x)
}
\arguments{
\item{x}{\code{data.frame} of daily weather, with columns named
\code{"date"}, \code{"mintemp"}, \code{"maxtemp"}, \code{"meantemp"}, 
\code{"precipitation"}, and \code{"newmoonnumber"}.}
}
\value{
\code{data.frame} of \code{x} summarized and arranged by
 \code{x$newmoonnumber}.
}
\description{
Summarizes a daily weather table by newmoons. Taking the 
 max date, min of the min temperatures, max of the max temperatures,
 mean of the mean temperatures, and sum of the precipitation.
}
\examples{
 \donttest{
  create_dir()
  fill_raw()
  "\%>\%" <- dplyr::"\%>\%"
  raw_path <- sub_paths(specific_subs = "raw")
  moons <- prep_moons()
  weather("daily", TRUE, raw_path) \%>\% 
  add_date_from_components() \%>\%
  dplyr::select(-c(year, month, day, battery_low, locally_measured))  \%>\%
  add_newmoons_from_date(moons) \%>\%
  summarize_daily_weather_by_newmoon()
 }

}
