% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{target_moons}
\alias{target_moons}
\alias{last_moon}
\title{Determine specific newmoon numbers}
\usage{
target_moons(moons = NULL, end_moon = NULL, lead_time = 12, date = Sys.Date())

last_moon(moons = NULL, date = Sys.Date())
}
\arguments{
\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the
number of timesteps forward a cast will cover.}

\item{date}{\code{Date} used for defining a window (starting with moons
subsequent to the date) or a previous moon (before or on the date).}

\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{control_files}{\code{list} of names of the folders and files within
the sub directories and saving strategies (save, overwrite, append, etc.).
Generally shouldn't need to be edited. See \code{\link{files_control}}.}
}
\value{
\code{target_moons}: \code{numeric} vector of the moon numbers 
 targeted by the date window.
 \code{last_newmoon}:  \code{numeric} value of the last moon number.
}
\description{
\code{target_moons} determines which moons are in a window, based on a
 date and window width. \cr \cr
 \code{last_moon} determines the most recently passed newmoon, based on
 a date.
}
\examples{
 \donttest{
  setup_dir()
  target_moons()
  last_moon()
 }

}
