% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{template_autoarima}
\alias{template_autoarima}
\title{Create template autoarima text}
\usage{
template_autoarima()
}
\value{
text of the template autoarima script
}
\description{
Creates the text for a template autoarima script
}
