% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{template_esss}
\alias{template_esss}
\title{Create template esss text}
\usage{
template_esss()
}
\value{
text of the template esss script
}
\description{
Creates the text for a template esss script
}
