% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_script_generation.R
\name{template_nbgarch}
\alias{template_nbgarch}
\title{Create template nbGARCH text}
\usage{
template_nbgarch(tree = dirtree())
}
\arguments{
\item{tree}{directory tree}
}
\value{
text of the template nbGARCH script
}
\description{
Creates the text for a template nbGARCH script
}
