% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{template_pevgarch}
\alias{template_pevgarch}
\title{Create template pevGARCH text}
\usage{
template_pevgarch()
}
\value{
text of the template pevGARCH script
}
\description{
Creates the text for a template pevGARCH script
}
