% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{transfer_hist_covariate_forecasts}
\alias{transfer_hist_covariate_forecasts}
\title{Transfer the historical covariate data forecasts to the data folder}
\usage{
transfer_hist_covariate_forecasts(options_data = data_options())
}
\arguments{
\item{options_data}{Class-\code{data_options} \code{list} of control 
options \code{list}s for setting up the data structures.}
}
\description{
Currently, the historical covariate covariate forecasts are 
  held in the \code{extdata} folder in the package directory, and the
  transfer is hard-wired to come from there. This will be relaxed in the 
  future.
}
