% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{transfer_historical_covariate_forecasts}
\alias{transfer_historical_covariate_forecasts}
\title{Transfer the historical covariate data forecasts to the data folder}
\usage{
transfer_historical_covariate_forecasts(tree = dirtree())
}
\arguments{
\item{tree}{directory tree}
}
\value{
nothing
}
\description{
Currently, the historical covariate data forecasts are being
  held in the \code{extdata} folder in the package directory, and are the
  transfer is hard-wired to come from there. In the future, we should 
  consider housing the data differently (they'll be continuously updated
  with the forecasts), and thus we'll want to un-hard-wire that.
}
