% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{transfer_trapping_table}
\alias{transfer_trapping_table}
\title{Transfer the trapping table to the data folder}
\usage{
transfer_trapping_table(options_data = data_options())
}
\arguments{
\item{options_data}{Class-\code{data_options} list containing available
for controlling the set up and population of the data folder in the 
portalcasting directory (see \code{\link{data_options}}).}
}
\description{
Move the trapping table from the PortalData sub directory to 
  the data subdirectory for use.
}
