% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{transpose_args}
\alias{transpose_args}
\title{Transpose argument lists for use in do.call}
\usage{
transpose_args(in_args = NULL, eval_args = NULL, enquote_args = NULL)
}
\arguments{
\item{in_args}{\code{list} of arguments that can be of varying lengths,
which need to be combined into a \code{list} of argument values that can 
be easily input into \code{\link{do.call}}.}

\item{eval_args}{\code{character} vector (or \code{NULL} if not needed)
of any arguments in \code{in_args} need to be wrapped in 
\code{\link{eval}} for replication purposes. \cr \cr
This is generally needed if, for example, a function is used to pass
a list of values (e.g., \code{species} in 
\code{\link{rodents_controls}}.}

\item{enquote_args}{\code{character} vector (or \code{NULL} if not needed)
of any arguments in \code{in_args} need to be wrapped in 
\code{\link{enquote}} for replication purposes. \cr \cr
This is generally needed if, for example, an argument's input is another
argument (e.g., \code{name} in 
\code{\link{rodents_controls}}.}
}
\value{
\code{list} of argument values that can be easily input into 
  \code{\link{do.call}} (each element is a \code{list} of arguments.
}
\description{
Arguments coming into functions, when captured using 
 \code{\link[DesignLibrary]{match.call.defaults}}, require expansion and
 transposition for use within other functions.
}
\details{
As an example, see the source code of 
 \code{\link{rodents_controls}}.
}
\examples{
 in_args <- list(x = 1, y = 1:2, z = NULL)
 transpose_args(in_args, NULL)

}
