% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{trim_moons}
\alias{trim_moons}
\title{Trim a moons table based on target moons}
\usage{
trim_moons(moons = NULL, target_moons = NULL,
  retain_target_moons = TRUE, target_cols = c("newmoonnumber",
  "newmoondate", "period", "censusdate"))
}
\arguments{
\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{target_moons}{\code{integer}-conformable newmoon numbers, usually
made through \code{\link{target_newmoons}} of the moons to either 
drop or retain based on \code{retain_target_moons}.}

\item{retain_target_moons}{\code{logical} value that dictates if 
\code{target_moons} are to be dropped (\code{retain_target_moons = FALSE}
or retained (\code{retain_target_moons = TRUE})}

\item{target_cols}{\code{character} vector of columns to retain.}
}
\value{
\code{moons} and a \code{data.frame} but trimmed.
}
\description{
Some functions require that a data table of time (the moons)
 table only includes specific time stamps (newmoon numbers). This function
 is a simple utility to reduce the moons table to that as well as the 
 columns.
}
