% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{trim_treatment}
\alias{trim_treatment}
\title{Remove the treatment column from the rodent data table}
\usage{
trim_treatment(rodents, level = "Site", treatment = NULL)
}
\arguments{
\item{rodents}{Class-\code{rodents} \code{data.table} of rodent data.}

\item{level}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{treatment}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}
}
\value{
Trimmed \code{data}.
}
\description{
Remove the treatment column from the rodent data table if 
  needed (in the case of controls, for example, where the data have been
  reduced already so the column is vestigial).
}
