% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tsglm.R
\name{tsglm models}
\alias{tsglm models}
\alias{meta_tsglm}
\alias{forecast.tsglm}
\title{Create, Run, and Forecast Multi-model Inference TSGLM Model Runs}
\usage{
meta_tsglm(
  ts,
  model,
  distr,
  link,
  lag,
  submodels,
  covariates,
  metadata,
  quiet = FALSE
)

\method{forecast}{tsglm}(object, h, level, ...)
}
\arguments{
\item{ts}{Non-negative \code{integer}-conformable vector of rodent abundances to use in forecasting. See \code{\link{prepare_abundance}}.}

\item{model}{A named \code{list} of model linear predictors. See \code{\link[tscount:tsglm]{tsglm}}.}

\item{distr}{\code{character} of the response distribution. See \code{\link[tscount:tsglm]{tsglm}}.}

\item{link}{\code{character} of the link function. See \code{\link[tscount:tsglm]{tsglm}}.}

\item{lag}{\code{integer}-conformable value of the number of timesteps used in a bulk lagging for all covariates in all submodels.}

\item{submodels}{\code{list} of \code{character} vectors defining the covariates to include in each of the submodels.}

\item{covariates}{\code{data.frame} of covariates used in modeling. See \code{\link{prepare_covariates}}.}

\item{metadata}{\code{list} of model control elements. See \code{\link{prepare_metadata}}.}

\item{quiet}{\code{logic} indicator controlling if messages are printed.}

\item{object}{A \code{tsglm}-class object.}

\item{h}{\code{integer}-conformable number of steps forward to forecast. Passed into \code{predict} as \code{n.ahead}.}

\item{level}{\code{numeric} of the confidence level to use in summarizing the predictions.}

\item{...}{Additional parameters passed into \code{predict}.}
}
\value{
\code{meta_tsglm}: An object of class \code{"tsglm"} with additional elements defining the submodel and lag. \cr \cr
\code{forecast.tsglm}: \code{list} with \code{"forecast"}-class with named elements including \code{"mean"}, \code{"lower"}, \code{"upper"}, and \code{"newxreg"} (if provided for prediction) as well as the other elements returned by \code{predict}.
}
\description{
Using the tscount (Liboschik et al. 2017) package to forecast time series of counts. \cr \cr
\code{meta_tsglm}: Combines the model running with the covariate preparation functions for a multi-model \code{\link[tscount:tsglm]{tsglm}} (from the tscount (Liboschik et al. 2017) package) model. \cr \cr
\code{forecast.tsglm}: A wrapper around the \code{predict} function for tsglm objects that produces a \code{"forecast"}-class object.
}
\references{
Liboschik T., K. Fokianos, and R. Fried. 2017. tscount: An R Package for Analysis of Count Time Series Following Generalized Linear Models. Journal of Statistical Software, 82:1-51. \href{https://doi.org/10.18637/jss.v082.i05}{URL}.
}
