% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{update_covariates}
\alias{update_covariates}
\title{Update the covariate data for a step in a hindcast}
\usage{
update_covariates(moons, options_covariates)
}
\arguments{
\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}
}
\value{
\code{covariates_options}: a \code{covariates_options} \code{list} 
  of settings controlling the covariates data creation.
}
\description{
Given an updated \code{options_covariates$hind_step}, update
  the covariate data.
}
