% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{update_covfcast_options}
\alias{update_covfcast_options}
\title{Update the data options for covariate forecasts based on existing 
  data}
\usage{
update_covfcast_options(options_covariates, hist_cov, moons)
}
\arguments{
\item{options_covariates}{A class-\code{covariates_options} \code{list} of 
settings controlling the covariates data creation.}

\item{hist_cov}{Historical covariate data table as a code{covariates}-class 
\code{data.frame}, returned from \code{\link{prep_hist_covariates}}.}

\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}
}
\value{
An updated \code{covariates_options} \code{list} of control options 
  for covariates.
}
\description{
Update the covariate data control options based on the 
  historical covariate data and moon data.
}
