% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{update_rodents}
\alias{update_rodents}
\title{Update the rodent data for a step in a hindcast}
\usage{
update_rodents(options_rodents)
}
\arguments{
\item{options_rodents}{rodents options list}
}
\value{
a list of two dataframes, all plots and control plots, updated
}
\description{
Given an updated hind_step, update the rodent data objects
}
