% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{update_rodents_list}
\alias{update_rodents_list}
\title{Update the rodent data for a step in a hindcast}
\usage{
update_rodents_list(options_rodents)
}
\arguments{
\item{options_rodents}{A class-\code{rodents_options} \code{list} of 
settings controlling the rodents data creation.}
}
\value{
A \code{rodents_list}-class \code{list} of two \code{rodents}-class
  \code{data.frames}: \code{"all"} (containing counts across all plots)
  \code{"controls"} (containing counts across just the control plots), 
  updated for the particular step in the hindcast.
}
\description{
Given an updated \code{options_rodents$hind_step}, update
  the rodent data.
}
