% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{verify}
\alias{verify}
\title{Verify that a required folder exists}
\usage{
verify(path = NULL, level = NULL)
}
\arguments{
\item{path}{\code{character} vector of the folder paths.}

\item{level}{\code{character} vector of the folder names.}
}
\description{
If any of the folders (specified by \code{path} and named by
 \code{level}) do not exist, the operation throws an error.
}
