% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{verify_and_create}
\alias{verify_and_create}
\alias{verify}
\alias{create}
\title{Verify that folders exist and create folders}
\usage{
verify(paths = NULL, names = NULL, arg_checks = TRUE)

create(paths = NULL, names = NULL, quiet = FALSE, verbose = FALSE,
  arg_checks = TRUE)
}
\arguments{
\item{paths}{\code{character} vector of the folder paths.}

\item{names}{\code{character} vector of the names of the levels in the
hierarchy at which the folders will exist or the folder names. For the
purposes of messaging and not inherently required to be non-\code{NULL}.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}

\item{quiet}{\code{logical} indicator if messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out
all of the information or not (and thus just the tidy messages).}
}
\value{
\code{verify}: throws an error if any of the folders do not exist, 
 otherwise \code{NULL}.
 \code{create}: \code{NULL}.
}
\description{
\code{verify} throws an error if any of the folders (specified by 
 \code{path} and named by \code{level}) do not exist. \cr \cr
 \code{create} creates a requested folder if it does not already exist.
}
\examples{
 \donttest{
  mainp <- main_path()
  create(mainp, "main")
  verify(mainp, "main")
 }

}
