% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{verify_PortalData}
\alias{verify_PortalData}
\title{Verify that the PortalData sub is present and has required data}
\usage{
verify_PortalData(tree = dirtree(), filename = "moon_dates.csv")
}
\arguments{
\item{tree}{directory tree}

\item{filename}{name of the file to specifically check}
}
\value{
nothing
}
\description{
Check that the PortalData subdirectory exists and has the
  needed file. If the file is not present, the directory is filled.
}
