% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{verify_raw_data}
\alias{verify_raw_data}
\title{Verify that the raw data folder exists}
\usage{
verify_raw_data(main = ".", raw_data = "PortalData", arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{raw_data}{\code{character} value indicating the name of the raw
data directory. A standard portalcasting directory downloads the raw data
files into from the PortalData repository, so 
\code{raw_data = "PortalData"}.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
\code{logical} indicator of whether or not the raw data folder
 exists.
}
\description{
Check that the raw data folder exists
}
\examples{
 \donttest{
  create_dir()
  fill_raw()
  verify_raw_data()
 }

}
