% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp_server.R
\name{web app server}
\alias{web app server}
\alias{portal_forecast_server}
\title{Generate the Server Code for the Web App}
\usage{
portal_forecast_server(main = ".", input, output, session)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{input}{\code{input} list for the UI.}

\item{output}{\code{output} list for the GenEst GUI.}

\item{session}{Environment for the GenEst GUI.}
}
\value{
\itemize{
\item \code{portal_forecast_server}: An observer reference class object (see \code{\link[shiny:observeEvent]{shiny::observeEvent}} and \code{\link[shiny:observe]{shiny::observe}}).
}
}
\description{
Generate the Server Code for the Web App
}
