% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp.R
\name{web app utilities}
\alias{web app utilities}
\alias{historic_end_newmoonnumber_list}
\alias{model_list}
\alias{species_list}
\title{Utilities for the Web App}
\usage{
historic_end_newmoonnumber_list(main = ".")

model_list()

species_list()
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
\code{historic_end_newmoonnumber_list}: \code{vector} of \code{integer} newmoonnumbers. \cr \cr
\code{model_list}: \code{vector} of \code{character} models named with print names from controls list. \cr \cr
\code{species_list}: \code{vector} of \code{character} species codes named with their Latin names.
}
\description{
Helper functions
}
