% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp.R
\name{web app}
\alias{web app}
\alias{run_web_app}
\title{Build and Launch the Portal Forecast Web Application}
\usage{
run_web_app(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\description{
Constructs and launches a local version of the web application by running \code{\link[shiny:runApp]{shiny::runApp}} pointed to the \code{app} subdirectory in the local \code{portalcasting} package folder.
}
