% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_output.R
\name{write_data}
\alias{write_data}
\title{Save Data Out to a File and Return It	(Invisibly)}
\usage{
write_data(
  x = NULL,
  main = ".",
  subdirectory = "data",
  save = TRUE,
  overwrite = TRUE,
  filename = NULL,
  quiet = FALSE
)
}
\arguments{
\item{x}{\code{data.frame} or \code{list} to be written out.}

\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{subdirectory}{\code{character} value defining the data subdirectory of the portalcasting directory tree.}

\item{save}{\code{logic} indicator controlling if \code{x} should be saved out.}

\item{overwrite}{\code{logic} indicator of whether or not file writing sould occur even if a local copy already exists.}

\item{filename}{\code{character} name of the file for saving \code{x}.}

\item{quiet}{\code{logic} indicator if messages should be quieted.}
}
\value{
\code{x} as input, \code{\link[base:invisible]{invisible}}-ly.
}
\description{
Save inputted data out to a data file if requested and return it to the console, \code{\link[base:invisible]{invisible}}-ly. Currently available for yaml, csv, and json file extensions.
}
\examples{
   \donttest{
     main1 <- file.path(tempdir(), "standard")
     create_dir(main = main1)
     write_data(main = main1, x = data.frame(rnorm(10)), filename = "xx.csv")
     unlink(main1, force = TRUE, recursive = TRUE)
   }

}
