% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{zenodo_downloads}
\alias{zenodo_downloads}
\title{Create a downloads list for zenodo downloads}
\usage{
zenodo_downloads(
  concept_rec_id = NULL,
  rec_version = "latest",
  rec_id = NULL,
  arg_checks = TRUE
)
}
\arguments{
\item{concept_rec_id}{Concept record identifier, a \code{character} value
corresponding to the Zenodo concept.}

\item{rec_version}{\code{character} value of the version number or 
\code{"latest"} for the data to be download.}

\item{rec_id}{Optional input record identifier, a \code{character} value
corresponding to the Zenodo record.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
\code{list} of \code{list}s of arguments to \code{\link{download}}.
}
\description{
Create a downloads \code{list} for downloads from Zenodo.
}
\examples{
 zenodo_downloads()

}
