% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_unzip_forecasts.R
\name{zip_unzip}
\alias{zip_unzip}
\title{Zip and unzip forecasts by forecast date}
\usage{
zip_unzip(type = NULL, forecast_path = "forecasts/", date = NULL)
}
\arguments{
\item{type}{either zip or unzip}

\item{forecast_path}{location of forecast directory}

\item{date}{forecast date to use while zipping}
}
\description{
Zip and unzip forecasts by forecast date
}
