% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{add to cast tab}
\alias{add to cast tab}
\alias{add_lead_to_cast_tab}
\alias{add_err_to_cast_tab}
\alias{add_covered_to_cast_tab}
\alias{add_obs_to_cast_tab}
\title{Add the Associated Values to a Cast Tab}
\usage{
add_lead_to_cast_tab(
  main = ".",
  settings = directory_settings(),
  cast_tab = NULL
)

add_err_to_cast_tab(
  main = ".",
  settings = directory_settings(),
  cast_tab = NULL
)

add_covered_to_cast_tab(
  main = ".",
  settings = directory_settings(),
  cast_tab = NULL
)

add_obs_to_cast_tab(
  main = ".",
  settings = directory_settings(),
  cast_tab = NULL
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{cast_tab}{A \code{data.frame} of a cast's output. See \code{\link{read_cast_tab}}.}
}
\value{
\code{data.frame} of \code{cast_tab} with an additional column or columns if needed.
}
\description{
Add values to a cast's cast tab. If necessary components are missing (such as no observations added yet and the user requests errors), the missing components are added. \cr \cr
 \code{add_lead_to_cast_tab} adds a column of lead times. \cr \cr
 \code{add_obs_to_cast_tab} appends a column of observations. \cr \cr
 \code{add_err_to_cast_tab} adds a column of raw error values. \cr \cr
 \code{add_covered_to_cast_tab} appends a \code{logical} column indicating if the observation was within the prediction interval.
}
\details{
If a model interpolated a data set, \code{add_obs_to_cast_tab} adds the true (non-interpolated) observations so that model predictions are all compared to the same data.
}
