% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{add_addl_future_moons}
\alias{add_addl_future_moons}
\title{Add more more moons to accomplish requested forecast length}
\usage{
add_addl_future_moons(future_moons, fdate)
}
\arguments{
\item{future_moons}{A \code{moons}-class \code{data.frame} forward from 
the historic moons table (produced by \code{\link{prep_moons}}), as
generated by \code{\link[portalr]{get_future_moons}}.}

\item{fdate}{\code{Date} from which future is defined, typically today's 
date (using \code{\link{today}}).}
}
\value{
appended moon table
}
\description{
Add extra future moons to the moon table to achieve requested
  forecast length. Because the moon table might not have the most recent 
  moons, more rows than initially requested may need to be added to the 
  table for it to be of the required coverage.
}
