% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_date_from_components}
\alias{add_date_from_components}
\title{Add a date to a table that has the year month and day as components}
\usage{
add_date_from_components(x)
}
\arguments{
\item{x}{\code{data.frame} with columns named \code{year}, \code{month},
and \code{day}.}
}
\value{
\code{data.frame} \code{x} with column of \code{Date}s 
 named \code{date} added.
}
\description{
Add a date (as a \code{Date}) column to a table that has the 
 year month and day as components.
}
\examples{
 df <- data.frame(year = 2010, month = 2, day = 1:10)
 add_date_from_components(df)

}
