% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{add_ensemble}
\alias{add_ensemble}
\title{Add ensemble model to forecasts}
\usage{
add_ensemble(main = ".", end_moon = NULL, cast_date = Sys.Date(),
  confidence_level = 0.9, quiet = FALSE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}

\item{confidence_level}{\code{numeric} confidence level used in 
summarizing model output. Must be between \code{0} and \code{1}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\value{
Forecast abundance table for the ensemble model.
}
\description{
Add the predictions of an ensemble model to the forecasting 
  files.
}
