% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{add_ensemble}
\alias{add_ensemble}
\title{Add ensemble model to forecasts}
\usage{
add_ensemble(options_cast = cast_options())
}
\arguments{
\item{options_cast}{Class-\code{cast_options} \code{list} containing the
hind- or forecasting options. See \code{\link{cast_options}}.}
}
\value{
Forecast abundance table for the ensemble model.
}
\description{
Add the predictions of an ensemble model to the forecasting 
  files.
}
