% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{add_ensemble}
\alias{add_ensemble}
\title{Add Ensemble Model to Forecasts}
\usage{
add_ensemble(model_metadata, temp_dir, pred_dir)
}
\arguments{
\item{options_cast}{casting options}
}
\value{
list of [1] the forecasts and [2] the model AIC values
}
\description{
Add ensembles to the forecast files
}
