% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{add_future_moons}
\alias{add_future_moons}
\title{Add future moons to the moon table}
\usage{
add_future_moons(moons = prep_moons(), data_options = moons_options())
}
\arguments{
\item{moons}{current newmoonnumber table}

\item{data_options}{control options list for moons}
}
\value{
appended moon table
}
\description{
Add future moons to the moon table, counting forward from
  \code{fdate}. Because the moon table might not have the most recent 
  moons, more rows than \code{n_future_moons} may be added to the table.
}
