% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{add_future_moons}
\alias{add_future_moons}
\title{Add future moons to the moon table}
\usage{
add_future_moons(moons = prep_moons(), options_moons = moons_options())
}
\arguments{
\item{moons}{Moons data table as a code{moons}-class \code{data.frame}.}

\item{options_moons}{control options list for moons}
}
\value{
Appended \code{moons} data table as a \code{moons}-class 
  \code{data.frame}.
}
\description{
Add future moons to the moon table, counting forward from
  the forecast date (\code{options_moons$fdate}). Because the \code{moons}
  table might not have the most recent moons, more rows than anticipated
  (\code{options_moons$n_future_moons}) may need to be added to the table.
}
