% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{add_moons_from_date}
\alias{add_moons_from_date}
\title{Add a newmoon number to a table that has the date}
\usage{
add_moons_from_date(df, moons = NULL, arg_checks = TRUE)
}
\arguments{
\item{df}{\code{data.frame} with column of newmoon \code{Date}s 
named \code{date}.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
\code{data.frame} \code{x} with column of \code{newmoonnumber}s 
 added.
}
\description{
Add a \code{moon} (newmoon number) column to a table that has 
 a \code{date} (as a \code{Date}) column.
}
\examples{
 \donttest{
  create_dir()
  fill_dir()
  moons <- prep_moons()
  raw_path <- raw_path()
  weather <- portalr::weather("daily", TRUE, raw_path)
  add_moons_from_date(weather, moons)
 }

}
