% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{all_options}
\alias{all_options}
\alias{dir_options}
\alias{PortalData_options}
\alias{data_options}
\alias{moons_options}
\alias{rodents_options}
\alias{covariates_options}
\alias{metadata_options}
\alias{predictions_options}
\alias{models_options}
\alias{cast_options}
\alias{check_options_args}
\title{Prepare options lists for a portalcasting directory}
\usage{
all_options(base = ".", main = "", subs = subdirs(type =
  "portalcasting"), quiet = FALSE, cast_date = today(),
  append_missing_to_raw = TRUE, m_save = TRUE,
  m_filename = "moons.csv", tmnt_type = NULL, start = 217,
  end = NULL, hind_step = 1, drop_spp = "PI", min_plots = 24,
  min_traps = 1, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  cov_hist = TRUE, cov_fcast = TRUE, yr = as.numeric(format(today(),
  "\%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL,
  nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", cast_type = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml",
  download_existing_predictions = FALSE, model = models(set =
  "prefab"), ensemble = TRUE, version = "latest", from_zenodo = TRUE,
  to_cleanup = c("tmp", "PortalData"))

dir_options(base = ".", main = "", subs = subdirs(type =
  "portalcasting"), quiet = FALSE, to_cleanup = c("tmp", "PortalData"))

PortalData_options(base = ".", main = "", subs = subdirs(type =
  "portalcasting"), quiet = FALSE, version = "latest",
  from_zenodo = TRUE)

data_options(base = ".", main = "", subs = subdirs(type =
  "portalcasting"), quiet = FALSE, cast_date = today(),
  append_missing_to_raw = TRUE, m_save = TRUE,
  m_filename = "moons.csv", tmnt_type = NULL, start = 217,
  end = NULL, hind_step = 1, drop_spp = "PI", min_plots = 24,
  level = "Site", treatment = NULL, plots = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  cov_hist = TRUE, cov_fcast = TRUE, yr = as.numeric(format(today(),
  "\%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL,
  nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", cast_type = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml")

moons_options(n_future_moons = 12, cast_date = today(),
  append_missing_to_raw = TRUE, save = TRUE, filename = "moons.csv",
  tree = dirtree(), quiet = FALSE)

rodents_options(cast_type = "forecasts", tmnt_type = NULL,
  start = 217, end = NULL, hind_step = 1, drop_spp = "PI",
  min_plots = 24, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", save = TRUE, filename = "all.csv",
  tree = dirtree(), quiet = FALSE)

covariates_options(cast_type = "forecasts", cov_hist = TRUE,
  cov_fcast = TRUE, cast_date = today(),
  yr = as.numeric(format(today(), "\%Y")), start = 217, end = NULL,
  hind_step = 1, lead_time = 12, min_lag = 6, fcast_nms = NULL,
  nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", save = TRUE,
  filename = "covariates.csv", tree = dirtree(), quiet = FALSE)

metadata_options(cast_date = today(), cast_type = "forecasts",
  confidence_level = 0.9, lead_time = 12, save = TRUE,
  filename = "metadata.yaml", quiet = FALSE, tree = dirtree())

predictions_options(base = ".", main = "", subs = subdirs(type =
  "portalcasting"), download_existing_predictions = FALSE,
  quiet = FALSE)

models_options(base = ".", main = "", subs = subdirs(type =
  "portalcasting"), quiet = FALSE, model = models(set = "prefab"))

cast_options(base = ".", main = "", subs = subdirs(type =
  "portalcasting"), quiet = FALSE, model = models(set = "prefab"),
  cast_type = "forecasts", cast_date = today(), ensemble = TRUE,
  start = 217, end = NULL, hind_step = 1, min_plots = 24,
  min_traps = 1)

check_options_args(base = ".", main = "", subs = subdirs(type =
  "portalcasting"), quiet = FALSE, cast_date = today(),
  append_missing_to_raw = TRUE, m_save = TRUE,
  m_filename = "moons.csv", tmnt_type = NULL, start = 217,
  end = NULL, hind_step = 1, drop_spp = "PI", min_plots = 24,
  min_traps = 1, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  cov_hist = TRUE, cov_fcast = TRUE, yr = as.numeric(format(today(),
  "\%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL,
  nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", cast_type = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml",
  download_existing_predictions = FALSE, model = models(set =
  "prefab"), ensemble = TRUE, version = "latest", from_zenodo = TRUE,
  n_future_moons = 12, save = TRUE, filename = "moons.csv",
  tree = dirtree(), to_cleanup = c("tmp", "PortalData"))
}
\arguments{
\item{base}{\code{character} name of the base folder where the directory 
should or does exist. Default \code{"."} (working directory). Will be 
created if it doesn't exist.}

\item{main}{\code{character} name of the main folder within the base that
contains the subdirectories. Default \code{""} (no main level included).}

\item{subs}{\code{character} vector naming the specific subdirectories
within the portalcasting directory tree. Default \code{subdirs(type = 
"portalcasting")} sets the subdirectories as \code{"predictions"}, 
\code{"models"}, \code{"PortalData"}, \code{"data"}, and \code{"tmp"}. 
It is generally not advised to change the subdirectories.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{cast_date}{\code{Date} from which future is defined, typically 
today's date (using \code{\link{today}}).}

\item{append_missing_to_raw}{\code{logical} indicator dictating if the 
missing moon dates should be appended to the raw data file (should be 
\code{TRUE} to allow the imported portalr functions to work properly).}

\item{m_save}{\code{logical} indicator controlling if the moons data should 
be saved out.}

\item{m_filename}{\code{character} name of the file for saving moons data.}

\item{tmnt_type}{Treatment type: \code{"all"} or \code{"controls"}.}

\item{start}{\code{integer} (or integer \code{numeric}) newmoon number of 
the first sample to be included. Default value is \code{217}, 
corresponding to \code{1995-01-01}.}

\item{end}{\code{integer} (or integer \code{numeric}) newmoon number of the
last sample to be included. Default value is \code{NULL}, which equates
to the most recently included sample. If \code{cast_type} is 
\code{"hindcasts"} and \code{end} is \code{NULL}, default becomes 
\code{490:403}, corresponding to \code{2017-01-01} back to
\code{2010-01-01}.}

\item{hind_step}{\code{integer} (or integer \code{numeric}) iteration 
parameter used to work across the input values for \code{end}. Default is
\code{1}, which is likely all it should be run using.}

\item{drop_spp}{\code{character}-valued vector of species names to drop 
from the forecasting.}

\item{min_plots}{\code{integer} (or integer \code{numeric}) of the minimum 
number of plots surveyed for a survey to be used.}

\item{min_traps}{\code{integer} (or integer \code{numeric}) of the minimum 
number of traps trapped for a plot to be used.}

\item{level}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{treatment}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{plots}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{output}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{r_save}{\code{logical} value indicating if the rodent data should be 
saved out.}

\item{r_filename}{\code{character} name of the file to save the rodent data
in.}

\item{cov_hist}{\code{logical} indicator of whether or not historical 
covariates are to be included.}

\item{cov_fcast}{\code{logical} indicator whether or not forecasted 
covariates are to be included.}

\item{yr}{\code{numeric} value of the year of today's date.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) number of moons
into the future the rodents are to be forecast.}

\item{min_lag}{\code{integer} (or integer \code{numeric}) of the minimum 
covariate lag time used in any model.}

\item{fcast_nms}{\code{integer} (or integer \code{numeric}) vector of 
newmoon numbers to be forecast for covariates.}

\item{nfcnm}{\code{integer} (or integer \code{numeric}) number of forecast 
newmoons for covariates.}

\item{append_fcast_csv}{\code{logical} indicator controlling if the new 
forecast should be appended to the historical forecasts for the purposes 
of hindcasting.}

\item{hist_fcast_file}{\code{character} name of the file where the 
historical covariate forecasts are held.}

\item{source_name}{\code{character} value for the name to give the 
covariaate forecast. Currently is \code{"current_archive"}. Previous to
\code{"current_archive"}, the data were retroactively filled in and are 
given the source name \code{"retroactive"}.}

\item{c_save}{\code{logical} indicator for if the covariate data should be 
saved out.}

\item{c_filename}{\code{character} value for the name of the covariate file 
for the saving if \code{c_save} is \code{TRUE}.}

\item{cast_type}{-cast type: \code{"forecasts"} or \code{"hindcasts"}.}

\item{confidence_level}{\code{numeric} confidence level used in 
summarizing model output. Must be between \code{0} and \code{1}.}

\item{meta_save}{\code{logical} indicator for if the metadata should be
saved out}

\item{meta_filename}{\code{character} value for the name of the metadata 
file for the saving if \code{c_save} is \code{TRUE}.}

\item{download_existing_predictions}{\code{logical} indicator for if the 
existing predictions files should be retrieved from the 
\href{https://github.com/weecology/portalPredictions}{portalPredictions 
repo}.}

\item{model}{\code{character} vector of the names of model scripts to 
include in a cast of the pipeline.}

\item{ensemble}{\code{logical} indicator of whether to create an ensemble
model.}

\item{version}{\code{character} values of the version number or 
\code{"latest"} (default) for the Portal Data to be download.}

\item{from_zenodo}{\code{logical} indicator of whether or not the Portal 
Data should come from Zenodo (come from GitHub if not).}

\item{to_cleanup}{\code{character} vector of subdirectory names to
cleanup after completion of casting.}

\item{n_future_moons}{\code{integer} (or integer \code{numeric}) value for 
the number of future moons to add.}

\item{save}{\code{logical} indicator if the specific data should be saved 
out.}

\item{filename}{\code{character} name of the file for the saving.}

\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}
}
\value{
\code{all_options}: class-\code{all_options} \code{list} of control
  options lists (\code{options_dir} generated by \code{dir_options}, 
  \code{options_PortalData} generated by 
  \code{PortalData_options}, \code{options_data} generated by 
  \code{data_options}, \code{options_predictions} generated by
  \code{predictions_options}, \code{options_models} generated by
  \code{models_options}, and \code{options_cast} generated by 
  \code{cast_options}.

\code{dir_options}: a \code{dir_options} \code{list} of options for 
  the directory structure creation.

\code{PortalData_options}: a \code{PortalData_options} \code{list} 
  of settings controlling the population of the PortalData subdirectory.

\code{data_options}: \code{data_options} \code{list} of control 
  options \code{list}s (\code{moons} generated by \code{moons_options}, 
  \code{rodents} generated by \code{rodents_options}, \code{covariates},
  generated by \code{covariates_options}, and \code{metadata} generated by
  \code{metadata_options}, as well as items \code{cast_type}, \code{tree},
  and \code{quiet} for simple use with all or ancillary data..

\code{moons_options}: a \code{moons_options} \code{list} of 
  settings controlling the moon data creation.

\code{rodents_options}: a \code{rodents_options} \code{list} of 
  settings controlling the rodents data creation.

\code{covariates_options}: a \code{covariates_options} \code{list} 
  of settings controlling the covariates data creation.

\code{metadata_options}: a \code{metadata_options} \code{list}
  of settings controlling the metadata creation.

\code{predictions_options}: a \code{predictions_options} 
  \code{list} of settings controlling the population of the predictions 
  subdirectory.

\code{models_options}: a \code{models_options} \code{list} of 
  settings controlling the population of the models subdirectory.

\code{cast_options}: a \code{cast_options} \code{list} of settings 
  controlling the running of \code{\link{portalcast}}.
}
\description{
Suite of functions used to generate options \code{list}s 
 that control the setup and use of a portalcasting directory: \cr \cr
 \code{all_options} creates an \code{all_options} \code{list} that is a
 full hierarchy of portalcasting options \code{list}s:
 \code{options_dir}, \code{options_PortalData}, \code{options_data}, 
 \code{options_predictions}, \code{options_models}, and 
 \code{options_cast}.

\code{dir_options} creates a \code{dir_options} \code{list} of
  control options for the directory set-up.

\code{PortalData_options} creates a \code{PortalData_options}
  \code{list} of options for filling the PortalData subdirectory.

\code{data_options} creates a \code{data_options} \code{list} 
  of \code{list}s of options for filling the data subdirectory.

\code{moons_options} creates a \code{moons_options} 
  \code{list}of options for the moons data.

\code{rodents_options} creates a \code{rodents_options} 
  \code{list} of options for the rodents data.

\code{covariates_options} creates a \code{covariates_options}
  \code{list}of options for the covariates data.

\code{metadata_options} creates a \code{metadata_options} 
  \code{list}of options for the metadata creation.

\code{predictions_options} creates a 
  \code{predictions_options} \code{list} of options for populating the 
  predictions subdirectory

\code{models_options} creates a \code{models_options} 
  \code{list} of options for populating the models subdirectory.

\code{cast_options} creates a \code{cast_options} \code{list} 
  of control options for running \code{\link{portalcast}}.

\code{check_options_args} verifies the validity (class, size,
  etc.) of the inputs to the options list functions.
}
\examples{
\dontrun{

setup_dir()
all_options()
dir_options()
PortalData_options()
data_options()
moons_options()
rodents_options()
covariates_options()
metadata_options()
predictions_options()
models_options()
cast_options()
check_options_args()
}

}
