% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{all_options}
\alias{all_options}
\title{Prepare all of the options for creating a portalcasting directory}
\usage{
all_options(base = "~", main = "forecasting", subs = subdirs(),
  quiet = FALSE, fdate = today(), append_missing_to_raw = TRUE,
  m_save = TRUE, m_filename = "moons.csv", type = NULL, start = 217,
  drop_spp = "PI", min_plots = 24, level = "Site", treatment = NULL,
  length = "all", output = "abundance", r_save = TRUE,
  r_filename = "all.csv", historical = TRUE, forecasts = TRUE,
  yr = as.numeric(format(today(), "\%Y")), lead_time = 12, min_lag = 6,
  fcast_nms = NULL, nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", filename_suffix = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml", models = models(), model = models())
}
\arguments{
\item{base}{name of the base directory where the forecasting 
directory should exist (will be created if it doesn't)}

\item{main}{name of the portalcasting directory}

\item{subs}{names of the portalcasting subdirectories}

\item{quiet}{logical indicator if progress messages should be quieted}

\item{fdate}{date from which future is defined, typically today's date.}

\item{append_missing_to_raw}{logical indicating if the missing moon dates 
should be appended to the raw data file (should be TRUE to allow the
imported portalr functions to work properly)}

\item{m_save}{logical if the moons data should be saved out}

\item{m_filename}{the name of the file for saving moons data}

\item{type}{"all" or "controls"}

\item{start}{newmoon number of the first sample to be included. Default 
value is \code{217}, corresponding to 1995-01-01.}

\item{drop_spp}{species names to drop from the table}

\item{min_plots}{minimum number of plots surveyed for a survey to be used}

\item{level}{for get_rodent_data, automatically set by type}

\item{treatment}{for get_rodent_data, automatically set by type}

\item{length}{for get_rodent_data, automatically set by type}

\item{output}{for get_rodent_data, automatically set by type}

\item{r_save}{logical if the data should be saved out}

\item{r_filename}{the name of the file to save the rodent data in}

\item{historical}{logical indicator whether or historical covariates are
to be included}

\item{forecasts}{logical indicator whether or forecasted covariates are
to be included}

\item{yr}{year of today}

\item{lead_time}{number of moons into the future the rodents are forecast}

\item{min_lag}{the minimum lag time used in any model}

\item{fcast_nms}{newmoon numbers to be forecast for covariates}

\item{nfcnm}{number of forecast newmoons for covariates}

\item{append_fcast_csv}{logical if the new forecast should be appended to
the historical forecasts for the purposes of hindcasting}

\item{hist_fcast_file}{name of the file where the historical 
covariate forecasts are held}

\item{source_name}{character value for the name to give the covariaate
forecast. Currently is "current_archive". Previous to "current_archive",
the data were retroactively filled in and are given the source name
"retroactive"}

\item{c_save}{logical if the covariate data should be saved out}

\item{c_filename}{the name of the covariate file for the saving}

\item{filename_suffix}{"forecasts" or "hindcasts"}

\item{confidence_level}{confidence level used in summarizing model output}

\item{meta_save}{logical if the metadata should be saved out}

\item{meta_filename}{the name of the metadata file for the saving}

\item{model}{names of model scripts to include}
}
\value{
list of control options lists
}
\description{
Create a list of lists (of lists) of control options for the 
  directory set-up
}
