% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{all_options}
\alias{all_options}
\alias{dir_options}
\alias{data_options}
\alias{moons_options}
\alias{rodents_options}
\alias{covariates_options}
\alias{metadata_options}
\alias{predictions_options}
\alias{models_options}
\alias{cast_options}
\title{Prepare options for a portalcasting directory}
\usage{
all_options(base = "~", main = "forecasting", subs = subdirs(),
  quiet = FALSE, fdate = today(), append_missing_to_raw = TRUE,
  m_save = TRUE, m_filename = "moons.csv", tmnt_type = NULL,
  start = 217, end = NULL, hind_step = 1, drop_spp = "PI",
  min_plots = 24, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  cov_hist = TRUE, cov_fcast = TRUE, yr = as.numeric(format(today(),
  "\%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL, nfcnm = 0,
  append_fcast_csv = TRUE, hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", cast_type = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml", download_existing_predictions = TRUE,
  model = models(), ensemble = TRUE)

dir_options(base = "~", main = "forecasting", subs = subdirs(),
  quiet = FALSE)

data_options(base = "~", main = "forecasting", subs = subdirs(),
  quiet = FALSE, fdate = today(), append_missing_to_raw = TRUE,
  m_save = TRUE, m_filename = "moons.csv", tmnt_type = NULL,
  start = 217, end = NULL, hind_step = 1, drop_spp = "PI",
  min_plots = 24, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  cov_hist = TRUE, cov_fcast = TRUE, yr = as.numeric(format(today(),
  "\%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL, nfcnm = 0,
  append_fcast_csv = TRUE, hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", cast_type = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml")

moons_options(n_future_moons = 12, fdate = today(),
  append_missing_to_raw = TRUE, save = TRUE, filename = "moons.csv",
  tree = dirtree(), quiet = FALSE)

rodents_options(cast_type = "forecasts", tmnt_type = NULL, start = 217,
  end = NULL, hind_step = 1, drop_spp = "PI", min_plots = 24,
  level = "Site", treatment = NULL, plots = "all", output = "abundance",
  save = TRUE, filename = "all.csv", tree = dirtree(), quiet = FALSE)

covariates_options(cast_type = "forecasts", cov_hist = TRUE,
  cov_fcast = TRUE, fdate = today(), yr = as.numeric(format(today(),
  "\%Y")), start = 217, end = NULL, hind_step = 1, lead_time = 12,
  min_lag = 6, fcast_nms = NULL, nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", save = TRUE,
  filename = "covariates.csv", tree = dirtree(), quiet = FALSE)

metadata_options(fdate = today(), cast_type = "forecasts",
  confidence_level = 0.9, lead_time = 12, save = TRUE,
  filename = "metadata.yaml", quiet = FALSE, tree = dirtree())

predictions_options(base = "~", main = "forecasting", subs = subdirs(),
  download_existing_predictions = TRUE, quiet = FALSE)

models_options(base = "~", main = "forecasting", subs = subdirs(),
  quiet = FALSE, model = models())

cast_options(base = "~", main = "forecasting", subs = subdirs(),
  quiet = FALSE, model = models(), cast_type = "forecasts",
  fdate = today(), ensemble = TRUE)
}
\arguments{
\item{base}{name of the base directory where the forecasting 
directory should exist (will be created if it doesn't)}

\item{main}{name of the portalcasting directory}

\item{subs}{names of the portalcasting subdirectories}

\item{quiet}{logical indicator if progress messages should be quieted}

\item{fdate}{date from which future is defined, typically today's date.}

\item{append_missing_to_raw}{logical indicating if the missing moon dates 
should be appended to the raw data file (should be TRUE to allow the
imported portalr functions to work properly)}

\item{m_save}{logical if the moons data should be saved out}

\item{m_filename}{the name of the file for saving moons data}

\item{tmnt_type}{"all" or "controls"}

\item{start}{newmoon number of the first sample to be included. Default 
value is \code{217}, corresponding to 1995-01-01}

\item{end}{newmoon number of the last sample to be included. Default
value is \code{NULL}, which equates to the most recently included
sample. If \code{cast_type} is "hindcasts" and \code{end} is \code{NULL},
default becomes \code{490:403}, corresponding to 2017-01-01 back to
2010-01-01}

\item{hind_step}{an iteration parameter to work across the input values 
for \code{end}. Default is 1, which is likely all it should be run using}

\item{drop_spp}{species names to drop from the table}

\item{min_plots}{minimum number of plots surveyed for a survey to be used}

\item{level}{for get_rodent_data, automatically set by tmnt_type}

\item{treatment}{for get_rodent_data, automatically set by tmnt_type}

\item{plots}{for get_rodent_data, automatically set by tmnt_type}

\item{output}{for get_rodent_data, automatically set by tmnt_type}

\item{r_save}{logical if the data should be saved out}

\item{r_filename}{the name of the file to save the rodent data in}

\item{cov_hist}{logical indicator whether or historical covariates are
to be included}

\item{cov_fcast}{logical indicator whether or not forecasted covariates are
to be included}

\item{yr}{year of today}

\item{lead_time}{number of moons into the future the rodents are forecast}

\item{min_lag}{the minimum lag time used in any model}

\item{fcast_nms}{newmoon numbers to be forecast for covariates}

\item{nfcnm}{number of forecast newmoons for covariates}

\item{append_fcast_csv}{logical if the new forecast should be appended to
the historical forecasts for the purposes of hindcasting}

\item{hist_fcast_file}{name of the file where the historical 
covariate forecasts are held}

\item{source_name}{character value for the name to give the covariaate
forecast. Currently is "current_archive". Previous to "current_archive",
the data were retroactively filled in and are given the source name
"retroactive"}

\item{c_save}{logical if the covariate data should be saved out}

\item{c_filename}{the name of the covariate file for the saving}

\item{cast_type}{"forecasts" or "hindcasts"}

\item{confidence_level}{confidence level used in summarizing model output}

\item{meta_save}{logical if the metadata should be saved out}

\item{meta_filename}{the name of the metadata file for the saving}

\item{download_existing_predictions}{logical indicator is the existing
predictions files should be retrieved from the portalPredictions repo}

\item{model}{names of model scripts to include}

\item{ensemble}{logical indicator of whether to create an ensemble model}

\item{n_future_moons}{integer value for the number of future moons to add}

\item{save}{logical if the specific data should be saved out}

\item{filename}{the name of the file for the saving}

\item{tree}{directory tree}
}
\value{
\code{all_options}: list of control options lists ("options_dir"
  generated by \code{dir_options}, "options_data" generated by 
  \code{data_options}, "options_predictions" generated by
  \code{predictions_options}, "options_models" generated by
  \code{models_options}, and "options_cast" generated by 
  \code{cast_options}.

\code{dir_options}: a list of settings controlling the directory
  structure creation

\code{data_options}: list of control options lists ("moons"
  generated by \code{moons_options}, "rodents" generated by 
  \code{rodents_options}, "covariates" generated by
  \code{covariates_options}, and "metadata" generated by
  \code{metadata_options}, as well as items "cast_type", "tree",  and
  "quiet" for simple use with all or ancillary data.

\code{moons_options}: a list of settings controlling the moon data
  creation

\code{rodents_options}: a list of settings controlling the rodents
  data creation

\code{covariates_options}: a list of settings controlling the 
  covariates data creation

\code{metadata_options}: a list of settings controlling the 
  metadata creation

\code{predictions_options}: a list of settings controlling the 
  population of the predictions subdirectory

\code{models_options}: a list of settings controlling the 
  population of the models subdirectory

\code{cast_options}: a list of settings controlling the running of
  \code{\link{portalcast}}
}
\description{
Suite of functions used to generate options lists: \cr \cr
 \code{all_options} creates the full hierarchy of lists

\code{dir_options} creates a list of control options for the
  directory set-up

\code{data_options} creates a list of lists of options for
  filling the data directory

\code{moons_options} creates a list of options for the moons
  data

\code{rodents_options} creates a list of options for the 
  rodents data

\code{covariates_options} creates a list of options for the 
  covariates data

\code{metadata_options} creates a list of options for the 
  metadata creation

\code{predictions_options} creates a list of options for 
  populating the predictions subdirectory

\code{models_options} creates a list of options for 
  populating the models subdirectory

\code{cast_options} creates a list of control options for 
  running \code{\link{portalcast}}
}
