% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{append_cov_fcast_csv}
\alias{append_cov_fcast_csv}
\title{Append a covariate forecast to existing covariate forecasts}
\usage{
append_cov_fcast_csv(new_forecast_covariates,
  options_covariates = covariates_options())
}
\arguments{
\item{new_forecast_covariates}{forecasted covariates}

\item{options_covariates}{control options list for covariates}
}
\value{
new_forecast_covariates as input
}
\description{
combining weather and ndvi forecasts to the existing forecasts
}
