% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{append_covariate_fcast}
\alias{append_covariate_fcast}
\title{Append a covariate forecast to historical covariate table}
\usage{
append_covariate_fcast(covariates, metadata, moons = prep_moons())
}
\arguments{
\item{covariates}{output from \code{get_covariate_data}}

\item{metadata}{model metadata}

\item{moons}{moon data table}
}
\value{
no value
}
\description{
combining weather and ndvi forecasts to the existing 
  covariates
}
