% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{append_csv}
\alias{append_csv}
\title{Save data out to a csv, appending the file if it already exists}
\usage{
append_csv(df, filename)
}
\arguments{
\item{df}{\code{data.frame} table to be written out.}

\item{filename}{\code{character} filename of existing \code{.csv} to be 
appended.}
}
\value{
\code{NULL}.
}
\description{
Appending a \code{.csv} without re-writing the header of the
 file. If the doesn't exist, it will be created.
}
