% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{append_observed_to_cast}
\alias{append_observed_to_cast}
\title{Append the observed values to a table of -casts}
\usage{
append_observed_to_cast(casts, tree = dirtree(), add_error = TRUE,
  add_in_window = TRUE, add_lead = TRUE)
}
\arguments{
\item{casts}{Class \code{casts} \code{data.frame} of requested fore- or 
hindcasts to have observations added to.}

\item{tree}{\code{dirtree}-class list. See \code{\link{dirtree}}.}

\item{add_error}{\code{logical} indicator if the \code{error} column should
be added to the output as well.}

\item{add_in_window}{\code{logical} indicator if the \code{in_window} 
column should be added to the output as well.}

\item{add_lead}{\code{logical} indicator if the \code{lead} column should
be added to the output as well.}
}
\value{
Class \code{casts} \code{data.frame} with additional columns.
}
\description{
Add a column of observed values and optionally raw error,
  in-prediction-window logical indicator, and lead time columns. Additional
  columns are added by default.
}
\examples{
\dontrun{

setup_dir(all_options(download_existing_predictions = TRUE))
casts <- read_casts()
casts_a <- select_casts(casts, level = "All")
casts_a <- append_observed_to_cast(casts_a) 
}

}
