% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{append_past_moons_to_raw}
\alias{append_past_moons_to_raw}
\title{Append missing past moons to the raw data}
\usage{
append_past_moons_to_raw(moons, options_moons = moons_options())
}
\arguments{
\item{moons}{Moons data table as a code{moons}-class \code{data.frame}.}

\item{options_moons}{A class-\code{moons_options} \code{list} of options 
for the moons data. See \code{\link{moons_options}}.}
}
\value{
\code{moons} as it was input (a \code{moons}-class 
  \code{data.frame}).
}
\description{
Sometimes the raw moon data table is not up-to-date. Because
  the \code{portalr} functions \code{\link[portalr]{weather}} and 
  \code{\link[portalr]{fcast_ndvi}} point to the raw moons data, that table
  needs to be updated to produce the correct current data table for 
  casting.
}
