% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoArima.R
\name{AutoArima}
\alias{AutoArima}
\title{Flexible Auto-Regressive Integrated Moving Average model for Portal 
 Predictions}
\usage{
AutoArima(abundances, metadata, level = "All", quiet = FALSE)
}
\arguments{
\item{abundances}{Class-\code{rodents} \code{data.frame} table of rodent 
abundances and time measures.}

\item{metadata}{Class-\code{metadata} model metadata \code{list}.}

\item{level}{\code{character} value name of the type of plots included 
(\code{"All"} or \code{"Controls"}).}

\item{quiet}{\code{logical} value indicating if the function should be 
quiet.}
}
\value{
\code{list} of [1] \code{"forecast"} (the forecasted abundances)
  and [2] \code{"all_model_aic"} (the model AIC values).
}
\description{
Fit an AutoARIMA model in the portalcasting pipeline.
}
\details{
Model "AutoARIMA" is a flexible auto-regressive integrated 
 moving average model fit to the data using \code{auto.arima} in the
 \href{http://pkg.robjhyndman.com/forecast/}{\code{forecast} package}
 (Hyndman \emph{et al}. 2018).

 Because the seasonality and sampling occurring with different frequencies,
 which \code{\link[forecast]{auto.arima}} cannot accommodate, seasonal 
 models are not included.

 Although the auto.arima model can handle missing data, the other models
 used currently cannot, so we interpolate missing data here for 
 comparison.
}
\references{
Hyndman R., Bergmeir C., Caceres G., Chhay L., O'Hara-Wild M., Petropoulos
 F., Razbash S., Wang E., and Yasmeen F. 2018. forecast: Forecasting 
 functions for time series and linear models. 
 \href{http://pkg.robjhyndman.com/forecast}{R package version 8.3}.
}
