% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab_models.R
\name{AutoArima}
\alias{AutoArima}
\title{Flexible Auto-Regressive Integrated Moving Average model for Portal 
 Predictions}
\usage{
AutoArima(main = ".", tmnt_type = "All", quiet = FALSE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{tmnt_type}{\code{character} value name of the type of plots included 
(\code{"All"} or \code{"Controls"}).}

\item{quiet}{\code{logical} value indicating if the function should be 
quiet.}
}
\value{
\code{list} of [1] \code{"cast"} (the casted abundances)
 and [2] \code{"all_model_aic"} (the model AIC values).
}
\description{
Fit an AutoArima model in the portalcasting pipeline.
}
\details{
Model "AutoArima" is a flexible auto-regressive integrated 
 moving average model fit to the data using 
 \code{\link[forecast]{auto.arima}} in the
 \href{http://pkg.robjhyndman.com/forecast/}{\code{forecast} package}
 (Hyndman \emph{et al}. 2018). \cr
 Because the seasonality and sampling occurring with different frequencies,
 which \code{\link[forecast]{auto.arima}} cannot accommodate, seasonal 
 models are not included. \cr
 Although \code{\link[forecast]{auto.arima}} can handle missing data, the
 other models used currently cannot, so we interpolate missing data here 
 for comparison.
}
\examples{
 \donttest{
  setup_dir()
  AutoArima()
 }

}
\references{
Hyndman R., Bergmeir C., Caceres G., Chhay L., O'Hara-Wild M., Petropoulos
 F., Razbash S., Wang E., and Yasmeen F. 2018. forecast: Forecasting 
 functions for time series and linear models. 
 \href{http://pkg.robjhyndman.com/forecast}{R package version 8.3}.
}
