% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{base_path}
\alias{base_path}
\alias{main_path}
\alias{sub_paths}
\alias{sub_path}
\title{Determine the path for a specific level of a portalcasting directory}
\usage{
base_path(tree = dirtree())

main_path(tree = dirtree())

sub_paths(tree = dirtree())

sub_path(tree = dirtree(), specific_sub = "tmp")
}
\arguments{
\item{tree}{\code{dirtree}-class list. See \code{\link{dirtree}}.}

\item{specific_sub}{\code{character}-value name of the specific 
subdirectory/subdirectories of interest.}
}
\value{
\code{base_path}: The normalized path of the main folder in the 
  directory tree as a character value (see \code{\link{normalizePath}}).

\code{main_path}: The normalized path of the main folder in the 
  directory tree as a character value (see \code{\link{normalizePath}}).

\code{sub_paths}: The normalized paths of the sub folders in the 
  directory tree as a character vector (see \code{\link{normalizePath}}).

\code{sub_paths}: The normalized path(s) of the \code{specific_sub}
  folder(s) in the directory tree as \code{character} value(s) (see 
  \code{\link{normalizePath}}).
}
\description{
\code{base_path}: Return the path for the \code{base} folder
  in the directory.

\code{main_path}: Return the path for the \code{main} folder
  in the directory.

\code{sub_paths}: Return the paths for the \code{subs} folders
  in the directory.

\code{sub_path}: Return the path for a specific \code{subs} 
  folder or folders in the directory.
}
\examples{
\dontrun{

setup_dir()
base_path()
main_path()
sub_paths()
sub_path(specific_sub = "models")
}

}
