% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{base_path}
\alias{base_path}
\alias{main_path}
\alias{sub_paths}
\title{Determine the path for a specific level of a forecasting directory}
\usage{
base_path(base = ".")

main_path(base = ".", main = "")

sub_paths(base = ".", main = "", subs = subdirs(),
  specific_subs = NULL)
}
\arguments{
\item{base}{\code{character} value of the name of the base component of
the directory tree.}

\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{subs}{\code{character} vector of the names of the sub components of
the directory tree.}

\item{specific_subs}{\code{character}-value name of the specific 
subdirectory/subdirectories of interest, or \code{NULL} (default)
for all.}
}
\value{
\code{character} value normalized paths 
  (see \code{\link{normalizePath}}) . \cr \cr
  \code{base_path} normalized path of the \code{base} folder. \cr \cr
  \code{main_path} normalized path of the \code{main} folder. \cr \cr
  \code{subs_paths} normalized paths of the \code{subs} folders. \cr \cr
}
\description{
Produce paths for a forecasting directory. \cr \cr
  \code{base_path} returns the path for the \code{base} folder. \cr \cr
  \code{main_path} returns the path for the \code{main} folder. \cr \cr
  \code{sub_paths} returns the path for the \code{subs} folders. \cr \cr
}
\examples{
 create_dir()
 base_path()
 main_path()
 sub_paths()
 sub_paths(specific_subs = "models")

}
