% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cast0}
\alias{cast0}
\title{Zero-abundance cast}
\usage{
cast0(nmoons = NULL, colname = "pred")
}
\arguments{
\item{nmoons}{\code{integer} number of forecast newmoons.}

\item{colname}{\code{character} name for the predictor column (to match
variable model output names).}
}
\value{
Two-element \code{list} of means and interval values for a 
  0-abundance forecast to be used as a filler when a model fails or there 
  is no non-0 historical abundance.
}
\description{
Create a 0-abundance cast for fill-in usage when a model 
  fails or there is no non-0 historical abundance.
}
\examples{
 cast0(12)

}
