% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_covariates.R
\name{cast_covariates}
\alias{cast_covariates}
\alias{prep_cast_covariates}
\alias{cast_ndvi}
\alias{cast_weather}
\alias{save_cast_cov_csv}
\title{Cast the covariates}
\usage{
cast_covariates(main = ".", moons = NULL, hist_cov = NULL,
  end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), control_files = files_control(),
  control_climate_dl = climate_dl_control(), quiet = TRUE,
  verbose = FALSE, arg_checks = TRUE)

prep_cast_covariates(main = ".", moons = NULL, hist_cov = NULL,
  end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), control_climate_dl = climate_dl_control(),
  control_files = files_control(), quiet = TRUE, verbose = FALSE,
  arg_checks = TRUE)

cast_ndvi(main = ".", hist_cov = NULL, lead_time = 12, min_lag = 6,
  arg_checks = TRUE)

cast_weather(main = ".", moons = NULL, hist_cov = NULL,
  end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), control_climate_dl = climate_dl_control(),
  control_files = files_control(), quiet = TRUE, verbose = TRUE,
  arg_checks = TRUE)

save_cast_cov_csv(main = ".", moons = NULL, end_moon = NULL,
  cast_date = Sys.Date(), cast_cov = NULL,
  control_files = files_control(), quiet = TRUE, verbose = FALSE,
  arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} name of the main directory tree component.}

\item{moons}{Lunar data \code{data.frame}. See \code{\link{prep_moons}}.}

\item{hist_cov}{\code{data.frame} of historic covariates. See 
\code{\link{prep_hist_covariates}}.}

\item{end_moon}{Forecast origin. \code{integer}-conformable newmoon number 
or \code{NULL}, which equates to the most recently included census'
newmoon number.}

\item{lead_time}{\code{integer}-conformable number of timesteps forward 
a cast will proceed.}

\item{min_lag}{\code{integer}-conformable minimum non-0 covariate lag time 
used in any model.}

\item{cast_date}{\code{Date} on which the cast happens. Generally should
be \code{\link{Sys.Date}}.}

\item{control_files}{\code{list} of names of the folders and files within
the sub directories and saving strategies (save, overwrite, append, etc.).
Generally shouldn't need to be edited. See \code{\link{files_control}}.}

\item{control_climate_dl}{\code{list} of specifications for the climate 
download. Sent to \code{\link{NMME_urls}} to create the specific URLs. See 
\code{\link{climate_dl_control}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be
shown.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}

\item{cast_cov}{\code{data.frame} of cast covariates. See
\code{\link{cast_covariates}}.}
}
\value{
\code{cast_covariates}: \code{data.frame} of the covariates cast for the
  moons as defined as needed for saving. \cr \cr
 \code{prep_cast_covariates}: \code{data.frame} of the cast covariates
  after saving, with excess columns removed. \cr \cr
 \code{cast_ndvi}: \code{data.frame} of -casted NDVI values. \cr \cr
 \code{cast_weather}: \code{data.frame} of -casted weather values. \cr \cr
 \code{save_cast_cov_csv}: \code{cast_cov} \code{data.frame} exactly as 
  input.
}
\description{
Cast the covariates for a model run. \cr \cr
 \code{cast_covariates} is the primary function, which produces the 
 covariates (min, mean, and max temperature; precipitation; and NDVI) 
 required for a model to be cast. \cr \cr
 \code{prep_cast_covariates} provides a wrapper on \code{cast_covariates}
 that includes saving the cast data out via \code{save_cast_cov_csv}
 and tidying the data for use within the model (by removing the vestigial
 columns that are only needed for saving).
}
\details{
If forecasting from the current newmoon, weather is forecast using 
 \code{\link{cast_weather}} and NDVI is forecast using 
 \code{\link{cast_ndvi}}. If casting from a previous newmoon, the
 historical forecasted weather and NDVI data are retrieved from the file 
 pointed to by \code{filename_cov_casts}.
}
\examples{
 \donttest{
  setup_dir()
  hist_cov <- prep_hist_covariates()
  prep_cast_covariates(hist_cov = hist_cov)
  cast_cov <- cast_covariates(hist_cov = hist_cov)
  save_cast_cov_csv(cast_cov = cast_cov)
  cast_ndvi_data <- cast_ndvi(hist_cov = hist_cov)
  cast_weather_data <- cast_weather(hist_cov = hist_cov)
 }

}
