% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_covariates.R
\name{cast_covariates}
\alias{cast_covariates}
\alias{prepare_forecast_covariates}
\alias{save_cast_cov_csv}
\title{Cast the covariates}
\usage{
prepare_forecast_covariates(
  main = ".",
  lead_time = 12,
  cast_date = Sys.Date(),
  settings = directory_settings(),
  quiet = TRUE,
  verbose = FALSE
)

save_cast_cov_csv(
  main = ".",
  moons = NULL,
  end_moon = NULL,
  cast_date = Sys.Date(),
  cast_cov = NULL,
  control_files = files_control(),
  quiet = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} name of the main directory tree component.}

\item{lead_time}{\code{integer}-conformable number of timesteps forward 
a cast will proceed.}

\item{cast_date}{\code{Date} on which the cast happens. Generally should
be \code{\link{Sys.Date}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be
shown.}

\item{moons}{Lunar data \code{data.frame}. See \code{\link{prep_moons}}.}

\item{end_moon}{Forecast origin. \code{integer}-conformable newmoon number 
or \code{NULL}, which equates to the most recently included census'
newmoon number.}

\item{cast_cov}{\code{data.frame} of cast covariates. See
\code{\link{cast_covariates}}.}

\item{control_files}{\code{list} of names of the folders and files within
the sub directories and saving strategies (save, overwrite, append, etc.).
Generally shouldn't need to be edited. See \code{\link{files_control}}.}

\item{hist_cov}{\code{data.frame} of historic covariates. See 
\code{\link{prep_hist_covariates}}.}

\item{min_lag}{\code{integer}-conformable minimum non-0 covariate lag time 
used in any model.}
}
\value{
\code{prep_cast_covariates}: \code{data.frame} of the cast covariates
  after saving, with excess columns removed. \cr \cr
 \code{save_cast_cov_csv}: \code{cast_cov} \code{data.frame} exactly as 
  input.
}
\description{
Cast the covariates for a model run. \cr \cr
 \code{prep_cast_covariates} is the primary function, which produces the 
 covariates (min, mean, and max temperature; precipitation; and NDVI) 
 required for a model to be cast. Saves the cast data out via
 \code{save_cast_cov_csv} and tidying the data for use within the model 
 (by removing the vestigial columns that are only needed for saving).
}
\details{
If forecasting from the current newmoon, casts the covariates 
 If casting from a previous newmoon, the
 historical forecasted weather and NDVI data are retrieved from the file 
 pointed to by \code{filename_cov_casts}.
}
\examples{
 \donttest{
  setup_dir()
  hist_cov <- prep_hist_covariates()
  cast_cov <- prep_cast_covariates(hist_cov = hist_cov)
  save_cast_cov_csv(cast_cov = cast_cov)
 }

}
