% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{cast_models}
\alias{cast_models}
\title{Source the scripts to run the selected forecast or hindcast models}
\usage{
cast_models(models = NULL, tree = dirtree(), quiet = FALSE,
  data_control = list())
}
\arguments{
\item{models}{\code{character} vector of the names of models to include in
the pipeline. Defaults to \code{NULL}, which retrieves the model names 
from \code{\link{model_names}}.}

\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{data_control}{A \code{list} of arguments to control the filling of
the data subdirectory. See \code{\link{data_control}}. 
Values not set assume defaults.}
}
\description{
Source the R scripts of the selected model(s) indicated by the
  \code{model} element in \code{options_cast}.
}
