% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{cast_models}
\alias{cast_models}
\title{Source the scripts to run the selected forecast or hindcast models}
\usage{
cast_models(options_cast = cast_options())
}
\arguments{
\item{options_cast}{Class-\code{cast_options} \code{list} containing the
hind- or forecasting options. See \code{\link{cast_options}}.}
}
\description{
Source the R scripts of the selected model(s) indicated by the
  \code{model} element in \code{options_cast}.
}
