% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cast_window}
\alias{cast_window}
\title{Determine the start and end calendar dates for a cast window}
\usage{
cast_window(main = ".", moons = NULL, cast_date = Sys.Date(),
  lead_time = 12, min_lag = 6, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the
number of timesteps forward a cast will cover.}

\item{min_lag}{\code{integer} (or integer \code{numeric}) of the minimum 
covariate lag time used in any model.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
Named \code{list} with elements \code{start} and \code{end},
 which are both \code{Dates}.
}
\description{
Based on the cast origin (\code{cast_date}), lead time
 (\code{lead_time}), and minimum non-0 lag (\code{min_lag}), determines
 the dates bracketing the requested window.
}
\examples{
 \donttest{
   create_dir()
   fill_raw()
   cast_window()
 }

}
