% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{casts}
\alias{casts}
\alias{cast}
\alias{step_casts}
\title{Run one or potentially multiple forecasts or hindcasts}
\usage{
casts(models = NULL, tree = dirtree(), quiet = FALSE,
  data_control = list(), ensemble = TRUE)

cast(models = NULL, tree = dirtree(), quiet = FALSE,
  data_control = list(), ensemble = TRUE)

step_casts(models = NULL, tree = dirtree(), quiet = FALSE,
  data_control = list(), ensemble = TRUE)
}
\arguments{
\item{models}{\code{character} vector of the names of models to include in
the pipeline. Defaults to \code{NULL}, which retrieves the model names 
from \code{\link{model_names}}.}

\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{data_control}{A \code{list} of arguments to control the filling of
the data subdirectory. See \code{\link{data_control}}. 
Values not set assume defaults.}

\item{ensemble}{\code{logical} indicator if the ensemble should be added.}
}
\description{
Given that the data and models are prepared accordingly, run 
  a set of one or more forecast or hindcasts, (optionally) add an ensemble,
  combine the results, and add them to the predictions folder. \cr \cr
  \code{casts}: run one or more -casts flexibly \cr \cr
  \code{cast}: run a single -cast
  \code{step_casts}: iterate through subsequent-to-the-first
  -cast(s) within \code{casts}.
}
