% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{casts}
\alias{casts}
\title{Multiple forecast or hindcast}
\usage{
casts(options_all = all_options())
}
\arguments{
\item{options_all}{full options list}
}
\value{
nothing
}
\description{
Given the data and models are prepared accordingly, run the
  forecast or hindcasts, (optionally) add an ensemble, combine the 
  results into the predictions folder.
}
