% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_arg}
\alias{check_arg}
\title{Check an argument's value for validity}
\usage{
check_arg(arg_name, arg_value, fun_name = NULL)
}
\arguments{
\item{arg_name}{\code{character} value of the argument name.}

\item{arg_value}{Input value for the argument.}

\item{fun_name}{\code{character} value of the function name or \code{NULL}.}
}
\description{
Check that an argument's value is valid, potentially 
  dependent upon the function. See \code{Details} for argument-specific
  rules.
}
\details{
Usage and rules for arguments are as follows: \cr \cr
  \code{add_obs}: must be a length-1 \code{logical} vector in
    \code{\link{plot_cast_ts}} \cr \cr
  \code{append_fcast_csv}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{data_options}},
    \code{\link{covariates_options}} \cr \cr
  \code{append_missing_to_raw}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{data_options}} , 
    \code{\link{moons_options}} \cr \cr
  \code{base} must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{dir_options}},
    \code{\link{PortalData_options}}, \code{\link{data_options}} 
    \code{\link{predictions_options}} 
    \code{\link{models_options}}, \code{\link{cast_options}},
    \code{\link{dirtree}} \cr \cr
  \code{c_filename}: must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{data_options}},
    \code{\link{covariates_options}}  \cr \cr
  \code{c_save}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{data_options}},
    \code{\link{covariates_options}}  \cr \cr
  \code{cast_date}: must be \code{NULL} or a length-1 \code{Date} or 
    \code{Date}-conformable vector in \code{\link{plot_cast_point}},  
    \code{\link{plot_cast_ts}}, \code{\link{all_options}}, 
    \code{\link{data_options}}, \code{\link{cast_options}},
    \code{\link{moons_options}}, \code{\link{metadata_options}},
    \code{\link{covariates_options}},
    \code{\link{add_addl_future_moons}} \cr \cr
  \code{cast_dates}: must be \code{NULL} or a \code{Date} or 
    \code{Date}-conformable vector in \code{\link{plot_cov_RMSE_mod_spp}}  
    \cr \cr
  \code{cast_type}: must be a length-1 \code{character} vector in
    \code{\link{plot_cov_RMSE_mod_spp}}, 
    \code{\link{plot_err_lead_spp_mods}}, \code{\link{plot_cast_point}},
    \code{\link{plot_cast_ts}} \cr \cr
  \code{confidence_level}: must be a length-1 \code{numeric} value between
    0 and 1 in \code{\link{all_options}}, \code{\link{data_options}}, 
    \code{\link{metadata_options}} \cr \cr
  \code{cov_fcast}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}},
    \code{\link{covariates_options}}, \code{\link{data_options}} \cr \cr
  \code{cov_hist}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}},
    \code{\link{covariates_options}}, \code{\link{data_options}} \cr \cr
  \code{covariates}: must be a \code{data.frame} of class \code{options} in
    \code{\link{forecast_covariates}}, \code{\link{forecast_ndvi}},
    \code{\link{prep_metadata}} \cr \cr
  \code{download_existing_predictions}: must be a length-1 \code{logical} 
    vector in \code{\link{all_options}}, 
    \code{\link{predictions_options}} \cr \cr
  \code{drop_spp}: must be a \code{character} vector in
    \code{\link{all_options}}, \code{\link{data_options}},
    \code{\link{rodents_options}} \cr \cr
  \code{end}: must be \code{NULL} or a positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{covariates_options}}, \code{\link{data_options}} 
    \code{\link{rodents_options}}, \code{\link{cast_options}}  
    \cr \cr
  \code{ensemble}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{cast_options}}  \cr \cr
  \code{extension}: must be a length-1 \code{character} vector with
    a single period in \code{\link{model_paths}} \cr \cr 
  \code{fcast_nms}: must be \code{NULL} or a non-negative \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{covariates_options}}, \code{\link{data_options}} 
    \cr \cr
  \code{filename}: must be a length-1 \code{character} vector in
    \code{\link{verify_PortalData}}  \cr \cr
  \code{from_date}: must be \code{NULL} or a length-1 \code{Date} or 
    \code{Date}-conformable vector in \code{\link{plot_cast_point}}  
    \cr \cr
  \code{from_zenodo}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{PortalData_options}} \cr \cr
  \code{future_moons}: must be a \code{data.frame} of class \code{moons} 
    in \code{\link{add_addl_future_moons}} \cr \cr
  \code{hind_step}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{covariates_options}}, \code{\link{data_options}} 
    \code{\link{rodents_options}}, \code{\link{cast_options}}   
    \cr \cr
  \code{hist_cov}: must be a \code{data.frame} of class \code{options} in
    \code{\link{prep_fcast_covariates}},
    \code{\link{update_covfcast_options}} \cr \cr
  \code{hist_fcast_file}: must be a length-1 \code{character} vector in
    \code{\link{all_options}}, 
    \code{\link{covariates_options}}, \code{\link{data_options}} \cr \cr
  \code{lag}: must be \code{NULL} or a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{model_options}}
    \cr \cr
  \code{lead}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{plot_cast_point}}  
    \cr \cr
  \code{lead_time}: must be a length-1 non-negative \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}}, 
    \code{\link{metadata_options}}, \code{\link{covariates_options}}, 
    \code{\link{data_options}} \cr \cr
  \code{level}: must be a length-1 \code{character} vector of value
    \code{"All"} or \code{"Controls"} in 
    \code{\link{plot_cov_RMSE_mod_spp}}, 
    \code{\link{plot_err_lead_spp_mods}}, \code{\link{plot_cast_point}},
    \code{\link{plot_cast_ts}}, \code{\link{plot_cast_ts_ylab}} 
    or a length-1 \code{character} vector of value \code{"Site"} or 
    \code{"Treatment"} in \code{\link{all_options}}, 
    \code{\link{rodents_options}}, \code{\link{data_options}}, 
    \code{\link{metadata_options}} \cr \cr
  \code{local_paths}: must be a \code{character} 
    vector in \code{\link{file_paths}} \cr \cr 
  \code{m_filename}: must be a length-1 \code{character} vector in
    \code{\link{all_options}},
    \code{\link{moons_options}}, \code{\link{data_options}}  \cr \cr
  \code{m_save}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}},
    \code{\link{moons_options}}, \code{\link{data_options}}  \cr \cr
  \code{meta_filename}: must be a length-1 \code{character} vector in
    \code{\link{all_options}},
    \code{\link{metadata_options}}, \code{\link{data_options}}  \cr \cr
  \code{meta_save}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}},
    \code{\link{metadata_options}}, \code{\link{data_options}} \cr \cr
  \code{main} must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{dir_options}},
    \code{\link{PortalData_options}}, \code{\link{data_options}} 
    \code{\link{predictions_options}} 
    \code{\link{models_options}}, \code{\link{cast_options}},
    \code{\link{dirtree}} \cr \cr
  \code{min_lag}: must be a length-1 non-negative \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{covariates_options}}, \code{\link{data_options}} \cr \cr
  \code{min_observed}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in 
    \code{\link{plot_cov_RMSE_mod_spp}}  \cr \cr
  \code{min_plots}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}}, 
    \code{\link{data_options}}, \code{\link{rodents_options}}, 
    \code{\link{cast_options}} \cr \cr
  \code{min_traps}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}}, 
    \code{\link{data_options}}, \code{\link{rodents_options}}, 
    \code{\link{cast_options}} \cr \cr
  \code{mod_covariates}: must be a length-1 \code{logical} vector in 
    \code{\link{model_options}} \cr \cr
  \code{mod_type}: must be \code{"pevGARCH"} in
    \code{\link{covariate_models}}  \cr \cr
  \code{model}: must be a length-1 \code{character} vector in
    \code{\link{plot_cast_point}}, \code{\link{plot_cast_ts}},
    \code{\link{plot_cast_ts_ylab}}  \cr \cr
  \code{models}: must be a \code{character} vector in
    \code{\link{all_options}}, \code{\link{models_options}},
    \code{\link{cast_options}}, \code{\link{model_paths}} \cr \cr
  \code{moons}: must be a \code{data.frame} of class \code{moons} in
    \code{\link{forecast_covariates}}, \code{\link{forecast_ndvi}},
    \code{\link{forecast_weather}}, \code{\link{trim_moons_fcast}},
    \code{\link{get_climate_forecasts}}, \code{\link{update_covariates}},
    \code{\link{prep_covariates}}, \code{\link{prep_fcast_covariates}},
    \code{\link{update_covfcast_options}},
    \code{\link{prep_metadata}}, \code{\link{append_past_moons_to_raw}},
    \code{\link{add_future_moons}}, \code{\link{format_moons}} \cr \cr
  \code{n_future_moons}: must be a length-1 non-negative \code{integer} or 
    \code{integer}-conformable vector in \code{\link{moons_options}}
    \cr \cr
  \code{name}: must be a length-1 \code{character} vector of the model name
    in \code{\link{model_options}} \cr \cr
  \code{ndates}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in 
    \code{\link{plot_err_lead_spp_mods}}  \cr \cr
  \code{nfcnm}: must be a length-1 non-negative \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{covariates_options}}, \code{\link{data_options}} \cr \cr
  \code{new_forecast_covariates}: must be a \code{data.frame} of class 
    \code{options} in \code{\link{append_cov_fcast_csv}} \cr \cr
  \code{options_all}: must be of class \code{all_options} in
    \code{\link{setup_dir}}, \code{\link{fill_dir}}, 
    \code{\link{cleanup_dir}}, \code{\link{portalcast}},
    \code{\link{casts}}, \code{\link{step_casts}},
    \code{\link{step_hind_forward}}  \cr \cr
  \code{options_cast}: must be of class \code{cast_options} in
    \code{\link{verify_models}}, \code{\link{cast_models}}, 
    \code{\link{models_to_cast}}, \code{\link{cast}},
    \code{\link{check_to_skip}} \cr \cr
  \code{options_covariates}: must be of class \code{covariates_options} in
    \code{\link{forecast_covariates}}, \code{\link{forecast_ndvi}},
    \code{\link{forecast_weather}}, \code{\link{trim_moons_fcast}},
    \code{\link{get_climate_forecasts}}, 
    \code{\link{append_cov_fcast_csv}}, \code{\link{update_covariates}} 
    \code{\link{prep_covariates}}, \code{\link{prep_hist_covariates}}, 
    \code{\link{prep_fcast_covariates}},
    \code{\link{update_covfcast_options}} \cr \cr
  \code{options_data}: must be of class \code{data_options} in
    \code{\link{fill_data}}, \code{\link{prep_data}},
    \code{\link{update_data}},
    \code{\link{transfer_hist_covariate_forecasts}} \cr \cr
  \code{options_dir}: must be of class \code{dir_options} in
    \code{\link{create_dir}}, \code{\link{create_main_dir}},
    \code{\link{create_sub_dirs}}  \cr \cr
  \code{options_metadata}: must be of class \code{metadata_options} in
    \code{\link{prep_metadata}} \cr \cr
  \code{options_models}: must be of class \code{models_options} in
    \code{\link{fill_models}} \cr \cr
  \code{options_moons}: must be of class \code{moons_options} in
    \code{\link{fill_moons}}, \code{\link{append_past_moons_to_raw}},
    \code{\link{add_future_moons}} \cr \cr
  \code{options_PortalData}: must be of class \code{PortalData_options} in
    \code{\link{fill_PortalData}} \cr \cr
  \code{options_predictions}: must be of class \code{predictions_options}
    in \code{\link{fill_predictions}} \cr \cr
  \code{options_rodents}: must be of class \code{rodents_options} in
    \code{\link{update_rodents_list}}  \cr \cr
  \code{output}: must be \code{"abundance"} in \code{\link{all_options}}, 
    \code{\link{data_options}}, \code{\link{rodents_options}}  
    \cr \cr
  \code{plots}: must be a length-1 \code{character} vector of value
    \code{"all"} or \code{"longerm"} in \code{\link{all_options}},
    \code{\link{data_options}}, \code{\link{rodents_options}} \cr \cr
  \code{quiet}: must be a length-1 \code{logical} vector in
    \code{\link{create_sub_dir}}, \code{\link{verify_PortalData}},
    \code{\link{model_options}}, \code{\link{all_options}},   
    \code{\link{dir_options}}, \code{\link{PortalData_options}}, 
    \code{\link{data_options}}, \code{\link{covariates_options}}, 
    \code{\link{predictions_options}}, \code{\link{models_options}}, 
    \code{\link{cast_options}}, \code{\link{metadata_options}},
    \code{\link{moons_options}}, \code{\link{rodents_options}} \cr \cr
  \code{r_filename}: must be a length-1 \code{character} vector in
    \code{\link{all_options}},
    \code{\link{rodents_options}}, \code{\link{data_options}} \cr \cr
  \code{r_save}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}},
    \code{\link{rodents_options}}, \code{\link{data_options}} \cr \cr
  \code{rangex}: must be a length-2 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{plot_cast_ts_xaxis}}  
    \cr \cr
  \code{rodents_list}: must be of class \code{rodents_list} in 
    \code{\link{prep_metadata}} \cr \cr
  \code{save}: must be a length-1 \code{logical} vector in
    \code{\link{covariates_options}}, \code{\link{metadata_options}},
    \code{\link{moons_options}}, \code{\link{rodents_options}} \cr \cr
  \code{source_name}: must be a length-1 \code{character} vector in
    \code{\link{all_options}},
    \code{\link{covariates_options}}, \code{\link{data_options}} \cr \cr
  \code{species}: must be \code{NULL} or a \code{character} vector in
    \code{\link{plot_cov_RMSE_mod_spp}}, 
    \code{\link{plot_err_lead_spp_mods}}, \code{\link{plot_cast_point}},
    \code{\link{plot_cast_point_yaxis}}; specifically length-1 in
    \code{\link{plot_cast_ts}}, \code{\link{plot_cast_ts_ylab}} \cr \cr
  \code{specific_sub}: must be \code{NULL} or a length-1 \code{character} 
    vector in \code{\link{sub_paths}} \cr \cr 
  \code{start}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{covariates_options}}, \code{\link{data_options}} 
    \code{\link{rodents_options}}, \code{\link{cast_options}}  
    \cr \cr
  \code{start_newmoon}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{plot_cast_ts}}  
    \cr \cr
  \code{sub_path}: must be \code{NULL} or a length-1 \code{character} 
    vector in \code{\link{create_sub_dir}}  \cr \cr
  \code{subs} must be a class-\code{subdirs} vector in
    \code{\link{all_options}}, \code{\link{dir_options}},
    \code{\link{PortalData_options}}, \code{\link{data_options}} 
    \code{\link{predictions_options}} 
    \code{\link{models_options}}, \code{\link{cast_options}},
    \code{\link{dirtree}} \cr \cr
  \code{subs_names}: must be \code{NULL} or a \code{character} 
    vector in \code{\link{subdirs}}  \cr \cr
  \code{subs_type}: must be \code{NULL} or a length-1 \code{character} 
    vector in \code{\link{subdirs}}  \cr \cr
  \code{tmnt_type}: must be \code{NULL} or a length-1 \code{character} 
    vector of value \code{"all"} or \code{"controls"} in 
    \code{\link{all_options}}, \code{\link{rodents_options}}, 
    \code{\link{data_options}} \cr \cr
  \code{to_cleanup}: must be \code{NULL} or a \code{character} vector in
    \code{\link{all_options}}, \code{\link{dir_options}} \cr \cr
  \code{treatment}: must be \code{NULL} or \code{"control"} in
    \code{\link{all_options}},
    \code{\link{rodents_options}}, \code{\link{data_options}} \cr \cr
  \code{tree}: must be of class \code{dirtree} in
    \code{\link{verify_PortalData}}, \code{\link{plot_cov_RMSE_mod_spp}},
    \code{\link{plot_err_lead_spp_mods}}, \code{\link{plot_cast_point}},
    \code{\link{plot_cast_point_yaxis}}, \code{\link{plot_cast_ts}},
    \code{\link{plot_cast_ts_xaxis}}, \code{\link{plot_cast_ts_ylab}},
    \code{\link{model_options}},
    \code{\link{rodents_options}}, \code{\link{metadata_options}}, 
    \code{\link{covariates_options}}, \code{\link{moons_options}},
    \code{\link{base_path}}, \code{\link{main_path}}, 
    \code{\link{sub_paths}}, 
    \code{\link{file_paths}}, \code{\link{model_paths}},
    \code{\link{create_tmp}}, \code{\link{clear_tmp}},
    \code{\link{prep_weather_data}} \cr \cr 
  \code{version}: must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{PortalData_options}} \cr \cr
  \code{with_census}: must be a length-1 \code{logical} vector in
    \code{\link{plot_cast_point}} \cr \cr
  \code{yr}: must be a length-1 \code{integer} or 
    \code{integer}-conformable value after 1970 in 
    \code{\link{all_options}},   
    \code{\link{covariates_options}}, \code{\link{data_options}} \cr \cr
}
