% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_args}
\alias{check_args}
\title{Check arguments}
\usage{
check_args(base = ".", main = "", subs = subdirs(type =
  "portalcasting"), quiet = FALSE, cast_date = today(),
  append_missing_to_raw = TRUE, m_save = TRUE,
  m_filename = "moons.csv", tmnt_type = NULL, start = 217,
  end = NULL, hind_step = 1, drop_spp = "PI", min_plots = 24,
  min_traps = 1, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  cov_hist = TRUE, cov_fcast = TRUE, yr = as.numeric(format(today(),
  "\%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL,
  nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", cast_type = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml",
  download_existing_predictions = FALSE, model = models(set =
  "prefab"), ensemble = TRUE, version = "latest", from_zenodo = TRUE,
  n_future_moons = 12, save = TRUE, filename = "moons.csv",
  tree = dirtree(), to_cleanup = c("tmp", "PortalData"),
  options_all = all_options(), options_dir = dir_options(),
  options_PortalData = PortalData_options(),
  options_data = data_options(),
  options_predictions = predictions_options(),
  options_models = models_options(), path = path)
}
\arguments{
\item{base}{\code{character} name of the base folder where the directory 
should or does exist. Default \code{"."} (working directory). Will be 
created if it doesn't exist.}

\item{main}{\code{character} name of the main folder within the base that
contains the subdirectories. Default \code{""} (no main level included).}

\item{subs}{\code{character} vector naming the specific subdirectories
within the portalcasting directory tree. Default \code{subdirs(type = 
"portalcasting")} sets the subdirectories as \code{"predictions"}, 
\code{"models"}, \code{"PortalData"}, \code{"data"}, and \code{"tmp"}. 
It is generally not advised to change the subdirectories.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{cast_date}{\code{Date} from which future is defined, typically 
today's date (using \code{\link{today}}).}

\item{append_missing_to_raw}{\code{logical} indicator dictating if the 
missing moon dates should be appended to the raw data file (should be 
\code{TRUE} to allow the imported portalr functions to work properly).}

\item{m_save}{\code{logical} indicator controlling if the moons data should 
be saved out.}

\item{m_filename}{\code{character} name of the file for saving moons data.}

\item{tmnt_type}{Treatment type: \code{"all"} or \code{"controls"}.}

\item{start}{\code{integer} (or integer \code{numeric}) newmoon number of 
the first sample to be included. Default value is \code{217}, 
corresponding to \code{1995-01-01}.}

\item{end}{\code{integer} (or integer \code{numeric}) newmoon number of the
last sample to be included. Default value is \code{NULL}, which equates
to the most recently included sample. If \code{cast_type} is 
\code{"hindcasts"} and \code{end} is \code{NULL}, default becomes 
\code{490:403}, corresponding to \code{2017-01-01} back to
\code{2010-01-01}.}

\item{hind_step}{\code{integer} (or integer \code{numeric}) iteration 
parameter used to work across the input values for \code{end}. Default is
\code{1}, which is likely all it should be run using.}

\item{drop_spp}{\code{character}-valued vector of species names to drop 
from the forecasting.}

\item{min_plots}{\code{integer} (or integer \code{numeric}) of the minimum 
number of plots surveyed for a survey to be used.}

\item{min_traps}{\code{integer} (or integer \code{numeric}) of the minimum 
number of traps trapped for a plot to be used.}

\item{level}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{treatment}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{plots}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{output}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{r_save}{\code{logical} value indicating if the rodent data should be 
saved out.}

\item{r_filename}{\code{character} name of the file to save the rodent data
in.}

\item{cov_hist}{\code{logical} indicator of whether or not historical 
covariates are to be included.}

\item{cov_fcast}{\code{logical} indicator whether or not forecasted 
covariates are to be included.}

\item{yr}{\code{numeric} value of the year of today's date.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) number of moons
into the future the rodents are to be forecast.}

\item{min_lag}{\code{integer} (or integer \code{numeric}) of the minimum 
covariate lag time used in any model.}

\item{fcast_nms}{\code{integer} (or integer \code{numeric}) vector of 
newmoon numbers to be forecast for covariates.}

\item{nfcnm}{\code{integer} (or integer \code{numeric}) number of forecast 
newmoons for covariates.}

\item{append_fcast_csv}{\code{logical} indicator controlling if the new 
forecast should be appended to the historical forecasts for the purposes 
of hindcasting.}

\item{hist_fcast_file}{\code{character} name of the file where the 
historical covariate forecasts are held.}

\item{source_name}{\code{character} value for the name to give the 
covariaate forecast. Currently is \code{"current_archive"}. Previous to
\code{"current_archive"}, the data were retroactively filled in and are 
given the source name \code{"retroactive"}.}

\item{c_save}{\code{logical} indicator for if the covariate data should be 
saved out.}

\item{c_filename}{\code{character} value for the name of the covariate file 
for the saving if \code{c_save} is \code{TRUE}.}

\item{cast_type}{-cast type: \code{"forecasts"} or \code{"hindcasts"}.}

\item{confidence_level}{\code{numeric} confidence level used in 
summarizing model output. Must be between \code{0} and \code{1}.}

\item{meta_save}{\code{logical} indicator for if the metadata should be
saved out}

\item{meta_filename}{\code{character} value for the name of the metadata 
file for the saving if \code{c_save} is \code{TRUE}.}

\item{download_existing_predictions}{\code{logical} indicator for if the 
existing predictions files should be retrieved from the 
\href{https://github.com/weecology/portalPredictions}{portalPredictions 
repo}.}

\item{model}{\code{character} vector of the names of model scripts to 
include in a cast of the pipeline.}

\item{ensemble}{\code{logical} indicator of whether to create an ensemble
model.}

\item{version}{\code{character} values of the version number or 
\code{"latest"} (default) for the Portal Data to be download.}

\item{from_zenodo}{\code{logical} indicator of whether or not the Portal 
Data should come from Zenodo (come from GitHub if not).}

\item{n_future_moons}{\code{integer} (or integer \code{numeric}) value for 
the number of future moons to add.}

\item{save}{\code{logical} indicator if the specific data should be saved 
out.}

\item{filename}{\code{character} name of the file for the saving.}

\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{to_cleanup}{\code{character} vector of subdirectory names to
cleanup after completion of casting.}

\item{options_all}{Class-\code{all_options} list containing all options 
available for controlling the set up and population of the directory (see 
\code{\link{all_options}}).}

\item{options_dir}{Class-\code{dir_options} list containing all options 
available for controlling the set up of the directory (see 
\code{\link{dir_options}}).}

\item{options_PortalData}{Class-\code{PortalData_options} list containing 
available for controlling the set up and population of the PortalData
folder in the portalcasting directory (see 
\code{\link{PortalData_options}}).}

\item{options_data}{Class-\code{data_options} list containing available
for controlling the set up and population of the data folder in the 
portalcasting directory (see \code{\link{data_options}}).}

\item{options_predictions}{Class-\code{predictions_options} list containing 
available for controlling the set up and population of the predictions
folder in the portalcasting directory (see 
\code{\link{predictions_options}}).}

\item{options_models}{Class-\code{models_options} list containing 
available for controlling the set up and population of the models
folder in the portalcasting directory (see 
\code{\link{models_options}}).}

\item{path}{The normalized path of the specific subdirectory folder to be 
created in the directory tree as a \code{character} value (see 
\code{\link{normalizePath}}, \code{\link{sub_path}}).}
}
\description{
Verify the validity (class, size, etc.) of the inputs to the 
  functions.
}
\examples{
\dontrun{

check_args()
}

}
