% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{check_args}
\alias{check_args}
\alias{check_arg}
\alias{check_arg_list}
\title{Check a function's arguments' values for validity}
\usage{
check_args()

check_arg(arg_name, arg_value, fun_name = NULL)

check_arg_list()
}
\arguments{
\item{arg_name}{\code{character} value of the argument name.}

\item{arg_value}{Input value for the argument.}

\item{fun_name}{\code{character} value of the function name or \code{NULL}.}
}
\value{
\code{check_arg}: \code{character} vector of error messages
  associated with the specific situation 
  (\code{<fun_name>(<arg_name> = arg_value)}).
}
\description{
\code{check_args} checks that all of the arguments to a given
 function have valid values within the pipeline to avoid naming collision
 and improper formatting, by wrapping around \code{check_arg} for each
 argument. \cr \cr
 \code{check_arg} returns \code{NULL} if the argument is valid or 
 \code{character} vector of error messages associated 
 with the specific situation (\code{<fun_name>(<arg_name> = arg_value)})
 if the argument is invalid. \cr \cr
 \code{check_args} should only be called within a function. \cr \cr
 \code{check_arg_list_maker} produces a list of argument details for 
 checking within \code{check_arg}. \cr  \cr 
 See \code{Details} for argument-specific rules.
}
\details{
Argument details used for checking are produced simply via
 \code{check_arg_list_maker}. If you are adding to the codebase and use
 novel arguments, add them to that function's production. \cr
 Each argument should have its own element that is a list with the 
 following elements: \code{class} (\code{character} value of its class), 
 \code{null} (\code{logical} indicator if the argument can be \code{NULL}),
 \code{na} (\code{logical} indicator if the argument can be \code{NA}),
 \code{length} (\code{numeric} value of how many elements the argument can
 be or \code{NULL} if unrestricted), and \code{addl} if needed (an
 unevaluated \code{expression} that can be evaluated and allows for 
 simple evaluation of, e.g., element names of a \code{list}; see the 
 source code).
}
\examples{
 check_arg_list()

}
