% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_args}
\alias{check_args}
\alias{check_arg}
\title{Check a function's arguments' values for validity}
\usage{
check_args()

check_arg(arg_name, arg_value, fun_name = NULL)
}
\arguments{
\item{arg_name}{\code{character} value of the argument name.}

\item{arg_value}{Input value for the argument.}

\item{fun_name}{\code{character} value of the function name or \code{NULL}.}
}
\value{
\code{check_arg}: \code{character} vector of error messages
  associated with the specific situation 
  (\code{<fun_name>(<arg_name> = arg_value)}).
}
\description{
\code{check_args} checks that all of the arguments to a given
  function have valid values within the pipeline to avoid naming collision
  and improper formatting, by wrapping around \code{check_arg} for each
  argument. \cr \cr
  \code{check_arg} produces (returns) a \code{NULL} if the 
  argument is valid or \code{character} vector of error messages associated 
  with the specific situation (\code{<fun_name>(<arg_name> = arg_value)})
  if the argument is invalid. \cr \cr
  \code{check_args} should only be called within a function. \cr \cr
  See \code{Details} for argument-specific rules.
}
\details{
Usage and rules for arguments are as follows: \cr \cr
  \code{add}: must be \code{NULL} or a \code{character} vector in
    \code{\link{model_names}}. \cr \cr
  \code{add_error}: must be a length-1 \code{logical} vector in
    \code{\link{append_observed_to_cast}}. \cr \cr
  \code{add_in_window}: must be a length-1 \code{logical} vector in
    \code{\link{append_observed_to_cast}}. \cr \cr
  \code{add_lead}: must be a length-1 \code{logical} vector in
    \code{\link{append_observed_to_cast}}. \cr \cr
  \code{add_obs}: must be a length-1 \code{logical} vector in
    \code{\link{plot_cast_ts}}. \cr \cr
  \code{all}: must be a \code{list} with elements named \code{forecast} 
    and \code{aic} in \code{\link{save_forecast_output}} \cr \cr
  \code{all_forecasts}: must be a \code{data.frame} in
    \code{\link{make_ensemble}}. \cr \cr
  \code{append_fcast_csv}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{covariates_options}}, and 
      \code{\link{data_options}}. \cr \cr
  \code{append_missing_to_raw}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{data_options}}, and
    \code{\link{moons_options}}. \cr \cr
  \code{base} must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{cast_options}},
    \code{\link{data_options}}, \code{\link{dirtree}}, 
    \code{\link{dir_options}}, \code{\link{models_options}}, 
    \code{\link{PortalData_options}}, and
    \code{\link{predictions_options}}. \cr \cr
  \code{cast}: must be a valid -cast \code{data.frame} as checked by
    \code{\link{cast_is_valid}} in \code{\link{verify_cast}}. \cr \cr
  \code{casts}: must be of class \code{casts} in
    \code{\link{append_observed_to_cast}} and 
    \code{\link{measure_cast_error}}. \cr \cr
  \code{cast_date}: must be \code{NULL} or a length-1 \code{Date} or 
    \code{Date}-conformable vector in \code{\link{add_addl_future_moons}},
    \code{\link{all_options}}, \code{\link{cast_options}}, 
    \code{\link{covariates_options}}, \code{\link{data_options}},
    \code{\link{metadata_options}}, \code{\link{moons_options}}, 
    \code{\link{plot_cast_point}}, \code{\link{plot_cast_ts}}, and 
    \code{\link{read_cast}}. \cr \cr
  \code{cast_dates}: must be \code{NULL} or a \code{Date} or 
    \code{Date}-conformable vector in \code{\link{plot_cov_RMSE_mod_spp}}, 
    \code{\link{read_casts}}, and \code{\link{select_most_ab_spp}}. \cr \cr
  \code{cast_to_check}: must be a \code{data.frame} in 
    \code{\link{cast_is_valid}}. \cr \cr
  \code{cast_type}: must be a length-1 \code{character} vector in
    \code{\link{most_recent_cast}}, \code{\link{plot_cast_point}},
    \code{\link{plot_cast_ts}}, \code{\link{plot_cov_RMSE_mod_spp}}, 
    \code{\link{plot_err_lead_spp_mods}}, \code{\link{read_cast}}, 
    \code{\link{read_casts}}, and \code{\link{select_most_ab_spp}}. \cr \cr
  \code{class}: must be \code{NULL} or a \code{character} vector in
    \code{\link{classy}}. \cr \cr
  \code{colname}: must be a length-1 \code{character} vector in
    \code{\link{fcast0}}, \code{\link{na_conformer}}, and
    \code{\link{remove_incompletes}}. \cr \cr
  \code{confidence_level}: must be a length-1 \code{numeric} value between
    0 and 1 in \code{\link{all_options}}, \code{\link{data_options}}, 
    \code{\link{make_ensemble}}, and \code{\link{metadata_options}}. 
    \cr \cr
  \code{controls}: must be a \code{list} with elements named 
    \code{forecast} and \code{aic} in \code{\link{save_forecast_output}}. 
    \cr \cr
  \code{covariates}: must be a \code{data.frame} of class \code{covariates}
    in \code{\link{forecast_covariates}}, \code{\link{forecast_ndvi}},
    and \code{\link{prep_metadata}}. \cr \cr
  \code{cov_fcast}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{covariates_options}}, and
    \code{\link{data_options}}. \cr \cr
  \code{cov_hist}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{covariates_options}}, and
    \code{\link{data_options}}. \cr \cr
  \code{c_filename}: must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{covariates_options}}, and 
    \code{\link{data_options}}. \cr \cr
  \code{c_save}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{covariates_options}}, and 
    \code{\link{data_options}}. \cr \cr
  \code{data_name}: must be a length-1 \code{character} vector of value 
    \code{"all"}, \code{"controls"}, \code{"covariates"}, \code{"moons"},
    or \code{"metadata"} in \code{\link{read_data}}. \cr \cr
  \code{dates}: must be \code{NULL} or a \code{Date} or 
    \code{Date}-conformable vector in \code{\link{foy}}. \cr \cr
  \code{df}: must be a \code{data.frame} in 
    \code{\link{append_csv}}, \code{\link{dataout}}, and
    \code{\link{remove_incompletes}}. \cr \cr
  \code{download}: must be a length-1 \code{logical} vector in
    \code{\link{download_predictions}}. \cr \cr 
  \code{download_existing_predictions}: must be a length-1 \code{logical} 
    vector in \code{\link{all_options}} and  
    \code{\link{predictions_options}}. \cr \cr
  \code{drop_spp}: must be a \code{character} vector in
    \code{\link{all_options}}, \code{\link{data_options}},
    \code{\link{remove_spp}}, and \code{\link{rodents_options}}. \cr \cr
  \code{end}: must be \code{NULL} or a positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{cast_options}}, \code{\link{covariates_options}}, 
    \code{\link{data_options}}, and \code{\link{rodents_options}}.
    \cr \cr
  \code{ensemble}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}} and \code{\link{cast_options}}. \cr \cr
  \code{extension}: must be a length-1 \code{character} vector with
    a single period in \code{\link{model_paths}}. \cr \cr
  \code{fcast_nms}: must be \code{NULL} or a non-negative \code{integer} or
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{covariates_options}}, and \code{\link{data_options}} 
    \cr \cr
  \code{filename}: must be a length-1 \code{character} vector in
    \code{\link{append_csv}} and \code{\link{verify_PortalData}}. \cr \cr
  \code{from_date}: must be \code{NULL} or a length-1 \code{Date} or 
    \code{Date}-conformable vector in \code{\link{plot_cast_point}} and
    \code{\link{select_most_ab_spp}}. \cr \cr
  \code{from_zenodo}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}} and \code{\link{PortalData_options}}. 
    \cr \cr
  \code{future_moons}: must be a \code{data.frame} of class \code{moons} 
    in \code{\link{add_addl_future_moons}}. \cr \cr
  \code{hind_step}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{cast_options}}, \code{\link{covariates_options}}, 
    \code{\link{data_options}}, and \code{\link{rodents_options}}. \cr \cr
  \code{hist_cov}: must be a \code{data.frame} of class \code{covariates} 
    in \code{\link{prep_fcast_covariates}} and
    \code{\link{update_covfcast_options}}. \cr \cr
  \code{hist_fcast_file}: must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{covariates_options}}, and
    \code{\link{data_options}}. \cr \cr
  \code{lag}: must be \code{NULL} or a length-1 non-negative \code{integer}
    or \code{integer}-conformable vector in \code{\link{lag_covariates}}, 
    \code{\link{model_options}}, and \code{\link{pevGARCH}}. \cr \cr
  \code{lead}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{plot_cast_point}} and
    \code{\link{select_most_ab_spp}}. \cr \cr
  \code{lead_time}: must be a length-1 non-negative \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}}, 
    \code{\link{covariates_options}}, \code{\link{data_options}}, and
    \code{\link{metadata_options}}. \cr \cr
  \code{level}: must be \code{NULL} or a length-1 \code{character} vector 
    of value \code{"All"} or \code{"Controls"} in 
    \code{\link{plot_cast_point}},
    \code{\link{plot_cast_ts}}, \code{\link{plot_cast_ts_ylab}},
    \code{\link{plot_cov_RMSE_mod_spp}}, 
    \code{\link{plot_err_lead_spp_mods}}, \code{\link{select_most_ab_spp}}, 
    and \code{\link{select_casts}} or a length-1 \code{character} vector 
    of value \code{"Site"} or \code{"Treatment"} in 
    \code{\link{AutoArima}}, \code{\link{all_options}}, 
    \code{\link{data_options}}, \code{\link{ESSS}}, 
    \code{\link{metadata_options}}, \code{\link{nbGARCH}},
    \code{\link{nbsGARCH}},
    \code{\link{pevGARCH}}, and \code{\link{rodents_options}}. \cr \cr
  \code{local_paths}: must be a \code{character} 
    vector in \code{\link{file_paths}}. \cr \cr 
  \code{main} must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{cast_options}},
    \code{\link{data_options}}, \code{\link{dirtree}}, 
    \code{\link{dir_options}}, \code{\link{PortalData_options}},  
    \code{\link{models_options}}, and \code{\link{predictions_options}}.
    \cr \cr
  \code{metadata}: must be of class \code{metadata} in
    \code{\link{save_forecast_output}}. \cr \cr
  \code{meta_filename}: must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{data_options}}, and
    \code{\link{metadata_options}}. \cr \cr
  \code{meta_save}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{data_options}}, and
    \code{\link{metadata_options}}. \cr \cr
  \code{min_lag}: must be a length-1 non-negative \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{covariates_options}}, and \code{\link{data_options}}. 
    \cr \cr
  \code{min_observed}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in 
    \code{\link{measure_cast_error}} and 
    \code{\link{plot_cov_RMSE_mod_spp}}. \cr \cr
  \code{min_plots}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}}, 
    \code{\link{cast_options}}, \code{\link{data_options}}, and
    \code{\link{rodents_options}}. \cr \cr
  \code{min_traps}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}}, 
    \code{\link{cast_options}}, \code{\link{data_options}}, and
    \code{\link{rodents_options}}. \cr \cr
  \code{model}: must be a length-1 \code{character} vector in
    \code{\link{model_options}}, \code{\link{plot_cast_point}}, 
    \code{\link{plot_cast_ts}}, \code{\link{plot_cast_ts_ylab}},
    \code{\link{save_forecast_output}}, and
    \code{\link{select_most_ab_spp}}. \cr \cr
  \code{models}: must be \code{NULL} or a \code{character} vector in
    \code{\link{all_options}}, \code{\link{cast_options}}, 
    \code{\link{models_options}}, \code{\link{model_paths}}, and 
    \code{\link{select_casts}}, \code{\link{plot_cov_RMSE_mod_spp}}, and
    \code{\link{plot_err_lead_spp_mods}}. \cr \cr
  \code{model_set}: must be \code{NULL} or a length-1 \code{character} 
    vector with value \code{"prefab"} or \code{"wEnsemble"} in 
    \code{\link{model_names}}. \cr \cr
  \code{mod_covariates}: must be a length-1 \code{logical} vector in 
    \code{\link{model_options}}. \cr \cr
  \code{mod_type}: must be \code{"pevGARCH"} in
    \code{\link{covariate_models}} . \cr \cr
  \code{moons}: must be a \code{data.frame} of class \code{moons} in
    \code{\link{add_future_moons}}, \code{\link{append_past_moons_to_raw}},
    \code{\link{forecast_covariates}}, \code{\link{forecast_ndvi}},
    \code{\link{forecast_weather}}, \code{\link{format_moons}},
    \code{\link{get_climate_forecasts}}, 
    \code{\link{prep_covariates}}, \code{\link{prep_fcast_covariates}},
    \code{\link{prep_metadata}}, \code{\link{prep_rodents}}, 
    \code{\link{prep_rodents_list}}, \code{\link{trim_moons_fcast}},
    \code{\link{update_covariates}}, and
    \code{\link{update_covfcast_options}}. \cr \cr
  \code{msg}: must be \code{NULL} or a \code{character} vector in
    \code{\link{messageq}}. \cr \cr
  \code{m_filename}: must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{data_options}}, and
    \code{\link{moons_options}}. \cr \cr
  \code{m_save}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{data_options}}, and
    \code{\link{moons_options}}. \cr \cr
  \code{nadot}: must be a length-1 \code{logical} vector in
    \code{\link{rodent_spp}}. \cr \cr
  \code{ndates}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in 
    \code{\link{plot_err_lead_spp_mods}}. \cr \cr
  \code{newmoonnumbers}: must be \code{NULL} or a length-1 non-negative 
    \code{integer} or \code{integer}-conformable vector in 
    \code{\link{select_casts}}. \cr \cr
  \code{new_forecast_covariates}: must be a \code{data.frame} of class 
    \code{covariates} in \code{\link{append_cov_fcast_csv}}. \cr \cr
  \code{nfcnm}: must be a length-1 non-negative \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{covariates_options}}, and \code{\link{data_options}}. 
    \cr \cr
  \code{n_future_moons}: must be a length-1 non-negative \code{integer} or 
    \code{integer}-conformable vector in \code{\link{moons_options}}.
    \cr \cr
  \code{options_all}: must be of class \code{all_options} in
    \code{\link{casts}}, \code{\link{cleanup_dir}}, 
    \code{\link{fill_dir}}, \code{\link{portalcast}},
    \code{\link{setup_dir}}, \code{\link{step_casts}},and 
    \code{\link{step_hind_forward}}. \cr \cr
  \code{options_cast}: must be of class \code{cast_options} in
    \code{\link{add_ensemble}}, \code{\link{cast}},
    \code{\link{cast_models}}, \code{\link{check_to_skip}}, 
    \code{\link{combine_forecasts}}, \code{\link{models_to_cast}}, and
    \code{\link{verify_models}}. \cr \cr
  \code{options_covariates}: must be of class \code{covariates_options} in
    \code{\link{append_cov_fcast_csv}}, 
    \code{\link{forecast_covariates}}, \code{\link{forecast_ndvi}},
    \code{\link{forecast_weather}}, \code{\link{get_climate_forecasts}}, 
    \code{\link{prep_covariates}}, \code{\link{prep_fcast_covariates}},
    \code{\link{prep_hist_covariates}}, \code{\link{trim_moons_fcast}}, 
    \code{\link{update_covariates}}, and
    \code{\link{update_covfcast_options}}. \cr \cr
  \code{options_data}: must be of class \code{data_options} in
    \code{\link{fill_data}}, \code{\link{prep_data}},
    \code{\link{transfer_hist_covariate_forecasts}},
    \code{\link{transfer_trapping_table}}, and 
    \code{\link{update_data}}. \cr \cr
  \code{options_dir}: must be of class \code{dir_options} in
    \code{\link{create_dir}}, \code{\link{create_main_dir}}, and
    \code{\link{create_sub_dirs}}. \cr \cr
  \code{options_metadata}: must be of class \code{metadata_options} in
    \code{\link{prep_metadata}}. \cr \cr
  \code{options_model}: must be of class \code{model_options} in
    \code{\link{model_template}} and \code{\link{write_model}}. \cr \cr
  \code{options_models}: must be of class \code{models_options} in
    \code{\link{fill_models}}. \cr \cr
  \code{options_moons}: must be of class \code{moons_options} in
    \code{\link{add_future_moons}}, \code{\link{append_past_moons_to_raw}},
    and \code{\link{prep_moons}}. \cr \cr
  \code{options_out}: must be an options \code{list} in
    \code{\link{dataout}}. \cr \cr
  \code{options_PortalData}: must be of class \code{PortalData_options} in
    \code{\link{fill_PortalData}}. \cr \cr
  \code{options_predictions}: must be of class \code{predictions_options}
    in \code{\link{fill_predictions}}. \cr \cr
  \code{options_rodents}: must be of class \code{rodents_options} in
    \code{\link{enforce_rodents_options}}, \code{\link{prep_rodents}}, 
    \code{\link{prep_rodents_list}}, \code{\link{trim_treatment}}, and
    \code{\link{update_rodents_list}}. \cr \cr
  \code{output}: must be \code{"abundance"} in \code{\link{all_options}}, 
    \code{\link{data_options}}, and \code{\link{rodents_options}}. \cr \cr
  \code{plots}: must be a length-1 \code{character} vector of value
    \code{"all"} or \code{"longerm"} in \code{\link{all_options}},
    \code{\link{data_options}}, and \code{\link{rodents_options}}. \cr \cr
  \code{pred_dir}: must be a length-1 \code{character} vector in
    \code{\link{compile_aic_weights}} and \code{\link{make_ensemble}}. 
    \cr \cr
  \code{quiet}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{AutoArima}}, 
    \code{\link{cast_options}}, \code{\link{covariates_options}},
    \code{\link{create_sub_dir}}, \code{\link{data_options}},
    \code{\link{download_predictions}}, \code{\link{dir_options}}, 
    \code{\link{ESSS}}, \code{\link{metadata_options}},
    \code{\link{models_options}}, \code{\link{model_options}},  
    \code{\link{predictions_options}}, \code{\link{PortalData_options}}, 
    \code{\link{moons_options}}, \code{\link{messageq}},
    \code{\link{nbGARCH}}, \code{\link{nbsGARCH}}, \code{\link{pevGARCH}},
    \code{\link{rodents_options}}, and \code{\link{verify_PortalData}}.
    \cr \cr 
  \code{rangex}: must be a length-2 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{plot_cast_ts_xaxis}}.  
    \cr \cr
  \code{rodents}: must be a \code{data.table} of class \code{rodents} in 
    \code{\link{interpolate_abundance}}, \code{\link{is.spcol}},  
    \code{\link{remove_spp}}, and \code{\link{trim_treatment}}. \cr \cr 
  \code{rodents_list}: must be of class \code{rodents_list} in 
    \code{\link{prep_metadata}}. \cr \cr 
  \code{r_filename}: must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{data_options}}, and
    \code{\link{rodents_options}}. \cr \cr
  \code{r_save}: must be a length-1 \code{logical} vector in
    \code{\link{all_options}}, \code{\link{data_options}}, and
    \code{\link{rodents_options}}. \cr \cr
  \code{save}: must be a length-1 \code{logical} vector in
    \code{\link{covariates_options}}, \code{\link{metadata_options}},
    \code{\link{moons_options}}, and \code{\link{rodents_options}}. \cr \cr
  \code{source_name}: must be a length-1 \code{character} vector in
    \code{\link{all_options}}, \code{\link{covariates_options}},
    and \code{\link{data_options}}. \cr \cr
  \code{species}: must be \code{NULL} or a \code{character} vector in
    \code{\link{plot_cast_point}}, \code{\link{plot_cast_point_yaxis}},
    \code{\link{plot_cov_RMSE_mod_spp}}, 
    \code{\link{plot_err_lead_spp_mods}}, 
    \code{\link{select_casts}}, and \code{\link{select_most_ab_spp}};
    specifically length-1 in
    \code{\link{plot_cast_ts}} and \code{\link{plot_cast_ts_ylab}}. \cr \cr
  \code{species_set}: must be \code{NULL} or a length-1 \code{character} 
    vector of value \code{"base"}, \code{"wtotal"}, or \code{"evalplot"} in 
    \code{\link{rodent_spp}}. \cr \cr
  \code{specific_subs}: must be \code{NULL} or a length-1 \code{character} 
    vector in \code{\link{sub_paths}}. \cr \cr 
  \code{spp_names}: must be a \code{character} vector in
    \code{\link{is.spcol}}. \cr \cr
  \code{start}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{all_options}},
    \code{\link{cast_options}}, \code{\link{covariates_options}}, 
    \code{\link{data_options}}, and \code{\link{rodents_options}}. \cr \cr
  \code{start_newmoon}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{plot_cast_ts}}.  
    \cr \cr
  \code{subs} must be a class-\code{character} vector in
    \code{\link{all_options}}, \code{\link{cast_options}}, 
    \code{\link{data_options}}, \code{\link{dirtree}}, 
    \code{\link{dir_options}}, \code{\link{models_options}}, 
    \code{\link{PortalData_options}}, and 
    \code{\link{predictions_options}}. \cr \cr
  \code{subs_names}: must be \code{NULL} or a \code{character} 
    vector in \code{\link{subdirs}}. \cr \cr
  \code{subs_type}: must be \code{NULL} or a length-1 \code{character} 
    vector in \code{\link{subdirs}}. \cr \cr
  \code{sub_path}: must be \code{NULL} or a length-1 \code{character} 
    vector in \code{\link{create_sub_dir}}. \cr \cr
  \code{tail}: must be a length-1 \code{logical} vector in
    \code{\link{lag_covariates}}. \cr \cr
  \code{temp_dir}: must be a length-1 \code{character} vector in
    \code{\link{save_forecast_output}}. \cr \cr
  \code{time}: must be a length-1 \code{logical} vector in
    \code{\link{today}}. \cr \cr
  \code{tmnt_type}: must be \code{NULL} or a length-1 \code{character} 
    vector of value \code{"all"} or \code{"controls"} in 
    \code{\link{all_options}}, \code{\link{data_options}}, 
    \code{\link{enforce_rodents_options}}, and
    \code{\link{rodents_options}}. \cr \cr
  \code{to_cleanup}: must be \code{NULL} or a \code{character} vector in
    \code{\link{all_options}} and \code{\link{dir_options}}. \cr \cr
  \code{topx}: must be a length-1 positive \code{integer} or 
    \code{integer}-conformable vector in \code{\link{select_most_ab_spp}}.
    \cr \cr
  \code{treatment}: must be \code{NULL} or \code{"control"} in
    \code{\link{all_options}}, \code{\link{data_options}}, and
    \code{\link{rodents_options}}. \cr \cr
  \code{tree}: must be of class \code{dirtree} in
    \code{\link{append_observed_to_cast}}, \code{\link{AutoArima}}, 
    \code{\link{base_path}}, \code{\link{clear_tmp}},
    \code{\link{covariates_options}}, \code{\link{create_tmp}}, 
    \code{\link{download_predictions}}, \code{\link{ESSS}}, 
    \code{\link{file_paths}}, \code{\link{main_path}}, 
    \code{\link{metadata_options}}, \code{\link{model_options}},
    \code{\link{model_paths}}, \code{\link{model_scripts}}, 
    \code{\link{moons_options}}, \code{\link{most_recent_cast}}, 
    \code{\link{most_recent_census}}, \code{\link{nbGARCH}}, 
    \code{\link{nbsGARCH}},
    \code{\link{pevGARCH}}, \code{\link{plot_cast_point}},
    \code{\link{plot_cast_point_yaxis}}, \code{\link{plot_cast_ts}},
    \code{\link{plot_cast_ts_xaxis}}, \code{\link{plot_cast_ts_ylab}},
    \code{\link{plot_cov_RMSE_mod_spp}},
    \code{\link{plot_err_lead_spp_mods}}, \code{\link{prep_weather_data}},
    \code{\link{read_all}}, \code{\link{read_cast}}, 
    \code{\link{read_casts}}, \code{\link{read_controls}},
    \code{\link{read_covariates}}, \code{\link{read_data}}, 
    \code{\link{read_metadata}}, \code{\link{read_moons}},  
    \code{\link{rodents_options}}, \code{\link{select_most_ab_spp}},    
    \code{\link{sub_paths}}, and \code{\link{verify_PortalData}}. \cr \cr
  \code{verbose}: must be a length-1 \code{logical} vector in
    \code{\link{cast_is_valid}}. \cr \cr
  \code{version}: must be a length-1 \code{character} vector in
    \code{\link{all_options}} and \code{\link{PortalData_options}}. \cr \cr
  \code{with_census}: must be a length-1 \code{logical} vector in
    \code{\link{most_recent_cast}} and \code{\link{plot_cast_point}}. 
    \cr \cr
  \code{x}: must be a \code{data.frame} or \code{vector} in
    \code{\link{na_conformer}} and must simply exist as an object that can
     have a class attribute in \code{\link{classy}}. \cr \cr
  \code{yr}: must be a length-1 \code{integer} or 
    \code{integer}-conformable value after 1970 in 
    \code{\link{all_options}}, \code{\link{covariates_options}}, 
    and \code{\link{data_options}}. \cr \cr
}
