% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_args}
\alias{check_args}
\title{Check a function's arguments' values for validity}
\usage{
check_args()
}
\description{
\code{check_args} checks that all of the arguments to a given
  function have valid values within the pipeline to avoid naming collision
  and improper formatting, by wrapping around \code{check_arg} for each
  argument. \cr \cr
  \code{check_arg} produces (returns) a \code{NULL} if the 
  argument is valid or \code{character} vector of error messages associated 
  with the specific situation (\code{<fun_name>(<arg_name> = arg_value)})
  if the argument is invalid. \cr \cr
  \code{check_args} should only be called within a function. \cr \cr
  See \code{Details} for argument-specific rules.
}
