% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_args}
\alias{check_args}
\title{Check arguments}
\usage{
check_args(toggle = NULL, base = ".", main = "", subs = NULL,
  tail = FALSE, quiet = FALSE, cast_date = today(),
  append_missing_to_raw = TRUE, m_save = TRUE,
  m_filename = "moons.csv", tmnt_type = NULL, start = 217,
  end = NULL, hind_step = 1, drop_spp = "PI", min_plots = 24,
  min_traps = 1, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  cov_hist = TRUE, cov_fcast = TRUE, yr = as.numeric(format(today(),
  "\%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL,
  nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", cast_type = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml",
  download_existing_predictions = FALSE, models = NULL,
  ensemble = TRUE, download = FALSE, version = "latest",
  from_zenodo = TRUE, n_future_moons = 12, save = TRUE,
  filename = "moons.csv", tree = NULL, to_cleanup = c("tmp",
  "PortalData"), options_all = NULL, options_moons = NULL,
  options_dir = NULL, options_PortalData = NULL, options_data = NULL,
  options_predictions = NULL, options_models = NULL,
  options_model = NULL, options_covariates = NULL,
  options_rodents = NULL, options_cast = NULL,
  options_metadata = NULL, path = NULL, species = rodent_spp(set =
  "evalplot"), cast_dates = NULL, min_observed = 1, ndates = 3,
  from_date = today(), with_census = FALSE, rangex = 2:3,
  start_newmoon = 300, add_obs = TRUE, moons = NULL,
  covariates = NULL, hist_cov = NULL, new_forecast_covariates = NULL,
  name = "AutoArima", lag = NULL, mod_covariates = FALSE,
  subs_names = NULL, subs_type = "portalcasting",
  specific_sub = NULL, extension = ".R", local_path = "data/all.csv",
  rodents = NULL, model = "AutoArima", set = "prefab", add = NULL,
  lead = 1, rodents_list = NULL, future_moons = NULL,
  data_name = "all")
}
\arguments{
\item{toggle}{\code{character} value indicating special aspects of 
checking.}

\item{base}{\code{character} name of the base folder where the directory 
should or does exist. Default \code{"."} (working directory). Will be 
created if it doesn't exist.}

\item{main}{\code{character} name of the main folder within the base that
contains the subdirectories. Default \code{""} (no main level included).}

\item{subs}{\code{character} vector naming the specific subdirectories
within the portalcasting directory tree. Default \code{subdirs(type = 
"portalcasting")} sets the subdirectories as \code{"predictions"}, 
\code{"models"}, \code{"PortalData"}, \code{"data"}, and \code{"tmp"}. 
It is generally not advised to change the subdirectories.}

\item{tail}{\code{logical} indicator if the data lagged to the tail end 
should be retained.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{cast_date}{\code{Date} from which future is defined, typically 
today's date (using \code{\link{today}}).}

\item{append_missing_to_raw}{\code{logical} indicator dictating if the 
missing moon dates should be appended to the raw data file (should be 
\code{TRUE} to allow the imported portalr functions to work properly).}

\item{m_save}{\code{logical} indicator controlling if the moons data should 
be saved out.}

\item{m_filename}{\code{character} name of the file for saving moons data.}

\item{tmnt_type}{Treatment type: \code{NULL}, \code{"all"} or 
\code{"controls"}.}

\item{start}{\code{integer} (or integer \code{numeric}) newmoon number of 
the first sample to be included. Default value is \code{217}, 
corresponding to \code{1995-01-01}.}

\item{end}{\code{integer} (or integer \code{numeric}) newmoon number of the
last sample to be included. Default value is \code{NULL}, which equates
to the most recently included sample. If \code{cast_type} is 
\code{"hindcasts"} and \code{end} is \code{NULL}, default becomes 
\code{490:403}, corresponding to \code{2017-01-01} back to
\code{2010-01-01}.}

\item{hind_step}{\code{integer} (or integer \code{numeric}) iteration 
parameter used to work across the input values for \code{end}. Default is
\code{1}, which is likely all it should be run using.}

\item{drop_spp}{\code{character}-valued vector of species names to drop 
from the forecasting.}

\item{min_plots}{\code{integer} (or integer \code{numeric}) of the minimum 
number of plots surveyed for a survey to be used.}

\item{min_traps}{\code{integer} (or integer \code{numeric}) of the minimum 
number of traps trapped for a plot to be used.}

\item{level}{For data collection, \code{character} input of \code{level} 
for \code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}. \cr \cr For -casts, \code{character} value of the 
\code{level} of interest  (\code{"All"} or \code{"Controls"}) in the 
-casts. \cr \cr Differentiation by \code{toggle} (\code{NULL} for
data collection and a toggle character value including \code{"plot"}
for -casts).}

\item{treatment}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{plots}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{output}{\code{character} input for 
\code{\link[portalr]{summarize_rodent_data}}, automatically set by 
\code{tmnt_type}.}

\item{r_save}{\code{logical} value indicating if the rodent data should be 
saved out.}

\item{r_filename}{\code{character} name of the file to save the rodent data
in.}

\item{cov_hist}{\code{logical} indicator of whether or not historical 
covariates are to be included.}

\item{cov_fcast}{\code{logical} indicator whether or not forecasted 
covariates are to be included.}

\item{yr}{\code{numeric} value of the year of today's date.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) number of moons
into the future the rodents are to be forecast.}

\item{min_lag}{\code{integer} (or integer \code{numeric}) of the minimum 
covariate lag time used in any model.}

\item{fcast_nms}{\code{integer} (or integer \code{numeric}) vector of 
newmoon numbers to be forecast for covariates.}

\item{nfcnm}{\code{integer} (or integer \code{numeric}) number of forecast 
newmoons for covariates.}

\item{append_fcast_csv}{\code{logical} indicator controlling if the new 
forecast should be appended to the historical forecasts for the purposes 
of hindcasting.}

\item{hist_fcast_file}{\code{character} name of the file where the 
historical covariate forecasts are held.}

\item{source_name}{\code{character} value for the name to give the 
covariaate forecast. Currently is \code{"current_archive"}. Previous to
\code{"current_archive"}, the data were retroactively filled in and are 
given the source name \code{"retroactive"}.}

\item{c_save}{\code{logical} indicator for if the covariate data should be 
saved out.}

\item{c_filename}{\code{character} value for the name of the covariate file 
for the saving if \code{c_save} is \code{TRUE}.}

\item{cast_type}{-cast type: \code{"forecasts"} or \code{"hindcasts"}.}

\item{confidence_level}{\code{numeric} confidence level used in 
summarizing model output. Must be between \code{0} and \code{1}.}

\item{meta_save}{\code{logical} indicator for if the metadata should be
saved out}

\item{meta_filename}{\code{character} value for the name of the metadata 
file for the saving if \code{c_save} is \code{TRUE}.}

\item{download_existing_predictions}{\code{logical} indicator for if the 
existing predictions files should be retrieved from the 
\href{https://github.com/weecology/portalPredictions}{portalPredictions 
repo}.}

\item{models}{\code{character} vector of the names of models to include 
in a cast of the pipeline.}

\item{ensemble}{\code{logical} indicator of whether to create an ensemble
model.}

\item{download}{\code{logical} indicator of whether the download should 
actually happen. Should be \code{TRUE} except for testing purposes.}

\item{version}{\code{character} values of the version number or 
\code{"latest"} (default) for the Portal Data to be download.}

\item{from_zenodo}{\code{logical} indicator of whether or not the Portal 
Data should come from Zenodo (come from GitHub if not).}

\item{n_future_moons}{\code{integer} (or integer \code{numeric}) value for 
the number of future moons to add.}

\item{save}{\code{logical} indicator if the specific data should be saved 
out.}

\item{filename}{\code{character} name of the file for the saving.}

\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{to_cleanup}{\code{character} vector of subdirectory names to
cleanup after completion of casting.}

\item{options_all}{Class-\code{all_options} list containing all options 
available for controlling the set up and population of the directory (see 
\code{\link{all_options}}).}

\item{options_moons}{A class-\code{moons_options} \code{list} of options 
for the moons data. See \code{\link{moons_options}}.}

\item{options_dir}{Class-\code{dir_options} list containing all options 
available for controlling the set up of the directory (see 
\code{\link{dir_options}}).}

\item{options_PortalData}{Class-\code{PortalData_options} list containing 
available for controlling the set up and population of the PortalData
folder in the portalcasting directory (see 
\code{\link{PortalData_options}}).}

\item{options_data}{Class-\code{data_options} list containing available
for controlling the set up and population of the data folder in the 
portalcasting directory (see \code{\link{data_options}}).}

\item{options_predictions}{Class-\code{predictions_options} list containing 
available for controlling the set up and population of the predictions
folder in the portalcasting directory (see 
\code{\link{predictions_options}}).}

\item{options_models}{Class-\code{models_options} list containing 
available for controlling the set up and population of the models
folder in the portalcasting directory (see 
\code{\link{models_options}}).}

\item{options_model}{A class-\code{model_options} \code{list} of
options used to set up a general model script. See 
\code{\link{model_options}}.}

\item{options_covariates}{Class-\code{covariates_options} list of 
options for the covariate data. See \code{\link{covariates_options}}.}

\item{options_rodents}{A class-\code{rodents_options} \code{list} of 
settings controlling the rodents data creation.}

\item{options_cast}{Class-\code{cast_options} \code{list} containing the
hind- or forecasting options. See \code{\link{cast_options}}.}

\item{options_metadata}{A class-\code{metadata_options} \code{list} of 
settings controlling the metadata data creation.}

\item{path}{The normalized path of the specific subdirectory folder to be 
created in the directory tree as a \code{character} value (see 
\code{\link{normalizePath}}, \code{\link{sub_path}}).}

\item{species}{\code{character} vector of the species codes (or 
\code{"total"} for the total across species) to be selected from or 
\code{NULL} to include all species and the total. Default set to 
\code{rodent_spp(set = "evalplot")}, the only setting for which the
plot is reliably coded/}

\item{cast_dates}{\code{Date}s the predictions were made. Used to select 
the files in the predictions subdirectory. Can be length 1 or more and if 
\code{NULL} (default), selects all available -casts.}

\item{min_observed}{\code{integer} value for the minimum number of observed
values needed for a -cast to be retained in the output table. Default is
\code{1}, which returns all -casts with any observations. To include all
-casts (even those without any evaluations), set to \code{0}.}

\item{ndates}{\code{integer} number of -cast issue dates to include.}

\item{from_date}{\code{Date} to be used as a reference of when to count
the \code{lead} from. If \code{NULL} (default), for 
\code{cast_type = "forecasts"}, \code{from_date = cast_date} and 
\code{plot_cast_point} is not yet reliable for 
\code{cast_type = "hindcasts"}.}

\item{with_census}{\code{logical} toggle if the plot should include the
observed data collected during the predicted census.}

\item{rangex}{\code{integer}-conformable vector of two values corresponding
to the minimum and maximum newmoonnumbers plotted.}

\item{start_newmoon}{\code{integer}-conformable newmoon number used as the
the minimum x value for the plot.}

\item{add_obs}{\code{logical} indicator if values observed during the 
-cast time should be added (default is \code{TRUE}).}

\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}. For input 
into \code{forecast_weather} and \code{forecast_ndvi}, \code{moons} 
should be trimmed by \code{\link{trim_moons_fcast}}.}

\item{covariates}{Class-\code{covariates} \code{data.frame} of 
historical covariate data generated by 
\code{\link{prep_hist_covariates}}.}

\item{hist_cov}{Historical covariate data table as a code{covariates}-class 
\code{data.frame}, returned from \code{\link{prep_hist_covariates}}.}

\item{new_forecast_covariates}{Class-\code{covariates} \code{data.frame} of
forecasted covariate data generated by 
\code{\link{forecast_covariates}}.}

\item{name}{\code{character}-valued name of the model (MUST match the 
function name).}

\item{lag}{\code{integer} (or integer \code{numeric}) lag time used for the
covariates or \code{NULL} if \code{covariates} is \code{FALSE}.}

\item{mod_covariates}{\code{logical} indicator for if the model requires 
covariates.}

\item{subs_names}{Either [1] names of additional subdirectories to add to
the vector set up by \code{type} or [2] an optional way to input all 
subdirectory names (requires \code{type = NULL}, which is the default).}

\item{subs_type}{\code{character} name for quick generation of subdirectory
vector. Presently only defined for \code{"portalcasting"}.}

\item{specific_sub}{\code{character}-value name of the specific 
subdirectory/subdirectories of interest.}

\item{extension}{\code{character} file extension (including the period).}

\item{local_path}{\code{character} file path within the \code{main} level 
of the portalcasting directory.}

\item{rodents}{Class-\code{rodents} \code{data.table} of rodent data.}

\item{model}{A singular \code{character} value for a \code{model} or 
\code{"Ensemble"} in certain situations.}

\item{set}{\code{characher} value of the type of model (currently only 
support for \code{"prefab"}). Use \code{NULL} to build a custom set
from scratch via \code{add}.}

\item{add}{\code{character} vector of name(s) of model(s) to add to the 
setup by \code{set}.}

\item{lead}{\code{integer}-conformable lead of the newmoon number used to
select the data plotted.}

\item{rodents_list}{A class-\code{rodents_list} \code{list} of two class-
\code{rodents} \code{data.frame}s, \code{all} (abundances on all plots)
and \code{controls} (abundances on control plots only).}

\item{future_moons}{A \code{moons}-class \code{data.frame} forward from 
the historic moons table (produced by \code{\link{prep_moons}}), as
generated by \code{\link[portalr]{get_future_moons}}.}

\item{data_name}{\code{character} representation of the data needed.
Current options include \code{"all"}, \code{"controls"},
\code{"covariates"}, \code{"moons"}, and \code{"metadata"}.}
}
\description{
Omnibus function to verify the validity (class, size, etc.) 
  of the inputs to the functions.
}
\examples{
\dontrun{

check_args()
}

}
