% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{check_to_skip}
\alias{check_to_skip}
\title{Check if the newmoon should be skipped}
\usage{
check_to_skip(options_cast)
}
\arguments{
\item{options_cast}{casting options list}
}
\value{
logical indicating if the cast should be skipped
}
\description{
For hindcasts, skip casting newmoons with incomplete surveys
}
