% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{check_to_skip}
\alias{check_to_skip}
\title{Check if the newmoon should be skipped during a set of casts}
\usage{
check_to_skip(main = ".", moons = prep_moons(main = main),
  end_moon = NULL, raw_path_data = "PortalData",
  raw_traps_file = "Rodents/Portal_rodent_trapping.csv",
  controls_r = rodents_controls(), quiet = FALSE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{raw_path_data}{\code{character} value indicating the folder path
to the data within the \code{raw} subdirectory but above the files. A 
standard portalcasting directory downloads the raw data files into 
\code{"raw\PortalData"}, so \code{raw_path_data = "PortalData"} (as
\code{"raw/"} is implied).}

\item{raw_traps_file}{\code{character} value indicating the path
to the trapping data file within \code{raw_path_data}. A standard 
portalcasting directory downloads the raw data files into 
\code{"raw\PortalData"}, so 
\code{raw_traps_file = "Rodents/Portal_rodent_trapping.csv"}.}

\item{controls_r}{Control \code{list} (from 
\code{\link{rodents_control}}) or \code{list} of control \code{list}s 
(from \code{\link{rodents_controls}}) specifying the structuring of the 
rodents tables. See \code{\link{rodents_control}} for details.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\value{
\code{logical} indicator of if the moon should be skipped.
}
\description{
Newmoons with incomplete surveys are skipped.
}
\examples{
 \donttest{
  setup_dir()
  check_to_skip()
 }

}
