% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{cleanup_dir}
\alias{cleanup_dir}
\title{Remove the temporary files in PortalData and tmp}
\usage{
cleanup_dir(to_cleanup = c("tmp", "PortalData"), tree = dirtree(),
  quiet = FALSE)
}
\arguments{
\item{to_cleanup}{\code{character} vector of subdirectory names to
cleanup after completion of casting. Defaults to the temp and PortalData
subdirectories: \code{c("tmp", "PortalData")}.}

\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\description{
Remove content in specified (by \code{options_dir$to_cleanup})
  subdirectories and the subdirectories themselves.
}
\examples{
\dontrun{

setup_dir()
portalcast()
cleanup_dir()
}

}
