% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{cleanup_dir}
\alias{cleanup_dir}
\title{Remove the temporary files in PortalData and tmp}
\usage{
cleanup_dir(tree = dirtree())
}
\arguments{
\item{tree}{directory tree}
}
\value{
nothing
}
\description{
Remove the PortalData and tmp subdirectories.
}
