% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{cleanup_dir}
\alias{cleanup_dir}
\title{Remove the temporary files in PortalData and tmp}
\usage{
cleanup_dir(options_all = all_options())
}
\arguments{
\item{options_all}{Class-\code{all_options} list containing all options 
available for controlling the set up and population of the directory (see 
\code{\link{all_options}}).}
}
\description{
Remove content in specified (by \code{options_dir$to_cleanup})
  subdirectories and the subdirectories themselves.
}
