% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clear_tmp}
\alias{clear_tmp}
\title{Remove Files from the tmp Subdirectory}
\usage{
clear_tmp(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or just tidy messages.}
}
\value{
\code{NULL} \code{\link[base]{invisible}}-ly, with the tmp subdirectory's files removed.
}
\description{
Clear the files from the tmp subdirectory.
}
\examples{
 \donttest{
  create_dir()
  clear_tmp
 }

}
