% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{clear_tmp}
\alias{clear_tmp}
\title{Clear tmp}
\usage{
clear_tmp(tree = dirtree())
}
\arguments{
\item{tree}{directory tree}
}
\value{
nothing
}
\description{
Remove all of the files from the tmp subdirectory specifically
}
