% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clear_tmp}
\alias{clear_tmp}
\title{Remove files from the tmp subdirectory}
\usage{
clear_tmp(main = ".", quiet = FALSE, cleanup = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{cleanup}{\code{logical} indicator if any files put into the tmp
subdirectory should be removed at the end of the process.}
}
\value{
\code{NULL}, with the tmp subdirectory's files removed.
}
\description{
Clear the files from the tmp subdirectory.
}
\examples{
 \donttest{
  create_dir()
  clear_tmp
 }

}
