% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{clear_tmp}
\alias{clear_tmp}
\title{Clean a tmp subdirectory}
\usage{
clear_tmp(tree = dirtree(), quiet = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\description{
Clean (\code{clean_tmp}) the tmp subdirectory specifically.
}
\details{
If the tmp subdirectory does not exist, it will be created if
  possible (if the basal directory exists).
}
