% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{column_conformer}
\alias{column_conformer}
\title{Conform column names for use internally}
\usage{
column_conformer(df = NULL, arg_checks = TRUE)
}
\arguments{
\item{df}{\code{data.frame} to have column names converted if needed.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{df} with column names converted if needed.
}
\description{
Existing cast files may have column names using "forecast"
 rather than the current generalized "cast". This function replaces the
 old with the new for internal use.
}
\examples{
 df <- data.frame(xforecastx = 1:10, fore = 2:11)
 column_conformer(df)

}
