% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{combine_casts}
\alias{combine_casts}
\title{Combine casts and append them to existing files}
\usage{
combine_casts(main = ".", moons = NULL, end_moon = NULL,
  cast_date = Sys.Date(), quiet = FALSE, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{list} of [1] \code{"casts"} (the casted abundances)
 and [2] \code{"all_model_aic"} (the model AIC values).
}
\description{
Combine all new casts (from the tmp subdirectory) into
 and append the results to the existing files.
}
