% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{combine_forecasts}
\alias{combine_forecasts}
\title{Combine Forecasts and Append to Existing Files}
\usage{
combine_forecasts(model_metadata, temp_dir, pred_dir)
}
\arguments{
\item{model_metadata}{model metadata list}

\item{temp_dir}{directory name for the temporary housing of predictions}

\item{pred_dir}{directory name where the saved model predictions reside}
}
\value{
list of [1] the forecasts and [2] the model AIC values
}
\description{
Combine all new forecasts (from the tmp directory), add 
  append the results to the existing files
}
