% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_hist_and_cast}
\alias{combine_hist_and_cast}
\title{Combine a historical table and a cast table}
\usage{
combine_hist_and_cast(hist_tab = NULL, cast_tab = NULL,
  winner = "hist")
}
\arguments{
\item{hist_tab, cast_tab}{A pair of \code{data.frame}s with the same columns
includig a code{date} column of \code{Date}s, which is used to align them.}

\item{winner}{\code{character} value either {"hist"} or \code{"cast"} to
decide who wins any ties. In the typical portalcasting space, this is 
kept at its default value throughout.}
}
\value{
\code{data.frame} combining \code{hist_tab} and \code{cast_tab}.
}
\description{
A simple utility for combining a table of historical data
 and a table of cast data that might need to be assigned to either one
 or the other.
}
\examples{
 hist_tab <- data.frame(date = Sys.Date(), x = 1:10)
 cast_tab <- data.frame(date = Sys.Date(), x = 101:110)
 combine_hist_and_cast(hist_tab, cast_tab, "hist") 
 combine_hist_and_cast(hist_tab, cast_tab, "cast")  

}
