% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{compile_aic_weights}
\alias{compile_aic_weights}
\title{Calculate model weights}
\usage{
compile_aic_weights(main = ".", arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{data.frame} of model weights.
}
\description{
Calculate AIC-based weights across the models for the full
  time series.
}
