% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{compile_aic_weights}
\alias{compile_aic_weights}
\title{Calculate model weights}
\usage{
compile_aic_weights(pred_dir)
}
\arguments{
\item{pred_dir}{\code{character} value of the path to the directory where 
the saved model predictions reside.}
}
\value{
\code{data.frame} of model weights.
}
\description{
Calculate AIC-based weights across the models for the full
  time series.
}
