% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{compile_aic_weights}
\alias{compile_aic_weights}
\title{Calculate Model Weights}
\usage{
compile_aic_weights(pred_dir)
}
\arguments{
\item{pred_dir}{directory name where the saved model predictions reside}
}
\value{
model weights
}
\description{
calculate akaike weights across the models
}
