% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{compile_aic_weights}
\alias{compile_aic_weights}
\title{Calculate model weights}
\usage{
compile_aic_weights(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}
}
\value{
\code{data.frame} of model weights.
}
\description{
Calculate AIC-based weights across the models for the full
  time series.
}
\examples{
 \donttest{
  setup_dir()
  compile_aic_weights()
 }

}
