% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_models.R
\name{covariate_models}
\alias{covariate_models}
\title{Create a covariate model list}
\usage{
covariate_models(type = "pevgarch")
}
\arguments{
\item{type}{type name for covariate models. Currently only "pevgarch" is
supported.}
}
\value{
list of covariate model structures
}
\description{
Convenience function for creating covariate model lists
}
