% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{covariate_models}
\alias{covariate_models}
\title{Create a covariate model list}
\usage{
covariate_models(model = "pevGARCH")
}
\arguments{
\item{model}{\code{character} name for covariate models. Currently only 
\code{"pevGARCH"} is supported.}
}
\value{
\code{list} of covariate model structures.
}
\description{
Convenience function for creating covariate model 
 \code{list}s.
}
\examples{
 covariate_models()

}
