% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{covariates_options}
\alias{covariates_options}
\title{Prepare the data options for covariates data}
\usage{
covariates_options(historical = TRUE, forecasts = TRUE, fdate = today(),
  yr = as.numeric(format(today(), "\%Y")), lead_time = 12, min_lag = 6,
  fcast_nms = NULL, nfcnm = 0, save = TRUE, filename = "covariates.csv")
}
\arguments{
\item{historical}{logical indicator whether or historical covariates are
to be included}

\item{forecasts}{logical indicator whether or forecasted covariates are
to be included}

\item{fdate}{date from which future is defined, typically today's date.}

\item{yr}{year of today}

\item{lead_time}{number of moons into the future the rodents are forecast}

\item{min_lag}{the minimum lag time used in any model}

\item{fcast_nms}{newmoon numbers to be forecast}

\item{nfcnm}{number of forecast newmoons}

\item{save}{logical if the data should be saved out}

\item{filename}{the name of the file for the saving}
}
\value{
control options list for covariates
}
\description{
Create a list of control options for the covariates data
}
