% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create}
\alias{create}
\title{Create a specific folder}
\usage{
create(path = NULL, dir_level = NULL, quiet = FALSE,
  verbose = FALSE, arg_checks = TRUE)
}
\arguments{
\item{path}{\code{character} vector of the folder paths.}

\item{dir_level}{\code{character} vector of the names of the levels in the
hierarchy at which the folders will exist or the folder names. For the
purposes of messaging and not inherently required to be non-\code{NULL}.}

\item{quiet}{\code{logical} indicator if messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out
all of the information or not (and thus just the tidy messages).}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr}
}
\value{
\code{NULL}.
}
\description{
If folders (specified by \code{path} and named by
 \code{level}) do not exist, they are created, with a message or not as
 indicated by \code{quiet}.
}
\examples{
 \donttest{
  main <- main_path()
  create(main, "main")
 }

}
