% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create}
\alias{create}
\title{Create a specific folder}
\usage{
create(path = NULL, level = NULL, quiet = FALSE)
}
\arguments{
\item{path}{\code{character} vector of the folder paths.}

\item{level}{\code{character} vector of the names of the levels in the
hierarchy at which the folders will exist or the folder names. For the
purposes of messaging and not inherently required to be non-\code{NULL}.}

\item{quiet}{\code{logical} indicator if messages should be quieted.}
}
\value{
\code{NULL}.
}
\description{
If folders (specified by \code{path} and named by
 \code{level}) do not exist, they are created, with a message or not as
 indicated by \code{quiet}.
}
\examples{
 main <- main_path()
 create(main, "main")

}
