% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create the structure of a portalcasting directory}
\usage{
create_dir(main = ".", quiet = FALSE)
}
\arguments{
\item{main}{\code{character} value defining the main component of the 
portalcasting directory tree. Default value (\code{"."}) 
puts the directory in the present location.}

\item{quiet}{\code{logical} indicator if messages should be quieted.}
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Creates the necessary folder structure for a portalcasting 
             directory as well as a YAML file that tracks the setup 
             configurations. Folder paths are created internally 
             such that the user only needs to input the main folder's 
             location components (a la \code{\link[base]{file.path}}).
             The subdirectories are hardcoded as \code{"casts"},
             \code{"data"}, \code{"fits"}, \code{"models"}, 
             and \code{"resources"}.
}
\examples{
 \donttest{

  create_dir("./portalcasting")

 }

}
