% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{create_dir}
\alias{create_dir}
\alias{create_base_dir}
\alias{create_main_dir}
\alias{create_sub_dirs}
\title{Create the structure of a portalcasting directory}
\usage{
create_dir(tree = dirtree(), quiet = FALSE)

create_base_dir(tree = dirtree(), quiet = FALSE)

create_main_dir(tree = dirtree(), quiet = FALSE)

create_sub_dirs(tree = dirtree(), quiet = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\description{
This suite of functions creates the necessary folder structure
 for a portalcasting directory. At each level of the hierarchy, the more
 basal folders are required to be present and the relevant folders at that
 level are created if not present. \cr \cr
 \code{create_dir}: Create a full portalcasting directory or 
 any missing components. \cr \cr
 \code{create_main_dir}: Create the main level of the 
 portalcasting directory. \cr \cr
 \code{create_sub_dirs}: Create the sub level folders of the 
 portalcasting directory. \cr \cr
 \code{create_sub_dir}: Create a specific subdirectory 
 folder if it does not already exist.
}
\examples{
  \dontrun{

  create_dir()
  create_base_dir()
  create_main_dir()
  create_sub_dirs()
  create_sub_dir()
  }

}
