% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{create_dir}
\alias{create_dir}
\title{Create the full structure of a portalcasting directory}
\usage{
create_dir(tree = dirtree(), quiet = FALSE)
}
\arguments{
\item{tree}{the name tree of the portalcasting directory}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
Nothing
}
\description{
Create a full portalcasting directory or any missing 
  components.
}
