% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{create_main_dir}
\alias{create_main_dir}
\title{Create the main directory folder of a portalcasting directory}
\usage{
create_main_dir(tree = dirtree(), quiet = FALSE)
}
\arguments{
\item{tree}{the name tree of the portalcasting directory}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
Nothing
}
\description{
Create the top-level portalcasting directory folder
}
