% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{create_sub_dirs}
\alias{create_sub_dirs}
\title{Create the sub directory folders of a portalcasting directory}
\usage{
create_sub_dirs(options_dir = dir_options())
}
\arguments{
\item{options_dir}{directory options}
}
\value{
Nothing
}
\description{
Create a the portalcasting directory subfolders
}
