% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dataout}
\alias{dataout}
\title{Save data out to a file and return it}
\usage{
dataout(df, tree = dirtree(), control = list())
}
\arguments{
\item{df}{\code{data.frame} table to be written out.}

\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{control}{\code{list} that includes control parameters.}
}
\value{
df (as input)
}
\description{
Save inputted data out to a data file if requested via an 
  options list (request is handled by the \code{save} element and the file
  name to use is handled by the \code{filename} element) and return it 
  to the console.
}
