% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dataout}
\alias{dataout}
\title{Save data out to a file and return it}
\usage{
dataout(df, options_out = moons_options())
}
\arguments{
\item{df}{\code{data.frame} table to be written out.}

\item{options_out}{an options list that includes a save element and 
a filename element, and potentially a class element.}
}
\value{
df (as input)
}
\description{
Save inputted data out to a data file if requested via an 
  options list (request is handled by the \code{save} element and the file
  name to use is handled by the \code{filename} element) and return it 
  to the console.
}
