% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{directory creation}
\alias{directory creation}
\alias{create_dir}
\alias{update_dir}
\alias{setup_dir}
\alias{setup_production}
\alias{setup_sandbox}
\title{Create the Structure of a Directory and Fill It with Content or Update an Existing Directory}
\usage{
create_dir(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

update_dir(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  new_datasets_controls = NULL,
  new_models_controls = NULL,
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

setup_dir(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  new_datasets_controls = NULL,
  new_models_controls = NULL,
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

setup_production(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  settings = production_settings(),
  quiet = FALSE,
  verbose = TRUE
)

setup_sandbox(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  new_datasets_controls = NULL,
  new_models_controls = NULL,
  settings = sandbox_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree. Default value (\code{"."}) roots the directory in the present location.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to produce all of the messages.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{datasets}{\code{character} vector of name(s) of rodent dataset(s) to be created.}

\item{new_datasets_controls}{Optional \code{list} of controls for new datasets. See \code{\link{datasets_controls}}. This argument is not available in \code{setup_production}.}

\item{new_models_controls}{Optional \code{list} of controls for new models. See \code{\link{models_controls}}. This argument is not available in \code{setup_production}.}
}
\value{
The \code{list} of directory settings \code{\link[base:invisible]{invisible}}-ly.
}
\description{
Instantiates the necessary folder structure for a directory, writes the setup configuration file, and fills the directory with content. \cr
Options for pre-defined setups include \code{setup_sandbox} for quick and flexible builds and \code{setup_production} for robust, rigid builds, as defined in \code{\link{directory_settings}}.
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "standard")
   main2 <- file.path(tempdir(), "sandbox")
   main3 <- file.path(tempdir(), "production")

   setup_dir(main = main1)
   setup_sandbox(main = main2)
   setup_production(main = main3)

   update_dir(main = main1)

   unlink(main1, recursive = TRUE)
   unlink(main2, recursive = TRUE)
   unlink(main3, recursive = TRUE)
}

}
\seealso{
Directory orchestration functions:
\code{\link{directory configuration file}},
\code{\link{directory paths}},
\code{\link{directory settings}}
}
\concept{orchestration}
