% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{directory creation}
\alias{directory creation}
\alias{create_dir}
\alias{update_dir}
\alias{setup_dir}
\alias{setup_production}
\alias{setup_sandbox}
\title{Create the Structure of a Directory and Fill It with Content or Update an Existing Directory}
\usage{
create_dir(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

update_dir(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  new_dataset_controls = NULL,
  new_model_controls = NULL,
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

setup_dir(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  new_dataset_controls = NULL,
  new_model_controls = NULL,
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

setup_production(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  settings = production_settings(),
  quiet = FALSE,
  verbose = TRUE
)

setup_sandbox(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  new_dataset_controls = NULL,
  new_model_controls = NULL,
  settings = sandbox_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree. Default value (\code{"."}) roots the directory in the present location.}

\item{settings}{\code{list} of controls for the directory, with defaults set in\code{\link{directory_settings}}.}

\item{quiet}{\code{logic} indicator if progress messages should be quieted.}

\item{verbose}{\code{logic} indicator of whether or not to produce all of the messages.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{datasets}{\code{character} vector of name(s) of rodent dataset(s) to be created.}

\item{new_dataset_controls}{Optional \code{list} of controls for new datasets. See \code{\link{dataset_controls}}. This argument is not available in \code{setup_production}.}

\item{new_model_controls}{Optional \code{list} of controls for new models. See \code{\link{model_controls}}. This argument is not available in \code{setup_production}.}
}
\value{
The \code{list} of directory settings \code{\link[base:invisible]{base::invisible}}-ly.
}
\description{
Instantiates the necessary folder structure for a directory, writes the setup configuration file, and fills the directory with content.
}
