% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{directory creation}
\alias{directory creation}
\alias{create_dir}
\title{Create the Structure of a Forecasting Directory}
\usage{
create_dir(main = ".", settings = directory_settings(), quiet = FALSE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree. Default value (\code{"."}) puts the forecasting directory in the present locations. Nesting the forecasting directory in a folder can be done by simply adding to the \code{main} input (see \code{Examples}).}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information (and thus just the tidy messages).}
}
\value{
The \code{list} of directory settings \code{\link[base]{invisible}}-ly.
}
\description{
Instantiates the necessary folder structure for a forecasting directory and writes a YAML file that tracks the setup configurations (see \code{\link{write_directory_config}}).
}
\examples{
\donttest{
 create_dir()
}

}
