% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dir.R
\name{directory filling}
\alias{directory filling}
\alias{fill_dir}
\alias{fill_data}
\alias{fill_resources}
\alias{fill_forecasts}
\alias{fill_fits}
\alias{fill_models}
\title{Fill a Portalcasting Directory with Basic Components}
\usage{
fill_dir(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

fill_data(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

fill_resources(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

fill_forecasts(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

fill_fits(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

fill_models(
  main = ".",
  models = prefab_models(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{datasets}{\code{character} vector of name(s) of rodent dataset(s) to be created.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or just the tidy messages.}
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Fill the directory with components including \enumerate{
  \item{Resources (\code{\link{fill_resources}}): \itemize{
    \item{raw data (\code{\link[portalr]{download_observations}})}
    \item{directory archive (\code{\link{download_archive}})}
    \item{climate forecasts (\code{\link{download_climate_forecasts}})}}}
  \item{Output: \itemize{
    \item{forecasts (\code{\link{fill_forecasts}})}
    \item{model fits (\code{\link{fill_fits}})}}}
  \item{Data (\code{\link{fill_data}}):  \itemize{
    \item{rodent datasets (\code{\link{prepare_rodents}})}
    \item{temporal (lunar) data (\code{\link{prepare_moons}})}
    \item{covariates (\code{\link{prepare_covariates}})}
    \item{metadata (\code{\link{prepare_metadata}})}}}
  \item{Models (\code{\link{fill_models}})}
 }
}
\examples{
 \donttest{

  create_dir("./portalcasting")
  fill_dir("./portalcasting")

  create_dir("./pc")
  fill_resources("./pc")
  fill_forecasts("./pc")
  fill_fits("./pc")
  fill_models("./pc")
 }

}
