% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dir.R
\name{directory filling}
\alias{directory filling}
\alias{fill_dir}
\alias{fill_data}
\alias{fill_resources}
\alias{fill_forecasts}
\alias{fill_fits}
\alias{fill_models}
\title{Fill a Portalcasting Directory with Basic Components}
\usage{
fill_dir(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  new_dataset_controls = NULL,
  new_model_controls = NULL
)

fill_data(
  main = ".",
  datasets = prefab_datasets(),
  new_dataset_controls = NULL
)

fill_resources(main = ".")

fill_forecasts(main = ".")

fill_fits(main = ".")

fill_models(main = ".", models = prefab_models(), new_model_controls = NULL)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{models}{\code{character} vector of name(s) of model(s) to include. Defaults to \code{\link{prefab_models}}. If controls are provided in \code{new_model_controls}, the model still needs to be named here to be included.}

\item{datasets}{\code{character} vector of name(s) of rodent dataset(s) to be created. Defaults to \code{\link{prefab_datasets}}. If controls are provided in \code{new_dataset_controls}, the dataset still needs to be named here to be included.}

\item{new_dataset_controls}{Optional named \code{list} of controls for new datasets. See \code{\link{dataset_controls}}.}

\item{new_model_controls}{Optional named \code{list} of controls for new models. See \code{\link{model_controls}}.}
}
\value{
\code{NULL}, \code{\link[base:invisible]{invisible}}-ly.
}
\description{
Fill the directory with components including:
\itemize{
\item Resources (\code{\link{fill_resources}})
\itemize{
\item raw data (\code{\link[portalr:download_observations]{download_observations}})
\item directory archive (\code{\link{download_archive}})
\item climate forecasts (\code{\link{download_climate_forecasts}})
}
\item Output
\itemize{
\item forecasts (\code{\link{fill_forecasts}})
\item model fits (\code{\link{fill_fits}})
}
\item Data (\code{\link{fill_data}})
\itemize{
\item rodent datasets (\code{\link{prepare_rodents}})
\item temporal (lunar) data (\code{\link{prepare_newmoons}})
\item covariates (\code{\link{prepare_covariates}})
\item metadata (\code{\link{prepare_metadata}})
}
\item Models (\code{\link{fill_models}})
\itemize{
\item model controls (\code{\link{write_model_controls}})
\item model scripts (if needed) (\code{\link{write_model_scripts}})
Additional, new models and datasets can be added to the directory at filling using the optional arguments \code{new_model_controls} and \code{new_dataset_controls}, but the model or dataset must still be listed in its respective main argument, as well.
}
}
}
