% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{directory settings}
\alias{directory settings}
\alias{directory_settings}
\alias{directory_files}
\alias{directory_subdirectories}
\alias{directory_resources}
\alias{production_settings}
\alias{sandbox_settings}
\title{Create a Directory Settings List}
\usage{
directory_settings(
  files = directory_files(),
  subdirectories = directory_subdirectories(),
  resources = directory_resources(),
  save = TRUE,
  overwrite = FALSE,
  unzip_pause = 30,
  download_timeout = getOption("timeout")
)

directory_files(
  directory_configuration = "directory_configuration.yaml",
  newmoons = "newmoon_dates.csv",
  covariates = "covariates.csv",
  historical_covariates = "historical_covariates.csv",
  forecast_covariates = "forecast_covariates.csv",
  dataset_controls = "dataset_controls.yaml",
  model_controls = "model_controls.yaml",
  cast_evaluations = "cast_evaluations.csv",
  forecast_metadata = "casts_metadata.csv",
  metadata = "metadata.yaml"
)

directory_subdirectories(
  forecasts = "forecasts",
  fits = "fits",
  models = "models",
  resources = "resources",
  data = "data"
)

directory_resources(
  PortalData = list(source = "github", version = "latest"),
  portalPredictions = list(source = "github", version = "latest"),
  climate_forecast = list(source = "NMME", version = as.character(Sys.Date()), data =
    c("tasmin", "tasmean", "tasmax", "pr"))
)

production_settings(download_timeout = max(getOption("timeout"), 600))

sandbox_settings()
}
\arguments{
\item{files}{\code{list} of \code{character} names of standard files, see \code{\link{directory_files}}.}

\item{subdirectories}{\code{list} of \code{character} names of standard subdirectories, see \code{\link{directory_subdirectories}}.}

\item{resources}{\code{character} name for the resources subdirectory.}

\item{save}{\code{logical} indicator controlling if the output should be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not existing files or folders (such as the archive) should be over-written if an up-to-date copy exists (most users should leave as \code{FALSE}).}

\item{unzip_pause}{Positive \code{integer} or integer \code{numeric} seconds for pausing during steps around unzipping that require time delay.}

\item{download_timeout}{Positive \code{integer} or integer \code{numeric} seconds for timeout on downloads. Temporarily overrides the \code{"timeout"} option in \code{\link[base]{options}}.}

\item{directory_configuration}{\code{character} name for the directory configuration YAML.}

\item{newmoons}{\code{character} name for the lunar data csv.}

\item{covariates}{\code{character} name for the combined historical and forecast covariates csv.}

\item{historical_covariates}{\code{character} name for the historical covariates csv.}

\item{forecast_covariates}{\code{character} name for the forecast covariates csv.}

\item{dataset_controls}{\code{character} name for the YAML of dataset control list(s).}

\item{model_controls}{\code{character} name for the YAML of model controls list(s).}

\item{cast_evaluations}{\code{character} name for the forecast evaluations csv.}

\item{forecast_metadata}{\code{character} name for the forecast metadata csv.}

\item{metadata}{\code{character} name for the Forecast metadata YAML.}

\item{forecasts}{\code{character} name for the forecasts subdirectory.}

\item{fits}{\code{character} name for the fits subdirectory.}

\item{models}{\code{character} name for the models subdirectory.}

\item{data}{\code{character} name for the data subdirectory.}

\item{PortalData}{\code{list} of \code{source} and \code{version} elements of \code{character} values for the Portal Data download. Default values retrieve the latest data from github}

\item{portalPredictions}{\code{list} of \code{source} and \code{version} elements of \code{character} values for the archive download. Default values point to github, but \code{version = NULL} indicates no download.}

\item{climate_forecast}{\code{list} of \code{source}, \code{version}, and \code{data} elements of \code{character} values for the climate forecasts download. Default values retrieve the current day's forecast of min, mean, and max temperature and precipitation from the Northwest Knowledge Network's North American Multi-Model Ensemble (NMME) climate forecasts.}
}
\value{
Named \code{list} of settings for the directory (for \code{directory_settings}) or \code{list} of settings components (for \code{directory_files}, \code{directory_subdirectories}, and \code{directory_resources}).
}
\description{
Most users will not want or need to change the directory folders and file names, but it is helpful to have them be flexible for certain circumstances, and this function gathers them into a list for pipeline functionality.
}
