% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_output.R
\name{directory_config}
\alias{directory_config}
\alias{write_config}
\title{Create, update, and read the directory configuration file}
\usage{
write_config(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}
}
\value{
\code{write_directory_config} and \code{update_directory_config}
 both write out the \code{dir_config.yaml} file and return \code{NULL}. 
 \cr \cr
 \code{read_directory_config}: \code{list} of directory configurations.
}
\description{
The configuration file is a special file within the 
             portalcasting directory that has its own set of functions. 
             \cr \cr
             \code{write_config} creates the file. It is (and should only 
             be) called from within \code{\link{create_dir}}, as
             it captures information about the compute environment used to
             instantiate the directory. \cr \cr
}
