% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_output.R
\name{directory_config}
\alias{directory_config}
\alias{write_config}
\alias{update_config}
\alias{read_config}
\title{Create, update, and read the directory configuration file}
\usage{
write_config(main = ".", filename = "config.yaml")

update_config(
  main = ".",
  filename_config = "dir_config.yaml",
  downloads_versions = NULL,
  quiet = FALSE,
  arg_checks = TRUE
)

read_config(
  main = ".",
  filename_config = "dir_config.yaml",
  quiet = FALSE,
  arg_checks = TRUE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{filename}{\code{character} value of the the directory config YAML.}

\item{downloads_versions}{\code{character} vector returned from 
\code{fill_raw} of the successfully downloaded versions of
\code{downloads}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\value{
\code{write_directory_config} and \code{update_directory_config}
 both write out the \code{dir_config.yaml} file and return \code{NULL}. 
 \cr \cr
 \code{read_directory_config}: \code{list} of directory configurations.
}
\description{
The configuration file is a special file within the 
             portalcasting directory that has its own set of functions. 
             \cr \cr
             \code{write_config} creates the file. It is (and should only 
             be) called from within \code{\link{create_dir}}, as
             it captures information about the compute environment used to
             instantiate the directory. \cr \cr

 \code{update_directory_config} adds key components to the 
 \code{dir_config.yaml} file; presently only adding the versions of the
 downloaded data from within \code{\link{fill_raw}} (the only place it is
 and should be called from. \cr \cr
 \code{read_directory_config} reads     - s of the \code{dir_config.yaml} 
 file into the R session.
}
