% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{directory_settings}
\alias{directory_settings}
\title{Create a Directory Settings List}
\usage{
directory_settings(
  directory_config_file = "dir_config.yaml",
  moons_file = "moon_dates.csv",
  covariates_file = "covariates.csv",
  historical_covariates_file = "historical_covariates.csv",
  forecast_covariates_file = "forecast_covariates.csv",
  dataset_controls_file = "dataset_controls.yaml",
  model_controls_file = "model_controls.yaml",
  cast_evaluations_file = "cast_evaluations.csv",
  metadata_file = "metadata.yaml",
  subdirectories = list(forecasts = "casts", fits = "fits", models = "models",
    resources = "raw", data = "data"),
  PortalData = list(source = "github", version = "latest"),
  portalPredictions = list(source = "github", version = NULL),
  climate_forecast = list(source = "NMME", version = as.character(Sys.Date()), data =
    c("tasmin", "tasmean", "tasmax", "pr")),
  save = TRUE,
  overwrite = TRUE,
  unzip_pause = 30,
  download_timeout = getOption("timeout")
)
}
\arguments{
\item{directory_config_file}{\code{character} value of the path to the directory config YAML.}

\item{moons_file}{\code{character} name of the file for saving the moons data.}

\item{covariates_file, historical_covariates_file, forecast_covariates_file}{\code{character} name of the files for saving the covariate data (combined historical and forecast) and each component.}

\item{dataset_controls_file}{\code{character} name of the file for saving the rodent dataset controls.}

\item{model_controls_file}{\code{character} name of the file for saving the model controls.}

\item{cast_evaluations_file}{\code{character} name of the file for saving the cast evaluations.}

\item{metadata_file}{\code{character} name of the file for saving the forecast metadata.}

\item{subdirectories}{\code{character} vector of the subdirectory names. Default includes \code{raw}, \code{data}, \code{models}, \code{fits}, and \code{casts}.}

\item{PortalData}{\code{list} with \code{source} and \code{version} elements that are \code{character} values for the source and version of the Portal Data to download. Default values retrieve the latest data from github. \cr \cr
See \code{\link[portalr]{download_observations}}.}

\item{portalPredictions}{\code{list} with \code{source} and \code{version} elements that are \code{character} values for the source and version of the archive to download. Default values point to github, but \code{version = NULL} indicates no download. \cr \cr 
See \code{\link{download_archive}}.}

\item{climate_forecast}{\code{list} with \code{source} and \code{version} elements that are \code{character} values for the source and version of the climate forecasts to download. Default values retrieve the current day's forecast from the Northwest Knowledge Network's North American Multi-Model Ensemble (NMME) climate forecasts. \cr \cr 
See \code{\link{download_climate_forecasts}}.}

\item{save}{\code{logical} indicator controlling if the output should be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing files should be updated (most users should leave as \code{TRUE}).}

\item{unzip_pause}{Positive \code{integer} or integer \code{numeric} seconds for pausing during steps around unzipping that require time delay.}

\item{download_timeout}{Positive \code{integer} or integer \code{numeric} seconds for timeout on downloads. Temporarily overrides the \code{"timeout"} option in \code{\link[base]{options}}.}
}
\value{
Named \code{list} of settings for the directory.
}
\description{
Most users will not want or need to change the directory folders and file names, but it is helpful to have them be flexible for certain circumstances, and this function gathers them into a list for pipeline functionality.
}
