% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{dirtree}
\alias{dirtree}
\title{Define the names of the directory tree for a forecasting directory}
\usage{
dirtree(base = ".", main = "", subs = subdirs())
}
\arguments{
\item{base}{name of the base directory where the directory should exist}

\item{main}{name of the forecasting directory}

\item{subs}{names of the subdirectories}
}
\value{
named list of representing the directory tree
}
\description{
A tidy way to pass the full tree structure down.
}
