% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download climate forecasts}
\alias{download climate forecasts}
\alias{download_climate_forecasts}
\title{Download Climate Forecasts}
\usage{
download_climate_forecasts(
  main = ".",
  resources_sub = "resources",
  version = Sys.Date(),
  source = "NMME",
  data = c("tasmin", "tasmean", "tasmax", "pr"),
  quiet = FALSE,
  verbose = FALSE,
  overwrite = FALSE,
  timeout = getOption("timeout")
)
}
\arguments{
\item{main}{\code{character} value defining the main component of the portalcasting directory tree.}

\item{resources_sub}{\code{character} value defining the resources subdirectory of the portalcasting directory tree.}

\item{version}{\code{Date}-coercible start of the climate cast. See \code{\link{NMME_urls}} (used as \code{start}).}

\item{source}{\code{character} indicator of the source for the download. Only \code{"NMME"} presently available.}

\item{data}{\code{character} vector of the data to be collected. Currently min, mean, and max temperatures and precipitation (\code{c("tasmin", "tasmean", "tasmax", "pr")}).}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be generated.}

\item{overwrite}{\code{logical} indicator of whether or not existing files or folders (such as the archive) should be over-written if an up-to-date copy exists (most users should leave as \code{FALSE}).}

\item{timeout}{Positive \code{integer} or integer \code{numeric} seconds for timeout on downloads. Temporarily overrides the \code{"timeout"} option in \code{\link[base]{options}}.}
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Downloads climate forecasts, presently only from NMME (see \code{\link{NMME_urls}}) into the \code{<main>/raw} sub.
}
