% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_archive}
\alias{download_archive}
\title{Download the Portal Predictions Repository Archive}
\usage{
download_archive(
  main = ".",
  version = "latest",
  source = "github",
  quiet = FALSE,
  verbose = FALSE,
  pause = 30,
  timeout = getOption("timeout")
)
}
\arguments{
\item{main}{\code{character} value defining the main component of the portalcasting directory tree.}

\item{version}{\code{character} version of the data to download. Default \code{"latest"} downloads the most recent (by date published). \code{NULL} means no download.}

\item{source}{\code{character} indicator of the source for the download. Either \code{"github"} (defualt) or \code{"github"}.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be printed.}

\item{pause}{Positive \code{integer} or integer \code{numeric} seconds for pausing during steps around unzipping that require time delayment.}

\item{timeout}{Positive \code{integer} or integer \code{numeric} seconds for timeout on downloads. Temporarily overrides the \code{"timeout"} option in \code{\link[base]{options}}.}
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Downloads a specific \code{version} of the Portal Predictions repository from either GitHub or Zenodo (based on \code{source}) into the \code{<main>/raw} sub.
}
\note{
There are two calls to \code{link[base]{Sys.sleep}} for \code{pause} seconds each to allow for the file unzipping, copying, and such to catch up.
}
