% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_climate_forecasts}
\alias{download_climate_forecasts}
\title{Download Climate Forecasts}
\usage{
download_climate_forecasts(
  main = ".",
  version = Sys.Date(),
  source = "NMME",
  quiet = FALSE,
  verbose = FALSE,
  timeout = getOption("timeout")
)
}
\arguments{
\item{main}{\code{character} value defining the main component of the 
portalcasting directory tree.}

\item{version}{\code{Date}-coercible start of the climate cast. See
\code{\link{NMME_urls}} (used as \code{start}).}

\item{source}{\code{character} indicator of the source for the download. 
Only \code{"NMME"} presently available.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{timeout}{\code{numeric} value passed to \code{\link[base]{options}}
to dictate the download timeout limit.}
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Downloads climate forecasts, presently only from NMME 
             (see \code{\link{NMME_urls}}) into the
              \code{<main>/resources} sub.
}
