% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictions.R
\name{download_predictions}
\alias{download_predictions}
\title{Download the predictions files from the portalPredictions repo}
\usage{
download_predictions(tree = dirtree(), download = TRUE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{download}{\code{logical} indicator of whether the download should 
actually happen. Should be \code{TRUE} except for testing purposes.}
}
\description{
Download the files into the predictions subdirectory in the
  the portalcasting directory tree from the 
  \href{https://github.com/weecology/portalPredictions/tree/master/predictions}{main repository}.
  Due to the volume of files, \code{\link{download.file}} is forced 
  \code{quiet = TRUE}.
}
