% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictions.R
\name{download_predictions}
\alias{download_predictions}
\title{Download the predictions files from the Portal Predictions repo}
\usage{
download_predictions(tree = dirtree(), download = TRUE,
  quiet = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{download}{\code{logical} indicator of whether the download should 
actually happen. Should be \code{TRUE} except for testing purposes.}

\item{quiet}{\code{logical} indicator of whether the download messages 
are returned to the console.}
}
\description{
Download the files into the predictions subdirectory in the
  the portalcasting directory tree from the 
  \href{https://bit.ly/2IFuPd7}{main repository}.
}
\examples{
\dontrun{

setup_dir()
download_predictions()
}

}
