% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictions.R
\name{download_predictions}
\alias{download_predictions}
\title{Download the predictions files from the Portal Predictions repo}
\usage{
download_predictions(tree = dirtree(), quiet = FALSE,
  control = list())
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator of whether the download messages 
are returned to the console.}

\item{control}{A \code{list} of arguments to control the 
filling of the predictions subdirectory. See 
\code{\link{predictions_control}}. Arguments not specified assume default
values.}
}
\description{
Download the files into the predictions subdirectory in the
  the portalcasting directory tree from 
  \href{https://zenodo.org/record/3333770}{Zenodo} (default) or from the 
  \href{https://bit.ly/2Lo2xnQ}{portalPredictions repo}.
}
\examples{
\dontrun{

setup_dir()
download_predictions()
}

}
