% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{enforce_rodents_options}
\alias{enforce_rodents_options}
\title{Enforce the specific data options for the two rodent data types}
\usage{
enforce_rodents_options(data_options, type = "all")
}
\arguments{
\item{data_options}{existing control options list for rodents}

\item{type}{"all" or "controls"}
}
\value{
control options list for rodents
}
\description{
Force a list of control options for the rodent data based on
  the rodent data type
}
