% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{enforce_rodents_options}
\alias{enforce_rodents_options}
\title{Enforce the specific options for a given rodent data type}
\usage{
enforce_rodents_options(options_rodents = rodents_options(),
  tmnt_type = NULL)
}
\arguments{
\item{options_rodents}{Class-\code{rodents_options} \code{list} of 
settings controlling the rodents data creation. See 
\code{\link{rodents_options}}.}

\item{tmnt_type}{\code{character}-value representation of treatment type 
(\code{"all"} or \code{"controls"}) used to enforce certain elements of 
the \code{options_rodents} \code{list}. Alternatively, use \code{NULL}
to not change the options at all.}
}
\value{
Updated \code{options_rodents} input, still a class-
  \code{rodents_options} \code{list}.
}
\description{
Enforce a list of control options for the rodent data 
  (\code{rodents_options}, see \code{\link{options_rodents}}) based on
  the requested data type. Specifically, the \code{tmnt_type}, 
  \code{level}, \code{length}, \code{plots}, \code{output}, 
  \code{treatment}, and \code{filename} elements are set by the selection.
}
\details{
If \code{tmnt_type = "all"}, the elements are \code{tmnt_type} as
  \code{"all"}, \code{level} as \code{"Site"}, \code{length} as 
  \code{"all"}, \code{plots} as \code{"all"}, \code{output} as 
  \code{"abundance"}, \code{treatment} as \code{NULL}, and \code{filename} 
  as \code{"all.csv"}. \cr \cr
  If \code{tmnt_type = "controls"}, the elements are \code{tmnt_type} as
  \code{"controls"}, \code{level} as \code{"Treatment"}, \code{length} as 
  \code{"Longterm"}, \code{plots} as \code{"Longterm"}, \code{output} as 
  \code{"abundance"}, \code{treatment} as \code{controls}, and 
  \code{filename} as \code{"controls.csv"},
}
