% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESSS.R
\name{ESSS}
\alias{ESSS}
\title{Exponential Smoothing State Space model for Portal Predictions}
\usage{
ESSS(abundances, metadata, level = "All")
}
\arguments{
\item{abundances}{table of rodent abundances and time measures}

\item{metadata}{model metadata list}

\item{level}{name of the type of plots included ("All" or "Controls")}
}
\value{
list of forecast and aic tables
}
\description{
Model "ESSS" is a flexible exponential smoothing state space 
 model fit using \code{ets} in the \code{forecast} package (Hyndman 
 \emph{et al}. 2018) with the possibility of multiplicative trends.  

 Because the seasonality and sampling occurring with different frequencies,
 which \code{ets} cannot accommodate, seasonal models are not included.
}
\references{
Hyndman R., Bergmeir C., Caceres G., Chhay L., O'Hara-Wild M., Petropoulos
 F., Razbash S., Wang E., and Yasmeen F. 2018. forecast: Forecasting 
 functions for time series and linear models. 
 \href{http://pkg.robjhyndman.com/forecast}{R package version 8.3}.
}
