% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esss.R
\name{esss}
\alias{esss}
\title{Exponential Smoothing State Space model for Portal Predictions}
\usage{
esss(abundances, metadata, level, CL = 0.9)
}
\arguments{
\item{abundances}{table of rodent abundances and time measures}

\item{metadata}{model metadata list}

\item{level}{name of the type of plots included ("All" or "Controls")}

\item{CL}{confidence interval level used for forecast envelope}
}
\value{
list of forecast and aic tables
}
\description{
Model "ESSS" is a flexible exponential smoothing state space 
 model fit using \code{ets} in \code{forecast} package (Hyndman et al.
 2018) with the possibility of multiplicative trends.  

 Because the seasonality and sampling occurring with different frequencies,
 which \code{ets} cannot accommodate, seasonal models are not included.
}
\references{
Hyndman R., Bergmeir C., Caceres G., Chhay L., O'Hara-Wild M., Petropoulos
 F., Razbash S., Wang E., and Yasmeen F. 2018. forecast: Forecasting 
 functions for time series and linear models. 
 \href{http://pkg.robjhyndman.com/forecast}{R package version 8.3}.
}
