% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fcast0}
\alias{fcast0}
\title{Zero-abundance forecast}
\usage{
fcast0(nfcnm, pred_name = "pred")
}
\arguments{
\item{nfcnm}{\code{integer} number of forecast newmoons.}

\item{pred_name}{\code{character} name for the predictor column (to match
variable model output names).}
}
\value{
Two-element \code{list} of means and interval values for a 
  0-abundance forecast to be used as a filler when a model fails or there 
  is no non-0 historical abundance.
}
\description{
Create a 0-abundance forecast for fill-in usage when a model 
  fails or there is no non-0 historical abundance.
}
