% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{fcast_covariates}
\alias{fcast_covariates}
\title{Build a Covariate Forecast for Model Predictions}
\usage{
fcast_covariates(covariate_data, moons, model_metadata)
}
\arguments{
\item{covariate_data}{historical covariate data table}

\item{moons}{moon data}

\item{model_metadata}{metadata list for the forecast models}
}
\value{
a data.frame with 12 new moons of weather values
}
\description{
combining weather and ndvi forecasts
}
