% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{file_ext}
\alias{file_ext}
\title{Determine a File's Extension}
\usage{
file_ext(path, sep_char = ".")
}
\arguments{
\item{path}{\code{character} value of the file path possibly with an extension.}

\item{sep_char}{\code{character} value of the separator that delineates the extension from the file path. \cr 
Generally, this will be \code{"."}, but for some API URLs, the extension is actually a query component, so the separator may sometimes need to be \code{"="}.}
}
\value{
\code{character} value of the extension (\code{file_ext}).
}
\description{
Based on the separating character, determine the file extension.
}
\examples{
 file_ext("home/folders.with.dots/stuff/ok.csv")
 file_ext(NMME_urls()[[1]])
 file_ext(NMME_urls()[[1]], "=")

}
