% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{file_path}
\alias{file_path}
\title{Determine the path for a file in the forecasting directory}
\usage{
file_path(tree = dirtree(), local_path = NULL)
}
\arguments{
\item{tree}{\code{dirtree}-class list.}

\item{local_path}{\code{character} file path within the \code{main} level 
of the portalcasting directory.}
}
\value{
The normalized path of the specified file (see 
  \code{\link{normalizePath}}).
}
\description{
Return the normalized path for a specific file within the
  directory.
}
