% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{file_paths}
\alias{file_paths}
\title{Determine the path for a file or files in the forecasting directory}
\usage{
file_paths(main = ".", local_paths = NULL, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{local_paths}{\code{character} file path(s) within the \code{main}
level of the portalcasting directory.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
The normalized path(s) of the specified file(s) (see 
  \code{\link{normalizePath}}) or \code{NULL} if \code{local_paths = NULL}.
}
\description{
Return the normalized path for a specific file or files 
  within the directory.
}
\examples{
 file_paths(local_paths = "PortalData/Rodents/Portal_rodent_species.csv")

}
