% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dir.R
\name{fill_casts_message}
\alias{fill_casts_message}
\title{Create the final message for filling casts}
\usage{
fill_casts_message(files = NULL, movedTF = NULL, verbose = FALSE,
  arg_checks = TRUE)
}
\arguments{
\item{files}{\code{character} vector of the base file names of the files
that may or may not have been moved.}

\item{movedTF}{\code{logical} vector indicating if each of the files in 
\code{files} has been moved or not.}

\item{verbose}{\code{logical} indicator of whether or not to print out
all of the files and whether they were moved or not.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
A \code{character} value of the message to be printed (if desired).
}
\description{
Create the final message to be used in 
 \code{\link{fill_casts}} that relays the number, and optionally the
 specific names, of the files moved.
}
\examples{
 fill_casts_message("xx.csv", TRUE)
 fill_casts_message(c("xx.csv", "yy.R"), c(TRUE, FALSE), 
                          verbose = TRUE)

}
