% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_data}
\alias{fill_data}
\title{Populate the for-use data of a portalcasting directory}
\usage{
fill_data(tree = dirtree(), all_options = all_data_options(),
  quiet = FALSE)
}
\arguments{
\item{tree}{the name tree of the portalcasting directory}

\item{all_options}{list of lists of data options}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
Nothing
}
\description{
Populate the for-use data folder
}
