% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_data}
\alias{fill_data}
\title{Populate the for-use data of a portalcasting directory}
\usage{
fill_data(tree = dirtree(), start = 217, quiet = FALSE)
}
\arguments{
\item{tree}{the name tree of the portalcasting directory}

\item{start}{newmoon number of the first sample to be included. Default 
value is \code{217}, corresponding to 1995-01-01.}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
Nothing
}
\description{
Populate the for-use data folder
}
