% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dir.R
\name{fill_dir}
\alias{fill_dir}
\alias{fill_raw}
\title{Fill a forecasting directory with basic components.}
\usage{
fill_dir(downloads = NULL, main = ".", quiet = FALSE,
  cleanup = TRUE)

fill_raw(downloads = NULL, main = ".", quiet = FALSE,
  cleanup = TRUE)
}
\arguments{
\item{downloads}{\code{list} or \code{list} of \code{list}s containing
inputs to \code{\link{download}} for each download to be put into the 
raw subdirectory.}

\item{main}{\code{character} value of the name of the main component of
the directory tree. See \code{Details}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{cleanup}{\code{logical} indicator if any files put into the tmp
subdirectory should be removed at the end of the process.}
}
\value{
All \code{fill_} functions return \code{NULL}.
}
\description{
Fill the forecasting directory with basic components.
}
\details{
Arguments input directly here take precedence over those in the 
 \code{downloads} \code{list}.
}
\examples{
 \donttest{
  data(PortalPredictions)
  create_dir()
  fill_raw(PortalPredictions$downloads)
 }

}
