% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_dir}
\alias{fill_dir}
\alias{fill_PortalData}
\alias{fill_data}
\alias{fill_predictions}
\alias{fill_models}
\title{Populate the components of a portalcasting directory}
\usage{
fill_dir(models = NULL, tree = dirtree(), quiet = FALSE,
  PortalData_control = list(), data_control = list(),
  predictions_control = list(), models_control = list())

fill_PortalData(tree = dirtree(), quiet = FALSE, control = list())

fill_data(tree = dirtree(), quiet = FALSE, control = list())

fill_predictions(tree = dirtree(), quiet = FALSE, control = list())

fill_models(models = NULL, tree = dirtree(), quiet = FALSE,
  control = list())
}
\arguments{
\item{models}{\code{character} vector of the names of models to include in
the pipeline. Defaults to \code{NULL}, which retrieves the model names 
from \code{models_control}, so the models returned by the default
settings of \code{\link{model_names}}.}

\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{PortalData_control}{A \code{list} of arguments to control the filling
of the PortalData subdirectory. See \code{\link{PortalData_control}}. 
 Arguments not specified assume default values.}

\item{data_control}{A \code{list} of arguments to control the filling of
the data subdirectory. See \code{\link{data_control}}. 
Arguments not specified assume default values.}

\item{predictions_control}{A \code{list} of arguments to control the 
filling of the predictions subdirectory. See 
\code{\link{predictions_control}}. Arguments not specified assume default
values.}

\item{models_control}{A \code{list} of arguments to control the 
filling of the models subdirectory. See \code{\link{models_control}}.
Arguments not specified assume default values.}

\item{control}{A \code{list} of controls specific to the function.}
}
\description{
Fill each of the components of the directory. \cr \cr
  \code{fill_dir}: populate the files of an existing 
  portalcasting directory structure in full. (Future flexibility will 
  allow for selective population of components.) \cr \cr
  \code{fill_portalData}: populate the components of the PortalData 
  folder. \cr \cr
 \code{fill_data}: populate the for-use data folder in the 
 portalcasting directory, including the historical covariates, trapping 
 table, moons, rodents, covariates, and metadata files. \cr \cr
 \code{fill_predictions}: populate the predictions folder with 
 existing predictions housed on the main portalPredictions repo site.
 \cr \cr
 \code{fill_models}: Populate the model folder with specified 
 model scripts to be run.
}
\examples{
\dontrun{

create_dir()
fill_dir()
fill_PortalData()
fill_data()
fill_predictions()
fill_models()
}

}
