% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_dir}
\alias{fill_dir}
\title{Populate the base components of a portalcasting directory}
\usage{
fill_dir(tree = dirtree(), model = models(), quiet = FALSE)
}
\arguments{
\item{tree}{the name tree of the portalcasting directory}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
Nothing
}
\description{
Populate an existing portalcasting directory structure in full
   or any missing components.
}
