% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_dir}
\alias{fill_dir}
\alias{fill_PortalData}
\alias{fill_data}
\alias{fill_predictions}
\alias{fill_models}
\title{Populate the components of a portalcasting directory}
\usage{
fill_dir(options_all = all_options())

fill_PortalData(options_PortalData = PortalData_options())

fill_data(options_data = data_options())

fill_predictions(options_predictions = predictions_options())

fill_models(options_models = models_options())
}
\arguments{
\item{options_all}{Class-\code{all_options} list containing all options 
available for controlling the set up and population of the directory (see 
\code{\link{all_options}}).}

\item{options_PortalData}{Class-\code{PortalData_options} list containing 
available for controlling the set up and population of the PortalData
folder in the portalcasting directory (see 
\code{\link{PortalData_options}}).}

\item{options_data}{Class-\code{data_options} list containing available
for controlling the set up and population of the data folder in the 
portalcasting directory (see \code{\link{data_options}}).}

\item{options_predictions}{Class-\code{predictions_options} list containing 
available for controlling the set up and population of the predictions
folder in the portalcasting directory (see 
\code{\link{predictions_options}}).}

\item{options_models}{Class-\code{models_options} list containing 
available for controlling the set up and population of the models
folder in the portalcasting directory (see 
\code{\link{models_options}}).}
}
\description{
\code{fill_dir}: Populate the files of an existing 
   portalcasting directory structure in full. (Future flexibility will 
   allow for selective population of components.)

\code{fill_portalData}: Populate the components of the 
  PortalData folder.

\code{fill_data}: Populate the for-use data folder in the 
  portalcasting directory, including the historical covariates, trapping 
  table, moons, rodents, covariates, and metadata files.

\code{fill_predictions}: Populate the predictions folder with 
  existing predictions housed on the main portalPredictions repo site.

\code{fill_models}: Populate the model folder with specified 
  model scripts to be run.
}
\examples{
\dontrun{

create_dir()
fill_dir()
fill_PortalData()
fill_data()
fill_predictions()
fill_models()
}

}
