% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dir.R
\name{fill_directory}
\alias{fill_directory}
\alias{fill_dir}
\alias{fill_data}
\alias{fill_raw}
\alias{fill_casts}
\alias{fill_fits}
\alias{fill_models}
\title{Fill a Portalcasting Directory with Basic Components}
\usage{
fill_dir(
  main = ".",
  models = prefab_models(),
  datasets = prefab_rodent_datasets(),
  lead_time = 12,
  cast_date = Sys.Date(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

fill_data(
  main = ".",
  models = prefab_models(),
  datasets = prefab_rodent_datasets(),
  lead_time = 12,
  cast_date = Sys.Date(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

fill_raw(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

fill_casts(main = ".", quiet = FALSE, verbose = FALSE)

fill_fits(main = ".", quiet = FALSE, verbose = FALSE)

fill_models(
  main = ".",
  models = prefab_models(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{datasets}{\code{list} of datasets to be created using \code{\link{do.call}} on the defined functions.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the number of timesteps forward a cast will cover.}

\item{cast_date}{\code{Date} of the cast, typically today's date (set using \code{\link{Sys.Date}}).}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or not (and thus just the tidy messages).}
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Fill the directory with components including (optionally) the raw data (\code{\link[portalr]{download_observations}}), portalPredictions archive (\code{\link{download_archive}}), climate forecasts (\code{\link{download_climate_forecasts}}), models (\code{\link{write_model}}).
}
\examples{
 \donttest{

  create_dir("./portalcasting")
  fill_dir("./portalcasting")

  create_dir("./pc")
  fill_raw("./pc")
  fill_casts("./pc")
  fill_fits("./pc")
  fill_models("./pc")
 }

}
