% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dir.R
\name{fill_directory}
\alias{fill_directory}
\alias{fill_dir}
\alias{fill_data}
\alias{fill_models}
\alias{fill_casts}
\alias{fill_fits}
\alias{fill_raw}
\title{Fill a portalcasting directory with basic components.}
\usage{
fill_dir(
  main = ".",
  PortalData_version = "latest",
  PortalData_source = "gitub",
  portalPredictions_version = NULL,
  portalPredictions_source = "github",
  quiet = FALSE
)

fill_data(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  controls_rodents = rodents_controls(),
  control_climate_dl = climate_dl_control(),
  control_files = files_control(),
  downloads = zenodo_downloads(c("1215988", "833438")),
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

fill_models(
  main = ".",
  models = prefab_models(),
  controls_model = NULL,
  control_files = files_control(),
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

fill_casts(
  main = ".",
  control_files = files_control(),
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

fill_fits(
  main = ".",
  control_files = files_control(),
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

fill_raw(
  main = ".",
  downloads = zenodo_downloads(c("1215988", "833438")),
  only_if_missing = FALSE,
  control_files = files_control(),
  quiet = FALSE,
  arg_checks = TRUE
)
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Fill the directory with foundational components.
}
\examples{
 \donttest{

  create_dir("./portalcasting")
  fill_dir("./portalcasting")

 }

}
