% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dir.R
\name{fill_directory}
\alias{fill_directory}
\alias{fill_dir}
\alias{fill_models}
\alias{fill_resources}
\alias{fill_casts}
\alias{fill_fits}
\alias{fill_data}
\title{Fill a Portalcasting Directory with Basic Components}
\usage{
fill_dir(
  main = ".",
  PortalData_version = "latest",
  PortalData_source = "gitub",
  portalPredictions_version = NULL,
  portalPredictions_source = "github",
  climate_forecast_version = Sys.Date(),
  climate_forecast_source = "NMME",
  models = prefab_models(),
  quiet = FALSE,
  verbose = FALSE
)

fill_models(
  main = ".",
  models = prefab_models(),
  quiet = FALSE,
  verbose = FALSE
)

fill_resources(
  main = ".",
  PortalData_version = "latest",
  PortalData_source = "gitub",
  portalPredictions_version = NULL,
  portalPredictions_source = "github",
  climate_forecast_version = Sys.Date(),
  climate_forecast_source = "NMME",
  quiet = FALSE,
  verbose = FALSE
)

fill_casts(main = ".", quiet = FALSE, verbose = FALSE)

fill_fits(main = ".", quiet = FALSE, verbose = FALSE)

fill_data(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  controls_rodents = rodents_controls(),
  control_climate_dl = climate_dl_control(),
  control_files = files_control(),
  downloads = zenodo_downloads(c("1215988", "833438")),
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Fill the directory with components including (optionally) 
             the raw data (\code{\link[portalr]{download_observations}}), 
             portalPredictions archive (\code{\link{download_archive}}), 
             climate forecasts (\code{\link{download_climate_forecasts}}), 
             models (\code{\link{write_model}}), 
             and model-ready data.
}
\examples{
 \donttest{

  create_dir("./portalcasting")
  fill_dir("./portalcasting")

  create_dir("./pc")
  fill_resources("./pc")
  fill_casts("./pc")
  fill_fits("./pc")
  fill_models("./pc")
 }

}
