% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dir.R
\name{fill_directory}
\alias{fill_directory}
\alias{fill_dir}
\alias{fill_data}
\alias{fill_resources}
\alias{fill_output}
\alias{fill_models}
\title{Fill a Portalcasting Directory with Basic Components}
\usage{
fill_dir(
  main = ".",
  models = prefab_models(),
  datasets = prefab_rodent_datasets(),
  lead_time = 12,
  origin = Sys.Date(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

fill_data(
  main = ".",
  models = prefab_models(),
  datasets = prefab_rodent_datasets(),
  lead_time = 12,
  origin = Sys.Date(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

fill_resources(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

fill_output(main = ".", quiet = FALSE, verbose = FALSE)

fill_models(
  main = ".",
  models = prefab_models(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{datasets}{\code{list} of datasets to be created using \code{\link{do.call}} on the defined functions.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the number of timesteps forward a cast will cover.}

\item{origin}{\code{Date} forecast origin, typically today's date (set using \code{\link{Sys.Date}}).}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or not (and thus just the tidy messages).}
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Fill the directory with components including (optionally) the resources data (\code{\link[portalr]{download_observations}}), portalPredictions archive (\code{\link{download_archive}}), climate forecasts (\code{\link{download_climate_forecasts}}), models (\code{\link{write_model}}).
}
