% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_models}
\alias{fill_models}
\title{Populate the model scripts of a portalcasting directory}
\usage{
fill_models(tree = dirtree(), model = models(), quiet = FALSE)
}
\arguments{
\item{tree}{the name tree of the portalcasting directory}

\item{model}{names of model scripts to include}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
Nothing
}
\description{
Populate the model folder with specified scripts to be run
}
