% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{fill_PortalData}
\alias{fill_PortalData}
\title{Populate the raw data of a portalcasting directory}
\usage{
fill_PortalData(options_PortalData = PortalData_options())
}
\arguments{
\item{options_PortalData}{PortalData subdirectory options}
}
\value{
Nothing
}
\description{
Populate the raw data folder
}
