% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{forecast_covariates}
\alias{forecast_covariates}
\alias{forecast_ndvi}
\alias{forecast_weather}
\title{Forecast covariates for forecasting or hindcasting}
\usage{
forecast_covariates(covariate_data, moons,
  options_covariates = covariates_options())

forecast_ndvi(covariate_data, moons, options_covariates)

forecast_weather(moons = prep_moons(),
  options_covariates = covariates_options())
}
\arguments{
\item{covariate_data}{\code{data.frame} of historical covariate data 
generated by \code{\link{prep_hist_covariates}}.}

\item{moons}{\code{data.frame} containing the historic and future newmoons,
as produced by \code{\link{prep_moons}}.}

\item{options_covariates}{Class-\code{covariates_options} list of 
options for the covariate data. See \code{\link{covariates_options}}.}
}
\value{
\code{forecast_covariates}: \code{data.frame} with needed 
  forecasted covariates.

\code{forecast_ndvi}: \code{data.frame} of needed forecasted NDVI
  values.

\code{forecast_weather}: \code{data.frame} of needed forecasted
  temperature and precipiation values.
}
\description{
\code{forecast_covariates}: Forecast covariates (NDVI and 
  weather) based on the forecast type (\code{"forecast"} or 
  \code{"hindcast"}) as indicated by \code{options_covariates$cast_type}.

\code{forecast_ndvi}: Wraps around 
  \code{\link[portalr]{fcast_ndvi}} to forecast specific NDVI values
  as requested by a portalcasting directory based on existing NDVI values
  (from \code{covariate_data}) and \code{moons} data, and controlled by
  \code{options_covariates} (see \code{\link{covariates_options}}).

\code{forecast_weather}: Wraps around 
  \code{\link{get_climate_forecasts}} to forecast weather covariates 
  (min, max, and mean of temperatures and total precipitation for each
  newmoon) as requested by a portalcasting directory and controlled through
  \code{options_covariates} (see \code{\link{covariates_options}}).
}
