% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{forecast_covariates}
\alias{forecast_covariates}
\title{Build a covariate forecast for model predictions}
\usage{
forecast_covariates(tree, covariate_data, moons,
  data_options = covariates_options())
}
\arguments{
\item{tree}{directory tree}

\item{covariate_data}{historical covariate data table}

\item{moons}{moon data}

\item{data_options}{covariate data options}
}
\value{
a data.frame with needed forecasted covariates
}
\description{
combining weather and ndvi forecasts
}
