% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{forecast_weather}
\alias{forecast_weather}
\title{Build a weather forecast for model predictions}
\usage{
forecast_weather(moons = prep_moons(),
  options_covariates = covariates_options())
}
\arguments{
\item{moons}{moon data}

\item{options_covariates}{covariate data options}
}
\value{
a data.frame with needed foecasted weather values
}
\description{
creating forecasts
}
