% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{format_moons}
\alias{format_moons}
\title{Format the moon data}
\usage{
format_moons(moons)
}
\arguments{
\item{moons}{Moons data table as a code{moons}-class \code{data.frame}.}
}
\value{
Moons data table as a code{moons}-class \code{data.frame} with
  force-formatted columns.
}
\description{
Enforce the \code{year} and \code{month} columns and encode 
  \code{newmoondate} as a \code{\link{Date}}.
}
