% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{get_climate_forecasts}
\alias{get_climate_forecasts}
\title{Download downscaled weather forecasts for Portal}
\usage{
get_climate_forecasts(moons = prep_moons(),
  options_covariates = covariates_options())
}
\arguments{
\item{moons}{Class-\code{moons} type of a \code{data.frame} containing the
moons and date data. See \code{\link{prep_moons}}.}

\item{options_covariates}{Class-\code{covariates_options} options 
\code{list}. See \code{\link{covariates_options}}.}
}
\value{
Class-\code{climate_forecast} \code{data.frame} with columns 
  of \code{newmoonnumber}, \code{mintemp}, \code{maxtemp}, \code{meantemp},
  and \code{precipitation}. Temperatures are in C, precipitation is in mm.
}
\description{
Lunar cycle-based climate forecasts obtained from 
  \href{https://climate.northwestknowledge.net/RangelandForecast/download.php}{Northwest Knowledge
  Network (NKN) at the University of Idaho}
  and downscaled to Portal, AZ (31.9555, -109.0744). The downscaled 
  forecasts are based on the 
  \href{http://www.cpc.ncep.noaa.gov/products/NMME/}{North American 
  Multi-Model Ensemble}, and specifically the ENSMEAN (ensemble mean) 
  model.
}
