% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{get_climate_forecasts}
\alias{get_climate_forecasts}
\title{Download downscaled weather data}
\usage{
get_climate_forecasts(lead_time = 6, moons)
}
\arguments{
\item{lead_time}{the newmoons into the future to obtain forecasts. Max of 7}

\item{moons}{newmoon data table}
}
\value{
a data.frame with precipitation(mm), temperature(C), year, and
  month. Temperature is the mean temperature for the month, while 
  precipitation is the total forecasted precip.
}
\description{
ENSMEAN (ensemble mean) obtained from 
  https://climate.northwestknowledge.net/RangelandForecast/download.php
  and downscaled to Portal, AZ (31.9555, -109.0744)
}
