% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ifnull}
\alias{ifnull}
\title{Replace if NULL}
\usage{
ifnull(x = NULL, alt = NULL)
}
\arguments{
\item{x}{Focal input.}

\item{alt}{Alternative value.}
}
\value{
\code{x} if not \code{NULL}, \code{alt} otherwise.
}
\description{
If the focal input is \code{NULL}, replace it with 
  alternative.
}
\examples{
 ifnull(NULL, 123)
 ifnull(TRUE, 123)
 ifnull(FALSE, 123)

}
