% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{interpolate_abundance}
\alias{interpolate_abundance}
\title{Interpolate missing rodent data}
\usage{
interpolate_abundance(rodents_tab)
}
\arguments{
\item{rodents_tab}{\code{data.frame} of rodents data with a 
\code{newmoonnumber} column.}
}
\value{
\code{data.frame} data table of interpolation-inclusive counts
 for each species and total and with the columns trimmed to just 
 the species, total, and newmoonnumber (\code{newmoon}).
}
\description{
Interpolation of missing data in the rodent abundance data 
 set using \code{\link[forecast]{na.interp}}. Each species is individually 
 linearly interpolated, then the total number of rodents is calculated 
 from the sum of the individual species.
}
