% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{interpolate_abundance}
\alias{interpolate_abundance}
\title{Interpolate Missing Rodent Data}
\usage{
interpolate_abundance(abundance)
}
\arguments{
\item{abundance}{data table with new moon column}
}
\value{
data table of interpolation-inclusive counts for each species and 
 total
}
\description{
Interpolation of missing data in the rodent abundance data 
  set. Each species is individually linearly interpolated, then the total 
  number of rodents is calculated from the sum of the individual species.
}
