% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{is.spcol}
\alias{is.spcol}
\title{Which columns are species columns}
\usage{
is.spcol(x, spp_names = rodent_spp())
}
\arguments{
\item{x}{data table of interest}

\item{spp_names}{species name abbreviations}
}
\value{
logical vector corresponding to the columns in x
}
\description{
Check column names to verify if they're species
}
