% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{is_sp_col}
\alias{is_sp_col}
\title{Determine if columns in a table are species columns}
\usage{
is_sp_col(rodents_tab, species = all_species())
}
\arguments{
\item{rodents_tab}{\code{data.frame} of columns to be checked.}

\item{species}{\code{character} vector indicating species names to 
use in determining if columns are species columns. Defaults to
\code{\link{all_species}}.}
}
\value{
\code{logical} vector indicating if each column is a species' 
 column or not.
}
\description{
Given a table, returns a \code{logical} vector indicating
 if each column is a species' column or not.
}
\examples{
 \donttest{
  create_dir()
  fill_raw()
  rodents_tab <- prep_rodents_table()
  is_sp_col(rodents_tab)
 }

}
