% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_jags.R
\name{jags_ss}
\alias{jags_ss}
\title{Run a single-species JAGS model}
\usage{
jags_ss(main = ".", data_set = "all",
  control_files = files_control(), control_runjags = runjags_control(),
  jags_model = NULL, monitor = NULL, inits = NULL, lag = NA,
  quiet = FALSE, verbose = FALSE, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{data_set}{\code{character} value name of the rodent data set, such as 
(\code{"all"} or \code{"controls"}).}

\item{control_files}{\code{list} of names of the folders and files within
the sub directories and saving strategies (save, overwrite, append, etc.).
Generally shouldn't need to be edited. See \code{\link{files_control}}.}

\item{control_runjags}{\code{list} of arguments passed to 
\code{\link[runjags]{run.jags}} via \code{\link{runjags_control}}}

\item{jags_model}{\code{character} value of the model. See \code{Examples}.}

\item{monitor}{\code{character} vector of the names of the variables to
monitor. See \code{Examples} and \code{\link[runjags]{run.jags}}.}

\item{inits}{\code{function}, \code{character} vector, or \code{list} of
values (including random number generation components) for initializing
runs of the model. See \code{Examples} and 
\code{\link[runjags]{run.jags}}.}

\item{lag}{\code{integer} (or integer \code{numeric}) of the lag time to
use for the covariates.}

\item{quiet}{\code{logical} value indicating if the function should be 
quiet.}

\item{verbose}{\code{logical} indicator of whether or not to print out
all of the information or not (and thus just the tidy messages).}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}.}
}
\value{
\code{list} of [1] model metadata \code{list} (\code{"metadata"}), 
 [2] cast summary \code{data.frame} (\code{"cast_tab"}),
 [3] \code{list} of model fit objects (\code{"model_fits"}), and 
 [4] \code{list} of model cast objects (\code{"model_casts"}).
}
\description{
Provides an API to \code{\link[runjags]{run.jags}} for
 single-species portalcasting models that requires the user only provide
 the JAGS model, initializer, monitor, and control list.
}
\examples{
 \donttest{
  setup_dir()
  monitor <- c("mu", "tau")
  inits <- function(chain = chain){
    rngs <- c("base::Wichmann-Hill", "base::Marsaglia-Multicarry",
              "base::Super-Duper", "base::Mersenne-Twister")
    list(.RNG.name = sample(rngs, 1),
         .RNG.seed = sample(1:1e+06, 1),
          mu = rnorm(1, 0, 1),
          tau = rgamma(1, shape = 0.1, rate = 0.1))
  }
  jags_model <- "model {  
    # priors
    mu ~ dnorm(log(1), 0.25); 
    tau ~ dgamma(0.1,0.1); 
  
    # initial state
    X[1] <- mu;
    pred_count[1] <- exp(X[1]);
    count[1] ~ dpois(exp(X[1])) T(0,ntraps[1]);

    # through time
    for(i in 2:N) {
      # Process model
      predX[i] <- X[i-1];
      checkX[i] ~ dnorm(predX[i], tau); 
      X[i] <- min(c(checkX[i], log(ntraps[i] + 1))); 
      pred_count[i] <- exp(X[i]);
  
      # observation model
      count[i] ~ dpois(exp(X[i])) T(0, ntraps[i]); 
    }
  }"

  jags_ss(jags_model = jags_model, monitor = monitor, inits = inits)
 }


}
\references{
Denwood, M. J. 2016. runjags: an R package providing interface
 utilities, model templates, parallel computing methods and additional
 distributions for MCMC models in JAGS. Journal of Statistical
 Software, 71:9. 
 \href{https://www.jstatsoft.org/article/view/v071i09}{URL}. 
 
 Plummer, M. 2003. JAGS: A program for analysis of Bayesian graphical 
 models using Gibbs Sampling. Proceedings of the 3rd International 
 \href{https://bit.ly/33aQ37Y}{URL}.
}
