% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{lag_covariates}
\alias{lag_covariates}
\title{Lag covariate data}
\usage{
lag_covariates(covariates, lag, tail = FALSE)
}
\arguments{
\item{covariates}{\code{data.frame} of covariate data to be lagged.}

\item{lag}{\code{integer} lag between rodent census and covariate data, in
new moons.}

\item{tail}{\code{logical} indicator if the data lagged to the tail end 
should be retained.}
}
\value{
\code{data.frame} with a \code{newmoonnumber} column reflecting
 the lag.
}
\description{
Lag the covariate data together based on the new moons
}
