% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{lag_covariates}
\alias{lag_covariates}
\title{Lag covariate data}
\usage{
lag_covariates(covariates, lag, tail = FALSE)
}
\arguments{
\item{covariates}{Class \code{covariates} \code{data.frame} of covariate
data to be lagged.}

\item{lag}{\code{integer} lag between rodent census and covariate data, in
new moons.}

\item{tail}{\code{logical} indicator if the data lagged to the tail end 
should be retained.}
}
\value{
\code{covariates} \code{data.frame} with a \code{newmoonnumber} 
  columnn reflecting the lag.
}
\description{
Lag the covariate data together based on the new moons
}
