% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{lag_data}
\alias{lag_data}
\title{Lag covariate data}
\usage{
lag_data(data, lag, tail = FALSE)
}
\arguments{
\item{data}{\code{data.frame} of covariate data to be lagged.}

\item{lag}{\code{integer} lag between rodent census and covariate data, in
new moons.}

\item{tail}{\code{logical} indicator if the data lagged to the tail end 
should be retained.}
}
\value{
\code{data} \code{data.frame} \code{newmoonnumber} columnn now 
  reflecting the lag.
}
\description{
Lag the weather data together based on the new moons
}
