% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{last_census}
\alias{last_census}
\title{Determine the most recent data collection}
\usage{
last_census(main = ".", control_files = files_control(), arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{control_files}{\code{list} of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See \code{\link{files_control}}.}

\item{arg_checks}{\code{logical} value of if the arguments should be checked using standard protocols via \code{\link{check_args}}. The default (\code{arg_checks = TRUE}) ensures that all inputs are formatted correctly and provides directed error messages if not.}
}
\value{
\code{Date} of the last census.
}
\description{
Determine the most recent census.
}
\examples{
 \donttest{
  setup_dir()
  last_census()
 }

}
