% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{main_path}
\alias{main_path}
\title{Determine the path for the main directory}
\usage{
main_path(tree = dirtree())
}
\arguments{
\item{tree}{the name tree of the forecasting directory}
}
\value{
the path of the main folder in the directory tree
}
\description{
Return the path for the main directory
}
