% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{make_model_combinations}
\alias{make_model_combinations}
\title{Determine All Requested Combinations of Model, Dataset, and Species to Cast}
\usage{
make_model_combinations(main = ".", settings = directory_settings())
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}
}
\value{
\code{character} vector of the path(s) of the R script file(s) to be run.
}
\description{
Translate model controls into a \code{data.frame} of model, dataset, and species columns, with a row for each combination.
}
