% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{measure_cast_error}
\alias{measure_cast_error}
\title{Summarize a table of -casts to -cast-level errors}
\usage{
measure_cast_error(casts, min_observed = 1)
}
\arguments{
\item{casts}{Class \code{casts} \code{data.frame} of requested fore- or 
hindcasts with observed values and errors included.}

\item{min_observed}{\code{integer} value for the minimum number of observed
values needed for a -cast to be retained in the output table. Default is
\code{1}, which returns all -casts with any observations. To include all
-casts (even those without any evaluations), set to \code{0}.}
}
\value{
\code{data.frame} of errors summarized to the -cast-level.
}
\description{
Summarize error across all of the data points within each 
  forecast or hindcast in a set of -casts.
}
