% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{measure_cast_level_error}
\alias{measure_cast_level_error}
\title{Measure error/fit metrics for forecasts}
\usage{
measure_cast_level_error(cast_tab = NULL, arg_checks = TRUE)
}
\arguments{
\item{cast_tab}{A \code{data.frame} of a cast's output. See 
\code{\link{read_cast_tab}}.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
\code{data.frame} of metrics for each \code{cast_id}.
}
\description{
Summarize the cast-level errors or fits to the observations. 
 \cr 
 Presently included are root mean square error and coverage.
}
\examples{
 \donttest{
  setup_dir()
  portalcast(models = c("AutoArima", "NaiveArima"), end_moons = 515:520)
  cast_tab <- read_cast_tab(cast_id = 1)
  cast_tab <- add_lead_to_cast_tab(cast_tab = cast_tab)
  cast_tab <- add_obs_to_cast_tab(cast_tab = cast_tab)
  cast_tab <- add_err_to_cast_tab(cast_tab = cast_tab)
  cast_tab <- add_covered_to_cast_tab(cast_tab = cast_tab)
  measure_cast_level_error(cast_tab = cast_tab)
}

}
