% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{model_control}
\alias{model_control}
\title{Create a model control list}
\usage{
model_control(name = NULL, covariates = FALSE, lag = NA,
  update = TRUE)
}
\arguments{
\item{name}{\code{character} value of the model name.}

\item{covariates}{\code{logical} indicator of if the model needs
covariates. Defaults to \code{FALSE}.}

\item{lag}{\code{integer}-conformable value of the lag to use with the
covariates (defaults to \code{NA} as no covariates are used).}

\item{update}{\code{logical} indicator of whether or not the predictions
should be updated.}
}
\value{
\code{list} of model metadata (presently: name, covariates, lag
  and update).
}
\description{
Creates a \code{list} of options for writing a specific model  
  script.
}
\examples{
  models_control()

}
