% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{model_control}
\alias{model_control}
\title{Create a control list for a model}
\usage{
model_control(
  name = "model",
  data_sets = prefab_data_sets(interpolate = FALSE),
  covariatesTF = FALSE,
  lag = NA,
  arg_checks = TRUE
)
}
\arguments{
\item{name}{\code{character} value of the name of the model.}

\item{data_sets}{\code{character} vector of the rodent data set names
that the model is applied to. Defaults to all non-interpolated data sets.}

\item{covariatesTF}{\code{logical} indicator for if the model requires 
covariates.}

\item{lag}{\code{integer} (or integer \code{numeric}) lag time used for the
covariates or \code{NULL} if \code{covariatesTF} is \code{FALSE}.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
Named \code{list} of model's script-generating controls, 
 for input as part of \code{controls_m} in \code{\link{write_model}}.
}
\description{
Provides a ready-to-use template for the 
 controls \code{list} used to define the specific arguments for a user's
 novel model, setting the formal arguments to the basic default 
 values.
}
