% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{model_names}
\alias{model_names}
\alias{prefab_models}
\alias{wEnsemble_models}
\title{Provide the names of models}
\usage{
model_names(models = NULL, set = NULL)

prefab_models()

wEnsemble_models()
}
\arguments{
\item{models}{\code{character} vector of name(s) of model(s) to add to the 
set created by \code{set}.}

\item{set}{\code{character} value of the model set(s) to include. Default
value is \code{NULL} which allows for full customization with 
\code{models}. Currently there is only support for the
only support for \code{"prefab"} (AutoArima, ESSS, nbGARCH, nbsGARCH,
and pevGARCH) and \code{"wEnsemble"} (the prefab models plus the basic
ensemble, a specialized case).}
}
\value{
\code{character} vector of model names.
}
\description{
Create a \code{character} vector of the names of the models
 to be included. \cr \cr
 \code{prefab_models} provides the names for the base pre-loaded models
 (\code{c("AutoArima", "ESSS", "nbGARCH", "nbsGARCH", "pevGARCH")}). 
 \cr \cr
 \code{wEnsemble_models} provides the names for the base pre-loaded models
 plus the ensemble.
}
\examples{
 model_names(c("model1", "model2"))
 model_names(c("model1", "model2"), "prefab")
 prefab_models()
 wEnsemble_models()

}
