% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{model_names}
\alias{model_names}
\title{Provide the names of models}
\usage{
model_names(add = NULL, model_set = "prefab")
}
\arguments{
\item{add}{\code{character} vector of name(s) of model(s) to add to the 
setup by \code{ model_set}.}

\item{model_set}{\code{characher} value of the type of model (currently 
only support for \code{"prefab"}). Use \code{NULL} to build a custom set
from scratch via \code{add}.}
}
\value{
\code{models}-class \code{character} vector of model names.
}
\description{
Based on a \code{set}, returns a character vectory of model
  names to be included. Currently only support for \code{set = "prefab"}.
}
