% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{model_names}
\alias{model_names}
\title{Provide the names of models}
\usage{
model_names(names = prefab, set = NULL)
}
\arguments{
\item{names}{\code{character} vector of name(s) of model(s) to add to the 
set created by \code{set}. Defaults to \code{prefab}, which is 
defined internally as
\code{c("AutoArima", "ESSS", "nbGARCH", "nbsGARCH", "pevGARCH")}.}

\item{set}{\code{character} value of the model set(s) to include. Default
value is \code{NULL} which allows for full customization with 
\code{names}. Currently there is only support for the
only support for \code{"prefab"} (AutoArima, ESSS, nbGARCH, nbsGARCH,
and pevGARCH) and \code{"wEnsemble"} (the prefab models plus the basic
ensemble, a specialized case).}
}
\value{
\code{character} vector of model names.
}
\description{
Based on model \code{names} or a \code{set}, returns a 
  character vector of model names to be included.
}
