% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_options.R
\name{model_options}
\alias{model_options}
\alias{AutoArima_options}
\alias{ESSS_options}
\alias{nbGARCH_options}
\alias{pevGARCH_options}
\title{Prepare options for a portalcasting model}
\usage{
model_options(tree = dirtree(), name = "AutoArima", covariates = FALSE,
  lag = NULL, quiet = FALSE)

AutoArima_options(tree = dirtree(), name = "AutoArima",
  covariates = FALSE, lag = NULL, quiet = FALSE)

ESSS_options(tree = dirtree(), name = "ESSS", covariates = FALSE,
  lag = NULL, quiet = FALSE)

nbGARCH_options(tree = dirtree(), name = "nbGARCH", covariates = FALSE,
  lag = NULL, quiet = FALSE)

pevGARCH_options(tree = dirtree(), name = "pevGARCH", covariates = TRUE,
  lag = 6, quiet = FALSE)
}
\arguments{
\item{tree}{directory tree}

\item{name}{character name of the model (matches function name)}

\item{covariates}{does the model require covariates}

\item{lag}{lag time used for the covariates (\code{NULL} if 
\code{covariates} is false)}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
\code{model_options}: a list of model options

\code{AutoArima_options}: a list of settings controlling the 
  AutoArima model creation

\code{ESSS_options}: a list of settings controlling the 
  ESSS model creation

\code{nbGARCH_options}: a list of settings controlling the 
  nbGARCH model creation

\code{pevGARCH_options}: a list of settings controlling the 
  pevGARCH model creation
}
\description{
Suite of functions used to generate model-writing options 
  lists: \cr \cr \code{model_options} is a basic template

\code{AutoArima_options} creates a list of control options for
  the AutoArima model

\code{ESSS_options} creates a list of control options for
  the ESSS model

\code{nbGARCH_options} creates a list of control options for
  the nbGARCH model

\code{pevGARCH_options} creates a list of control options for
  the pevGARCH model
}
