% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{model_path}
\alias{model_path}
\title{Determine the path for a model in the model sub directory}
\usage{
model_path(tree = dirtree(), model = NULL, extension = ".R")
}
\arguments{
\item{tree}{the name tree of the forecasting directory}

\item{model}{the name of the specific model}

\item{extension}{file extension}
}
\value{
the path of specific model script
}
\description{
Return the path for a specific model
}
