% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{model_paths}
\alias{model_paths}
\title{Determine the path for one or more models in the model sub directory}
\usage{
model_paths(main = ".", models = NULL, extension = ".R",
  arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{models}{\code{character} name of the specific model(s).}

\item{extension}{\code{character} file extension (including the period).}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
The normalized path of the specified model script (see 
  \code{\link{normalizePath}}) or \code{NULL} if \code{models = NULL}.
}
\description{
Return the normalized path(s) for a specific model or models.
}
\examples{
 model_paths(models = "AutoArima")

}
