% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{model_paths}
\alias{model_paths}
\title{Determine the path for one or more models in the model sub directory}
\usage{
model_paths(tree = dirtree(), models = NULL, extension = ".R")
}
\arguments{
\item{tree}{\code{dirtree}-class list.}

\item{models}{\code{character} name of the specific model(s).}

\item{extension}{\code{character} file extension (including the period).}
}
\value{
The normalized path of the specified model script (see 
  \code{\link{normalizePath}}).
}
\description{
Return the normalized path(s) for a specific model or models.
}
\examples{
\dontrun{

setup_dir()
model_paths(models = "AutoArima")
}

}
