% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{model_template}
\alias{model_template}
\title{Create the text to be written out to a script for a model}
\usage{
model_template(options_model = model_options())
}
\arguments{
\item{options_model}{list of options for the specific model}
}
\value{
Nothing
}
\description{
Create the text for a model script to be housed in the model 
  directory
}
