% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_script_generation.R
\name{models}
\alias{models}
\title{Provide the names of models}
\usage{
models(type = "portalcasting")
}
\arguments{
\item{type}{type of model options (currently only support for 
"portalcasting")}
}
\value{
vector of model names
}
\description{
Based on a \code{type}, returns a character vectory of model
  names to be included. Currently only support for 
  \code{type = "portalcasting"}.
}
