% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{models_to_cast}
\alias{models_to_cast}
\title{Generate the file path(s) to the script(s) of the model(s) forecast
  or hindcast with}
\usage{
models_to_cast(models = NULL, tree = dirtree(),
  models_control = list())
}
\arguments{
\item{models}{\code{character} vector of the names of models to include in
the pipeline. Defaults to \code{NULL}, which retrieves the model names 
from \code{models_control}, so the models returned by the default
settings of \code{\link{model_names}}.}

\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{models_control}{A \code{list} of arguments to control the 
filling of the models subdirectory. See \code{\link{models_control}}.
Arguments not specified assume default values.}
}
\value{
\code{character} vector of the path(s) of the R script file(s) to 
  be run.
}
\description{
Translate a \code{models} vector of model name(s) into a 
  \code{character} vector of file path(s) corresponding to the model 
  scripts.
}
