% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{models_to_cast}
\alias{models_to_cast}
\title{Determine the Paths to the Scripts for Models to Cast with}
\usage{
models_to_cast(
  main = ".",
  models = prefab_models(),
  settings = directory_settings()
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}
}
\value{
\code{character} vector of the path(s) of the R script file(s) to be run.
}
\description{
Translate a \code{character} vector of model name(s) into a \code{character} vector of file path(s) corresponding to the model scripts.
}
\examples{
 \donttest{
  create_dir()
  fill_models()
  models_to_cast()
 }

}
