% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{most_recent_cast}
\alias{most_recent_cast}
\title{Determine the most recent forecast or hindcast}
\usage{
most_recent_cast(tree = dirtree(), cast_type = "forecasts")
}
\arguments{
\item{tree}{\code{dirtree}-class list. See \code{\link{dirtree}}.}

\item{cast_type}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory.}
}
\value{
\code{Date} \code{data.frame} of the most recent cast.
}
\description{
Determine the most recently produced forecast or hindcast
  in a predictions folder.
}
