% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{most_recent_cast}
\alias{most_recent_cast}
\title{Determine the most recent forecast or hindcast}
\usage{
most_recent_cast(tree = dirtree(), cast_type = "forecasts",
  with_census = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class list. See \code{\link{dirtree}}.}

\item{cast_type}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory.}

\item{with_census}{\code{logical} toggle if the plot should include the
observed data collected during the predicted census.}
}
\value{
\code{Date} of the most recent cast.
}
\description{
Determine the date of the most recently produced forecast or 
  hindcast in a predictions folder.
}
