% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{named_null_list}
\alias{named_null_list}
\title{Create a Named Empty List}
\usage{
named_null_list(element_names = NULL)
}
\arguments{
\item{element_names}{\code{character} vector of names for the elements in the list.}
}
\value{
\code{list} with names \code{element_names} and values \code{NULL}.
}
\description{
Produces a list with \code{NULL} for each element named according to \code{element_names}.
}
\examples{
   named_null_list(c("a", "b", "c"))

}
\seealso{
Other utilities: 
\code{\link{file_ext}()},
\code{\link{foy}()},
\code{\link{ifnull}()},
\code{\link{messages}},
\code{\link{package_version_finder}()},
\code{\link{round_na.interp}()},
\code{\link{update_list}()}
}
\concept{utilities}
