% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbsGARCH.R
\name{nbsGARCH}
\alias{nbsGARCH}
\title{negative binomial seasonal Generalized AutoRegressive Conditional
  Heteroscedasticity model for Portal Predictions}
\usage{
nbsGARCH(tree = dirtree(), level = "All", quiet = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{level}{\code{character} value name of the type of plots included 
(\code{"All"} or \code{"Controls"}).}

\item{quiet}{\code{logical} value indicating if the function should be 
quiet.}
}
\value{
\code{list} of [1] \code{"forecast"} (the forecasted abundances)
  and [2] \code{"all_model_aic"} (the model AIC values).
}
\description{
Fit an nbsGARCH model in the portalcasting pipeline.
}
\details{
Model "nbsGARCH" is a seasonal generalized autoregresive 
  conditional heteroscedasticity model with overdispersion 
  (\emph{i.e.}, a negative binomial response variable) fit to the data
  using \code{\link[tscount]{tsglm}} in the
  \href{http://tscount.r-forge.r-project.org/}{\code{tscount} package}
  (Liboschik \emph{et al}. 2017). Seasonal dynamics are modeled using
  a two-term Fourier transformation.
}
\examples{
\dontrun{

setup_dir()
nbsGARCH()
}

}
\references{
Liboschik T., Fokianos K., and Fried R. 2017. tscount: An R Package for 
  Analysis of Count Time Series Following Generalized Linear Models. 
  \emph{Journal of Statistical Software} \strong{82}:5, 1-51. 
  \href{http://doi.org/10.18637/jss.v082.i05}{URL}.
}
