% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\alias{main_path}
\alias{sub_path}
\alias{resources_path}
\alias{casts_path}
\alias{fits_path}
\alias{data_path}
\alias{models_path}
\title{Determine the path for a specific level of a portalcasting directory}
\usage{
main_path(main = ".")

sub_path(main = ".", sub = NULL)

resources_path(main = ".")

casts_path(main = ".")

fits_path(main = ".")

data_path(main = ".")

models_path(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{sub}{\code{character} vector of the name(s) of the sub component(s)
of the directory tree.}
}
\value{
\code{character} \code{\link[port{normalized_file_path}}(s).
         \cr \cr
         \code{main_path} path of the \code{main} folder.
         \cr \cr
         \code{subs_path} paths of the \code{sub} folder(s).
}
\description{
Produce paths for a portalcasting directory. 
             \cr \cr
             \code{main_path} returns the path for the \code{main} folder.
             \cr \cr
             \code{sub_path} returns the path(s) for the \code{subs}
             folder(s). \cr \cr
}
