% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pevgarch.R
\name{pevgarch}
\alias{pevgarch}
\title{poisson environmental variable Generalized Auto-Regressive 
  Conditional Heteroscedasticity model for Portal Predictions}
\usage{
pevgarch(abundances, covariates, metadata, level = "All", lag = 6)
}
\arguments{
\item{abundances}{table of rodent abundances and time measures}

\item{covariates}{table of historical and forecast covariate data and time 
measures}

\item{metadata}{model metadata list}

\item{level}{name of the type of plots included ("All" or "Controls")}

\item{lag}{the lag used for the covariate data}
}
\value{
list of forecast and aic tables
}
\description{
Model "pevGARCH" is a generalized autoregresive conditional 
  heteroscedasticity model with a Poisson response variable  fit to the 
  data using \code{tsglm} in the \code{tscount} package (Liboschik 
  \emph{et al}. 2017). 

  For this model, environmental data are included as predictors of ,
  abundance but with a 6 month lag between the covariate values and the 
  abundances.
}
\references{
Liboschik T., Fokianos K., and Fried R. 2017. tscount: An R Package for 
  Analysis of Count Time Series Following Generalized Linear Models. 
  \emph{Journal of Statistical Software} \strong{82}:5, 1-51. 
  \href{http://doi.org/10.18637/jss.v082.i05}{URL}.
}
