% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pevGARCH.R
\name{pevGARCH}
\alias{pevGARCH}
\title{poisson environmental variable Generalized Auto-Regressive 
  Conditional Heteroscedasticity model for Portal Predictions}
\usage{
pevGARCH(abundances, covariates, metadata, level = "All", lag = 6,
  quiet = FALSE)
}
\arguments{
\item{abundances}{Class-\code{rodents} \code{data.frame} table of rodent 
abundances and time measures.}

\item{covariates}{Class-\code{covaraites} \code{data.frame} table of 
historical and forecast covariate data and time measures.}

\item{metadata}{Class-\code{metadata} model metadata \code{list}.}

\item{level}{\code{character} value name of the type of plots included 
(\code{"All"} or \code{"Controls"}).}

\item{lag}{\code{integer} (or integer \code{numeric}) of the lag time to
use for the covariates.}

\item{quiet}{\code{logical} value indicating if the function should be 
quiet.}
}
\value{
\code{list} of [1] \code{"forecast"} (the forecasted abundances)
  and [2] \code{"all_model_aic"} (the model AIC values).
}
\description{
Fit a pevGARCH model in the portalcasting pipeline.
}
\details{
Model "pevGARCH" is a generalized autoregresive conditional 
  heteroscedasticity model with a Poisson response variable  fit to the 
  data using \code{\link[tscount]{tsglm}} in the
  \href{http://tscount.r-forge.r-project.org/}{\code{tscount} package}
  (Liboschik \emph{et al}. 2017). 

  For this model, environmental data are included as predictors of 
  abundance with an optional \code{lag} between the covariate values and 
  abundances (defaults to 6 newmoons).
}
\references{
Liboschik T., Fokianos K., and Fried R. 2017. tscount: An R Package for 
  Analysis of Count Time Series Following Generalized Linear Models. 
  \emph{Journal of Statistical Software} \strong{82}:5, 1-51. 
  \href{http://doi.org/10.18637/jss.v082.i05}{URL}.
}
