% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_cast_point}
\alias{plot_cast_point}
\title{Plot Predictions for a Given Point in Time Across Multiple Species}
\usage{
plot_cast_point(
  main = ".",
  cast_id = NULL,
  cast_groups = NULL,
  dataset = NULL,
  model = NULL,
  historic_end_newmoonnumber = NULL,
  species = base_species(total = TRUE),
  highlight_sp = NULL,
  newmoonnumber = NULL,
  with_census = FALSE,
  settings = directory_settings(),
  quiet = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{cast_id}{\code{integer} (or integer \code{numeric}) value representing the cast of interest, as indexed within the directory in the \code{casts} sub folder. See the casts metadata file (\code{casts_metadata.csv}) for summary information.}

\item{cast_groups}{\code{integer} (or integer \code{numeric}) value of the cast group to include.}

\item{dataset}{\code{character} value of the rodent data set to include Default value is \code{NULL}, which equates to no selection with respect to \code{dataset}.}

\item{model}{\code{character} value of the name of the model to include. Default value is \code{NULL}, which equates to no selection with respect to \code{model}.}

\item{historic_end_newmoonnumber}{\code{integer} (or integer \code{numeric}) newmoon number of the forecast origin. Default value is \code{NULL}, which equates to no selection with respect to \code{end_moon}.}

\item{species}{\code{character} vector of the species codes (or \code{"total"} for the total across species) to be plotted or \code{NULL} (default) to plot all species in \code{dataset}.}

\item{highlight_sp}{\code{character} vector of the species codes (or \code{"total"} for the total across species) to be highlighted or \code{NULL} (default) to not highlight anything.}

\item{newmoonnumber}{\code{integer} (or integer \code{numeric}) newmoon number for the plot.}

\item{with_census}{\code{logical} toggle if the plot should include the observed data collected during the predicted census.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}
}
\value{
\code{NULL}. Plot is generated.
}
\description{
Plot the point value with confidence interval for a time point across multiple species. Casts can be selected either by supplying a \code{cast_id} number or any combination of \code{dataset}, \code{model}, and \code{end_moon}, which filter the available casts in unison. This plot type can only handle output from a single cast, so if multiple casts still remain, the one with the highest number is selected. To be more certain about cast selection, use the \code{cast_id} input.
}
\details{
The resulting plot shows predictions as points (open white circles) with error, where the point represents the \code{estimate} and the bounds of the error are \code{lower_pi} and \code{upper_pi} in the \code{cast_table} saved output from a model. \cr
 As of \code{portalcasting v0.9.0}, this represents the mean and the 95\% prediction interval. If \code{with_census = TRUE}, the observations from the associated moon are plotted as blue filled squares.
}
