% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_cast_ts}
\alias{plot_cast_ts}
\title{Visualize a forecast or hindcast}
\usage{
plot_cast_ts(tree = dirtree(), species = "total", level = "Controls",
  cast_type = "forecasts", cast_date = NULL, model = "Ensemble",
  start_newmoon = 300)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{species}{\code{character} value of the species code or \code{"total"}
for the total across species.}

\item{level}{\code{character} value of the level of interest (\code{"All"} 
or \code{"Controls"}).}

\item{cast_type}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"forecasts"}.}

\item{cast_date}{\code{Date} the predictions were made. Used to select the
file in the predictions subdirectory.}

\item{model}{\code{character} value of the name (or \code{"Ensemble"}) of
the model to be plotted.}

\item{start_newmoon}{\code{integer}-conformable newmoon number used as the
the minimum x value for the plot.}
}
\description{
Plot an observed timeseries and forecast or hindcast 
  timeseries with a prediction interval. Currently only reliable for 
  forecasts.
}
