% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_cast_ts}
\alias{plot_cast_ts}
\title{Visualize a Time Series Cast of a Species}
\usage{
plot_cast_ts(
  main = ".",
  cast_id = NULL,
  cast_group = NULL,
  dataset = NULL,
  model = NULL,
  historic_start_newmoonnumber = NULL,
  historic_end_newmoonnumber = NULL,
  species = NULL
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{cast_id}{\code{integer} (or integer \code{numeric}) value representing the cast of interest, as indexed within the directory in the \code{casts} sub folder. See the casts metadata file (\code{casts_metadata.csv}) for summary information.}

\item{cast_group}{\code{integer} (or integer \code{numeric}) value of the cast group to include. Default value is \code{NULL}, which equates to no selection with respect to \code{cast_group}.}

\item{dataset}{\code{character} value of the rodent data set to include. Default value is \code{NULL}, which equates to no selection with respect to \code{dataset}.}

\item{model}{\code{character} value of the name of the model to include. Default value is \code{NULL}, which equates to no selection with respect to model.}

\item{historic_start_newmoonnumber}{\code{integer} (or integer \code{numeric}) newmoon number for the beginning of the x-axis of the plot. \cr
Does not influence the fit of the models, just the presentation.}

\item{historic_end_newmoonnumber}{\code{integer} (or integer \code{numeric}) newmoon number of the forecast origin. Default value is \code{NULL}, which equates to no selection.}

\item{species}{\code{character} value of the species codes (or \code{"total"} for the total across species) to be plotted. \code{NULL} (default) also gives the total.}
}
\value{
\code{NULL}. Plot is generated.
}
\description{
Plot an observed timeseries and cast timeseries with a prediction interval. \cr
Casts can be selected either by supplying a \code{cast_id} number or any combination of \code{dataset}, \code{model}, and \code{historic_end_newmoonnumber}, which filter the available casts in unison. This plot type can only handle output from a single cast, so if multiple casts still remain, the one with the highest number is selected. To be more certain about cast selection, use the \code{cast_id} input.
Casts can be selected either by supplying a \code{cast_id} number or any combination of \code{dataset}, \code{model}, and \code{historic_end_newmoonnumber}, which filter the available casts in unison. This plot type can only handle output from a single cast, so if multiple casts still remain, the one with the highest number is selected. To be more certain about cast selection, use the \code{cast_id} input.
}
\details{
The resulting plot shows observations as a solid black line and predictions as a blue polygon with the bounds represent the error given by \code{lower_pi} and \code{upper_pi} and the bisecting blue line representing the \code{estimate} in the \code{cast_table} saved output from a model. \cr
As of \verb{portalcasting v0.9.0}, this represents the mean and the 95\\% prediction interval. \cr
Observations that occurred after the cast are shown connected directly to the pre-cast observation data (as the black solid line).
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "production")
   setup_sandbox(main = main1)

   portalcast(main = main1, models = "AutoArima")
   plot_cast_ts(main = main1)

   unlink(main1, recursive = TRUE)
}

}
