% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_casts_cov_RMSE}
\alias{plot_casts_cov_RMSE}
\title{Plot the Forecast Coverage and RMSE}
\usage{
plot_casts_cov_RMSE(
  main = ".",
  forecast_ids = NULL,
  historic_end_newmoonnumbers = NULL,
  models = NULL,
  datasets = NULL,
  species = NULL
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{forecast_ids}{\code{integer} (or integer \code{numeric}) values representing the casts of interest for restricting plotting, as indexed within the directory in the \code{casts} sub folder. See the casts metadata file (\code{forecasts_metadata.csv}) for summary information.}

\item{historic_end_newmoonnumbers}{\code{integer} (or integer \code{numeric}) newmoon number(s) of the forecast origin. Default value is \code{NULL}, which equates to no selection.}

\item{models}{\code{character} value(s) of the name of the model to include. Default value is \code{NULL}, which equates to no selection with respect to \code{model}. \code{NULL} translates to all \code{models} in the table.}

\item{datasets}{\code{character} value of the rodent data set to include Default value is \code{NULL}, which equates to no selection with respect to \code{dataset}.}

\item{species}{\code{character} vector of the species code(s) or \code{"total"} for the total across species) to be plotted \code{NULL} translates to the species defined by \code{\link[portalr:rodent_species]{portalr::forecasting_species}}.}
}
\value{
\code{NULL}. Plot is generated.
}
\description{
Plot the coverage (fraction of predictions within the CI) and RMSE (root mean squared error) of each model among multiple species.
}
\details{
A pre-loaded table of casts can be input, but if not (default), the table will be efficiently (as defined by the inputs) loaded and trimmed. \cr
The casts can be trimmed specifically using the \code{forecast_ids} input, otherwise, all relevant casts will be plotted.
}
