% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_casts_err_lead}
\alias{plot_casts_err_lead}
\title{Plot the Forecast Error as a Function of Lead Time}
\usage{
plot_casts_err_lead(
  main = ".",
  cast_ids = NULL,
  historic_end_newmoonnumbers = NULL,
  models = NULL,
  datasets = NULL,
  species = NULL
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{cast_ids}{\code{integer} (or integer \code{numeric}) values representing the casts of interest for restricting plotting, as indexed within the directory in the \code{casts} sub folder. See the casts metadata file (\code{casts_metadata.csv}) for summary information.}

\item{historic_end_newmoonnumbers}{\code{integer} (or integer \code{numeric}) newmoon number(s) of the forecast origin. Default value is \code{NULL}, which equates to no selection.}

\item{models}{\code{character} value(s) of the name of the model to include. Default value is \code{NULL}, which equates to no selection with respect to \code{model}. \code{NULL} translates to all \code{models} in the table.}

\item{datasets}{\code{character} value of the rodent data set(s) to include. Default value is \code{"controls"}.}

\item{species}{\code{character} vector of the species code(s) or \code{"total"} for the total across species) to be plotted \code{NULL} translates to the species defined by \code{\link[portalr]{forecasting_species}}.}
}
\value{
\code{NULL}. Plot is generated.
}
\description{
Plot the raw error (estimate - observation) as a function of lead time across model runs from different forecast origins for multiple models and multiple species (or total) within a data set.
}
