% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_species_casts}
\alias{plot_species_casts}
\title{Plot forecast or hindcast predictions for a given point in time 
  across multiple species.}
\usage{
plot_species_casts(tree = dirtree(), species = NULL,
  level = "Controls", cast_type = "forecasts", cast_date = NULL,
  model = "Ensemble", lead = 1, from_date = NULL,
  with_census = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{species}{\code{character} vector of the species codes (or 
\code{"total"} for the total across species) to be plotted or 
\code{NULL} (default) to plot all species and the total.}

\item{level}{\code{character} value of the level of interest (\code{"All"} 
or \code{"Controls"}).}

\item{cast_type}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"forecasts"}.}

\item{cast_date}{\code{Date} the predictions were made. Used to select the
file in the predictions subdirectory. If \code{NULL} (default), selects
the most recent -casts (if \code{with_census = FALSE}) or the most 
recent -casts with a subsequent census (if \code{with_census = TRUE}).}

\item{model}{\code{character} value of the name (or \code{"Ensemble"}) of
the model to be plotted.}

\item{lead}{\code{integer}-conformable lead of the newmoon number used to
select the data plotted.}

\item{from_date}{\code{Date} to be used as a reference of when to count
the \code{lead} from. If \code{NULL} (default), for 
\code{cast_type = "forecasts"}, \code{from_date = cast_date} and 
\code{plot_species_casts} is not yet reliable for 
\code{cast_type = "hindcasts"}.}

\item{with_census}{\code{logical} toggle if the plot should include the
observed data collected during the predicted census.}
}
\description{
Plot the point value with confidence interval for a step in
  a forecast or hindcast across multiple species. Currently only reliable
  for forecasts.
}
