% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{portalcast}
\alias{portalcast}
\alias{xcast}
\title{Forecast Portal Rodents Models}
\usage{
portalcast(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

xcast(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or not (and thus just the tidy messages).}
}
\value{
Results are saved to files, \code{NULL} is returned \code{\link[base]{invisible}}-ly.
}
\description{
Forecast the Portal rodent population data using the data and models in a portalcasting directory. \cr \cr
 \code{portalcast} wraps around \code{cast} to allow multiple runs of multiple models. \cr \cr
 \code{cast} runs a single cast of a single model on one species of one data set.
}
