% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{portalcast}
\alias{portalcast}
\title{Forecast or hindcast Portal rodents}
\usage{
portalcast(options_all = all_options())
}
\arguments{
\item{options_all}{top-level list of options controlling the portalcasting 
directory. Of particular importance is the \code{options_cast} list,
which contains the options controlling the (fore- or hind-)casting 
\cr \cr To run a default hindcast, use 
\code{options_all = all_options(cast_type = "hindcasts")}}
}
\value{
Nothing
}
\description{
Main function for controlling the running of potentially 
  multiple models for either a forecast or a hindcast.
}
