% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{portalcast}
\alias{portalcast}
\title{Forecast Portal rodents}
\usage{
portalcast(type = "forecast", model_dir = "models",
  models = c("autoarima", "esss", "nbgarch", "pevgarch"), ensemble = TRUE)
}
\arguments{
\item{type}{"forecast" or "hindcast"}

\item{model_dir}{directory name where the model files reside}

\item{models}{names of models to run or "all" to run them all}

\item{ensemble}{logical indicator of whether to create an ensemble model}
}
\value{
Nothing
}
\description{
Main function for controlling the running of potentially 
  multiple models for either a forecast or a hindcast
}
