% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{portalcast}
\alias{portalcast}
\title{Forecast Portal rodents}
\usage{
portalcast(options_all = all_options())
}
\arguments{
\item{options_all}{top-level list of options controlling the portalcasting 
directory. Of particular importance is the \code{options_cast} list,
which contains the options controlling the (fore- or hind-)casting}
}
\value{
Nothing
}
\description{
Main function for controlling the running of potentially 
  multiple models for either a forecast or a hindcast. \cr \cr 
  Note: currently \code{portalcast} can only run "forecasts".
}
