% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{portalcast}
\alias{portalcast}
\alias{cast}
\alias{prep_data}
\title{Forecast or hindcast Portal rodents models}
\usage{
portalcast(main = ".", models = prefab_models(), ensemble = TRUE,
  end_moons = NULL, lead_time = 12, cast_date = Sys.Date(),
  start_moon = 217, confidence_level = 0.9, hist_covariates = TRUE,
  cast_covariates = TRUE, raw_path_archive = "portalPredictions",
  raw_path_data = "PortalData",
  raw_path_predictions = "portalPredictions/predictions",
  raw_cov_cast_file = "data/covariate_casts.csv",
  raw_path_cov_cast = "cov_casts",
  raw_moons_file = "Rodents/moon_dates.csv",
  raw_traps_file = "Rodents/Portal_rodent_trapping.csv",
  source_name = "current_archive", append_cast_csv = TRUE,
  controls_m = NULL, controls_r = rodents_controls(),
  control_cdl = climate_dl_control(),
  downloads = zenodo_downloads(c("1215988", "833438")), quiet = FALSE,
  verbose = FALSE, save = TRUE, overwrite = TRUE,
  filename_moons = "moon_dates.csv", filename_cov = "covariates.csv",
  filename_meta = "metadata.yaml", cleanup = TRUE)

cast(main = ".", models = prefab_models(), ensemble = TRUE,
  cast_date = Sys.Date(), moons = prep_moons(main = main),
  end_moon = NULL, raw_path_data = "PortalData",
  raw_traps_file = "Rodents/Portal_rodent_trapping.csv",
  controls_r = rodents_controls(), quiet = FALSE)

prep_data(main = ".", end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), start_moon = 217, confidence_level = 0.9,
  hist_covariates = TRUE, cast_covariates = TRUE,
  raw_path_archive = "portalPredictions", raw_path_data = "PortalData",
  raw_cov_cast_file = "data/covariate_casts.csv",
  raw_path_cov_cast = "cov_casts",
  raw_moons_file = "Rodents/moon_dates.csv",
  source_name = "current_archive", append_cast_csv = TRUE,
  controls_r = rodents_controls(), control_cdl = climate_dl_control(),
  downloads = zenodo_downloads(c("1215988", "833438")), quiet = FALSE,
  verbose = FALSE, save = TRUE, overwrite = TRUE,
  filename_moons = "moon_dates.csv", filename_cov = "covariates.csv",
  filename_meta = "metadata.yaml", cleanup = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{models}{\code{character} vector of name(s) of model(s) to 
include.}

\item{ensemble}{\code{logical} indicator if the ensemble should be added.}

\item{end_moons, end_moon}{\code{integer} (or integer \code{numeric}) 
newmoon number(s) of the last sample(s) to be included. Default value is 
\code{NULL}, which equates to the most recently included sample. \cr
\strong{\code{end_moons} allows for multiple moons, \code{end_moon}
can only be one value}.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the
number of timesteps forward a cast will cover.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}

\item{start_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the first sample to be included. Default value is \code{217}, 
corresponding to \code{1995-01-01}.}

\item{confidence_level}{\code{numeric} confidence level used in 
summarizing model output. Must be between \code{0} and \code{1}.}

\item{hist_covariates}{\code{logical} indicator of whether or not 
historical covariates are to be included.}

\item{cast_covariates}{\code{logical} indicator whether or not -casted 
covariates are to be included.}

\item{raw_path_archive}{\code{character} value of the path to the archive
within the raw sub folder.}

\item{raw_path_data}{\code{character} value indicating the folder path
to the data within the \code{raw} subdirectory but above the files. A 
standard portalcasting directory downloads the raw data files into 
\code{"raw\PortalData"}, so \code{raw_path_data = "PortalData"} (as
\code{"raw/"} is implied).}

\item{raw_path_predictions}{\code{character} value of the path to the
predictions folder within the raw sub folder (via the archive).}

\item{raw_cov_cast_file}{\code{character} value of the path to the
covariate cast file within \code{raw_path_archive}.}

\item{raw_path_cov_cast}{\code{character} value of the path to the folder
that holds any downloaded covariate cast files within the raw sub folder.}

\item{raw_moons_file}{\code{character} value indicating the path
to the moons data file within \code{raw_path_data}. A standard 
portalcasting directory downloads the raw data files into 
\code{"raw\PortalData"}, so 
\code{raw_moons_file = "Rodents/moon_dates.csv"}.}

\item{raw_traps_file}{\code{character} value indicating the path
to the trapping data file within \code{raw_path_data}. A standard 
portalcasting directory downloads the raw data files into 
\code{"raw\PortalData"}, so 
\code{raw_traps_file = "Rodents/Portal_rodent_trapping.csv"}.}

\item{source_name}{\code{character} value for the name to give the 
covariate forecast. Currently is \code{"current_archive"}. Previous to
\code{"current_archive"}, the data were retroactively filled in and are 
given the source name \code{"retroactive"}.}

\item{append_cast_csv}{\code{logical} indicator controlling if the new 
cast covariates should be appended to the historical casts for the
purposes of hindcasting later.}

\item{controls_m}{Additional controls for models not in the prefab set. 
\cr 
A \code{list} of a single model's script-writing controls or a
\code{list} of \code{list}s, each of which is a single model's 
script-writing controls. \cr 
Presently, each model's script writing controls
should include three elements: \code{name} (a \code{character} value of 
the model name), \code{covariates} (a \code{logical} indicator of if the 
model needs covariates), and \code{lag} (an \code{integer}-conformable 
value of the lag to use with the covariates or \code{NA} if 
\code{covariates = FALSE}). \cr 
If only a single model is added, the name of 
the model from the element \code{name} will be used to name the model's
\code{list} in the larger \code{list}. If multiple models are added, each
element \code{list} must be named according to the model and the
\code{name} element. \cr}

\item{controls_r}{Control \code{list} (from 
\code{\link{rodents_control}}) or \code{list} of control \code{list}s 
(from \code{\link{rodents_controls}}) specifying the structuring of the 
rodents tables. See \code{\link{rodents_control}} for details.}

\item{control_cdl}{\code{list} of specifications for the download, which
are sent to \code{\link{NMME_urls}} to create the specific URLs. See
\code{\link{climate_dl_control}}.}

\item{downloads}{\code{list} of arguments to pass to \code{\link{download}}
for raw file downloading.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out
all of the information or not (and thus just the tidy messages).}

\item{save}{\code{logical} indicator controlling if the output should 
be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing
files should be updated (most users should leave as \code{TRUE}).}

\item{filename_moons}{\code{character} name of the file for saving the 
data output.}

\item{filename_cov}{\code{character} filename for saving the output.}

\item{filename_meta}{\code{character} filename for saving the metadata.}

\item{cleanup}{\code{logical} indicator if any files put into the tmp
subdirectory should be removed at the end of the process.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{min_lag}{\code{integer} (or integer \code{numeric}) of the minimum 
covariate lag time used in any model.}
}
\value{
Results are saved to files, \code{NULL} is returned.
}
\description{
Forecast or hindcast the Portal rodent population data using
 the (updated if needed) data and models in a portalcasting directory. \cr
 \code{portalcast} wraps around \code{cast} to allow multiple runs of 
 multiple models, with data preparation as needed between runs occuring
 via \code{prep_data}. \cr \cr
 \code{prep_data} makes sure that the data subdirectory has the updated 
 files for a casting run, using the \code{metadata.yaml} file to verify 
 validity. If data are out of date, runs \code{\link{fill_data}}. \cr \cr
 \code{cast} runs a single cast of multiple models, using the data as
 prepared by \code{prep_data}.
}
\details{
Multiple models can be run together on the same data and 
 multiple runs of a model (e.g., with a rolling origin) can be initiated
 together using \code{portalcast} and \code{cast} runs the individual
 instances.
}
\examples{
 \donttest{
  setup_dir()
  portalcast()
  prep_data()
  cast()
 }

}
