% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{portalcast}
\alias{portalcast}
\title{Forecast or hindcast Portal rodents models}
\usage{
portalcast(models = NULL, cast_type = "forecasts", end = NULL,
  tree = dirtree(), quiet = FALSE, ensemble = TRUE,
  data_control = list(), models_control = list())
}
\arguments{
\item{models}{\code{character} vector of the names of models to include in
the pipeline. Defaults to \code{NULL}, which retrieves the model names 
from \code{models_control}, so the models returned by the default
settings of \code{\link{model_names}}.}

\item{cast_type}{-cast type: \code{"forecasts"} or \code{"hindcasts"}.
\strong{NOTE:} takes precendence over the value of 
\code{cast_type} in the control lists.}

\item{end}{\code{integer} (or integer \code{numeric}) newmoon number of the
last sample to be included. Default value is \code{NULL}, which equates
to the most recently included sample. If \code{cast_type} is 
\code{"hindcasts"} and \code{end} is \code{NULL}, default becomes 
\code{490:403}, corresponding to \code{2017-01-01} back to
\code{2010-01-01}.
\strong{NOTE:} takes precendence over the value of 
\code{end} in the control lists.}

\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{ensemble}{\code{logical} indicator if the ensemble should be added.}

\item{data_control}{A \code{list} of arguments to control the filling of
the data subdirectory. See \code{\link{data_control}}. 
Values not set assume defaults.}

\item{models_control}{A \code{list} of arguments to control the 
filling of the models subdirectory. See \code{\link{models_control}}.
Arguments not specified assume default values.}
}
\description{
Run potentially multiple models for either a forecast or a 
  hindcast.
}
\examples{
\dontrun{

setup_dir()
portalcast()
}

}
