% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{portalcast}
\alias{portalcast}
\alias{cast}
\alias{make_model_combinations}
\title{Forecast Portal Rodents Models}
\usage{
portalcast(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  species = NULL
)

cast(main = ".", dataset = NULL, species = NULL, model = NULL)

make_model_combinations(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  species = forecasting_species(total = TRUE)
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{models, model}{\code{character} vector of name(s) of model(s) to include in the forecast. In \code{cast}, \code{model} can only be length-one.}

\item{datasets, dataset}{\code{character} vector of datasets to be forecast. In \code{cast}, \code{dataset} can only be length-one.}

\item{species}{\code{character} vector of species to be forecast. In \code{cast}, \code{species} can only be length-one. See \code{\link[portalr:rodent_species]{rodent_species}}.}
}
\value{
\code{portalcast}: \code{data.frame} of model combinations with a \code{logical} column added for fit success, \code{\link[base:invisible]{invisible}}-ly. \cr
\code{cast}: \code{list} of model outputs from \code{\link{process_model_output}}. \cr
\code{make_model_combinations}: \code{data.frame} of the model combinations.
}
\description{
Forecast Portal rodent populations using the data and models in a portalcasting directory. \cr \cr
\code{portalcast} wraps around \code{cast} to allow multiple runs of model - dataset - species combinations. It returns and saves out the model combinations table with fit success added as a column.\cr
\code{cast} runs a single forecast of a single model on one species of one dataset. \cr
\code{make_model_combinations} translates model controls into a \code{data.frame} of model, dataset, and species columns, with a row for each combination.
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "portalcast")
   setup_dir(main = main1)

   make_model_combinations(main = main1)
   portalcast(main = main1, models = "AutoArima")
   cast(main = main1, model = "AutoArima", dataset = "controls", species = "DM")

   unlink(main1, recursive = TRUE)
}

}
\seealso{
Core forecasting functions:
\code{\link{ensemble}},
\code{\link{evaluate forecasts}},
\code{\link{process forecast output}}
}
\concept{core}
