% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{portalcast}
\alias{portalcast}
\alias{cast}
\title{Cast Portal Rodents Models}
\usage{
portalcast(
  main = ".",
  models = prefab_models(),
  end_moons = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  controls_rodents = rodents_controls(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

cast(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  controls_rodents = rodents_controls(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{end_moons, end_moon}{\code{integer} (or integer \code{numeric}) newmoon number(s) of the last sample(s) to be included. Default value is \code{NULL}, which equates to the most recently included sample. \cr
\strong{\code{end_moons} allows for multiple moons, \code{end_moon} can only be one value}.}

\item{start_moon}{\code{integer} (or integer \code{numeric}) newmoon number of the first sample to be included. Default value is \code{217}, corresponding to \code{1995-01-01}.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the number of timesteps forward a cast will cover.}

\item{confidence_level}{\code{numeric} confidence level used in summarizing model output. Must be between \code{0} and \code{1}.}

\item{cast_date}{\code{Date} from which future is defined (the origin of the cast). In the recurring forecasting, is set to today's date using \code{\link{Sys.Date}}.}

\item{controls_model}{Additional controls for models not in the prefab set or for overriding controls of a prefab model (for example when applying a model to a new data set). \cr 
A \code{list} of a single model's script-writing controls or a \code{list} of \code{list}s, each of which is a single model's script-writing controls. \cr 
Presently, each model's script writing controls should include four elements: 
\itemize{
 \item \code{name}: a \code{character} value of the model name.
 \item \code{data_sets}: a \code{character} vector of the data set names
  that the model is applied to. 
 \item \code{covariatesTF}: a \code{logical} indicator of if the 
  model needs covariates.
 \item \code{lag}: an \code{integer}-conformable value of the lag to use 
  with the covariates or \code{NA} if \code{covariatesTF = FALSE}.
}
If only a single model's controls are included, the name of the model from the element \code{name} will be used to name the model's \code{list} in the larger \code{list}. If multiple models are added, each element \code{list} must be named according to the model and the\code{name} element.}

\item{controls_rodents}{Control \code{list} or \code{list} of \code{list}s (from \code{\link{rodents_controls}}) specifying the structuring of the rodents tables. See \code{\link{rodents_controls}} for details.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or not (and thus just the tidy messages).}
}
\value{
Results are saved to files, \code{NULL} is returned \code{\link[base]{invisible}}-ly.
}
\description{
Cast the Portal rodent population data using the (updated if needed) data and models in a portalcasting directory. \cr \cr
 \code{portalcast} wraps around \code{cast} to allow multiple runs of multiple models, with data preparation as needed between runs occurring via \code{prep_data}. \cr \cr
 \code{cast} runs a single cast of multiple models across data sets.
}
\details{
Multiple models can be run together on the multiple, varying data sets for a single new moon using \code{cast}. \code{portalcast} wraps around \code{cast}, providing updating, verification, and resetting utilities as well as facilitating multiple runs of \code{cast} across new moons (e.g., when using a rolling origin).
}
\examples{
 \donttest{
  setup_dir()
  portalcast()
  cast()
 }

}
