% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{portalcast_welcome}
\alias{portalcast_welcome}
\title{Directory welcome}
\usage{
portalcast_welcome(quiet = FALSE)
}
\arguments{
\item{quiet}{\code{logical} indicator if progress message should be
quieted.}
}
\value{
\code{NULL}, as message is printed.
}
\description{
Create a welcome message for the directory based on the 
 package version.
}
\examples{
 portalcast_welcome()

}
