% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab_models.R
\name{prefab_model_functions}
\alias{prefab_model_functions}
\alias{AutoArima}
\alias{NaiveArima}
\alias{ESSS}
\alias{nbGARCH}
\alias{nbsGARCH}
\alias{pevGARCH}
\title{Functions for the prefab portalcasting models}
\usage{
AutoArima(main = ".", data_set = "all", quiet = FALSE,
  verbose = FALSE, arg_checks = TRUE)

NaiveArima(main = ".", data_set = "all", quiet = FALSE,
  verbose = FALSE, arg_checks = TRUE)

ESSS(main = ".", data_set = "all_interp", quiet = FALSE,
  verbose = FALSE, arg_checks = TRUE)

nbGARCH(main = ".", data_set = "all_interp", quiet = FALSE,
  verbose = FALSE, arg_checks = TRUE)

nbsGARCH(main = ".", data_set = "all_interp", quiet = FALSE,
  verbose = FALSE, arg_checks = TRUE)

pevGARCH(main = ".", data_set = "all_interp", lag = 6,
  verbose = FALSE, quiet = FALSE, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{data_set}{\code{character} value name of the rodent data set, such as 
(\code{"all"} or \code{"controls"}).}

\item{quiet}{\code{logical} value indicating if the function should be 
quiet.}

\item{verbose}{\code{logical} indicator of whether or not to print out
all of the information or not (and thus just the tidy messages).}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}.}

\item{lag}{\code{integer} (or integer \code{numeric}) of the lag time to
use for the covariates.}
}
\value{
\code{list} of [1] model metadata \code{list} (\code{"metadata"}), 
 [2] cast summary \code{data.frame} (\code{"cast_tab"}),
 [3] \code{list} of model fit objects (\code{"model_fits"}), and 
 [4] \code{list} of model cast objects (\code{"model_casts"}).
}
\description{
Functions used for fitting the prefab models. For detailed
 descriptions of the models, see the
 \href{https://bit.ly/2xP9jKI}{model vignette}. \cr \cr
 \code{AutoArima} fits a flexible auto-regressive integrated 
 moving average model using \code{\link[forecast]{auto.arima}}. \cr \cr
 \code{NaiveArima} fits a baseline random walk with no drift model as an
 ARIMA(0,1,0) using \code{\link[forecast]{Arima}}. \cr \cr
 \code{ESSS} fits a flexible exponential smoothing state space 
 model with the possibility of multiplicative trends using 
 \code{\link[forecast]{ets}}. \cr \cr
 \code{nbGARCH} fits a generalized autoregresive conditional 
 heteroscedasticity model with overdispersion (\emph{i.e.}, a negative 
 binomial response variable) using \code{\link[tscount]{tsglm}}. \cr \cr
 \code{nbsGARCH} fits a seasonal generalized autoregresive conditional 
 heteroscedasticity model with overdispersion (\emph{i.e.}, a negative 
 binomial response variable) using \code{\link[tscount]{tsglm}}. \cr \cr
 \code{pevGARCH} fits a set of generalized autoregressive conditional
 heteroscedasticity models with Poisson response variables and 
 environmental covariates (max temp, mean temp, min temp, precipitation,
 and NDVI) using \code{\link[tscount]{tsglm}}.
}
\details{
\code{AutoArima} \cr
 Because the seasonality and sampling occurring with different 
 frequencies, which \code{\link[forecast]{auto.arima}} cannot accommodate,
 seasonal AutoArima models are not included.  \cr \cr
 \code{pevGARCH} \cr
 Environmental predictors have an optional \code{lag} time, which is 
 currently structured to impact all covariates the same and defaults to
 \code{6}. \cr
 All covariate models are fit and the best model (by AIC) is selected for
 use.
}
\examples{
 \donttest{
  setup_dir()
  AutoArima()
  NaiveArima()
  ESSS()
  nbGARCH()
  nbsGARCH()
  pevGARCH()
 }

}
\references{
Hyndman R., Bergmeir C., Caceres G., Chhay L., O'Hara-Wild M., Petropoulos
 F., Razbash S., Wang E., and Yasmeen F. 2018. forecast: Forecasting 
 functions for time series and linear models. 
 \href{http://pkg.robjhyndman.com/forecast}{R package version 8.3}. 

 Liboschik T., Fokianos K., and Fried R. 2017. tscount: An R Package for 
 Analysis of Count Time Series Following Generalized Linear Models. 
 \emph{Journal of Statistical Software} \strong{82}:5, 1-51. 
 \href{http://doi.org/10.18637/jss.v082.i05}{URL}.
}
