% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{prefab_models}
\alias{prefab_models}
\title{Provide the names of the prefabricated models}
\usage{
prefab_models()
}
\value{
\code{character} vector of model names.
}
\description{
Create a \code{character} vector of the names of the models
             to be included including the pre-fabricated (prefab) models, 
             extracting them from the \code{prefab_model_controls} 
             internal function. 
             \cr \cr
             The currently prefabricated models include \code{"AutoArima"}, 
             \code{"ESSS"}, \code{"NaiveArima"}, \code{"nbGARCH"}, 
             \code{"nbsGARCH"}, \code{"pevGARCH"}, \code{"simplexEDM"}, 
             \code{"GPEDM"}, and \code{"jags_RW"}.
}
\examples{
 prefab_models()

}
