% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{prefabricated_models}
\alias{prefabricated_models}
\title{Provide the Names of the Prefabricated Models}
\usage{
prefabricated_models()
}
\value{
\code{character} vector of model names.
}
\description{
Create a \code{character} vector of the prefabricated models. \cr \cr
}
\examples{
 prefab_models()

}
