% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefabricated_rodent_datasets.R
\name{prefabricated_rodent_datasets}
\alias{prefabricated_rodent_datasets}
\title{Prefabricated Rodent Datasets}
\usage{
prefabricated_rodent_datasets()
}
\value{
\code{character} names of the data sets.
}
\description{
Extract the names of the prefabricated datasets from the \code{\link{rodent_dataset_controls}}. \cr \cr 
             The currently prefabricated datasets include \code{"all"}, \code{"all_interp"}, \code{"controls"}, \code{"controls_interp"}, \code{"exclosures"}, \code{"exclosures_interp"}, \code{"dm_controls"}, and \code{"dm_controls_interp"}.
}
