% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_covariates}
\alias{prep_covariates}
\title{Prepare covariate data table}
\usage{
prep_covariates(moons = prep_moons(),
  options_covariates = covariates_options())
}
\arguments{
\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}

\item{options_covariates}{A class-\code{covariates_options} \code{list} of 
settings controlling the covariates data creation.}
}
\value{
Covariate data table as a code{covariates}-class \code{data.frame}.
}
\description{
Prepare covariate data for forecasting or hindcasting, 
  including saving out the data and appending the new forecasts of 
  covariates to the existing covariate forecast table.
}
