% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_covariates}
\alias{prep_covariates}
\title{Prepare covariate data for casting}
\usage{
prep_covariates(
  main = ".",
  moons = NULL,
  min_lag = 183,
  origin = Sys.Date(),
  t1 = as.Date("1995-01-01"),
  lead = 365,
  control_files = files_control(),
  quiet = TRUE,
  verbose = FALSE,
  arg_checks = TRUE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{min_lag}{\code{integer} (or integer \code{numeric}) of the minimum 
covariate lag time used in any model.}

\item{origin}{\code{Date} forecast origin, typically today's date (set using \code{\link{Sys.Date}}).}

\item{t1}{\code{Date} for the beginning of the rodent time series to include within the model, typically \code{1995-01-01}, corresponding to \code{newmoonnumber = 217}.}

\item{lead}{\code{integer} (or integer \code{numeric}) value for the number of days forward a cast will cover.}

\item{control_files}{\code{list} of names of the folders and files within
the sub directories and saving strategies (save, overwrite, append, etc.).
Generally shouldn't need to be edited. See \code{\link{files_control}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be
shown.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}

\item{start_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the first sample to be included. Default value is \code{217}, 
corresponding to \code{1995-01-01}.}
}
\value{
\code{data.frame} of historical and -casted covariates, combined
 and saved out to \code{filename} if indicated by \code{save}.
}
\description{
Combine the historical and cast covariate data for a model
 run. \cr \cr See \code{\link{prep_hist_covariates}} and
 \code{\link{prep_cast_covariates}} for the historical and -casted
 covariate preparation details, respectively.
}
\examples{
 \donttest{
  setup_dir()
  prep_covariates()
 }
 
}
