% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_covariates}
\alias{prep_covariates}
\title{Prepare covariate data table}
\usage{
prep_covariates(moons = prep_moons(),
  options_covariates = covariates_options())
}
\arguments{
\item{moons}{current newmoon table}

\item{options_covariates}{control options list for covariates}
}
\value{
covariate data table
}
\description{
Prepare covariate data for forecasting, including saving out
  the data and appending the new forecasts of covariates to the existing
  covariate forecast table
}
