% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_covariates}
\alias{prep_covariates}
\title{Prepare covariate data table}
\usage{
prep_covariates(tree = dirtree(), moons = prep_moons(),
  data_options = covariates_options(), quiet = FALSE)
}
\arguments{
\item{tree}{directory tree}

\item{moons}{current newmoon table}

\item{data_options}{control options list for covariates}

\item{quiet}{logical indicator of printing messages}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
covariate data table
}
\description{
Prepare covariate data for forecasting, including saving out
  the data and appending the new forecasts of covariates to the existing
  covariate forecast table
}
