% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_covariates}
\alias{prep_covariates}
\title{Prepare covariate data for casting}
\usage{
prep_covariates(main = ".", moons = prep_moons(main = main),
  end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), hist_covariates = TRUE,
  cast_covariates = TRUE, raw_path_archive = "portalPredictions",
  raw_cov_cast_file = "data/covariate_casts.csv",
  raw_path_cov_cast = "cov_casts", source_name = "current_archive",
  append_cast_csv = TRUE, control_cdl = list(), quiet = FALSE,
  save = TRUE, overwrite = TRUE, filename_cov = "covariates.csv")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the
number of timesteps forward a cast will cover.}

\item{min_lag}{\code{integer} (or integer \code{numeric}) of the minimum 
covariate lag time used in any model.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}

\item{hist_covariates}{\code{logical} indicator of whether or not 
historical covariates are to be included.}

\item{cast_covariates}{\code{logical} indicator whether or not -casted 
covariates are to be included.}

\item{raw_path_archive}{\code{character} value of the path to the archive
within the raw sub folder.}

\item{raw_cov_cast_file}{\code{character} value of the path to the
covariate cast file within \code{raw_path_archive}.}

\item{raw_path_cov_cast}{\code{character} value of the path to the folder
that holds any downloaded covariate cast files within the raw sub folder.}

\item{source_name}{\code{character} value for the name to give the 
covariate forecast. Currently is \code{"current_archive"}. Previous to
\code{"current_archive"}, the data were retroactively filled in and are 
given the source name \code{"retroactive"}.}

\item{append_cast_csv}{\code{logical} indicator controlling if the new 
cast covariates should be appended to the historical casts for the
purposes of hindcasting later.}

\item{control_cdl}{\code{list} of specifications for the download, which
are sent to \code{\link{NMME_urls}} to create the specific URLs. See
\code{\link{climate_dl_control}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{save}{\code{logical} indicator controlling if the output should 
be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing
files should be updated (most users should leave as \code{TRUE}).}

\item{filename_cov}{\code{character} filename for saving the output.}
}
\value{
\code{data.frame} of historical and -casted covariates, combined
 and saved out to \code{filename} if indicated by \code{save}.
}
\description{
Combine the historical and forecast or hindcast covariate 
 data for a model run. See \code{link{prep_hist_covariates}} and
 \code{\link{prep_cast_covariates}} for the historical and -casted
 covariate preparation details, respectively.
}
\examples{
 \donttest{
  create_dir()
  fill_dir()
  prep_covariates()
 }
 
}
