% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_fcast_covariates}
\alias{prep_fcast_covariates}
\title{Prepare forecast covariate data table}
\usage{
prep_fcast_covariates(hist_data = prep_hist_covariates(), tree = dirtree(),
  moons = prep_moons(), data_options = covariates_options())
}
\arguments{
\item{hist_data}{historical covariate data}

\item{tree}{directory tree}

\item{moons}{current newmoon table}

\item{data_options}{control options list for covariates}
}
\value{
covariate data table
}
\description{
Prepare forecasts of covariate data as needed for rodent
  forecasting, including appending the new forecasts of covariates to the
  existing covariate forecast table
}
