% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_fcast_covariates}
\alias{prep_fcast_covariates}
\title{Prepare forecast covariate data table}
\usage{
prep_fcast_covariates(hist_cov = prep_hist_covariates(),
  moons = prep_moons(), options_covariates = covariates_options())
}
\arguments{
\item{hist_cov}{historical covariate data}

\item{moons}{current newmoon table}

\item{options_covariates}{control options list for covariates}
}
\value{
covariate data table
}
\description{
Prepare forecasts of covariate data as needed for rodent
  forecasting, including appending the new forecasts of covariates to the
  existing covariate forecast table
}
