% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_hist_covariates}
\alias{prep_hist_covariates}
\title{Prepare historical covariates data}
\usage{
prep_hist_covariates(tree = dirtree())
}
\arguments{
\item{tree}{directory tree}
}
\value{
historical covariate table
}
\description{
Create a data table of historical weather and ndvi data
}
