% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_hist_covariates}
\alias{prep_hist_covariates}
\alias{prep_weather_data}
\title{Prepare historical covariates data}
\usage{
prep_hist_covariates(main = ".", end_moon = NULL, quiet = FALSE)

prep_weather_data(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\value{
\code{prep_hist_covariates}: historical covariate data table as a 
  \code{data.frame}. \cr \cr
 \code{prep_weather_data}: \code{data.frame} of historical weather data.
}
\description{
\code{prep_hist_covariates} creates a \code{data.frame} of historical 
 weather and NDVI data. Automatically goes all the way back to the 
 beginning of the available data, as the table is automatically trimmed 
 later. \cr \cr
 \code{prep_weather_data} creates a \code{data.frame} of historical 
  weather data using \code{\link[portalr]{weather}}. \cr \cr
 NDVI data are generated by \code{\link[portalr]{ndvi}}.
}
\examples{
 \donttest{
  create_dir()
  fill_raw()
  prep_hist_covariates()
  prep_weather_data()
 }

}
