% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prep_metadata}
\alias{prep_metadata}
\title{Prepare Model Metadata}
\usage{
prep_metadata(rodents, moons, covariates)
}
\arguments{
\item{rodents}{rodent data table}

\item{moons}{newmoon data table (with future moons)}

\item{covariates}{covariate data table}
}
\value{
model metadata as a list
}
\description{
Set up the model metadata, primarily the forecast timeperiods
}
