% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prep_metadata}
\alias{prep_metadata}
\title{Prepare model metadata}
\usage{
prep_metadata(rodents, covariates, forecast_date = Sys.Date(),
  filename_suffix = "forecasts", confidence_level = 0.9, lead_time = 12,
  data_dir = pc_path("data", "~"))
}
\arguments{
\item{rodents}{rodent data table}

\item{covariates}{covariate data table}

\item{forecast_date}{date of the forecast (i.e., today)}

\item{filename_suffix}{"forecasts" or "hindcasts"}

\item{confidence_level}{confidence level used in summarizing model output}

\item{lead_time}{number of newmoons forward to forecast abundances}

\item{data_dir}{directory name where the data files will reside}
}
\value{
model metadata as a list
}
\description{
Set up the model metadata, primarily the forecast timeperiods
}
