% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prep_metadata}
\alias{prep_metadata}
\title{Prepare a model-running metadata list}
\usage{
prep_metadata(main = ".", models = prefab_models(), data_sets = NULL,
  moons = NULL, rodents = NULL, covariates = NULL, end_moon = NULL,
  lead_time = 12, min_lag = NULL, cast_date = Sys.Date(),
  start_moon = 217, confidence_level = 0.95, controls_r = NULL,
  controls_m = NULL, filename_config = "dir_config.yaml",
  quiet = TRUE, verbose = FALSE, save = TRUE, overwrite = TRUE,
  filename_meta = "metadata.yaml", arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{rodents}{Rodents \code{list}. See \code{\link{prep_rodents}},}

\item{covariates}{Covariates \code{data.frame}. See 
\code{\link{prep_covariates}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the
number of timesteps forward a cast will cover.}

\item{min_lag}{\code{integer} (or integer \code{numeric}) of the minimum 
covariate lag time used in any model.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}

\item{start_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the first sample to be included. Default value is \code{217}, 
corresponding to \code{1995-01-01}.}

\item{confidence_level}{\code{numeric} confidence level used in 
summarizing model output. Must be between \code{0} and \code{1}.}

\item{controls_r}{Control \code{list} or \code{list} of control 
\code{list}s from \code{\link{rodents_controls}} specifying the 
structuring of the rodents tables. See \code{\link{rodents_controls}} for 
details.}

\item{controls_m}{Additional controls for models not in the prefab set. 
\cr 
A \code{list} of a single model's script-writing controls or a
\code{list} of \code{list}s, each of which is a single model's 
script-writing controls. \cr 
Presently, each model's script writing controls
should include three elements: \code{name} (a \code{character} value of 
the model name), \code{covariates} (a \code{logical} indicator of if the 
model needs covariates), and \code{lag} (an \code{integer}-conformable 
value of the lag to use with the covariates or \code{NA} if 
\code{covariates = FALSE}). \cr 
If only a single model is added, the name of 
the model from the element \code{name} will be used to name the model's
\code{list} in the larger \code{list}. If multiple models are added, each
element \code{list} must be named according to the model and the
\code{name} element. \cr}

\item{filename_config}{\code{character} value of the path to the directory
config YAML.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be
shown.}

\item{save}{\code{logical} indicator controlling if the output should 
be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing
files should be updated (most users should leave as \code{TRUE}).}

\item{filename_meta}{\code{character} filename for saving the metadata.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{list} of casting metadata, which is also saved out as a 
 YAML file (\code{.yaml}) if desired.
}
\description{
Sets up the metadata used for casting, in particular the 
 matching of time period across the data sets. This should always be run
 after \code{\link{prep_moons}}, \code{\link{prep_rodents}}, and 
 \code{\link{prep_covariates}} before any model is run.
}
\examples{
 \donttest{
  setup_dir()
  moons <- prep_moons()
  rodents <- prep_rodents(moons = moons)
  covariates <- prep_covariates(moons = moons)
  prep_metadata(moons = moons, rodents = rodents, covariates = covariates)
 }

}
