% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prep_metadata}
\alias{prep_metadata}
\title{Prepare a metadata list}
\usage{
prep_metadata(main = ".", moons = NULL, rodents = NULL,
  covariates = NULL, end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), start_moon = 217, confidence_level = 0.9,
  quiet = FALSE, save = TRUE, overwrite = TRUE,
  filename_meta = "metadata.yaml", arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{rodents}{Rodents \code{list}. See \code{\link{prep_rodents}},}

\item{covariates}{Covariates \code{data.frame}. See 
\code{\link{prep_covariates}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the
number of timesteps forward a cast will cover.}

\item{min_lag}{\code{integer} (or integer \code{numeric}) of the minimum 
covariate lag time used in any model.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}

\item{start_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the first sample to be included. Default value is \code{217}, 
corresponding to \code{1995-01-01}.}

\item{confidence_level}{\code{numeric} confidence level used in 
summarizing model output. Must be between \code{0} and \code{1}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{save}{\code{logical} indicator controlling if the output should 
be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing
files should be updated (most users should leave as \code{TRUE}).}

\item{filename_meta}{\code{character} filename for saving the metadata.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{list} of casting metadata, which is also saved out as a 
 YAML file (\code{.yaml}) if desired.
}
\description{
Sets up the metadata used for casting, in particular the 
 matching of time period across the data sets. This should always be run
 after \code{\link{prep_moons}}, \code{\link{prep_rodents}}, and 
 \code{\link{prep_covariates}} before any model is run.
}
\examples{
 \donttest{
  setup_dir()
  moons <- prep_moons()
  rodents <- prep_rodents(moons = moons)
  covariates <- prep_covariates(moons = moons)
  prep_metadata(moons = moons, rodents = rodents, covariates = covariates)
 }

}
