% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prep_metadata}
\alias{prep_metadata}
\title{Prepare model metadata}
\usage{
prep_metadata(moons = prep_moons(), rodents = prep_rodents(),
  covariates = prep_covariates(),
  options_metadata = metadata_options())
}
\arguments{
\item{moons}{moons data table}

\item{rodents}{rodent data table}

\item{covariates}{covariate data table}

\item{options_metadata}{metadata options list}
}
\value{
model metadata as a list
}
\description{
Set up the model metadata, primarily the forecast timeperiods
}
