% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prep_metadata}
\alias{prep_metadata}
\title{Prepare model metadata}
\usage{
prep_metadata(moons = prep_moons(), rodents = prep_rodents(),
  covariates = prep_covariates(),
  options_metadata = metadata_options())
}
\arguments{
\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}

\item{rodents}{rodent data table}

\item{covariates}{Covariate data table as a code{covariates}-class 
\code{data.frame} generated by \code{\link{prep_covariates}}}

\item{options_metada}{A class-\code{metadata_options} \code{list} of 
settings controlling the metadata data creation.}
}
\value{
Model metadata as a \code{metadata}-class \code{list}.
}
\description{
Set up the metadata list used for -casting, primarily the 
  time periods.
}
