% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prep_metadata}
\alias{prep_metadata}
\title{Prepare Model Metadata}
\usage{
prep_metadata(rodents, covariates, forecast_date = Sys.Date(),
  filename_suffix = "forecasts", confidence_level = 0.9)
}
\arguments{
\item{rodents}{rodent data table}

\item{covariates}{covariate data table}

\item{forecast_date}{date of the forecast (i.e., today)}

\item{filename_suffix}{"forecasts" or "hindcasts"}

\item{confidence_level}{confidence level used in summarizing model output}
}
\value{
model metadata as a list
}
\description{
Set up the model metadata, primarily the forecast timeperiods
}
