% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{prep_moons}
\alias{prep_moons}
\title{Prepare temporal (lunar) data}
\usage{
prep_moons(tree = dirtree(), n_future_moons = 0, fdate = today())
}
\arguments{
\item{fdate}{date from which future is defined, typically today's date.
Required if \code{future > 0}}

\item{future_moons}{integer value for the number of future moons to add}

\item{data_dir}{directory name where the data files will reside}
}
\value{
data table of time variables for all surveys
}
\description{
Get time information (calendar dates, census periods, 
  newmoon numbers) associated with trapping events (achieved and missed)
  based on a lunar survey schedule.
}
