% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{prep_moons}
\alias{prep_moons}
\title{Prepare temporal (lunar) data}
\usage{
prep_moons(tree = dirtree(), data_options = moons_options())
}
\arguments{
\item{tree}{directory tree}

\item{data_options}{control options list for moons}
}
\value{
data table of time variables for all surveys
}
\description{
Get time information (calendar dates, census periods, 
  newmoon numbers) associated with trapping events (achieved and missed)
  based on a lunar survey schedule.
}
