% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{prep_moons}
\alias{prep_moons}
\title{Prepare temporal (lunar) data}
\usage{
prep_moons(options_moons = moons_options())
}
\arguments{
\item{options_moons}{control options list for moons}
}
\value{
data table of time variables for all surveys
}
\description{
Get time information (calendar dates, census periods, 
  newmoon numbers) associated with trapping events (achieved and missed)
  based on a lunar survey schedule. If needed, additional moons will be
  added to the data table, both the in-use and raw versions
}
