% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{prep_moons}
\alias{prep_moons}
\title{Prepare temporal (lunar) data}
\usage{
prep_moons(options_moons = moons_options())
}
\arguments{
\item{options_moons}{A class-\code{moons_options} \code{list} of options 
for the moons data. See \code{\link{moons_options}}.}
}
\value{
Moons data table as a code{moons}-class \code{data.frame}.
}
\description{
Get time information (calendar dates, census periods, 
  newmoon numbers) associated with trapping events (achieved and missed)
  based on a lunar survey schedule. If needed, additional moons will be
  added to both the in-use and raw versions of the data table.
}
