% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{prep_rodents}
\alias{prep_rodents}
\title{Prepare a list of rodents data tables for forecasting}
\usage{
prep_rodents(controls = rodents_controls(), moons = prep_moons(main =
  main), main = ".", quiet = FALSE, overwrite = TRUE,
  ref_species = all_species(), save = TRUE)
}
\arguments{
\item{controls}{Control \code{list} (from \code{\link{rodents_control}}) or
\code{list} of control \code{list}s (from \code{\link{rodents_controls}})
specifying the structuring of the rodents tables. See
\code{\link{rodents_control}} for details.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{overwrite}{\code{logical} indicator of whether or not the existing
files should be updated (most users should leave as \code{TRUE}).}

\item{ref_species}{\code{character}-valued vector of all possible species 
names that could be used should be set up via \code{\link{all_species}}.}

\item{save}{\code{logical} indicator controlling if the output should 
be saved out.}
}
\value{
\code{list} of \code{data.frame}s, one for each data subset
 specified, as generated by \code{\link{prep_rodents_table}}.
}
\description{
This function wraps around \code{\link{prep_rodents_table}} to 
 produce a list of rodents \code{data.frame}s associated with multiple
 sets of specifications. The default settings are for a general 
 portalcasting directory, which models the "all" and "controls" data 
 subsets.
}
\examples{
 \donttest{
   create_dir()
   fill_raw()
   prep_rodents()
 }

}
