% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{prep_rodents}
\alias{prep_rodents}
\title{Prepare a list of rodents data tables for forecasting}
\usage{
prep_rodents(
  main = ".",
  moons = NULL,
  data_sets = NULL,
  end_moon = NULL,
  controls_rodents = NULL,
  control_files = files_control(),
  ref_species = all_species(),
  quiet = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{data_sets}{\code{character} values of the treatment 
types. If left as \code{NULL} (default), it is replaced with
the output from \code{\link{prefab_data_sets}}. \cr
Used to set arguments for \code{\link{prep_rodents_table}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{controls_rodents}{Control \code{list} or \code{list} of control 
\code{list}s from \code{\link{rodents_controls}} specifying the 
structuring of the rodents tables. See \code{\link{rodents_controls}}.}

\item{control_files}{\code{list} of names of the folders and files within
the sub directories and saving strategies (save, overwrite, append, etc.).
Generally shouldn't need to be edited. See \code{\link{files_control}}.}

\item{ref_species}{\code{character}-valued vector of all possible species 
names that could be used. Should be set up via \code{\link{all_species}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be
shown.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
\code{list} of \code{data.frame}s, one for each data subset
 specified, as generated by \code{\link{prep_rodents_table}}.
}
\description{
This function loops around \code{\link{prep_rodents_table}},
 which wraps around \code{\link[portalr]{summarize_rodent_data}} to 
 produce a list of rodents \code{data.frame}s associated with multiple
 sets of specifications. The default settings are for a general 
 portalcasting directory, which models the "all" and "controls" data 
 subsets.
}
\examples{
 \donttest{
   create_dir()
   fill_raw()
   prep_rodents()
 }

}
