% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{prep_rodents}
\alias{prep_rodents}
\title{Prepare historic rodent data}
\usage{
prep_rodents(moons = prep_moons(), options_rodents = rodents_options())
}
\arguments{
\item{moons}{current newmoon table}

\item{options_rodents}{data options list for rodents, which is forced to 
conform internally for the two treatment types (all and controls)}
}
\value{
a list of two dataframes, all plots and control plots
}
\description{
Prepare the rodent dataset, tailored for forecasting. 
  Specifically, return data for all plots and control plots only, for each 
  species except PI and the total
}
