% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{prep_rodents}
\alias{prep_rodents}
\title{Prepare historic rodent data}
\usage{
prep_rodents(moons = prep_moons(), options_rodents = rodents_options())
}
\arguments{
\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}

\item{options_rodents}{Class-\code{rodents_options} \code{list} of 
settings controlling the rodents data creation. See 
\code{\link{rodents_options}}.}
}
\value{
A class-\code{rodents_list} \code{list} of two class-
  \code{rodents} \code{data.frame}s, \code{all} (abundances on all plots)
  and \code{controls} (abundances on control plots only).
}
\description{
Prepare the rodent dataset, tailored for forecasting. 
  Specifically, return data for all plots and control plots only, for each 
  species except PI and the total
}
