% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{prep_rodents_table}
\alias{prep_rodents_table}
\alias{trim_time}
\alias{add_moons}
\alias{add_total}
\alias{trim_species}
\alias{trim_treatment}
\title{Prepare a rodents data table for forecasting}
\usage{
prep_rodents_table(species = base_species(), total = TRUE,
  moons = prep_moons(main = main), main = ".", start_moon = 217,
  end_moon = NULL, level = "Site", treatment = NULL, plots = "all",
  min_plots = 24, min_traps = 1, output = "abundance",
  quiet = TRUE)

trim_time(rodents, start_moon = NULL, end_moon = NULL)

add_moons(rodents, moons = prep_moons())

add_total(rodents, total = TRUE)

trim_species(rodents, species = base_species(),
  ref_species = all_species())

trim_treatment(rodents, level = "Site", treatment = NULL)
}
\arguments{
\item{species}{\code{character}-valued vector of species names to include 
in the forecasting data table. Defaults to all species but \code{"PI"}
via \code{\link{base_species}}}

\item{total}{\code{logical} value indicating if a total (sum across 
species) should be added or not. Only available if more than one species
is included.}

\item{moons}{Moons \code{data.frame}.}

\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{start_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the first sample to be included. Default value is \code{217}, 
corresponding to \code{1995-01-01}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{level}{\code{character} indicating the type of summary:
\code{"Plot"}, \code{"Treatment"}, or \code{"Site"}. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}

\item{treatment}{\code{character} indicating the specific treatment(s) to
trim to if \code{level = "Treatment"}: \code{"control"},
\code{"exclosure"}, \code{"removal"}, or \code{"spectabs"} .}

\item{plots}{Specification of subset of plots. Can be a vector of 
\code{numeric} plots indicators or specific sets indicated by
\code{character} values: \code{"all"} plots or \code{"Longterm"} plots
(plots that have had the same treatment for the entire time series).}

\item{min_plots}{\code{integer} (or integer \code{numeric}) of the minimum 
number of plots surveyed for a survey to be used.Pipes directly to 
\code{\link[portalr]{summarize_rodent_data}}.}

\item{min_traps}{\code{integer} (or integer \code{numeric}) of the minimum 
number of traps collected for a plot to be used. Pipes directly to 
\code{\link[portalr]{summarize_rodent_data}}.}

\item{output}{\code{character} indicating the type of data:
\code{"abundance"}, \code{"biomass"}, or \code{"energy"}. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{pipes}{directly to \code{\link[portalr]{summarize_rodent_data}}.}
}
\value{
\code{data.frame} rodents table, with varying levels of editing
 for use in forecasting.
}
\description{
This set of functions provides a convenient wrapper on
 \code{\link[portalr]{summarize_rodent_data}} to generate a specific
 data table for use in the forecasting pipeline. Many arguments pipe
 directly to \code{\link[portalr]{summarize_rodent_data}}.
}
\examples{
 \donttest{
  create_dir()
  fill_raw()
  prep_rodents_table()
  raw_path <- sub_paths(".", specific_subs = "raw")
  dat1 <- summarize_rodent_data(path = raw_path, clean = FALSE)
  dat2 <- trim_species(dat1)
  dat3 <- add_total(dat2)
  dat4 <- trim_treatment(dat3)
  dat5 <- add_moons(dat4)
  dat6 <- trim_time(dat5, 217)
 }

}
