% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prep_weather_data}
\alias{prep_weather_data}
\title{Collect Historic Weather Data}
\usage{
prep_weather_data()
}
\value{
weather data table
}
\description{
Gather the downloaded historical weather data, select the 
  columns needed, and remove the incomplete data
}
