% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prep_weather_data}
\alias{prep_weather_data}
\title{Prepare historical weather data}
\usage{
prep_weather_data(tree = dirtree())
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}
}
\value{
\code{data.frame} of historical weather data.
}
\description{
Create a data table of historical weather data using 
  \code{\link[portalr]{weather}}.
}
