% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prepare covariates}
\alias{prepare covariates}
\alias{prepare_covariates}
\title{Prepare Covariate Data for Casting}
\usage{
prepare_covariates(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
\code{data.frame} of historical and forecasted covariates that is also saved out to \code{settings$files$covariates} if indicated by \code{settings$save}.
}
\description{
Prepare and combine the historical and forecast covariate data for a model run, according to the \code{\link{directory_settings}}.
}
