% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prepare metadata}
\alias{prepare metadata}
\alias{prepare_metadata}
\title{Prepare a Model-Running Metadata List}
\usage{
prepare_metadata(
  main = ".",
  datasets = prefab_datasets(),
  new_datasets_controls = NULL
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{datasets}{\code{character} vector of name(s) of dataset(s) to include.}

\item{new_datasets_controls}{\code{list} of controls for any new datasets (not in the prefab datasets) listed in \code{datasets} that are to be added to the control list and file.}
}
\value{
\code{list} of forecasting metadata, which is also saved out as a YAML file (\code{.yaml}) if desired.
}
\description{
Sets up the metadata list used for forecasting, in particular the matching of time period across the datasets, according to the \code{\link{directory_settings}}.
}
\examples{
\dontrun{
   main1 <- file.path(tempdir(), "metadata")

   create_dir(main = main1)
   fill_resources(main = main1)
   fill_forecasts(main = main1)
   fill_fits(main = main1)
   fill_models(main = main1)

   prepare_newmoons(main   = main1)
   prepare_rodents(main    = main1) 
   prepare_covariates(main = main1)
   prepare_metadata(main   = main1)

   unlink(main1, recursive = TRUE)
}

}
\seealso{
Content preparation functions:
\code{\link{directory filling}},
\code{\link{prepare covariates}},
\code{\link{prepare models}},
\code{\link{prepare newmoons}},
\code{\link{prepare rodents}}
}
\concept{content-prep}
