% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{prepare newmoons}
\alias{prepare newmoons}
\alias{prepare_newmoons}
\alias{add_forecast_newmoons}
\title{Prepare Lunar Data for the Portalcasting Repository}
\usage{
prepare_newmoons(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

add_forecast_newmoons(
  main = ".",
  newmoons = NULL,
  settings = directory_settings()
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or not (and thus just the tidy messages).}

\item{newmoons}{\code{data.frame} of newmoon data.}
}
\value{
Some version of a newmoons \code{data.frame}. \cr \cr. 
        \code{prepare_newmoons}: fully appended and formatted \code{data.frame} (also saved out if \code{settings$save = TRUE}). \cr 
        \code{add_forecast_newmoons}: fully appended and formatted \code{data.frame}. \cr
}
\description{
Get time information (calendar dates, census periods, and newmoon numbers) associated with trapping events (achieved and missed) based on a lunar survey schedule. 
             \code{add_forecast_newmoons} adds future newmoon dates to the newmoon table from \code{prepare_newmoons} associated with the forecast.
}
