% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{prepare rodents}
\alias{prepare rodents}
\alias{prep_rodents}
\title{Prepare Rodents Data for the Portalcasting Repository}
\usage{
prep_rodents(
  main = ".",
  datasets = prefab_rodent_datasets(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{datasets}{\code{character} vector of name(s) of rodent dataset(s) to include.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or not (and thus just the tidy messages).}
}
\value{
\code{list} of prepared \code{datasets}.
}
\description{
Create specified \code{datasets} using their associated function and arguments.
}
