% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prepare_covariates}
\alias{prepare_covariates}
\alias{prepare_historic_covariates}
\alias{prepare_forecast_covariates}
\title{Prepare Covariate Data for Casting}
\usage{
prepare_covariates(
  main = ".",
  settings = directory_settings(),
  origin = Sys.Date(),
  quiet = TRUE,
  verbose = FALSE
)

prepare_historic_covariates(
  main = ".",
  settings = directory_settings(),
  quiet = TRUE,
  verbose = FALSE
)

prepare_forecast_covariates(
  main = ".",
  origin = Sys.Date(),
  settings = directory_settings(),
  quiet = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}
}
\value{
\code{data.frame} of historical and/or -casted covariates that is also saved out to \code{settings$files$covariates} if indicated by \code{settings$save}.
}
\description{
Prepare and combine the historical and cast covariate data for a model run. \cr \cr 
             \code{prepare_historical_covariates} and \code{prepare_forecast_covariates} ready the historical and -casted covariates, respectively. \cr \cr
             \code{prepare_covariates} combines \code{prepare_historical_covariates} and \code{prepare_forecast_covariates}.
}
\details{
\code{prepare_historical_covariates} creates a \code{data.frame} of historical weather (\code{\link[portalr]{weather}}) and NDVI data (\code{\link[portalr]{ndvi}}). Automatically goes all the way back to the beginning of the available data, as the table is trimmed later as needed. NDVI data are generated by \code{\link[portalr]{ndvi}}. \cr \cr
         \code{prepare_forecast_covariates} combines the NMME climate forecasts (\code{\link{download_climate_forecasts}}) with NDVI forecasts using a seasonal auto-ARIMA model on the most resolved and raw data available via (\code{\link[portalr]{ndvi}}).
}
\examples{
 \donttest{
  setup_dir()
  prep_covariates()
 }
 
}
