% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prepare_historical_covariates}
\alias{prepare_historical_covariates}
\alias{prep_hist_covariates}
\alias{prep_weather_data}
\title{Prepare historical covariates data}
\usage{
prep_hist_covariates(main = ".", end_moon = NULL, quiet = TRUE,
  arg_checks = TRUE)

prep_weather_data(main = ".", arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{prep_hist_covariates}: historical covariate data table as a 
  \code{data.frame}. \cr \cr
 \code{prep_weather_data}: \code{data.frame} of historical weather data.
}
\description{
\code{prep_hist_covariates} creates a \code{data.frame} of historical 
 weather and NDVI data. Automatically goes all the way back to the 
 beginning of the available data, as the table is automatically trimmed 
 later. \cr \cr
 \code{prep_weather_data} creates a \code{data.frame} of historical 
  weather data using \code{\link[portalr]{weather}}. \cr \cr
 NDVI data are generated by \code{\link[portalr]{ndvi}}.
}
\examples{
 \donttest{   
  create_dir()
  fill_raw()
  prep_hist_covariates()
  prep_weather_data()
 }

}
