% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prepare_metadata}
\alias{prepare_metadata}
\title{Prepare a Model-Running Metadata List}
\usage{
prepare_metadata(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  cast_date = Sys.Date(),
  confidence_level = 0.95,
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{datasets}{\code{character} vector of name(s) of dataset(s) to include.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number of the last sample to be included. Default value is \code{NULL}, which equates to the most recently included sample.}

\item{start_moon}{\code{integer} (or integer \code{numeric}) newmoon number of the first sample to be included. Default value is \code{217}, corresponding to \code{1995-01-01}.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the number of timesteps forward a cast will cover.}

\item{cast_date}{\code{Date} from which future is defined (the origin of the cast). In the recurring forecasting, is set to today's date using \code{\link{Sys.Date}}.}

\item{confidence_level}{\code{numeric} confidence level used in summarizing model output. Must be between \code{0} and \code{1}.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or not (and thus just the tidy messages).}
}
\value{
\code{list} of casting metadata, which is also saved out as a YAML file (\code{.yaml}) if desired.
}
\description{
Sets up the metadata used for casting, in particular the matching of time period across the data sets. This should always be run after \code{\link{prepare_moons}}, \code{\link{prepare_rodents}}, and \code{\link{prepare_covariates}} before any model is run.
}
