% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{read_cast}
\alias{read_cast}
\alias{read_casts}
\title{Read in a cast file or multiple cast files and set the class}
\usage{
read_cast(main = ".", cast_type = "forecast", cast_date = NULL,
  verbose = FALSE, arg_checks = TRUE)

read_casts(main = ".", cast_type = "forecast", cast_dates = NULL,
  verbose = FALSE, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{cast_type}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"forecasts"}.}

\item{cast_date}{\code{Date} the predictions were made. Used to select the
file in the predictions subdirectory. Can only be length 1 or \code{NULL}. 
If \code{NULL} (default), selects the most recent -casts.}

\item{verbose}{\code{logical} indicator if all validation errors should
be reported.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}

\item{cast_dates}{\code{Date}s the predictions were made. Used to select 
the files in the predictions subdirectory. Can be length 1 or more and if 
\code{NULL} (default), selects all available -casts.}
}
\value{
\code{data.frame} of requested fore- or hindcast(s).
}
\description{
Read in a specified forecast or hindcast data file and ensure 
 its class attribute is appropriate for usage within the portalcasting 
 pipeline. Current not reliably coded for hindcasts. \cr \cr 
 \code{read_cast} only allows for a single date and defaults to
 the most recent. \cr \cr 
 \code{read_casts} allows for multiple dates and defaults to all.
}
\examples{
 \donttest{
  setup_dir()
  read_cast()
  read_casts()
 }

}
