% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{read_cast_output}
\alias{read_cast_output}
\alias{read_cast_tab}
\alias{read_cast_tabs}
\alias{read_cast_metadata}
\alias{read_model_fits}
\alias{read_model_casts}
\title{Read in cast output from a given cast}
\usage{
read_cast_tab(main = ".", cast_id = NULL, arg_checks = TRUE)

read_cast_tabs(main = ".", cast_ids = NULL, arg_checks = TRUE)

read_cast_metadata(main = ".", cast_id = NULL, arg_checks = TRUE)

read_model_fits(main = ".", cast_id = NULL, arg_checks = TRUE)

read_model_casts(main = ".", cast_id = NULL, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}

\item{cast_ids, cast_id}{\code{integer} (or integer \code{numeric}) value(s) 
representing the cast(s) of interest, as indexed within the directory in
the \code{casts} sub folder. See the casts metadata file 
(\code{casts_metadata.csv}) for summary information. If \code{NULL} (the
default), the most recently generated cast's output is read in. \cr
\code{cast_ids} can be NULL, one value, or more than one values,
\code{cast_id} can only be NULL or one value.}
}
\value{
\code{read_cast_tab}: \code{data.frame} of the \code{cast_tab}. \cr \cr
 \code{read_cast_tabs}: \code{data.frame} of the \code{cast_tab}s with
 a \code{cast_id} column added to distinguish among casts. \cr \cr
 \code{read_cast_metadata}: \code{list} of \code{cast_metadata}. \cr \cr
 \code{read_model_fits}: \code{list} of \code{model_fits}. \cr \cr
 \code{read_model_casts}: \code{list} of \code{model_casts}.
}
\description{
Read in the various output files of a cast or casts in the 
 casts sub directory. \cr \cr
 \code{read_cast_tab} retrieves the \code{cast_tab}. \cr \cr
 \code{read_cast_tabs} combines one or more \code{cast_tab}s. \cr \cr
 \code{read_cast_metadata} retrieves the \code{cast_metdata}. \cr \cr
 \code{read_model_fits} retrieves the \code{model_fits}. \cr \cr
 \code{read_model_casts} retrieves the \code{model_casts}.
}
\examples{
 \donttest{
  setup_dir()
  portalcast(models = c("AutoArima", "NaiveArima"), end_moons = 515:520)
  read_cast_tab(cast_id = 1)
  read_cast_tabs(cast_ids = 1:2)
  read_cast_metadata(cast_id = 1)
  read_model_fits(cast_id = 1)
  read_model_casts(cast_id = 1)
}

}
