% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{read_casts_metadata}
\alias{read_casts_metadata}
\title{Read in the casts metadata file}
\usage{
read_casts_metadata(main = ".", quiet = FALSE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\value{
Data requested.
}
\description{
Read in the casts metadata file. If the data file does not
 exist, an effort is made to create the file.
}
\examples{
 \donttest{
  setup_dir()
  read_casts_metadata()
 }

}
