% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{read_data}
\alias{read_data}
\title{Read in a data file and format it for specific class}
\usage{
read_data(tree, data_name)
}
\arguments{
\item{tree}{\code{dirtree}-class list. See \code{\link{dirtree}}.}

\item{data_name}{\code{character} representation of the data needed.
Current options include \code{"all"}, \code{"controls"},
\code{"covariates"}, \code{"moons"}, and \code{"metadata"}.}
}
\value{
Data requested with appropriate classes.
}
\description{
Read in a specified data file and ensure its class attribute
  is appropriate for usage within the portalcasting pipeline. Current 
  options include \code{"all"}, \code{"controls"}, \code{"covariates"},
  \code{"moons"}, and \code{"metadata"}.
}
\examples{
\dontrun{

setup_dir()
read_data(data_name = "all")
read_data(data_name = "controls")
read_data(data_name = "covariates")
read_data(data_name = "moons")
read_data(data_name = "metadata")
}

}
