% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{read_data}
\alias{read_data}
\alias{read_rodents}
\alias{read_covariates}
\alias{read_covariate_casts}
\alias{read_moons}
\alias{read_metadata}
\title{Read in a data file and format it for specific class}
\usage{
read_data(main = ".", data_name = NULL, level = "all")

read_rodents(main = ".", level = "all")

read_covariates(main = ".")

read_covariate_casts(main = ".")

read_moons(main = ".")

read_metadata(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{data_name}{\code{character} representation of the data needed.
Current options include \code{"rodents"}, \code{"covariates"}, 
\code{"covariate_forecasts"}, \code{"moons"}, and \code{"metadata"}.}

\item{level}{\code{character} representation of the grouping name used to
define the rodents. Standard options are \code{"all"} and 
\code{"controls"}.}
}
\value{
Data requested.
}
\description{
Read in a specified data file. \cr \cr
 Current options include \code{"rodents"}, \code{"covariates"},
 \code{"covariate_casts"}, \code{"moons"}, and \code{"metadata"}, which
 are available as calls to \code{read_data} with a specified 
 \code{data_name} or as calls to the specific \code{read_<data_name>} 
 functions (like \code{read_moons}).
}
\examples{
\donttest{
 setup_dir()
 read_data(data_name = "rodents")
 read_data(data_name = "controls")
 read_data(data_name = "covariates")
 read_data(data_name = "covariate_casts")
 read_data(data_name = "moons")
 read_data(data_name = "metadata")

 read_rodents(level = "all")
 read_controls()
 read_covariates()
 read_covariate_casts()
 read_moons()
 read_metadata()
}

}
