% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_output.R
\name{read_data}
\alias{read_data}
\alias{read_rodents_table}
\alias{read_rodents}
\alias{read_newmoons}
\alias{read_covariates}
\alias{read_climate_forecasts}
\alias{read_metadata}
\title{Read in and Format a Portalcasting Data File}
\usage{
read_data(
  main = ".",
  data_name = NULL,
  dataset = "all",
  datasets = prefab_datasets(),
  settings = directory_settings()
)

read_rodents_table(
  main = ".",
  dataset = "all",
  settings = directory_settings()
)

read_rodents(
  main = ".",
  datasets = prefab_datasets(),
  settings = directory_settings()
)

read_newmoons(main = ".", settings = directory_settings())

read_covariates(main = ".", settings = directory_settings())

read_climate_forecasts(main = ".", settings = directory_settings())

read_metadata(main = ".", settings = directory_settings())
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{data_name}{\code{character} representation of the data needed. Current options include \code{"rodents"}, \code{"rodents_table"}, \code{"covariates"}, \code{"forecast_covariates"},  \code{"historical_covariates"}, \code{"covariate_forecasts"}, \code{"newmoons"}, and \code{"metadata"}.}

\item{dataset, datasets}{\code{character} representation of the grouping name(s) used to define the rodents. Standard options are \code{"all"}, \code{"controls"}, and \code{"exclosures"}. \code{dataset} can only be length 1, \code{datasets} is not restricted in length.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}
}
\value{
Data requested.
}
\description{
Read in a specified data file.
}
