% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_output.R
\name{read_data}
\alias{read_data}
\alias{read_rodents_table}
\alias{read_rodents}
\alias{read_moons}
\alias{read_covariates}
\alias{read_forecast_covariates}
\alias{read_historical_covariates}
\alias{read_climate_forecasts}
\alias{read_metadata}
\title{Read in and Format a Portalcasting Data File}
\usage{
read_data(
  main = ".",
  data_name = NULL,
  rodent_dataset = "all",
  rodent_datasets = c("all", "controls"),
  settings = directory_settings()
)

read_rodents_table(
  main = ".",
  rodent_dataset = "all",
  settings = directory_settings()
)

read_rodents(
  main = ".",
  rodent_datasets = c("all", "controls"),
  settings = directory_settings()
)

read_moons(main = ".", settings = directory_settings())

read_covariates(main = ".", settings = directory_settings())

read_forecast_covariates(main = ".", settings = directory_settings())

read_historical_covariates(main = ".", settings = directory_settings())

read_climate_forecasts(main = ".", settings = directory_settings())

read_metadata(main = ".", settings = directory_settings())
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{data_name}{\code{character} representation of the data needed. Current options include \code{"rodents"}, \code{"rodents_table"}, \code{"covariates"}, \code{"covariate_forecasts"}, \code{"moons"}, and \code{"metadata"}.}

\item{rodent_dataset, rodent_datasets}{\code{character} representation of the grouping name(s) used to define the rodents. Standard options are \code{"all"} and \code{"controls"}. \code{dataset} can only be length 1, \code{datasets} is not restricted in length.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}
}
\value{
Data requested.
}
\description{
Read in a specified data file. \cr \cr
             Current options include \code{"rodents"} (produces a list), \code{"rodents_table"} (produces a rodents table), \code{"covariates"}, \code{"climate_forecasts"}, \code{"moons"}, and \code{"metadata"}, which  are available as calls to \code{read_data} with a specified  \code{data_name} or as calls to the specific \code{read_<data_name>}  functions (like \code{read_moons}). \cr \cr
             \code{read_cov_casts} reads in the current or (if no current version) local archived version of the covariate casts.
}
