% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_incompletes}
\alias{remove_incompletes}
\title{Remove any specific incomplete entries as noted by an NA}
\usage{
remove_incompletes(df, col_to_check)
}
\arguments{
\item{df}{\code{data.frame} table to be written out.}

\item{col_to_check}{A single \code{character} value of the column to use
to remove incomplete entries.}
}
\value{
df without any incomplete entries.
}
\description{
Remove any incomplete entries in a table, as determined by
  the presence of an \code{NA} entry in a specific column 
  (\code{col_to_check}).
}
