% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{remove_spp}
\alias{remove_spp}
\title{Remove species from the rodent data table}
\usage{
remove_spp(rodents, drop_spp = "PI")
}
\arguments{
\item{rodents}{Class-\code{rodents} \code{data.table} of rodent data.}

\item{drop_spp}{\code{character} vector of species name(s) (column header)
to drop from \code{data} or \code{NULL} to not drop any species.}
}
\value{
Trimmed \code{rodents}.
}
\description{
Remove species from the rodent data table
}
