% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodent_species.R
\name{rodent_species}
\alias{rodent_species}
\alias{all_species}
\alias{base_species}
\title{Rodent species abbreviations}
\usage{
rodent_species(set = NULL, nadot = FALSE, total = FALSE, type = "abbreviation")

all_species(nadot = FALSE, total = FALSE, type = "abbreviation")

base_species(nadot = FALSE, total = FALSE, type = "abbreviation")
}
\arguments{
\item{set}{\code{character} input of a specified set of species. Options include \code{all} (all species included) and \code{base} (all but PI).}

\item{nadot}{\code{logical} indicator if the dot should be added to the \code{"NA"} species name. Defaults to \code{FALSE}.}

\item{total}{\code{logical} value indicating if \code{"total"} should be added or not.}

\item{type}{\code{character} value indicating the output type. Current options include \code{'abbreviation'} (default, two-letter abbreviation), \code{'Latin'} (full scientific names), and  \code{'table'} a \code{data.frame} of the abbreviations and Latin names.}
}
\value{
\code{character} vector of species abbreviations.
}
\description{
Creates a simple \code{character} vector of abbreviations for the Portal \href{https://portal.naturecast.org/profiles.html}{Rodents}.
}
\examples{
 rodent_species()
 rodent_species(type = "Latin")
 rodent_species(type = "table")
 rodent_species(type = "table", total = TRUE)
 rodent_species(set = "all")
 all_species()
 all_species(nadot = TRUE)
 all_species(total = TRUE)
 base_species()

}
