% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{rodent_species}
\alias{rodent_species}
\alias{all_species}
\alias{base_species}
\alias{evalplot_species}
\title{Rodent species abbreviations}
\usage{
rodent_species(species = NULL, set = NULL, nadot = FALSE, total = FALSE)

all_species(species = NULL, nadot = FALSE, total = FALSE)

base_species(species = NULL, nadot = FALSE, total = FALSE)

evalplot_species(species = NULL, nadot = FALSE, total = FALSE)
}
\arguments{
\item{species}{\code{character}-valued vector of species names to include.}

\item{set}{\code{character} input of a specified set of species. Default entry (\code{"base"}) returns the standard set of all species included. Other options include \code{"wtotal"} (same as \code{"base"} but with "total" as well) and \code{"evalplot"} which only returns a subset of common species to be included in the evaluations plots.}

\item{nadot}{\code{logical} indicator if the dot should be added to the \code{"NA"} species name. Defaults to \code{FALSE}.}

\item{total}{\code{logical} value indicating if \code{"total"} should be added or not.}
}
\value{
\code{character} vector of species abbreviations.
}
\description{
Creates a simple \code{character} vector of abbreviations for the Portal \href{https://portal.naturecast.org/profiles.html}{Rodents}.
}
\examples{
 rodent_species()
 rodent_species(set = "all")
 rodent_species("BA")
 rodent_species("BA", set = "all")
 all_species()
 all_species(nadot = TRUE)
 all_species(total = TRUE)
 base_species()
 evalplot_species()

}
