% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{rodents_control}
\alias{rodents_control}
\alias{rodents_controls}
\title{Create control lists for generating rodents data tables}
\usage{
rodents_control(tmnt_type = NULL, name = tmnt_type,
  species = base_species(), total = TRUE,
  filename = "rodents_all.csv", start_moon = 217, end_moon = NULL,
  level = "Site", treatment = NULL, plots = "all", min_plots = 24,
  min_traps = 1, output = "abundance")

rodents_controls(tmnt_type = c("all", "controls"), name = tmnt_type,
  species = base_species(), total = TRUE, filename = NULL,
  start_moon = 217, end_moon = NULL, level = NULL,
  treatment = NULL, plots = NULL, min_plots = 24, min_traps = 1,
  output = "abundance")
}
\arguments{
\item{tmnt_type}{\code{character} value of the treatment type (currently
\code{"all"} or \code{"controls"}) used to enforce certain arguments.
Default within \code{rodents_control} (\code{tmnt_type = NULL}) allows 
easy full specification of user types.}

\item{name}{\code{character} value used to name the output element.
Default (\code{name = tmnt_type}) takes the name from \code{tmnt_type}.}

\item{species}{\code{character}-valued vector of species names to include 
in the forecasting data table. Defaults to all species but \code{"PI"}
via \code{\link{base_species}}}

\item{total}{\code{logical} value indicating if a total (sum across 
species) should be added or not. Only available if more than one species
is included.}

\item{filename}{\code{character} name of the file for saving the output.}

\item{start_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the first sample to be included. Default value is \code{217}, 
corresponding to \code{1995-01-01}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{level}{\code{character} indicating the type of summary:
\code{"Plot"}, \code{"Treatment"}, or \code{"Site"}. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}

\item{treatment}{\code{character} indicating the specific treatment(s) to
trim to if \code{level = "Treatment"}: \code{"control"},
\code{"exclosure"}, \code{"removal"}, or \code{"spectabs"} .}

\item{plots}{Specification of subset of plots. Can be a vector of 
\code{numeric} plots indicators or specific sets indicated by
\code{character} values: \code{"all"} plots or \code{"Longterm"} plots
(plots that have had the same treatment for the entire time series).}

\item{min_plots}{\code{integer} (or integer \code{numeric}) of the minimum 
number of plots surveyed for a survey to be used.Pipes directly to 
\code{\link[portalr]{summarize_rodent_data}}.}

\item{min_traps}{\code{integer} (or integer \code{numeric}) of the minimum 
number of traps collected for a plot to be used. Pipes directly to 
\code{\link[portalr]{summarize_rodent_data}}.}

\item{output}{\code{character} indicating the type of data:
\code{"abundance"}, \code{"biomass"}, or \code{"energy"}. Pipes 
directly to \code{\link[portalr]{summarize_rodent_data}}.}
}
\value{
\code{rodents_control}: \code{list} of specific arguments for 
 \code{\link{prep_rodents_table}}. \cr \cr 
 \code{rodents_controls}: \code{list} of \code{list}s (each produced by
 \code{rodents_control}) for input as \code{controls} in 
 \code{\link{prep_rodents}}.
}
\description{
Given the number of arguments into 
 \code{\link{prep_rodents_table}}, it helps to have a control \code{list}  
 to organize them. These functions produce those \code{list}s. \cr \cr
 \code{rodents_control} produces a control \code{list} for a single data 
 table (for input into \code{\link{prep_rodents_table}}). \cr \cr
 \code{rodents_controls} produces a \code{list} of control \code{list}s
 for multiple data tables (for input into 
 \code{\link{prep_rodents}}).
}
\examples{
 rodents_control("all")
 rodents_control("controls")
 rodents_controls(c("all", "controls"))

}
