% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{rodents_data}
\alias{rodents_data}
\title{Collect and process historic rodent data for -casting}
\usage{
rodents_data(moons = prep_moons(), options_rodents = rodents_options())
}
\arguments{
\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}

\item{options_rodents}{Class-\code{rodents_options} \code{list} of 
settings controlling the rodents data creation. See 
\code{\link{rodents_options}} for general creation and 
\code{\link{enforce_rodents_options}} for bulk setting of options for
each of the possible data table outputs.}
}
\value{
One of two class-\code{rodents} \code{data.frame}s, 
  \code{all} (abundances on all plots) or \code{controls} (abundances on
  control plots only).
}
\description{
Prepare a particular version of the rodent dataset, tailored 
  for forecasting or hindcasting. Available versions of the data currently
  include \code{all} (abundances on all plots) and \code{controls} 
  (abundances on control plots only)
}
