% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{rodents_data}
\alias{rodents_data}
\title{Collect and process historic rodent data}
\usage{
rodents_data(tree = dirtree(), moons = prep_moons(),
  data_options = rodents_options())
}
\arguments{
\item{tree}{the name tree of the portalcasting directory}

\item{moons}{current newmoon table}

\item{data_options}{}
}
\value{
a list of two dataframes, all plots and control plots
}
\description{
Prepare a particular version of the rodent dataset, tailored 
  for forecasting.
}
