% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{sandbox_welcome}
\alias{sandbox_welcome}
\title{Sandbox welcome}
\usage{
sandbox_welcome(tree = dirtree(), quiet = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress message should be
quieted.}
}
\description{
Create a welcome message for the sandbox
}
