% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{sandbox_welcome}
\alias{sandbox_welcome}
\title{Sandbox Welcome}
\usage{
sandbox_welcome(main = ".", quiet = FALSE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{quiet}{\code{logical} indicator if message should be quieted.}
}
\value{
Message is put to the console; \code{NULL} is returned.
}
\description{
Generates a special welcome message for sandboxes,
             displayed upon completion of \code{\link{setup_sandbox}}.
}
\examples{
 sandbox_welcome()

}
