% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{sandbox_welcome}
\alias{sandbox_welcome}
\title{Sandbox welcome}
\usage{
sandbox_welcome(main = ".", quiet = FALSE, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{quiet}{\code{logical} indicator if progress message should be
quieted.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{NULL}, as message is printed.
}
\description{
Create a welcome message for the sandbox
}
\examples{
 sandbox_welcome()

}
