% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{save_cast_output}
\alias{save_cast_output}
\title{Save cast output to files}
\usage{
save_cast_output(all, controls, model, main)
}
\arguments{
\item{all}{Output from a model function (e.g., 
\code{\link{AutoArima}}) run on the "all" data. Required to be a 
\code{list} with two elements named \code{forecast} and \code{aic}.}

\item{controls}{Output from a model function (e.g., 
\code{\link{AutoArima}}) run on the "controls" data. Required to be a 
\code{list} with two elements named \code{forecast} and \code{aic}.}

\item{model}{\code{character} value of the name of the model.}

\item{main}{\code{character} value of the name of the main component of
the directory tree.}
}
\description{
Save out the cast abundances and AIC tables from a given 
  set of models.
}
