% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{save_cast_output}
\alias{save_cast_output}
\title{Save cast output to files}
\usage{
save_cast_output(cast = NULL, main = ".", quiet = FALSE,
  arg_checks = TRUE)
}
\arguments{
\item{cast}{Output from a model function (e.g., 
\code{\link{AutoArima}}) run on the rodents data. Required to be a 
\code{list}.}

\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\description{
Save out the cast abundances and AIC tables from a given 
  set of models.
}
