% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{save_forecast_output}
\alias{save_forecast_output}
\title{Save forecast output to files}
\usage{
save_forecast_output(all, controls, name, metadata, temp_dir = pc_path("tmp",
  "~"))
}
\arguments{
\item{all}{ouput from the forecasting function for all data}

\item{controls}{ouput from the forecasting function for controls data}

\item{name}{model name for saving in files}

\item{metadata}{model metadata}

\item{temp_dir}{directory name for the temporary housing of predictions}
}
\value{
Nothing
}
\description{
Save out the forecast and AIC values from a given set of 
  forecasts on all and control-only rodent data
}
