% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{select_cast}
\alias{select_cast}
\title{Select the specific fore- or hindcast from a casts table}
\usage{
select_cast(casts, species = "total", level = "Controls",
  model = "Ensemble")
}
\arguments{
\item{species}{\code{character} value of the species code or \code{"total"}
for the total across species.}

\item{level}{\code{character} value of the level of interest (\code{"All"} 
or \code{"Controls"}).}

\item{model}{\code{character} value of the name (or \code{"Ensemble"}) of
the model to be plotted.}

\item{casttype}{\code{character} value of the type of -cast of model. Used
to select the file in the predictions subdirectory. Currently only 
reliably coded for \code{"forecasts"}.}
}
\description{
Given a \code{casts}-class \code{data.frame} of many models'
  predictions, select a specific one for use.
}
