% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{select_casts}
\alias{select_casts}
\title{Find casts that fit specifications}
\usage{
select_casts(main = ".", cast_ids = NULL, cast_groups = NULL,
  end_moons = NULL, models = NULL, data_sets = NULL, quiet = FALSE,
  include_interp = FALSE, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{cast_ids}{\code{integer} (or integer \code{numeric}) values 
representing the casts of interest, as indexed within the directory in
the \code{casts} sub folder. See the casts metadata file 
(\code{casts_metadata.csv}) for summary information.}

\item{cast_groups}{\code{integer} (or integer \code{numeric}) value
of the cast group to combine with an ensemble. If \code{NULL} (default),
the most recent cast group is ensembled.}

\item{end_moons}{\code{integer} (or integer \code{numeric}) 
newmoon numbers of the forecast origin. Default value is 
\code{NULL}, which equates to no selection with respect to 
\code{end_moon}.}

\item{models}{\code{character} values of the names of the models to 
include. Default value is \code{NULL}, which equates to no selection with 
respect to \code{model}.}

\item{data_sets}{\code{character} values of the rodent data sets to include
Default value is \code{NULL}, which equates to no selection with 
respect to \code{data_set}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{include_interp}{\code{logical} indicator of if the basic data set
names should also be inclusive of the associated interpolated data sets.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not.}
}
\value{
\code{data.frame} of the \code{cast_tab}.
}
\description{
Determines the casts that match user specifications. \cr
 Functionally, a wrapper on \code{\link{read_casts_metadata}} with 
 filtering for specifications that provides a simple user interface to
 the large set of available casts via the metadata.
}
\examples{
 \donttest{
  setup_dir()
  portalcast(models = c("AutoArima", "NaiveArima"), end_moons = 515:520)
  select_casts(models = "AutoArima")
}

}
