% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_casts.R
\name{select_casts}
\alias{select_casts}
\title{Find Casts that Fit Specifications}
\usage{
select_casts(
  main = ".",
  settings = directory_settings(),
  cast_ids = NULL,
  cast_groups = NULL,
  models = NULL,
  datasets = NULL,
  species = NULL,
  historic_end_newmoonnumbers = NULL,
  quiet = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{cast_ids}{\code{integer} (or integer \code{numeric}) values representing the casts of interest, as indexed within the directory in the \code{casts} sub folder. See the casts metadata file (\code{casts_metadata.csv}) for summary information.}

\item{cast_groups}{\code{integer} (or integer \code{numeric}) value of the cast groups to include. Default value is \code{NULL}, which equates to no selection with respect to \code{cast_group}.}

\item{models}{\code{character} values of the names of the models to include. Default value is \code{NULL}, which equates to no selection with respect to \code{model}.}

\item{datasets}{\code{character} values of the rodent data sets to include Default value is \code{NULL}, which equates to no selection with respect to \code{dataset}.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{end_moons}{\code{integer} (or integer \code{numeric}) newmoon numbers of the forecast origin. Default value is \code{NULL}, which equates to no selection with respect to \code{end_moon}.}
}
\value{
\code{data.frame} of the \code{cast_tab}.
}
\description{
Determines the casts that match user specifications. \cr
 Functionally, a wrapper on \code{\link{read_casts_metadata}} with filtering for specifications that provides a simple user interface to the large set of available casts via the metadata.
}
