% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{setup_dir}
\alias{setup_dir}
\alias{setup_production}
\alias{setup_sandbox}
\title{Create and Fill a Forecasting Directory}
\usage{
setup_dir(
  main = ".",
  settings = directory_settings(),
  models = prefab_models(),
  datasets = prefab_rodent_datasets(),
  quiet = FALSE,
  verbose = FALSE
)

setup_production(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  control_files = files_control(),
  PortalData_source = "gitub",
  PortalData_version = "latest",
  portalPredictions_source = "github",
  portalPredictions_version = "latest",
  climate_forecast_source = "NMME",
  climate_forecast_version = Sys.Date(),
  quiet = FALSE,
  verbose = TRUE,
  datasets = prefab_rodent_datasets()
)

setup_sandbox(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  control_files = files_control(),
  PortalData_source = "gitub",
  PortalData_version = "latest",
  portalPredictions_source = "github",
  portalPredictions_version = NULL,
  climate_forecast_source = "NMME",
  climate_forecast_version = Sys.Date(),
  datasets = prefab_rodent_datasets(),
  quiet = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{models}{\code{character} vector of name(s) of model(s) to include.}

\item{datasets}{\code{list} of datasets to be created using \code{\link{do.call}} on the defined functions.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or not (and thus just the tidy messages).}
}
\value{
\code{NULL}
}
\description{
Combines \code{\link{create_dir}} and \code{\link{fill_dir}} to create a ready-to-run (via \code{\link{portalcast}}) directory where indicated. \cr \cr
             \code{setup_production} creates a standard production directory. \cr \cr
             \code{setup_sandbox} creates a sandboxing directory. \cr \cr
}
\examples{
\donttest{
 setup_dir()
 setup_sandbox()
 setup_production()
}

}
