% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{setup_dir}
\alias{setup_dir}
\title{Set up a full forecasting directory}
\usage{
setup_dir(base = "~", main = "forecasting", subs = subdirs(),
  model = models(), quiet = FALSE)
}
\arguments{
\item{base}{name of the base directory where the forecasting 
directory should exist (will be created if it doesn't)}

\item{main}{name of the portalcasting directory}

\item{subs}{names of the portalcasting subdirectories}

\item{model}{names of model scripts to include}

\item{quiet}{logical indicator if progress messages should be quieted}
}
\value{
Nothing
}
\description{
Create and populate a full forecasting directory or any 
  missing components.
}
