% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{setup_dir}
\alias{setup_dir}
\title{Set up a full forecasting directory}
\usage{
setup_dir(options_all = all_options())
}
\arguments{
\item{options_all}{list containing all options available for controlling
the set up and population of the forecast directory (see 
\code{\link{all_options}})}
}
\value{
Nothing
}
\description{
Create and populate a full forecasting directory or any 
  missing components.
}
