% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{setup_dir}
\alias{setup_dir}
\alias{setup_sandbox}
\title{Set up a full forecasting directory}
\usage{
setup_dir(models = NULL, tree = dirtree(), quiet = FALSE,
  PortalData_control = list(), data_control = list(),
  predictions_control = list(), models_control = list())

setup_sandbox(models = NULL, tree = dirtree(), quiet = FALSE,
  PortalData_control = list(), data_control = list(download = TRUE),
  predictions_control = list(download = TRUE), models_control = list())
}
\arguments{
\item{models}{\code{character} vector of the names of models to include in
the pipeline. Defaults to \code{NULL}, which retrieves the model names 
from \code{models_control}, so the models returned by the default
settings of \code{\link{model_names}}.}

\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{PortalData_control}{A \code{list} of arguments to control the filling
of the PortalData subdirectory. See \code{\link{PortalData_control}}. 
Values not set assume defaults.}

\item{data_control}{A \code{list} of arguments to control the filling of
the data subdirectory. See \code{\link{data_control}}. 
Values not set assume defaults.}

\item{predictions_control}{A \code{list} of arguments to control the 
filling of the predictions subdirectory. See 
\code{\link{predictions_control}}. Values not set assume defaults.}

\item{models_control}{A \code{list} of arguments to control the 
filling of the models subdirectory. See \code{\link{models_control}}.
Arguments not specified assume default values.}
}
\description{
Create (via \code{\link{create_dir}}) and populate (via 
 \code{\link{fill_dir}}) a full portalcasting directory, with 
 specifications controlled via list inputs. Default behavior is
 to setup the directory for using all of the "prefab" models
 (AutoArima, ESSS, nbGARCH, nbsGARCH, and pevGARCH; see the
 \href{https://bit.ly/2xP9jKI}{model description site}). \cr \cr
 \code{setup_sandbox} creates a portalcasting directory with defaults set
 for optimal sandboxing experience (see Details).
}
\details{
The default behavior is to over-write existing files, which can
 be prevented for specific components by setting the control argument
 \code{update} to \code{FALSE}. \cr \cr
 Sandboxing: \code{setup_dir}'s defaults are designed for working on the
 forecasting server, namely it doesn't download the most recent set of 
 forecasting files because it has them already. However, sandbox users
 will want all files (including previous covariate forecasts and previous
 rodent forecasts) that the forecasting server has. Thus, default 
 \code{setup_sandbox} is functionally the same as \code{setup_dir} with the
 two \code{download} arguments (in \code{predictions_control} and
 \code{data_control}) set to \code{TRUE}.
}
\examples{
\dontrun{

 setup_dir()
 setup_sandbox()
}

}
