% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{setup_dir}
\alias{setup_dir}
\alias{setup_production}
\alias{setup_sandbox}
\title{Create and fill a forecasting directory}
\usage{
setup_dir(
  main = ".",
  PortalData_version = "latest",
  PortalData_source = "github",
  portalPredictions_version = NULL,
  portalPredictions_source = "github",
  climate_forecast_source = "NMME",
  quiet = FALSE
)

setup_production(...)

setup_sandbox(...)
}
\arguments{
\item{main}{\code{character} value defining the main component of the 
portalcasting directory tree. Default value (\code{"."}) 
puts the directory in the present location.}

\item{quiet}{\code{logical} indicator if messages should be quieted.}
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Combines \code{\link{create_dir}} and \code{\link{fill_dir}}
             to create a ready-to-run (via \code{\link{portalcast}})
             directory where indicated. 
             \cr \cr
             \code{setup_production} creates a standard production 
             directory for use in the automated pipeline. 
             \cr \cr
             \code{setup_sandbox} creates a sandboxing directory for 
             exploration, tinkering, etc. with the code outside of 
             productions. 
             \cr \cr
}
\examples{
 \donttest{

  setup_dir("./portalcasting")
  setup_sandbox("./sandbox")
  setup_production("./production")

 }

}
