% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{setup_dir}
\alias{setup_dir}
\title{Set up a full forecasting directory}
\usage{
setup_dir(options_all = all_options())
}
\arguments{
\item{options_all}{Class-\code{all-options} list containing all options 
available for controlling the set up and population of the directory (see 
\code{\link{all_options}}).}
}
\description{
Create (via \code{\link{create_dir}}) and populate (via 
  \code{\link{fill_dir}}) a full portalcasting directory. Presently, this
  function overwrites any existing components of the directory. Future
  iterations will allow for updating, overwriting, or neither.
}
