% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{setup_dir}
\alias{setup_dir}
\alias{setup_production}
\alias{setup_sandbox}
\title{Create and fill a forecasting directory}
\usage{
setup_dir(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  controls_rodents = rodents_controls(),
  control_climate_dl = climate_dl_control(),
  control_files = files_control(),
  PortalData_source = "gitub",
  PortalData_version = "latest",
  portalPredictions_source = "github",
  portalPredictions_version = NULL,
  climate_forecast_source = "NMME",
  climate_forecast_version = Sys.Date(),
  data_sets = prefab_data_sets(),
  bline = TRUE,
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

setup_production(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  controls_rodents = rodents_controls(),
  control_climate_dl = climate_dl_control(),
  control_files = files_control(),
  PortalData_source = "gitub",
  PortalData_version = "latest",
  portalPredictions_source = "github",
  portalPredictions_version = "latest",
  climate_forecast_source = "NMME",
  climate_forecast_version = Sys.Date(),
  quiet = FALSE,
  verbose = TRUE,
  data_sets = prefab_data_sets(),
  arg_checks = TRUE
)

setup_sandbox(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  controls_rodents = rodents_controls(),
  control_climate_dl = climate_dl_control(),
  control_files = files_control(),
  PortalData_source = "gitub",
  PortalData_version = "latest",
  portalPredictions_source = "github",
  portalPredictions_version = NULL,
  climate_forecast_source = "NMME",
  climate_forecast_version = Sys.Date(),
  data_sets = prefab_data_sets(),
  quiet = FALSE,
  verbose = TRUE,
  arg_checks = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{models}{\code{character} vector of name(s) of model(s) to 
include.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{start_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the first sample to be included. Default value is \code{217}, 
corresponding to \code{1995-01-01}.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the
number of timesteps forward a cast will cover.}

\item{confidence_level}{\code{numeric} confidence level used in 
summarizing model output. Must be between \code{0} and \code{1}.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}

\item{controls_model}{Additional controls for models not in the prefab
set. \cr 
A \code{list} of a single model's script-writing controls or a
\code{list} of \code{list}s, each of which is a single model's 
script-writing controls. \cr 
Presently, each model's script writing controls
should include three elements: \code{name} (a \code{character} value of 
the model name), \code{covariates} (a \code{logical} indicator of if the 
model needs covariates), and \code{lag} (an \code{integer}-conformable 
value of the lag to use with the covariates or \code{NA} if 
\code{covariates = FALSE}). \cr 
If only a single model is added, the name of 
the model from the element \code{name} will be used to name the model's
\code{list} in the larger \code{list}. If multiple models are added, each
element \code{list} must be named according to the model and the
\code{name} element. \cr}

\item{controls_rodents}{Control \code{list} or \code{list} of \code{list}s 
(from \code{\link{rodents_controls}}) specifying the structuring of the 
rodents tables. See \code{\link{rodents_controls}} for details.}

\item{control_climate_dl}{\code{list} of specifications for the download, 
which are sent to \code{\link{NMME_urls}} to create the specific URLs. See
\code{\link{climate_dl_control}}.}

\item{control_files}{\code{list} of names of the folders and files within
the sub directories and saving strategies (save, overwrite, append, etc.).
Generally shouldn't need to be edited. See \code{\link{files_control}}.}

\item{PortalData_source, PortalData_version}{\code{character} values for the
source and version of the Portal Data to download. 
\cr \cr 
Default values retrieve the latest data from github.
\cr \cr 
See \code{\link[portalr]{download_observations}}.}

\item{portalPredictions_source, portalPredictions_version}{\code{character} 
values for the source and version of the archive to download.
\cr \cr 
Default values point to github, but \code{version = NULL} indicates
no download.
\cr \cr 
See \code{\link{download_archive}}.}

\item{data_sets}{\code{list} of data_sets to be created using 
\code{\link{do.call}} on the defined functions. 
\cr \cr
For example, \code{\link{prefab_data_sets}} use 
\code{\link{prepare_rodent_data}}.}

\item{bline}{\code{logical} indicator if a horizontal break lines should be
included in messages.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out
all of the information or not (and thus just the tidy messages).}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}

\item{climate_source, climate_version}{\code{character} values for the
source and version of the climate forecasts to download.
\cr \cr 
Default values retrieve the current day's forecast from the
Northwest Knowledge Network's North American Multi-Model Ensemble 
(NMME) climate forecasts.
\cr \cr 
See \code{\link{download_climate_forecasts}}.}
}
\value{
\code{NULL}
}
\description{
Combines \code{\link{create_dir}} and \code{\link{fill_dir}}
 to create a ready-to-run (via \code{\link{portalcast}}) directory where
 indicated. \cr \cr
 \code{setup_production} creates a standard production directory. \cr \cr
 \code{setup_sandbox} creates a sandboxing directory. \cr \cr
}
\examples{
\donttest{
 setup_dir()
 setup_sandbox()
 setup_production()
}

}
