% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{step_hind_forward}
\alias{step_hind_forward}
\title{Step the hind_step values forward}
\usage{
step_hind_forward(options_all = all_options())
}
\arguments{
\item{options_all}{all options control list}
}
\value{
updated options_all
}
\description{
Step all of the hind_step values in a full options control 
  list forward one
}
