% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{step_hind_forward}
\alias{step_hind_forward}
\title{Step the hind_step values forward}
\usage{
step_hind_forward(options_all = all_options())
}
\arguments{
\item{options_all}{\code{all_options}-class \code{list} of options 
controlling the portalcasting directory. See \code{\link{all_options}}.}
}
\value{
Updated \code{options_all} \code{list} for the next -cast.
}
\description{
Iterate the \code{hind_step} elements of options \code{list}s
  through subsequent-to-the-first -cast(s) within \code{\link{casts}}.
}
