% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{subdirs}
\alias{subdirs}
\title{Define the names of the subdirectories in a forecasting directory}
\usage{
subdirs(type = "portalcasting", subs = NULL)
}
\arguments{
\item{type}{\code{charachter} name for quick generation of subdirectory
vector. Presently only defined for the default input 
\code{"portalcasting"}.}

\item{subs}{Either [1] names of additional subdirectories to add to the
vector set up by \code{type} or [2] an optional way to input all 
subdirectory names (requires setting \code{type = NULL}).}
}
\value{
Class-\code{subdirs} vector of character elements of 
  subdirectory names.
}
\description{
Produces a vector of subdirectory names that is of class
  \code{subdirs}, and which should generally contain the elements
  \code{"predictions"}, \code{"models"}, \code{"PortalData"}, 
  \code{"data"}, and \code{"tmp"}. It is generally not advised to change
  the subdirectory structure at this time.
}
