% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{subdirs}
\alias{subdirs}
\title{Define the names of the subdirectories in a forecasting directory}
\usage{
subdirs(subs_names = NULL, subs_type = "prefab", arg_checks = TRUE)
}
\arguments{
\item{subs_names}{Either [1] names of additional subdirectories to add to
the vector set up by \code{type} or [2] an optional way to input all 
subdirectory names (requires \code{subs_type = NULL}).}

\item{subs_type}{\code{character} name for quick generation of subdirectory
vector. Presently only defined for \code{"prefab"}, or the setting
\code{NULL} which allows for complete customization.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{character} vector of subdirectory names.
}
\description{
Produces a vector of subdirectory names that is of class
  \code{subdirs}, and which should generally contain the elements
  \code{"predictions"}, \code{"models"}, \code{"PortalData"}, 
  \code{"data"}, and \code{"tmp"}. It is generally not advised to change
  the subdirectory structure at this time.
}
\examples{
 subdirs()

}
