% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{target_newmoons}
\alias{target_newmoons}
\title{Determine which newmoon numbers fall into a range of times for 
 casting}
\usage{
target_newmoons(main = ".", moons = prep_moons(main = main),
  end_moon = NULL, lead_time = 12, cast_date = Sys.Date())
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the
number of timesteps forward a cast will cover.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}
}
\value{
\code{numeric} vector of the newmoon numbers targeted by the date
 window.
}
\description{
Based on the forecast origin and the lead time, determine 
 which moons should be included.
}
\examples{
 \donttest{
  create_dir()
  fill_dir()
  target_newmoons()
 }

}
