% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_moons.R
\name{target_newmoons}
\alias{target_newmoons}
\alias{last_newmoon}
\title{Determine specific newmoon numbers}
\usage{
target_newmoons(main = ".", moons = NULL, end_moon = NULL,
  lead_time = 12, cast_date = Sys.Date(), arg_checks = TRUE)

last_newmoon(main = ".", moons = NULL, cast_date = Sys.Date(),
  arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{moons}{Moons \code{data.frame}. See \code{\link{prep_moons}}.}

\item{end_moon}{\code{integer} (or integer \code{numeric}) newmoon number 
of the last sample to be included. Default value is \code{NULL}, which 
equates to the most recently included sample.}

\item{lead_time}{\code{integer} (or integer \code{numeric}) value for the
number of timesteps forward a cast will cover.}

\item{cast_date}{\code{Date} from which future is defined (the origin of
the cast). In the recurring forecasting, is set to today's date
using \code{\link{Sys.Date}}.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{target_newmoons}: \code{numeric} vector of the newmoon numbers 
 targeted by the date window.
 \code{last_newmoon}:  \code{numeric} value of the last newmoon number.
}
\description{
\code{target_newmoons}: based on the forecast origin and the lead time, 
 determines which moons should be included. \cr \cr
 \code{last_newmoon}: based on the forecast origin date, determine its
 newmoon number.
}
\examples{
 \donttest{
  create_dir()
  fill_dir()
  target_newmoons()
  last_newmoon()
 }

}
