% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{template_nbgarch}
\alias{template_nbgarch}
\title{Create template nbGARCH text}
\usage{
template_nbgarch()
}
\value{
text of the template nbGARCH script
}
\description{
Creates the text for a template nbGARCH script
}
