% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_script_generation.R
\name{template_pevgarch}
\alias{template_pevgarch}
\title{Create template pevGARCH text}
\usage{
template_pevgarch(tree = dirtree())
}
\arguments{
\item{tree}{directory tree}
}
\value{
text of the template pevGARCH script
}
\description{
Creates the text for a template pevGARCH script
}
