% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{transfer_trapping_table}
\alias{transfer_trapping_table}
\title{Transfer the trapping table to the data folder}
\usage{
transfer_trapping_table(options_data = data_options())
}
\arguments{
\item{options_data}{data options}
}
\value{
nothing
}
\description{
Move the trapping table from the raw sub directory to the
  for-use sub directory
}
