% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{transfer_trapping_table}
\alias{transfer_trapping_table}
\title{Transfer the trapping table to the data folder}
\usage{
transfer_trapping_table(tree = dirtree(), quiet = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}. An optional format for inputting the file tree
structure. If used, overrides \code{base}, \code{main}, and \code{subs}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\description{
Move the trapping table from the PortalData sub directory to 
  the data subdirectory for use.
}
