% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{trim_moons_fcast}
\alias{trim_moons_fcast}
\title{Trim the moons table for covariate forecasting}
\usage{
trim_moons_fcast(moons, data_options)
}
\arguments{
\item{moons}{moon data}

\item{data_options}{covariate data options}
}
\value{
a trimmed moons data table
}
\description{
creating forecasts
}
