% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{trim_moons_fcast}
\alias{trim_moons_fcast}
\title{Trim the moons table for covariate forecasting}
\usage{
trim_moons_fcast(moons, control = list())
}
\arguments{
\item{moons}{Class-\code{moons} type of a \code{data.frame} containing the
moons and date data. See \code{\link{prep_moons}}.}

\item{control}{\code{list} of options for the covariate data. 
See \code{\link{covariates_control}}.}
}
\value{
A trimmed \code{moons} \code{data.frame}.
}
\description{
Covariate forecasting requires a moons data table that is
  trimmed to the moons of interest only.
}
