% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_covariates.R
\name{trim_moons_fcast}
\alias{trim_moons_fcast}
\title{Trim the moons table for covariate forecasting}
\usage{
trim_moons_fcast(moons, options_covariates)
}
\arguments{
\item{moons}{Class-\code{moons} type of a \code{data.frame} containing the
moons and date data. See \code{\link{prep_moons}}.}

\item{options_covariates}{Class \code{covariates_options} options 
\code{list}. See \code{\link{covariates_options}}.}
}
\value{
A trimmed \code{moons} \code{data.frame}.
}
\description{
Covariate forecasting requires a moons data table that is
  trimmed to the moons of interest only.
}
