% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{trim_treatment}
\alias{trim_treatment}
\title{Remove the treatment column from the rodent data table}
\usage{
trim_treatment(rodents, options_rodents = rodents_options())
}
\arguments{
\item{rodents}{Class-\code{rodents} \code{data.table} of rodent data.}

\item{options_rodents}{Class-\code{rodents_options} \code{list} of 
settings controlling the rodents data creation. See 
\code{\link{rodents_options}}}
}
\value{
Trimmed \code{data}.
}
\description{
Remove the treatment column from the rodent data table if 
  needed (in the case of controls, for example, where the data have been
  reduced already so the column is vestigial).
}
