% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rodents.R
\name{trim_treatment}
\alias{trim_treatment}
\title{Remove the treatment column from the rodent data table}
\usage{
trim_treatment(data, options_rodents = rodents_options())
}
\arguments{
\item{data}{rodent data table}

\item{options_rodents}{rodent data options list}
}
\value{
trimmed data
}
\description{
Remove the treatment column from the rodent data table if 
  needed (in the case of controls, for example, where the data have been
  reduced already so the column is vestigial)
}
