% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{update_covariates}
\alias{update_covariates}
\title{Update the covariate data for a step in a hindcast}
\usage{
update_covariates(moons, tree = dirtree(), quiet = FALSE,
  control = list())
}
\arguments{
\item{moons}{Class-\code{moons} \code{data.frame} containing the historic 
and future newmoons, as produced by \code{\link{prep_moons}}.}

\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{control}{A \code{list} of arguments to control the coviariates
data. See \code{\link{covariates_control}}. Values not set assume 
defaults.}
}
\value{
An updated \code{covariates}-class \code{data.frame} with needed 
  forecasted covariates associated with the focal forecast or hindcast 
  newmoon.
}
\description{
Given an updated \code{options_covariates$hind_step}, update
  the covariate data.
}
