% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{update_covariates}
\alias{update_covariates}
\title{Update the covariate data for a step in a hindcast}
\usage{
update_covariates(moons, options_covariates)
}
\arguments{
\item{moons}{current newmoon table}

\item{options_covariates}{covariates options list}
}
\value{
nothing
}
\description{
Given an updated hind_step, update the covariate data
}
