% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{update_covfcast_options}
\alias{update_covfcast_options}
\title{Update the data options for covariate forecasts based on existing 
  data}
\usage{
update_covfcast_options(data_options, hist_data, moons)
}
\arguments{
\item{data_options}{covariate data options}

\item{hist_data}{historical covariate data}

\item{moons}{current newmoon table}
}
\value{
control options list for covariates
}
\description{
Update the covariate data control options based on the 
  historical covariate data and moon data
}
