% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{update_data}
\alias{update_data}
\title{Update the data for each subsequent step in a hindcast}
\usage{
update_data(options_data)
}
\arguments{
\item{options_data}{data options list}
}
\value{
nothing
}
\description{
Rather than re-create the full rodent and covariate files
  again, simply step back in time through the hindcast's end vector 
  and trim the last "historical" sample off
}
