% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{update_data}
\alias{update_data}
\title{Update the data for each subsequent step in a hindcast}
\usage{
update_data(tree = dirtree, quiet = FALSE, control = list())
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. Consisting of 
\code{base}, \code{main}, and \code{subs} elements. See
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{control}{A \code{list} of arguments to control the filling of
the data subdirectory. See \code{\link{data_control}}. 
Values not set assume defaults.}
}
\description{
Rather than re-create the full rodent and covariate files
  again, simply step back in time through the hindcast's end vector 
  and trim the last "historical" sample off.
}
