% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{verify}
\alias{verify}
\title{Verify that a required folder exists}
\usage{
verify(path = NULL, level = NULL)
}
\arguments{
\item{path}{\code{character} vector of the folder paths.}

\item{level}{\code{character} vector of the names of the levels in the
hierarchy at which the folders will exist or the folder names. For the
purposes of messaging and not inherently required to be non-\code{NULL}.}
}
\value{
Throws an error if any of the folders do not exist, otherwise
 \code{NULL}.
}
\description{
If any of the folders (specified by \code{path} and named by
 \code{level}) do not exist, the operation throws an error.
}
\examples{
 tree <- dirtree()
 base <- base_path(tree)
 verify(base, "base")

}
