% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{verify}
\alias{verify}
\title{Verify that folders exist}
\usage{
verify(paths = NULL)
}
\arguments{
\item{paths}{\code{character} vector of the folder paths.}
}
\value{
throws an error if any of the folders do not exist, 
 otherwise \code{NULL}.
}
\description{
throws an error if any of the folders (specified by 
 \code{path} and named by \code{level}) do not exist. \cr \cr
}
\examples{
  verify(".")

}
