% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{verify_and_create}
\alias{verify_and_create}
\alias{verify}
\alias{create}
\title{Verify that folders exist and create folders}
\usage{
verify(paths = NULL)

create(paths = NULL)
}
\arguments{
\item{paths}{\code{character} vector of the folder paths.}
}
\value{
\code{verify}: throws an error if any of the folders do not exist, 
 otherwise \code{NULL}.
 \code{create}: \code{NULL}.
}
\description{
\code{verify} throws an error if any of the folders (specified by 
 \code{path} and named by \code{level}) do not exist. \cr \cr
 \code{create} creates a requested folder if it does not already exist.
}
\examples{
 \donttest{
  mainp <- main_path()
  create(mainp, "main")
  verify(mainp, "main")
 }

}
