% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_forecasts.R
\name{verify_cast}
\alias{verify_cast}
\alias{cast_is_valid}
\title{Verify that a read-in cast file is formatted appropriately}
\usage{
verify_cast(cast, verbose = FALSE)

cast_is_valid(cast, verbose = FALSE)
}
\arguments{
\item{cast}{\code{data.frame} -cast file read in.}

\item{verbose}{\code{logical} indicator if details of failure should
be printed}
}
\value{
\code{verify_cast}: \code{cast} as read in (as long as it is 
  valid). \cr \cr
  \code{cast_is_valid}: \code{logical} of if the -cast is formatted 
  properly.
}
\description{
Ensure that a -cast file that has been read in is formatted
  according to \href{https://bit.ly/2H2z3Jb}{specifications.}
}
