% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{verify_models}
\alias{verify_models}
\title{Verify that models requested to forecast or hindcast with exist}
\usage{
verify_models(options_cast = cast_options())
}
\arguments{
\item{options_cast}{Class-\code{cast_options} \code{list} containing the
hind- or forecasting options. See \code{\link{cast_options}}.}
}
\description{
Verify that models requested have scripts in the models 
  subdirectory.
}
