% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{verify_models}
\alias{verify_models}
\title{Verify that models requested to forecast or hindcast with exist}
\usage{
verify_models(main = ".", models = prefab_models(), quiet = FALSE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{models}{\code{character} vector of the names of models to verify.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}
}
\description{
Verify that models requested have scripts in the models 
  subdirectory.
}
\examples{
 \donttest{
  create_dir()
  fill_models()
  verify_models()
 }

}
