% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{verify_models}
\alias{verify_models}
\title{Verify that models requested to forecast or hindcast with exist}
\usage{
verify_models(main = ".", models = prefab_models(), quiet = FALSE,
  arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{models}{\code{character} vector of the names of models to verify.}

\item{quiet}{\code{logical} indicator if progress messages should be
quieted.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\description{
Verify that models requested have scripts in the models 
  subdirectory.
}
\examples{
 \donttest{
  create_dir()
  fill_models()
  verify_models()
 }

}
