% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalcast.R
\name{verify_models}
\alias{verify_models}
\title{Verify that models to forecast or hindcast with exist}
\usage{
verify_models(options_cast = cast_options())
}
\arguments{
\item{options_cast}{casting options}
}
\value{
nothing
}
\description{
Verify that models requested are available
}
