% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{verify_PortalData}
\alias{verify_PortalData}
\title{Verify that the PortalData subdirectory is present and has required 
  data}
\usage{
verify_PortalData(tree = dirtree(), foldername = "Rodents",
  filename = "moon_dates.csv", quiet = FALSE, control = list())
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{foldername}{\code{character}-valued vector of name(s) of the 
folder(s) to use for specific checking. Default (\code{Rodents}) 
settings use the moon dates table in the Rodents folder.}

\item{filename}{\code{character}-valued vector of name(s) of the file(s) to
use for specific checking. Default (\code{moon_dates.csv})
settings use the moon dates table in the Rodents folder.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{control}{A \code{list} of arguments to control the filling
of the PortalData subdirectory. See \code{\link{PortalData_control}}. 
Arguments not specified assume default values.}
}
\description{
Check that the PortalData subdirectory exists and has the
  needed specified file(s) within the "Rodents" subdirectory. If any of
   the file(s) is(/are) not present, the (entire) directory is filled.
}
\examples{
\dontrun{

setup_dir()
verify_PortalData()
}

}
