% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{verify_PortalData}
\alias{verify_PortalData}
\title{Verify that the PortalData subdirectory is present and has required 
  data}
\usage{
verify_PortalData(tree = dirtree(), filename = "moon_dates.csv",
  quiet = FALSE)
}
\arguments{
\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{filename}{\code{character}-valued vector of name(s) of the file(s) to
specifically check.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}
}
\description{
Check that the PortalData subdirectory exists and has the
  needed specified file(s) within the "Rodents" subdirectory. If any of
   the file(s) is(/are) not present, the (entire) directory is filled.
}
