% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp_ui.R
\name{web app ui}
\alias{web app ui}
\alias{portal_forecast_ui}
\alias{page_main_panel}
\alias{page_title_panel}
\alias{page_subtitle_panel}
\alias{about_tab}
\alias{models_tab}
\alias{profiles_tab}
\alias{forecast_tab}
\alias{forecast_tab_input_selection_checks_row}
\alias{forecast_tab_input_selection_row}
\alias{forecast_tab_input_selection_row_species}
\alias{forecast_tab_input_selection_row_dataset}
\alias{forecast_tab_input_selection_row_model}
\alias{forecast_tab_input_selection_row_historic_end_newmoonnumber}
\alias{evaluation_tab}
\alias{evaluation_tab_input_selection_checks_row}
\alias{evaluation_tab_input_selection_row}
\alias{evaluation_tab_input_selection_row_species}
\alias{evaluation_tab_input_selection_row_dataset}
\alias{evaluation_tab_input_selection_row_model}
\alias{evaluation_tab_input_selection_row_historic_end_newmoonnumber}
\alias{evaluation_tab_input_selection_row_newmoonnumber}
\alias{global_list}
\title{Generate the User Interface for the Web App}
\usage{
portal_forecast_ui(global = global_list())

page_main_panel(global = global_list())

page_title_panel()

page_subtitle_panel()

about_tab()

models_tab()

profiles_tab()

forecast_tab(global = global_list())

forecast_tab_input_selection_checks_row()

forecast_tab_input_selection_row(global = global_list())

forecast_tab_input_selection_row_species(global = global_list())

forecast_tab_input_selection_row_dataset(global = global_list())

forecast_tab_input_selection_row_model(global = global_list())

forecast_tab_input_selection_row_historic_end_newmoonnumber(
  global = global_list()
)

evaluation_tab(global = global_list())

evaluation_tab_input_selection_checks_row()

evaluation_tab_input_selection_row(global = global_list())

evaluation_tab_input_selection_row_species(global = global_list())

evaluation_tab_input_selection_row_dataset(global = global_list())

evaluation_tab_input_selection_row_model(global = global_list())

evaluation_tab_input_selection_row_historic_end_newmoonnumber(
  global = global_list()
)

evaluation_tab_input_selection_row_newmoonnumber(global = global_list())

global_list(main = ".")
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{`global`}{A \code{list} of global values for the app.}
}
\value{
A UI definition or component shiny tags.
}
\description{
\code{portal_forecast_ui} constructs the user interface (UI) for the web application by updating the static pages (models and rodent profiles) then running \code{\link[shiny:fluidPage]{fluidPage}} on the UI components. \cr \cr
See \code{Details} for hierarchy of functions.
}
\details{
The UI is hierarchical built as:
\itemize{
\item portal_forecast_ui
\itemize{
\item page_title_panel
\item page_subtitle_panel
\item page_main_panel
\itemize{
\item forecast_tab
\itemize{
\item forecast_tab_input_selection_row
\itemize{
\item forecast_tab_input_selection_row_species
\item forecast_tab_input_selection_row_dataset
\item forecast_tab_input_selection_row_model
\item forecast_tab_input_selection_row_historic_end_newmoonnumber
}
\item forecast_tab_input_selection_checks_row  # commented out, but available for checking reactive inputs in dev
\item plot_cast_ts
\item plot_cast_point
}
\item evaluation_tab
\itemize{
\item evaluation_tab_input_selection_row
\itemize{
\item evaluation_tab_input_selection_row_species
\item evaluation_tab_input_selection_row_dataset
\item evaluation_tab_input_selection_row_model
\item evaluation_tab_input_selection_row_historic_end_newmoonnumber
\item evaluation_tab_input_selection_row_newmoonnumber
}
\item evaluation_tab_input_selection_checks_row  # commented out, but available for checking reactive inputs in dev
\item plot_cast_point
\item plot_casts_cov_RMSE
}
\item about_tab
\itemize{
\item htmltools::includeMarkdown
}
\item models_tab
\itemize{
\item htmltools::includeHTML
}
\item profiles_tab
\itemize{
\item htmltools::includeHTML
}
}
}
}
}
