% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp_ui.R
\name{web app ui}
\alias{web app ui}
\alias{portal_forecast_ui}
\alias{page_main_panel}
\alias{page_title_panel}
\alias{page_subtitle_panel}
\alias{about_tab}
\alias{models_tab}
\alias{profiles_tab}
\alias{forecast_tab}
\alias{forecast_tab_input_selection_checks_row}
\alias{forecast_tab_input_selection_row}
\alias{forecast_tab_input_selection_row_species}
\alias{forecast_tab_input_selection_row_dataset}
\alias{forecast_tab_input_selection_row_model}
\alias{forecast_tab_input_selection_row_historic_end_newmoonnumber}
\alias{evaluation_tab}
\alias{evaluation_tab_input_selection_checks_row}
\alias{evaluation_tab_input_selection_row}
\alias{evaluation_tab_input_selection_row_species}
\alias{evaluation_tab_input_selection_row_dataset}
\alias{evaluation_tab_input_selection_row_model}
\alias{evaluation_tab_input_selection_row_historic_end_newmoonnumber}
\alias{evaluation_tab_input_selection_row_newmoonnumber}
\alias{covariates_tab}
\alias{data_sources_section}
\alias{portal_project_href}
\alias{portal_data_href}
\alias{portal_weather_href}
\alias{portal_ndvi_href}
\alias{write_rodent_profiles_tab_html}
\alias{write_models_tab_html}
\title{Generate the User Interface for the Web App}
\usage{
portal_forecast_ui(global = global_list())

page_main_panel(global = global_list())

page_title_panel()

page_subtitle_panel()

about_tab()

models_tab()

profiles_tab()

forecast_tab(global = global_list())

forecast_tab_input_selection_checks_row()

forecast_tab_input_selection_row(global = global_list())

forecast_tab_input_selection_row_species(global = global_list())

forecast_tab_input_selection_row_dataset(global = global_list())

forecast_tab_input_selection_row_model(global = global_list())

forecast_tab_input_selection_row_historic_end_newmoonnumber(
  global = global_list()
)

evaluation_tab(global = global_list())

evaluation_tab_input_selection_checks_row()

evaluation_tab_input_selection_row(global = global_list())

evaluation_tab_input_selection_row_species(global = global_list())

evaluation_tab_input_selection_row_dataset(global = global_list())

evaluation_tab_input_selection_row_model(global = global_list())

evaluation_tab_input_selection_row_historic_end_newmoonnumber(
  global = global_list()
)

evaluation_tab_input_selection_row_newmoonnumber(global = global_list())

covariates_tab(global = global_list())

data_sources_section()

portal_project_href(text = "The Portal Project")

portal_data_href(text = "The Portal Data Repository")

portal_weather_href(text = "weather data")

portal_ndvi_href(text = "NDVI data")

write_rodent_profiles_tab_html(main = ".")

write_models_tab_html(main = ".")
}
\arguments{
\item{global}{A \code{list} of global values for the app.}

\item{text}{\code{character} value of the text used in \code{\link[htmltools:builder]{htmltools::a}}.}

\item{main}{\code{character} value of the name of the main component of the directory tree.}
}
\value{
A UI definition or component shiny tags.
}
\description{
\code{portal_forecast_ui} constructs the user interface (UI) for the web application by updating the static pages (models and rodent profiles) then running \code{\link[shiny:fluidPage]{fluidPage}} on the UI components. \cr \cr
\code{write_rodent_profiles_tab_html} and \code{write_model_tab_html} build and write-out static html files for the rodent profiles and models tabs during \code{\link{fill_app}}. \cr \cr
The \verb{<_>_href} functions provide simplified calls to hyperlinked texts that are repeatedly used. \cr \cr
See \code{Details} for hierarchy of functions.
}
\details{
The UI is hierarchical built as:
\itemize{
\item \code{portal_forecast_ui}
\itemize{
\item \code{page_title_panel}
\item \code{page_subtitle_panel}
\item \code{page_main_panel}
\itemize{
\item \code{forecast_tab}
\itemize{
\item \code{forecast_tab_input_selection_row}
\itemize{
\item \code{forecast_tab_input_selection_row_species}
\item \code{forecast_tab_input_selection_row_dataset}
\item \code{forecast_tab_input_selection_row_model}
\item \code{forecast_tab_input_selection_row_historic_end_newmoonnumber}
}
\item \code{forecast_tab_input_selection_checks_row}  # commented out, but available for checking reactive inputs in dev
\item \code{plot_cast_ts}
\item \code{plot_cast_point}
}
\item \code{evaluation_tab}
\itemize{
\item \code{evaluation_tab_input_selection_row}
\itemize{
\item \code{evaluation_tab_input_selection_row_species}
\item \code{evaluation_tab_input_selection_row_dataset}
\item \code{evaluation_tab_input_selection_row_model}
\item \code{evaluation_tab_input_selection_row_historic_end_newmoonnumber}
\item \code{evaluation_tab_input_selection_row_newmoonnumber}
}
\item \code{evaluation_tab_input_selection_checks_row}  # commented out, but available for checking reactive inputs in dev
\item \code{plot_cast_point}
\item \code{plot_casts_cov_RMSE}
}
\item \code{about_tab}
\itemize{
\item \code{htmltools::includeMarkdown}
}
\item \code{models_tab}
\itemize{
\item \code{htmltools::includeHTML}
}
\item \code{profiles_tab}
\itemize{
\item \code{htmltools::includeHTML}
}
\item \code{covariates_tab}
\itemize{
\item \code{data_sources_section}
}
}
}
}
}
