% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_output.R
\name{write_data}
\alias{write_data}
\title{Save Data Out to a File and Return It	Invisibly}
\usage{
write_data(
  x = NULL,
  main = ".",
  data_sub = "data",
  save = TRUE,
  filename = NULL,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{\code{data.frame} or YAML \code{list} to be written out.}

\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{data_sub}{\code{character} value defining the data subdirectory of the portalcasting directory tree.}

\item{save}{\code{logical} indicator controlling if \code{x} should be saved out.}

\item{filename}{\code{character} name of the file for saving \code{x}.}

\item{overwrite}{\code{logical} indicator of if the file should be overwritten if it exists.}

\item{quiet}{\code{logical} indicator if messages should be quieted.}
}
\value{
\code{x} as input, \code{\link[base]{invisible}}-ly.
}
\description{
Save inputted data out to a data file if requested and return it to the console, \code{\link[base]{invisible}}-ly..
}
