% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{write_model}
\alias{write_model}
\alias{model_template}
\title{Write the template for a model into model subdirectory}
\usage{
write_model(name = "AutoArima", covariates = FALSE, lag = NA,
  main = ".", quiet = FALSE, overwrite = TRUE)

model_template(name = "AutoArima", covariates = FALSE, lag = NULL,
  main = ".", quiet = FALSE)
}
\arguments{
\item{name}{\code{character} value of the name of the model.}

\item{covariates}{\code{logical} indicator for if the model requires 
covariates.}

\item{lag}{\code{integer} (or integer \code{numeric}) lag time used for the
covariates or \code{NA} if \code{covariates = FALSE}.}

\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{overwrite}{\code{logical} indicator of whether or not the model
script should be updated.}
}
\value{
\code{write_mode} \code{\link{write}}s the model script out
 and returns \code{NULL}. \cr \cr
 \code{model_template}: \code{character}-valued text for a model script 
 to be housed in the model directory
}
\description{
\code{write_model} creates a template script (as written by 
  \code{model_template}) for a given model. \cr \cr
  \code{model_template} creates the \code{character}-valued
  text for a model script to be housed in the model directory, as written
  out by \code{write_model}. \cr \cr
}
\examples{
 create_dir()
 write_model()
 model_template()

}
