% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{write_model}
\alias{write_model}
\alias{model_template}
\alias{control_list_arg}
\title{Write Model Function Script into Directory}
\usage{
write_model(main = ".", model = NULL, quiet = FALSE)

model_template(
  name = NULL,
  data_sets = NULL,
  covariatesTF = FALSE,
  lag = NULL,
  main = ".",
  control_files = files_control(),
  max_E = NULL,
  control_runjags = NULL,
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

control_list_arg(control_list = NULL, list_function = NULL, arg_checks = TRUE)
}
\arguments{
\item{main}{\code{character} value defining the main component of the portalcasting directory tree.}

\item{model}{\code{character} name of a model function, the \code{function} itself, or its symbol (backquoted name).}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}
}
\value{
\code{NULL}, \code{\link[base]{invisible}}-ly.
}
\description{
Writes a model's function as a script into the defined directory for use in forecasting. \cr \cr
             \code{model} can be input as a \code{character} string, symbol (backquoted name), or \code{function}, as \code{\link{match.fun}}
}
