% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{write_model}
\alias{write_model}
\alias{model_template}
\title{Write the template for a model into model subdirectory}
\usage{
write_model(name = "AutoArima", covariates = FALSE, lag = NA,
  update = TRUE, tree = dirtree(), quiet = FALSE)

model_template(name = "AutoArima", covariates = FALSE, lag = NULL,
  tree = dirtree(), quiet = FALSE)
}
\arguments{
\item{name}{\code{character} value of the name of the model.}

\item{covariates}{\code{logical} indicator for if the model requires 
covariates.}

\item{lag}{\code{integer} (or integer \code{numeric}) lag time used for the
covariates or \code{NA} if \code{covariates} is \code{FALSE}.}

\item{update}{\code{logical} indicator of whether or not the predictions
should be updated.}

\item{tree}{\code{dirtree}-class directory tree list. See 
\code{\link{dirtree}}.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}
}
\description{
\code{write_model}: Create template script (as written by 
  \code{model_template}) for a given model. \cr \cr
  \code{model_template}: create the \code{character}-valued
  text for a model script to be housed in the model directory, as written
  out by \code{write_model}. \cr \cr
}
