% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_script_generation.R
\name{write_model}
\alias{write_model}
\title{Write the template for a model into the model sub directory}
\usage{
write_model(model = NULL, tree = tree())
}
\arguments{
\item{model}{the name of the specific model}

\item{tree}{the name tree of the forecasting directory}
}
\value{
Nothing
}
\description{
Create template script for a given model
}
