% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_models.R
\name{write_model}
\alias{write_model}
\alias{model_template}
\title{Write the template for a model into model subdirectory}
\usage{
write_model(name = NULL, data_sets = NULL, covariatesTF = NULL,
  lag = NULL, main = ".", quiet = FALSE, verbose = TRUE,
  overwrite = TRUE, control = NULL, arg_checks = TRUE)

model_template(name = NULL, data_sets = prefab_data_sets(),
  covariatesTF = FALSE, lag = NULL, main = ".", quiet = FALSE,
  arg_checks = TRUE)
}
\arguments{
\item{name}{\code{character} value of the name of the model.}

\item{covariatesTF}{\code{logical} indicator for if the model requires 
covariates.}

\item{lag}{\code{integer} (or integer \code{numeric}) lag time used for the
covariates or \code{NULL} if \code{covariatesTF} is \code{FALSE}.}

\item{main}{\code{character} value of the name of the main component of
the directory tree.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{overwrite}{\code{logical} indicator of whether or not the model
script should be updated.}

\item{control}{\code{list} of model-level controls, including
\code{name}, a \code{character} value of the model's name;
\code{covariatesTF}, a \code{logical} indicator for if the model requires 
covariates; and \code{lag}, a \code{integer} (or integer \code{numeric}) 
lag time used for the covariates or \code{NA} if 
\code{covariatesTF = FALSE}. Only used if the specific valued argument
is \code{NULL}.}

\item{arg_checks}{\code{logical} value of if the arguments should be
checked using standard protocols via \code{\link{check_args}}. The 
default (\code{arg_checks = TRUE}) ensures that all inputs are 
formatted correctly and provides directed error messages if not. \cr
However, in sandboxing, it is often desirable to be able to deviate from 
strict argument expectations. Setting \code{arg_checks = FALSE} triggers
many/most/all enclosed functions to not check any arguments using 
\code{\link{check_args}}, and as such, \emph{caveat emptor}.}
}
\value{
\code{write_mode} \code{\link{write}}s the model script out
 and returns \code{NULL}. \cr \cr
 \code{model_template}: \code{character}-valued text for a model script 
 to be housed in the model directory
}
\description{
\code{write_model} creates a template script (as written by 
 \code{model_template}) for a given model. \cr \cr
 \code{model_template} creates the \code{character}-valued
 text for a model script to be housed in the model directory, as written
 out by \code{write_model}. \cr \cr
}
\examples{
 \donttest{
  create_dir()
  write_model("AutoArima")
  model_template()
 }

}
